build the project
