--- wwwroot/assets/index-5Mdl4-AY.js
@@ -1,15 +1,0 @@
-(function(){const t=document.createElement("link").relList;if(t&&t.supports&&t.supports("modulepreload"))return;for(const i of document.querySelectorAll('link[rel="modulepreload"]'))r(i);new MutationObserver(i=>{for(const o of i)if(o.type==="childList")for(const s of o.addedNodes)s.tagName==="LINK"&&s.rel==="modulepreload"&&r(s)}).observe(document,{childList:!0,subtree:!0});function n(i){const o={};return i.integrity&&(o.integrity=i.integrity),i.referrerPolicy&&(o.referrerPolicy=i.referrerPolicy),i.crossOrigin==="use-credentials"?o.credentials="include":i.crossOrigin==="anonymous"?o.credentials="omit":o.credentials="same-origin",o}function r(i){if(i.ep)return;i.ep=!0;const o=n(i);fetch(i.href,o)}})();const Jo="modulepreload",Zo=function(e){return"/"+e},Qr={},kn=function(t,n,r){let i=Promise.resolve();if(n&&n.length>0){const o=document.getElementsByTagName("link");i=Promise.all(n.map(s=>{if(s=Zo(s),s in Qr)return;Qr[s]=!0;const a=s.endsWith(".css"),l=a?'[rel="stylesheet"]':"";if(!!r)for(let f=o.length-1;f>=0;f--){const h=o[f];if(h.href===s&&(!a||h.rel==="stylesheet"))return}else if(document.querySelector(`link[href="${s}"]${l}`))return;const u=document.createElement("link");if(u.rel=a?"stylesheet":Jo,a||(u.as="script",u.crossOrigin=""),u.href=s,document.head.appendChild(u),a)return new Promise((f,h)=>{u.addEventListener("load",f),u.addEventListener("error",()=>h(new Error(`Unable to preload CSS for ${s}`)))})}))}return i.then(()=>t()).catch(o=>{const s=new Event("vite:preloadError",{cancelable:!0});if(s.payload=o,window.dispatchEvent(s),!s.defaultPrevented)throw o})},N={context:void 0,registry:void 0};function wt(e){N.context=e}const es=(e,t)=>e===t,on=Symbol("solid-proxy"),ts=Symbol("solid-track"),sn={equals:es};let Oi=Li;const Pe=1,an=2,Ri={owned:null,cleanups:null,context:null,owner:null},On={};var M=null;let Rn=null,ns=null,q=null,te=null,ke=null,xn=0;function ot(e,t){const n=q,r=M,i=e.length===0,o=t===void 0?r:t,s=i?Ri:{owned:null,cleanups:null,context:o?o.context:null,owner:o},a=i?e:()=>e(()=>J(()=>Cn(s)));M=s,q=null;try{return Ee(a,!0)}finally{q=n,M=r}}function L(e,t){t=t?Object.assign({},sn,t):sn;const n={value:e,observers:null,observerSlots:null,comparator:t.equals||void 0},r=i=>(typeof i=="function"&&(i=i(n.value)),Bi(n,i));return[ji.bind(n),r]}function Jr(e,t,n){const r=En(e,t,!0,Pe);dt(r)}function B(e,t,n){const r=En(e,t,!1,Pe);dt(r)}function rs(e,t,n){Oi=us;const r=En(e,t,!1,Pe);(!n||!n.render)&&(r.user=!0),ke?ke.push(r):dt(r)}function U(e,t,n){n=n?Object.assign({},sn,n):sn;const r=En(e,t,!0,0);return r.observers=null,r.observerSlots=null,r.comparator=n.equals||void 0,dt(r),ji.bind(r)}function is(e){return e&&typeof e=="object"&&"then"in e}function os(e,t,n){let r,i,o;arguments.length===2&&typeof t=="object"||arguments.length===1?(r=!0,i=e,o=t||{}):(r=e,i=t,o=n||{});let s=null,a=On,l=null,c=!1,u="initialValue"in o,f=typeof r=="function"&&U(r);const h=new Set,[p,v]=(o.storage||L)(o.initialValue),[d,y]=L(void 0),[g,m]=L(void 0,{equals:!1}),[b,k]=L(u?"ready":"unresolved");if(N.context){l=`${N.context.id}${N.context.count++}`;let w;o.ssrLoadFrom==="initial"?a=o.initialValue:N.load&&(w=N.load(l))&&(a=w)}function _(w,R,P,A){return s===w&&(s=null,A!==void 0&&(u=!0),(w===a||R===a)&&o.onHydrated&&queueMicrotask(()=>o.onHydrated(A,{value:R})),a=On,S(R,P)),R}function S(w,R){Ee(()=>{R===void 0&&v(()=>w),k(R!==void 0?"errored":u?"ready":"unresolved"),y(R);for(const P of h.keys())P.decrement();h.clear()},!1)}function x(){const w=as,R=p(),P=d();if(P!==void 0&&!s)throw P;return q&&!q.user&&w&&Jr(()=>{g(),s&&(w.resolved||h.has(w)||(w.increment(),h.add(w)))}),R}function T(w=!0){if(w!==!1&&c)return;c=!1;const R=f?f():r;if(R==null||R===!1){_(s,J(p));return}const P=a!==On?a:J(()=>i(R,{value:p(),refetching:w}));return is(P)?(s=P,"value"in P?(P.status==="success"?_(s,P.value,void 0,R):_(s,void 0,void 0,R),P):(c=!0,queueMicrotask(()=>c=!1),Ee(()=>{k(u?"refreshing":"pending"),m()},!1),P.then(A=>_(P,A,void 0,R),A=>_(P,void 0,Ki(A),R)))):(_(s,P,void 0,R),P)}return Object.defineProperties(x,{state:{get:()=>b()},error:{get:()=>d()},loading:{get(){const w=b();return w==="pending"||w==="refreshing"}},latest:{get(){if(!u)return x();const w=d();if(w&&!s)throw w;return p()}}}),f?Jr(()=>T(!1)):T(!1),[x,{refetch:T,mutate:v}]}function J(e){if(q===null)return e();const t=q;q=null;try{return e()}finally{q=t}}function Ti(e,t,n){const r=Array.isArray(e);let i,o=n&&n.defer;return s=>{let a;if(r){a=Array(e.length);for(let c=0;c<e.length;c++)a[c]=e[c]()}else a=e();if(o){o=!1;return}const l=J(()=>t(a,i,s));return i=a,l}}function ut(e){return M===null||(M.cleanups===null?M.cleanups=[e]:M.cleanups.push(e)),e}function ln(){return M}function $r(e,t){const n=M,r=q;M=e,q=null;try{return Ee(t,!0)}catch(i){Or(i)}finally{M=n,q=r}}function ss(e){const t=q,n=M;return Promise.resolve().then(()=>{q=t,M=n;let r;return Ee(e,!1),q=M=null,r?r.done:void 0})}function Ii(e,t){const n=Symbol("context");return{id:n,Provider:fs(n),defaultValue:e}}function Ar(e){return M&&M.context&&M.context[e.id]!==void 0?M.context[e.id]:e.defaultValue}function Sn(e){const t=U(e),n=U(()=>zn(t()));return n.toArray=()=>{const r=n();return Array.isArray(r)?r:r!=null?[r]:[]},n}let as;function ji(){if(this.sources&&this.state)if(this.state===Pe)dt(this);else{const e=te;te=null,Ee(()=>un(this),!1),te=e}if(q){const e=this.observers?this.observers.length:0;q.sources?(q.sources.push(this),q.sourceSlots.push(e)):(q.sources=[this],q.sourceSlots=[e]),this.observers?(this.observers.push(q),this.observerSlots.push(q.sources.length-1)):(this.observers=[q],this.observerSlots=[q.sources.length-1])}return this.value}function Bi(e,t,n){let r=e.value;return(!e.comparator||!e.comparator(r,t))&&(e.value=t,e.observers&&e.observers.length&&Ee(()=>{for(let i=0;i<e.observers.length;i+=1){const o=e.observers[i],s=Rn&&Rn.running;s&&Rn.disposed.has(o),(s?!o.tState:!o.state)&&(o.pure?te.push(o):ke.push(o),o.observers&&Di(o)),s||(o.state=Pe)}if(te.length>1e6)throw te=[],new Error},!1)),t}function dt(e){if(!e.fn)return;Cn(e);const t=xn;ls(e,e.value,t)}function ls(e,t,n){let r;const i=M,o=q;q=M=e;try{r=e.fn(t)}catch(s){return e.pure&&(e.state=Pe,e.owned&&e.owned.forEach(Cn),e.owned=null),e.updatedAt=n+1,Or(s)}finally{q=o,M=i}(!e.updatedAt||e.updatedAt<=n)&&(e.updatedAt!=null&&"observers"in e?Bi(e,r):e.value=r,e.updatedAt=n)}function En(e,t,n,r=Pe,i){const o={fn:e,state:r,updatedAt:null,owned:null,sources:null,sourceSlots:null,cleanups:null,value:t,owner:M,context:M?M.context:null,pure:n};return M===null||M!==Ri&&(M.owned?M.owned.push(o):M.owned=[o]),o}function cn(e){if(e.state===0)return;if(e.state===an)return un(e);if(e.suspense&&J(e.suspense.inFallback))return e.suspense.effects.push(e);const t=[e];for(;(e=e.owner)&&(!e.updatedAt||e.updatedAt<xn);)e.state&&t.push(e);for(let n=t.length-1;n>=0;n--)if(e=t[n],e.state===Pe)dt(e);else if(e.state===an){const r=te;te=null,Ee(()=>un(e,t[0]),!1),te=r}}function Ee(e,t){if(te)return e();let n=!1;t||(te=[]),ke?n=!0:ke=[],xn++;try{const r=e();return cs(n),r}catch(r){n||(ke=null),te=null,Or(r)}}function cs(e){if(te&&(Li(te),te=null),e)return;const t=ke;ke=null,t.length&&Ee(()=>Oi(t),!1)}function Li(e){for(let t=0;t<e.length;t++)cn(e[t])}function us(e){let t,n=0;for(t=0;t<e.length;t++){const r=e[t];r.user?e[n++]=r:cn(r)}if(N.context){if(N.count){N.effects||(N.effects=[]),N.effects.push(...e.slice(0,n));return}else N.effects&&(e=[...N.effects,...e],n+=N.effects.length,delete N.effects);wt()}for(t=0;t<n;t++)cn(e[t])}function un(e,t){e.state=0;for(let n=0;n<e.sources.length;n+=1){const r=e.sources[n];if(r.sources){const i=r.state;i===Pe?r!==t&&(!r.updatedAt||r.updatedAt<xn)&&cn(r):i===an&&un(r,t)}}}function Di(e){for(let t=0;t<e.observers.length;t+=1){const n=e.observers[t];n.state||(n.state=an,n.pure?te.push(n):ke.push(n),n.observers&&Di(n))}}function Cn(e){let t;if(e.sources)for(;e.sources.length;){const n=e.sources.pop(),r=e.sourceSlots.pop(),i=n.observers;if(i&&i.length){const o=i.pop(),s=n.observerSlots.pop();r<i.length&&(o.sourceSlots[s]=r,i[r]=o,n.observerSlots[r]=s)}}if(e.owned){for(t=e.owned.length-1;t>=0;t--)Cn(e.owned[t]);e.owned=null}if(e.cleanups){for(t=e.cleanups.length-1;t>=0;t--)e.cleanups[t]();e.cleanups=null}e.state=0}function Ki(e){return e instanceof Error?e:new Error(typeof e=="string"?e:"Unknown error",{cause:e})}function Or(e,t=M){throw Ki(e)}function zn(e){if(typeof e=="function"&&!e.length)return zn(e());if(Array.isArray(e)){const t=[];for(let n=0;n<e.length;n++){const r=zn(e[n]);Array.isArray(r)?t.push.apply(t,r):t.push(r)}return t}return e}function fs(e,t){return function(r){let i;return B(()=>i=J(()=>(M.context={...M.context,[e]:r.value},Sn(()=>r.children))),void 0),i}}const hs=Symbol("fallback");function Zr(e){for(let t=0;t<e.length;t++)e[t]()}function ds(e,t,n={}){let r=[],i=[],o=[],s=0,a=t.length>1?[]:null;return ut(()=>Zr(o)),()=>{let l=e()||[],c,u;return l[ts],J(()=>{let h=l.length,p,v,d,y,g,m,b,k,_;if(h===0)s!==0&&(Zr(o),o=[],r=[],i=[],s=0,a&&(a=[])),n.fallback&&(r=[hs],i[0]=ot(S=>(o[0]=S,n.fallback())),s=1);else if(s===0){for(i=new Array(h),u=0;u<h;u++)r[u]=l[u],i[u]=ot(f);s=h}else{for(d=new Array(h),y=new Array(h),a&&(g=new Array(h)),m=0,b=Math.min(s,h);m<b&&r[m]===l[m];m++);for(b=s-1,k=h-1;b>=m&&k>=m&&r[b]===l[k];b--,k--)d[k]=i[b],y[k]=o[b],a&&(g[k]=a[b]);for(p=new Map,v=new Array(k+1),u=k;u>=m;u--)_=l[u],c=p.get(_),v[u]=c===void 0?-1:c,p.set(_,u);for(c=m;c<=b;c++)_=r[c],u=p.get(_),u!==void 0&&u!==-1?(d[u]=i[c],y[u]=o[c],a&&(g[u]=a[c]),u=v[u],p.set(_,u)):o[c]();for(u=m;u<h;u++)u in d?(i[u]=d[u],o[u]=y[u],a&&(a[u]=g[u],a[u](u))):i[u]=ot(f);i=i.slice(0,s=h),r=l.slice(0)}return i});function f(h){if(o[u]=h,a){const[p,v]=L(u);return a[u]=v,t(l[u],p)}return t(l[u])}}}function $(e,t){return J(()=>e(t||{}))}function Nt(){return!0}const Hn={get(e,t,n){return t===on?n:e.get(t)},has(e,t){return t===on?!0:e.has(t)},set:Nt,deleteProperty:Nt,getOwnPropertyDescriptor(e,t){return{configurable:!0,enumerable:!0,get(){return e.get(t)},set:Nt,deleteProperty:Nt}},ownKeys(e){return e.keys()}};function Tn(e){return(e=typeof e=="function"?e():e)?e:{}}function ps(){for(let e=0,t=this.length;e<t;++e){const n=this[e]();if(n!==void 0)return n}}function fn(...e){let t=!1;for(let s=0;s<e.length;s++){const a=e[s];t=t||!!a&&on in a,e[s]=typeof a=="function"?(t=!0,U(a)):a}if(t)return new Proxy({get(s){for(let a=e.length-1;a>=0;a--){const l=Tn(e[a])[s];if(l!==void 0)return l}},has(s){for(let a=e.length-1;a>=0;a--)if(s in Tn(e[a]))return!0;return!1},keys(){const s=[];for(let a=0;a<e.length;a++)s.push(...Object.keys(Tn(e[a])));return[...new Set(s)]}},Hn);const n={},r=Object.create(null);for(let s=e.length-1;s>=0;s--){const a=e[s];if(!a)continue;const l=Object.getOwnPropertyNames(a);for(let c=l.length-1;c>=0;c--){const u=l[c];if(u==="__proto__"||u==="constructor")continue;const f=Object.getOwnPropertyDescriptor(a,u);if(!r[u])r[u]=f.get?{enumerable:!0,configurable:!0,get:ps.bind(n[u]=[f.get.bind(a)])}:f.value!==void 0?f:void 0;else{const h=n[u];h&&(f.get?h.push(f.get.bind(a)):f.value!==void 0&&h.push(()=>f.value))}}}const i={},o=Object.keys(r);for(let s=o.length-1;s>=0;s--){const a=o[s],l=r[a];l&&l.get?Object.defineProperty(i,a,l):i[a]=l?l.value:void 0}return i}function ms(e,...t){if(on in e){const i=new Set(t.length>1?t.flat():t[0]),o=t.map(s=>new Proxy({get(a){return s.includes(a)?e[a]:void 0},has(a){return s.includes(a)&&a in e},keys(){return s.filter(a=>a in e)}},Hn));return o.push(new Proxy({get(s){return i.has(s)?void 0:e[s]},has(s){return i.has(s)?!1:s in e},keys(){return Object.keys(e).filter(s=>!i.has(s))}},Hn)),o}const n={},r=t.map(()=>({}));for(const i of Object.getOwnPropertyNames(e)){const o=Object.getOwnPropertyDescriptor(e,i),s=!o.get&&!o.set&&o.enumerable&&o.writable&&o.configurable;let a=!1,l=0;for(const c of t)c.includes(i)&&(a=!0,s?r[l][i]=o.value:Object.defineProperty(r[l],i,o)),++l;a||(s?n[i]=o.value:Object.defineProperty(n,i,o))}return[...r,n]}function Rr(e){let t,n;const r=i=>{const o=N.context;if(o){const[a,l]=L();N.count||(N.count=0),N.count++,(n||(n=e())).then(c=>{wt(o),N.count--,l(()=>c.default),wt()}),t=a}else if(!t){const[a]=os(()=>(n||(n=e())).then(l=>l.default));t=a}let s;return U(()=>(s=t())&&J(()=>{if(!o)return s(i);const a=N.context;wt(o);const l=s(i);return wt(a),l}))};return r.preload=()=>n||((n=e()).then(i=>t=()=>i.default),n),r}const Ni=e=>`Stale read from <${e}>.`;function Re(e){const t="fallback"in e&&{fallback:()=>e.fallback};return U(ds(()=>e.each,e.children,t||void 0))}function G(e){const t=e.keyed,n=U(()=>e.when,void 0,{equals:(r,i)=>t?r===i:!r==!i});return U(()=>{const r=n();if(r){const i=e.children;return typeof i=="function"&&i.length>0?J(()=>i(t?r:()=>{if(!J(n))throw Ni("Show");return e.when})):i}return e.fallback},void 0,void 0)}function ei(e){let t=!1;const n=(o,s)=>(t?o[1]===s[1]:!o[1]==!s[1])&&o[2]===s[2],r=Sn(()=>e.children),i=U(()=>{let o=r();Array.isArray(o)||(o=[o]);for(let s=0;s<o.length;s++){const a=o[s].when;if(a)return t=!!o[s].keyed,[s,a,o[s]]}return[-1]},void 0,{equals:n});return U(()=>{const[o,s,a]=i();if(o<0)return e.fallback;const l=a.children;return typeof l=="function"&&l.length>0?J(()=>l(t?s:()=>{if(J(i)[0]!==o)throw Ni("Match");return a.when})):l},void 0,void 0)}function Mt(e){return e}const ys=["allowfullscreen","async","autofocus","autoplay","checked","controls","default","disabled","formnovalidate","hidden","indeterminate","inert","ismap","loop","multiple","muted","nomodule","novalidate","open","playsinline","readonly","required","reversed","seamless","selected"],gs=new Set(["className","value","readOnly","formNoValidate","isMap","noModule","playsInline",...ys]),vs=new Set(["innerHTML","textContent","innerText","children"]),bs=Object.assign(Object.create(null),{className:"class",htmlFor:"for"}),ws=Object.assign(Object.create(null),{class:"className",formnovalidate:{$:"formNoValidate",BUTTON:1,INPUT:1},ismap:{$:"isMap",IMG:1},nomodule:{$:"noModule",SCRIPT:1},playsinline:{$:"playsInline",VIDEO:1},readonly:{$:"readOnly",INPUT:1,TEXTAREA:1}});function _s(e,t){const n=ws[e];return typeof n=="object"?n[t]?n.$:void 0:n}const ks=new Set(["beforeinput","click","dblclick","contextmenu","focusin","focusout","input","keydown","keyup","mousedown","mousemove","mouseout","mouseover","mouseup","pointerdown","pointermove","pointerout","pointerover","pointerup","touchend","touchmove","touchstart"]),xs={xlink:"http://www.w3.org/1999/xlink",xml:"http://www.w3.org/XML/1998/namespace"};function Ss(e,t,n){let r=n.length,i=t.length,o=r,s=0,a=0,l=t[i-1].nextSibling,c=null;for(;s<i||a<o;){if(t[s]===n[a]){s++,a++;continue}for(;t[i-1]===n[o-1];)i--,o--;if(i===s){const u=o<r?a?n[a-1].nextSibling:n[o-a]:l;for(;a<o;)e.insertBefore(n[a++],u)}else if(o===a)for(;s<i;)(!c||!c.has(t[s]))&&t[s].remove(),s++;else if(t[s]===n[o-1]&&n[a]===t[i-1]){const u=t[--i].nextSibling;e.insertBefore(n[a++],t[s++].nextSibling),e.insertBefore(n[--o],u),t[i]=n[o]}else{if(!c){c=new Map;let f=a;for(;f<o;)c.set(n[f],f++)}const u=c.get(t[s]);if(u!=null)if(a<u&&u<o){let f=s,h=1,p;for(;++f<i&&f<o&&!((p=c.get(t[f]))==null||p!==u+h);)h++;if(h>u-a){const v=t[s];for(;a<u;)e.insertBefore(n[a++],v)}else e.replaceChild(n[a++],t[s++])}else s++;else t[s++].remove()}}}const ti="_$DX_DELEGATE";function Es(e,t,n,r={}){let i;return ot(o=>{i=o,t===document?e():C(t,e(),t.firstChild?null:void 0,n)},r.owner),()=>{i(),t.textContent=""}}function K(e,t,n){let r;const i=()=>{const s=document.createElement("template");return s.innerHTML=e,n?s.content.firstChild.firstChild:s.content.firstChild},o=t?()=>J(()=>document.importNode(r||(r=i()),!0)):()=>(r||(r=i())).cloneNode(!0);return o.cloneNode=o,o}function pe(e,t=window.document){const n=t[ti]||(t[ti]=new Set);for(let r=0,i=e.length;r<i;r++){const o=e[r];n.has(o)||(n.add(o),t.addEventListener(o,Rs))}}function Yn(e,t,n){N.context||(n==null?e.removeAttribute(t):e.setAttribute(t,n))}function Cs(e,t,n,r){N.context||(r==null?e.removeAttributeNS(t,n):e.setAttributeNS(t,n,r))}function F(e,t){N.context||(t==null?e.removeAttribute("class"):e.className=t)}function Pn(e,t,n,r){if(r)Array.isArray(n)?(e[`$$${t}`]=n[0],e[`$$${t}Data`]=n[1]):e[`$$${t}`]=n;else if(Array.isArray(n)){const i=n[0];e.addEventListener(t,n[0]=o=>i.call(e,n[1],o))}else e.addEventListener(t,n)}function Ps(e,t,n={}){const r=Object.keys(t||{}),i=Object.keys(n);let o,s;for(o=0,s=i.length;o<s;o++){const a=i[o];!a||a==="undefined"||t[a]||(ni(e,a,!1),delete n[a])}for(o=0,s=r.length;o<s;o++){const a=r[o],l=!!t[a];!a||a==="undefined"||n[a]===l||!l||(ni(e,a,!0),n[a]=l)}return n}function Gn(e,t,n){if(!t)return n?Yn(e,"style"):t;const r=e.style;if(typeof t=="string")return r.cssText=t;typeof n=="string"&&(r.cssText=n=void 0),n||(n={}),t||(t={});let i,o;for(o in n)t[o]==null&&r.removeProperty(o),delete n[o];for(o in t)i=t[o],i!==n[o]&&(r.setProperty(o,i),n[o]=i);return n}function Mi(e,t={},n,r){const i={};return r||B(()=>i.children=ft(e,t.children,i.children)),B(()=>t.ref&&t.ref(e)),B(()=>As(e,t,n,!0,i,!0)),i}function $s(e,t,n){return J(()=>e(t,n))}function C(e,t,n,r){if(n!==void 0&&!r&&(r=[]),typeof t!="function")return ft(e,t,r,n);B(i=>ft(e,t(),i,n),r)}function As(e,t,n,r,i={},o=!1){t||(t={});for(const s in i)if(!(s in t)){if(s==="children")continue;i[s]=ri(e,s,null,i[s],n,o)}for(const s in t){if(s==="children"){r||ft(e,t.children);continue}const a=t[s];i[s]=ri(e,s,a,i[s],n,o)}}function Os(e){return e.toLowerCase().replace(/-([a-z])/g,(t,n)=>n.toUpperCase())}function ni(e,t,n){const r=t.trim().split(/\s+/);for(let i=0,o=r.length;i<o;i++)e.classList.toggle(r[i],n)}function ri(e,t,n,r,i,o){let s,a,l,c,u;if(t==="style")return Gn(e,n,r);if(t==="classList")return Ps(e,n,r);if(n===r)return r;if(t==="ref")o||n(e);else if(t.slice(0,3)==="on:"){const f=t.slice(3);r&&e.removeEventListener(f,r),n&&e.addEventListener(f,n)}else if(t.slice(0,10)==="oncapture:"){const f=t.slice(10);r&&e.removeEventListener(f,r,!0),n&&e.addEventListener(f,n,!0)}else if(t.slice(0,2)==="on"){const f=t.slice(2).toLowerCase(),h=ks.has(f);if(!h&&r){const p=Array.isArray(r)?r[0]:r;e.removeEventListener(f,p)}(h||n)&&(Pn(e,f,n,h),h&&pe([f]))}else if(t.slice(0,5)==="attr:")Yn(e,t.slice(5),n);else if((u=t.slice(0,5)==="prop:")||(l=vs.has(t))||!i&&((c=_s(t,e.tagName))||(a=gs.has(t)))||(s=e.nodeName.includes("-"))){if(u)t=t.slice(5),a=!0;else if(N.context)return n;t==="class"||t==="className"?F(e,n):s&&!a&&!l?e[Os(t)]=n:e[c||t]=n}else{const f=i&&t.indexOf(":")>-1&&xs[t.split(":")[0]];f?Cs(e,f,t,n):Yn(e,bs[t]||t,n)}return n}function Rs(e){const t=`$$${e.type}`;let n=e.composedPath&&e.composedPath()[0]||e.target;for(e.target!==n&&Object.defineProperty(e,"target",{configurable:!0,value:n}),Object.defineProperty(e,"currentTarget",{configurable:!0,get(){return n||document}}),N.registry&&!N.done&&(N.done=_$HY.done=!0);n;){const r=n[t];if(r&&!n.disabled){const i=n[`${t}Data`];if(i!==void 0?r.call(n,i,e):r.call(n,e),e.cancelBubble)return}n=n._$host||n.parentNode||n.host}}function ft(e,t,n,r,i){if(N.context){!n&&(n=[...e.childNodes]);let a=[];for(let l=0;l<n.length;l++){const c=n[l];c.nodeType===8&&c.data.slice(0,2)==="!$"?c.remove():a.push(c)}n=a}for(;typeof n=="function";)n=n();if(t===n)return n;const o=typeof t,s=r!==void 0;if(e=s&&n[0]&&n[0].parentNode||e,o==="string"||o==="number"){if(N.context)return n;if(o==="number"&&(t=t.toString()),s){let a=n[0];a&&a.nodeType===3?a.data!==t&&(a.data=t):a=document.createTextNode(t),n=tt(e,n,r,a)}else n!==""&&typeof n=="string"?n=e.firstChild.data=t:n=e.textContent=t}else if(t==null||o==="boolean"){if(N.context)return n;n=tt(e,n,r)}else{if(o==="function")return B(()=>{let a=t();for(;typeof a=="function";)a=a();n=ft(e,a,n,r)}),()=>n;if(Array.isArray(t)){const a=[],l=n&&Array.isArray(n);if(Xn(a,t,n,i))return B(()=>n=ft(e,a,n,r,!0)),()=>n;if(N.context){if(!a.length)return n;if(r===void 0)return[...e.childNodes];let c=a[0],u=[c];for(;(c=c.nextSibling)!==r;)u.push(c);return n=u}if(a.length===0){if(n=tt(e,n,r),s)return n}else l?n.length===0?ii(e,a,r):Ss(e,n,a):(n&&tt(e),ii(e,a));n=a}else if(t.nodeType){if(N.context&&t.parentNode)return n=s?[t]:t;if(Array.isArray(n)){if(s)return n=tt(e,n,r,t);tt(e,n,null,t)}else n==null||n===""||!e.firstChild?e.appendChild(t):e.replaceChild(t,e.firstChild);n=t}}return n}function Xn(e,t,n,r){let i=!1;for(let o=0,s=t.length;o<s;o++){let a=t[o],l=n&&n[o],c;if(!(a==null||a===!0||a===!1))if((c=typeof a)=="object"&&a.nodeType)e.push(a);else if(Array.isArray(a))i=Xn(e,a,l)||i;else if(c==="function")if(r){for(;typeof a=="function";)a=a();i=Xn(e,Array.isArray(a)?a:[a],Array.isArray(l)?l:[l])||i}else e.push(a),i=!0;else{const u=String(a);l&&l.nodeType===3&&l.data===u?e.push(l):e.push(document.createTextNode(u))}}return i}function ii(e,t,n=null){for(let r=0,i=t.length;r<i;r++)e.insertBefore(t[r],n)}function tt(e,t,n,r){if(n===void 0)return e.textContent="";const i=r||document.createTextNode("");if(t.length){let o=!1;for(let s=t.length-1;s>=0;s--){const a=t[s];if(i!==a){const l=a.parentNode===e;!o&&!s?l?e.replaceChild(i,a):e.insertBefore(i,n):l&&a.remove()}else o=!0}}else e.insertBefore(i,n);return[i]}const Ts=!1,Is="http://www.w3.org/2000/svg";function js(e,t=!1){return t?document.createElementNS(Is,e):document.createElement(e)}function Bs(e){const{useShadow:t}=e,n=document.createTextNode(""),r=()=>e.mount||document.body,i=ln();let o,s=!!N.context;return rs(()=>{s&&(ln().user=s=!1),o||(o=$r(i,()=>U(()=>e.children)));const a=r();if(a instanceof HTMLHeadElement){const[l,c]=L(!1),u=()=>c(!0);ot(f=>C(a,()=>l()?f():o(),null)),ut(u)}else{const l=js(e.isSVG?"g":"div",e.isSVG),c=t&&l.attachShadow?l.attachShadow({mode:"open"}):l;Object.defineProperty(l,"_$host",{get(){return n.parentNode},configurable:!0}),C(c,o),a.appendChild(l),e.ref&&e.ref(l),ut(()=>a.removeChild(l))}},void 0,{render:!s}),n}function Fi(){let e=new Set;function t(i){return e.add(i),()=>e.delete(i)}let n=!1;function r(i,o){if(n)return!(n=!1);const s={to:i,options:o,defaultPrevented:!1,preventDefault:()=>s.defaultPrevented=!0};for(const a of e)a.listener({...s,from:a.location,retry:l=>{l&&(n=!0),a.navigate(i,{...o,resolve:!1})}});return!s.defaultPrevented}return{subscribe:t,confirm:r}}let Qn;function Tr(){(!window.history.state||window.history.state._depth==null)&&window.history.replaceState({...window.history.state,_depth:window.history.length-1},""),Qn=window.history.state._depth}Tr();function Ls(e){return{...e,_depth:window.history.state&&window.history.state._depth}}function Ds(e,t){let n=!1;return()=>{const r=Qn;Tr();const i=r==null?null:Qn-r;if(n){n=!1;return}i&&t(i)?(n=!0,window.history.go(-i)):e()}}const Ks=/^(?:[a-z0-9]+:)?\/\//i,Ns=/^\/+|(\/)\/+$/g,qi="http://sr";function Ve(e,t=!1){const n=e.replace(Ns,"$1");return n?t||/^[?#]/.test(n)?n:"/"+n:""}function Gt(e,t,n){if(Ks.test(t))return;const r=Ve(e),i=n&&Ve(n);let o="";return!i||t.startsWith("/")?o=r:i.toLowerCase().indexOf(r.toLowerCase())!==0?o=r+i:o=i,(o||"/")+Ve(t,!o)}function Ms(e,t){if(e==null)throw new Error(t);return e}function Fs(e,t){return Ve(e).replace(/\/*(\*.*)?$/g,"")+Ve(t)}function Ui(e){const t={};return e.searchParams.forEach((n,r)=>{t[r]=n}),t}function qs(e,t,n){const[r,i]=e.split("/*",2),o=r.split("/").filter(Boolean),s=o.length;return a=>{const l=a.split("/").filter(Boolean),c=l.length-s;if(c<0||c>0&&i===void 0&&!t)return null;const u={path:s?"":"/",params:{}},f=h=>n===void 0?void 0:n[h];for(let h=0;h<s;h++){const p=o[h],v=l[h],d=p[0]===":",y=d?p.slice(1):p;if(d&&In(v,f(y)))u.params[y]=v;else if(d||!In(v,p))return null;u.path+=`/${v}`}if(i){const h=c?l.slice(-c).join("/"):"";if(In(h,f(i)))u.params[i]=h;else return null}return u}}function In(e,t){const n=r=>r.localeCompare(e,void 0,{sensitivity:"base"})===0;return t===void 0?!0:typeof t=="string"?n(t):typeof t=="function"?t(e):Array.isArray(t)?t.some(n):t instanceof RegExp?t.test(e):!1}function Us(e){const[t,n]=e.pattern.split("/*",2),r=t.split("/").filter(Boolean);return r.reduce((i,o)=>i+(o.startsWith(":")?2:3),r.length-(n===void 0?0:1))}function Vi(e){const t=new Map,n=ln();return new Proxy({},{get(r,i){return t.has(i)||$r(n,()=>t.set(i,U(()=>e()[i]))),t.get(i)()},getOwnPropertyDescriptor(){return{enumerable:!0,configurable:!0}},ownKeys(){return Reflect.ownKeys(e())}})}function Wi(e){let t=/(\/?\:[^\/]+)\?/.exec(e);if(!t)return[e];let n=e.slice(0,t.index),r=e.slice(t.index+t[0].length);const i=[n,n+=t[1]];for(;t=/^(\/\:[^\/]+)\?/.exec(r);)i.push(n+=t[1]),r=r.slice(t[0].length);return Wi(r).reduce((o,s)=>[...o,...i.map(a=>a+s)],[])}const Vs=100,zi=Ii(),Ir=Ii(),jr=()=>Ms(Ar(zi),"<A> and 'use' router primitives can be only used inside a Route."),Ws=()=>Ar(Ir)||jr().base,zs=e=>{const t=Ws();return U(()=>t.resolvePath(e()))},Hs=e=>{const t=jr();return U(()=>{const n=e();return n!==void 0?t.renderPath(n):n})},Ys=()=>jr().location;function Gs(e,t=""){const{component:n,load:r,children:i,info:o}=e,s=!i||Array.isArray(i)&&!i.length,a={key:e,component:n,load:r,info:o};return Hi(e.path).reduce((l,c)=>{for(const u of Wi(c)){const f=Fs(t,u);let h=s?f:f.split("/*",1)[0];h=h.split("/").map(p=>p.startsWith(":")||p.startsWith("*")?p:encodeURIComponent(p)).join("/"),l.push({...a,originalPath:c,pattern:h,matcher:qs(h,!s,e.matchFilters)})}return l},[])}function Xs(e,t=0){return{routes:e,score:Us(e[e.length-1])*1e4-t,matcher(n){const r=[];for(let i=e.length-1;i>=0;i--){const o=e[i],s=o.matcher(n);if(!s)return null;r.unshift({...s,route:o})}return r}}}function Hi(e){return Array.isArray(e)?e:[e]}function Yi(e,t="",n=[],r=[]){const i=Hi(e);for(let o=0,s=i.length;o<s;o++){const a=i[o];if(a&&typeof a=="object"){a.hasOwnProperty("path")||(a.path="");const l=Gs(a,t);for(const c of l){n.push(c);const u=Array.isArray(a.children)&&a.children.length===0;if(a.children&&!u)Yi(a.children,c.pattern,n,r);else{const f=Xs([...n],r.length);r.push(f)}n.pop()}}}return n.length?r:r.sort((o,s)=>s.score-o.score)}function jn(e,t){for(let n=0,r=e.length;n<r;n++){const i=e[n].matcher(t);if(i)return i}return[]}function Qs(e,t){const n=new URL(qi),r=U(l=>{const c=e();try{return new URL(c,n)}catch{return console.error(`Invalid path ${c}`),l}},n,{equals:(l,c)=>l.href===c.href}),i=U(()=>r().pathname),o=U(()=>r().search,!0),s=U(()=>r().hash),a=()=>"";return{get pathname(){return i()},get search(){return o()},get hash(){return s()},get state(){return t()},get key(){return a()},query:Vi(Ti(o,()=>Ui(r())))}}let ve;function Js(){return ve}function Zs(e,t,n,r={}){const{signal:[i,o],utils:s={}}=e,a=s.parsePath||(A=>A),l=s.renderPath||(A=>A),c=s.beforeLeave||Fi(),u=Gt("",r.base||"");if(u===void 0)throw new Error(`${u} is not a valid base path`);u&&!i().value&&o({value:u,replace:!0,scroll:!1});const[f,h]=L(!1),p=async A=>{h(!0);try{await ss(A)}finally{h(!1)}},[v,d]=L(i().value),[y,g]=L(i().state),m=Qs(v,y),b=[],k=L([]),_=U(()=>typeof r.transformUrl=="function"?jn(t(),r.transformUrl(m.pathname)):jn(t(),m.pathname)),S=Vi(()=>{const A=_(),I={};for(let D=0;D<A.length;D++)Object.assign(I,A[D].params);return I}),x={pattern:u,path:()=>u,outlet:()=>null,resolvePath(A){return Gt(u,A)}};return B(()=>{const{value:A,state:I}=i();J(()=>{p(()=>{ve="native",A!==v()&&d(A),g(I),k[1]([])}).then(()=>{ve=void 0})})}),{base:x,location:m,params:S,isRouting:f,renderPath:l,parsePath:a,navigatorFactory:w,matches:_,beforeLeave:c,preloadRoute:P,singleFlight:r.singleFlight===void 0?!0:r.singleFlight,submissions:k};function T(A,I,D){J(()=>{if(typeof I=="number"){I&&(s.go?s.go(I):console.warn("Router integration does not support relative routing"));return}const{replace:fe,resolve:he,scroll:H,state:oe}={replace:!1,resolve:!0,scroll:!0,...D},ie=he?A.resolvePath(I):Gt("",I);if(ie===void 0)throw new Error(`Path '${I}' is not a routable path`);if(b.length>=Vs)throw new Error("Too many redirects");const et=v();if((ie!==et||oe!==y())&&!Ts){if(c.confirm(ie,D)){const Qo=b.push({value:et,replace:fe,scroll:H,state:y()});p(()=>{ve="navigate",d(ie),g(oe),k[1]([])}).then(()=>{b.length===Qo&&(ve=void 0,R({value:ie,state:oe}))})}}})}function w(A){return A=A||Ar(Ir)||x,(I,D)=>T(A,I,D)}function R(A){const I=b[0];I&&((A.value!==I.value||A.state!==I.state)&&o({...A,replace:I.replace,scroll:I.scroll}),b.length=0)}function P(A,I){const D=jn(t(),A.pathname),fe=ve;ve="preload";for(let he in D){const{route:H,params:oe}=D[he];H.component&&H.component.preload&&H.component.preload();const{load:ie}=H;I&&ie&&$r(n(),()=>ie({params:oe,location:{pathname:A.pathname,search:A.search,hash:A.hash,query:Ui(A),state:null,key:""},intent:"preload"}))}ve=fe}}function ea(e,t,n,r){const{base:i,location:o,params:s}=e,{pattern:a,component:l,load:c}=r().route,u=U(()=>r().path);l&&l.preload&&l.preload();const f=c?c({params:s,location:o,intent:ve||"initial"}):void 0;return{parent:t,pattern:a,path:u,outlet:()=>l?$(l,{params:s,location:o,data:f,get children(){return n()}}):n(),resolvePath(p){return Gt(i.path(),p,u())}}}const ta=e=>t=>{const{base:n}=t,r=Sn(()=>t.children),i=U(()=>Yi(r(),t.base||""));let o;const s=Zs(e,i,()=>o,{base:n,singleFlight:t.singleFlight,transformUrl:t.transformUrl});return e.create&&e.create(s),$(zi.Provider,{value:s,get children(){return $(na,{routerState:s,get root(){return t.root},get load(){return t.rootLoad},get children(){return[U(()=>(o=ln())&&null),$(ra,{routerState:s,get branches(){return i()}})]}})}})};function na(e){const t=e.routerState.location,n=e.routerState.params,r=U(()=>e.load&&J(()=>{e.load({params:n,location:t,intent:Js()||"initial"})}));return $(G,{get when(){return e.root},keyed:!0,get fallback(){return e.children},children:i=>$(i,{params:n,location:t,get data(){return r()},get children(){return e.children}})})}function ra(e){const t=[];let n;const r=U(Ti(e.routerState.matches,(i,o,s)=>{let a=o&&i.length===o.length;const l=[];for(let c=0,u=i.length;c<u;c++){const f=o&&o[c],h=i[c];s&&f&&h.route.key===f.route.key?l[c]=s[c]:(a=!1,t[c]&&t[c](),ot(p=>{t[c]=p,l[c]=ea(e.routerState,l[c-1]||e.routerState.base,oi(()=>r()[c+1]),()=>e.routerState.matches()[c])}))}return t.splice(i.length).forEach(c=>c()),s&&a?s:(n=l[0],l)}));return oi(()=>r()&&n)()}const oi=e=>()=>$(G,{get when(){return e()},keyed:!0,children:t=>$(Ir.Provider,{value:t,get children(){return t.outlet()}})}),Ne=e=>{const t=Sn(()=>e.children);return fn(e,{get children(){return t()}})};function ia([e,t],n,r){return[n?()=>n(e()):e,r?i=>t(r(i)):t]}function oa(e){if(e==="#")return null;try{return document.querySelector(e)}catch{return null}}function sa(e){let t=!1;const n=i=>typeof i=="string"?{value:i}:i,r=ia(L(n(e.get()),{equals:(i,o)=>i.value===o.value&&i.state===o.state}),void 0,i=>(!t&&e.set(i),i));return e.init&&ut(e.init((i=e.get())=>{t=!0,r[1](n(i)),t=!1})),ta({signal:r,create:e.create,utils:e.utils})}function aa(e,t,n){return e.addEventListener(t,n),()=>e.removeEventListener(t,n)}function la(e,t){const n=oa(`#${e}`);n?n.scrollIntoView():t&&window.scrollTo(0,0)}const ca=new Map;function ua(e=!0,t=!1,n="/_server",r){return i=>{const o=i.base.path(),s=i.navigatorFactory(i.base);let a={};function l(d){return d.namespaceURI==="http://www.w3.org/2000/svg"}function c(d){if(d.defaultPrevented||d.button!==0||d.metaKey||d.altKey||d.ctrlKey||d.shiftKey)return;const y=d.composedPath().find(S=>S instanceof Node&&S.nodeName.toUpperCase()==="A");if(!y||t&&!y.hasAttribute("link"))return;const g=l(y),m=g?y.href.baseVal:y.href;if((g?y.target.baseVal:y.target)||!m&&!y.hasAttribute("state"))return;const k=(y.getAttribute("rel")||"").split(/\s+/);if(y.hasAttribute("download")||k&&k.includes("external"))return;const _=g?new URL(m,document.baseURI):new URL(m);if(!(_.origin!==window.location.origin||o&&_.pathname&&!_.pathname.toLowerCase().startsWith(o.toLowerCase())))return[y,_]}function u(d){const y=c(d);if(!y)return;const[g,m]=y,b=i.parsePath(m.pathname+m.search+m.hash),k=g.getAttribute("state");d.preventDefault(),s(b,{resolve:!1,replace:g.hasAttribute("replace"),scroll:!g.hasAttribute("noscroll"),state:k&&JSON.parse(k)})}function f(d){const y=c(d);if(!y)return;const[g,m]=y;typeof r=="function"&&(m.pathname=r(m.pathname)),a[m.pathname]||i.preloadRoute(m,g.getAttribute("preload")!=="false")}function h(d){const y=c(d);if(!y)return;const[g,m]=y;typeof r=="function"&&(m.pathname=r(m.pathname)),!a[m.pathname]&&(a[m.pathname]=setTimeout(()=>{i.preloadRoute(m,g.getAttribute("preload")!=="false"),delete a[m.pathname]},200))}function p(d){const y=c(d);if(!y)return;const[,g]=y;typeof r=="function"&&(g.pathname=r(g.pathname)),a[g.pathname]&&(clearTimeout(a[g.pathname]),delete a[g.pathname])}function v(d){if(d.defaultPrevented)return;let y=d.submitter&&d.submitter.hasAttribute("formaction")?d.submitter.getAttribute("formaction"):d.target.getAttribute("action");if(!y)return;if(!y.startsWith("https://action/")){const m=new URL(y,qi);if(y=i.parsePath(m.pathname+m.search),!y.startsWith(n))return}if(d.target.method.toUpperCase()!=="POST")throw new Error("Only POST forms are supported for Actions");const g=ca.get(y);if(g){d.preventDefault();const m=new FormData(d.target);d.submitter&&d.submitter.name&&m.append(d.submitter.name,d.submitter.value),g.call({r:i,f:d.target},m)}}pe(["click","submit"]),document.addEventListener("click",u),e&&(document.addEventListener("mouseover",h),document.addEventListener("mouseout",p),document.addEventListener("focusin",f),document.addEventListener("touchstart",f)),document.addEventListener("submit",v),ut(()=>{document.removeEventListener("click",u),e&&(document.removeEventListener("mouseover",h),document.removeEventListener("mouseout",p),document.removeEventListener("focusin",f),document.removeEventListener("touchstart",f)),document.removeEventListener("submit",v)})}}function fa(e){const t=()=>{const r=window.location.pathname+window.location.search;return{value:e.transformUrl?e.transformUrl(r)+window.location.hash:r+window.location.hash,state:window.history.state}},n=Fi();return sa({get:t,set({value:r,replace:i,scroll:o,state:s}){i?window.history.replaceState(Ls(s),"",r):window.history.pushState(s,"",r),la(window.location.hash.slice(1),o),Tr()},init:r=>aa(window,"popstate",Ds(r,i=>{if(i&&i<0)return!n.confirm(i);{const o=t();return!n.confirm(o.value,{state:o.state})}})),create:ua(e.preload,e.explicitLinks,e.actionBase,e.transformUrl),utils:{go:r=>window.history.go(r),beforeLeave:n}})(e)}var ha=K("<a>");function si(e){e=fn({inactiveClass:"inactive",activeClass:"active"},e);const[,t]=ms(e,["href","state","class","activeClass","inactiveClass","end"]),n=zs(()=>e.href),r=Hs(n),i=Ys(),o=U(()=>{const s=n();if(s===void 0)return[!1,!1];const a=Ve(s.split(/[?#]/,1)[0]).toLowerCase(),l=Ve(i.pathname).toLowerCase();return[e.end?a===l:l.startsWith(a),a===l]});return(()=>{var s=ha();return Mi(s,fn(t,{get href(){return r()||e.href},get state(){return JSON.stringify(e.state)},get classList(){return{...e.class&&{[e.class]:!0},[e.inactiveClass]:!o()[0],[e.activeClass]:o()[0],...t.classList}},link:"",get"aria-current"(){return o()[1]?"page":void 0}}),!1,!1),s})()}var Jn=function(e,t){return Jn=Object.setPrototypeOf||{__proto__:[]}instanceof Array&&function(n,r){n.__proto__=r}||function(n,r){for(var i in r)Object.prototype.hasOwnProperty.call(r,i)&&(n[i]=r[i])},Jn(e,t)};function Br(e,t){if(typeof t!="function"&&t!==null)throw new TypeError("Class extends value "+String(t)+" is not a constructor or null");Jn(e,t);function n(){this.constructor=e}e.prototype=t===null?Object.create(t):(n.prototype=t.prototype,new n)}function da(e,t,n,r){function i(o){return o instanceof n?o:new n(function(s){s(o)})}return new(n||(n=Promise))(function(o,s){function a(u){try{c(r.next(u))}catch(f){s(f)}}function l(u){try{c(r.throw(u))}catch(f){s(f)}}function c(u){u.done?o(u.value):i(u.value).then(a,l)}c((r=r.apply(e,t||[])).next())})}function Gi(e,t){var n={label:0,sent:function(){if(o[0]&1)throw o[1];return o[1]},trys:[],ops:[]},r,i,o,s;return s={next:a(0),throw:a(1),return:a(2)},typeof Symbol=="function"&&(s[Symbol.iterator]=function(){return this}),s;function a(c){return function(u){return l([c,u])}}function l(c){if(r)throw new TypeError("Generator is already executing.");for(;s&&(s=0,c[0]&&(n=0)),n;)try{if(r=1,i&&(o=c[0]&2?i.return:c[0]?i.throw||((o=i.return)&&o.call(i),0):i.next)&&!(o=o.call(i,c[1])).done)return o;switch(i=0,o&&(c=[c[0]&2,o.value]),c[0]){case 0:case 1:o=c;break;case 4:return n.label++,{value:c[1],done:!1};case 5:n.label++,i=c[1],c=[0];continue;case 7:c=n.ops.pop(),n.trys.pop();continue;default:if(o=n.trys,!(o=o.length>0&&o[o.length-1])&&(c[0]===6||c[0]===2)){n=0;continue}if(c[0]===3&&(!o||c[1]>o[0]&&c[1]<o[3])){n.label=c[1];break}if(c[0]===6&&n.label<o[1]){n.label=o[1],o=c;break}if(o&&n.label<o[2]){n.label=o[2],n.ops.push(c);break}o[2]&&n.ops.pop(),n.trys.pop();continue}c=t.call(e,n)}catch(u){c=[6,u],i=0}finally{r=o=0}if(c[0]&5)throw c[1];return{value:c[0]?c[1]:void 0,done:!0}}}function $t(e){var t=typeof Symbol=="function"&&Symbol.iterator,n=t&&e[t],r=0;if(n)return n.call(e);if(e&&typeof e.length=="number")return{next:function(){return e&&r>=e.length&&(e=void 0),{value:e&&e[r++],done:!e}}};throw new TypeError(t?"Object is not iterable.":"Symbol.iterator is not defined.")}function Zn(e,t){var n=typeof Symbol=="function"&&e[Symbol.iterator];if(!n)return e;var r=n.call(e),i,o=[],s;try{for(;(t===void 0||t-- >0)&&!(i=r.next()).done;)o.push(i.value)}catch(a){s={error:a}}finally{try{i&&!i.done&&(n=r.return)&&n.call(r)}finally{if(s)throw s.error}}return o}function er(e,t,n){if(n||arguments.length===2)for(var r=0,i=t.length,o;r<i;r++)(o||!(r in t))&&(o||(o=Array.prototype.slice.call(t,0,r)),o[r]=t[r]);return e.concat(o||Array.prototype.slice.call(t))}function st(e){return this instanceof st?(this.v=e,this):new st(e)}function pa(e,t,n){if(!Symbol.asyncIterator)throw new TypeError("Symbol.asyncIterator is not defined.");var r=n.apply(e,t||[]),i,o=[];return i={},s("next"),s("throw"),s("return"),i[Symbol.asyncIterator]=function(){return this},i;function s(h){r[h]&&(i[h]=function(p){return new Promise(function(v,d){o.push([h,p,v,d])>1||a(h,p)})})}function a(h,p){try{l(r[h](p))}catch(v){f(o[0][3],v)}}function l(h){h.value instanceof st?Promise.resolve(h.value.v).then(c,u):f(o[0][2],h)}function c(h){a("next",h)}function u(h){a("throw",h)}function f(h,p){h(p),o.shift(),o.length&&a(o[0][0],o[0][1])}}function ma(e){if(!Symbol.asyncIterator)throw new TypeError("Symbol.asyncIterator is not defined.");var t=e[Symbol.asyncIterator],n;return t?t.call(e):(e=typeof $t=="function"?$t(e):e[Symbol.iterator](),n={},r("next"),r("throw"),r("return"),n[Symbol.asyncIterator]=function(){return this},n);function r(o){n[o]=e[o]&&function(s){return new Promise(function(a,l){s=e[o](s),i(a,l,s.done,s.value)})}}function i(o,s,a,l){Promise.resolve(l).then(function(c){o({value:c,done:a})},s)}}function Y(e){return typeof e=="function"}function ya(e){var t=function(r){Error.call(r),r.stack=new Error().stack},n=e(t);return n.prototype=Object.create(Error.prototype),n.prototype.constructor=n,n}var Bn=ya(function(e){return function(n){e(this),this.message=n?n.length+` errors occurred during unsubscription:
-`+n.map(function(r,i){return i+1+") "+r.toString()}).join(`
-  `):"",this.name="UnsubscriptionError",this.errors=n}});function ai(e,t){if(e){var n=e.indexOf(t);0<=n&&e.splice(n,1)}}var Lr=function(){function e(t){this.initialTeardown=t,this.closed=!1,this._parentage=null,this._finalizers=null}return e.prototype.unsubscribe=function(){var t,n,r,i,o;if(!this.closed){this.closed=!0;var s=this._parentage;if(s)if(this._parentage=null,Array.isArray(s))try{for(var a=$t(s),l=a.next();!l.done;l=a.next()){var c=l.value;c.remove(this)}}catch(d){t={error:d}}finally{try{l&&!l.done&&(n=a.return)&&n.call(a)}finally{if(t)throw t.error}}else s.remove(this);var u=this.initialTeardown;if(Y(u))try{u()}catch(d){o=d instanceof Bn?d.errors:[d]}var f=this._finalizers;if(f){this._finalizers=null;try{for(var h=$t(f),p=h.next();!p.done;p=h.next()){var v=p.value;try{li(v)}catch(d){o=o??[],d instanceof Bn?o=er(er([],Zn(o)),Zn(d.errors)):o.push(d)}}}catch(d){r={error:d}}finally{try{p&&!p.done&&(i=h.return)&&i.call(h)}finally{if(r)throw r.error}}}if(o)throw new Bn(o)}},e.prototype.add=function(t){var n;if(t&&t!==this)if(this.closed)li(t);else{if(t instanceof e){if(t.closed||t._hasParent(this))return;t._addParent(this)}(this._finalizers=(n=this._finalizers)!==null&&n!==void 0?n:[]).push(t)}},e.prototype._hasParent=function(t){var n=this._parentage;return n===t||Array.isArray(n)&&n.includes(t)},e.prototype._addParent=function(t){var n=this._parentage;this._parentage=Array.isArray(n)?(n.push(t),n):n?[n,t]:t},e.prototype._removeParent=function(t){var n=this._parentage;n===t?this._parentage=null:Array.isArray(n)&&ai(n,t)},e.prototype.remove=function(t){var n=this._finalizers;n&&ai(n,t),t instanceof e&&t._removeParent(this)},e.EMPTY=function(){var t=new e;return t.closed=!0,t}(),e}();Lr.EMPTY;function Xi(e){return e instanceof Lr||e&&"closed"in e&&Y(e.remove)&&Y(e.add)&&Y(e.unsubscribe)}function li(e){Y(e)?e():e.unsubscribe()}var Qi={onUnhandledError:null,onStoppedNotification:null,Promise:void 0,useDeprecatedSynchronousErrorHandling:!1,useDeprecatedNextContext:!1},Ji={setTimeout:function(e,t){for(var n=[],r=2;r<arguments.length;r++)n[r-2]=arguments[r];return setTimeout.apply(void 0,er([e,t],Zn(n)))},clearTimeout:function(e){var t=Ji.delegate;return(t?.clearTimeout||clearTimeout)(e)},delegate:void 0};function Zi(e){Ji.setTimeout(function(){throw e})}function ci(){}function ga(e){e()}var Dr=function(e){Br(t,e);function t(n){var r=e.call(this)||this;return r.isStopped=!1,n?(r.destination=n,Xi(n)&&n.add(r)):r.destination=_a,r}return t.create=function(n,r,i){return new tr(n,r,i)},t.prototype.next=function(n){this.isStopped||this._next(n)},t.prototype.error=function(n){this.isStopped||(this.isStopped=!0,this._error(n))},t.prototype.complete=function(){this.isStopped||(this.isStopped=!0,this._complete())},t.prototype.unsubscribe=function(){this.closed||(this.isStopped=!0,e.prototype.unsubscribe.call(this),this.destination=null)},t.prototype._next=function(n){this.destination.next(n)},t.prototype._error=function(n){try{this.destination.error(n)}finally{this.unsubscribe()}},t.prototype._complete=function(){try{this.destination.complete()}finally{this.unsubscribe()}},t}(Lr),va=Function.prototype.bind;function Ln(e,t){return va.call(e,t)}var ba=function(){function e(t){this.partialObserver=t}return e.prototype.next=function(t){var n=this.partialObserver;if(n.next)try{n.next(t)}catch(r){Ft(r)}},e.prototype.error=function(t){var n=this.partialObserver;if(n.error)try{n.error(t)}catch(r){Ft(r)}else Ft(t)},e.prototype.complete=function(){var t=this.partialObserver;if(t.complete)try{t.complete()}catch(n){Ft(n)}},e}(),tr=function(e){Br(t,e);function t(n,r,i){var o=e.call(this)||this,s;if(Y(n)||!n)s={next:n??void 0,error:r??void 0,complete:i??void 0};else{var a;o&&Qi.useDeprecatedNextContext?(a=Object.create(n),a.unsubscribe=function(){return o.unsubscribe()},s={next:n.next&&Ln(n.next,a),error:n.error&&Ln(n.error,a),complete:n.complete&&Ln(n.complete,a)}):s=n}return o.destination=new ba(s),o}return t}(Dr);function Ft(e){Zi(e)}function wa(e){throw e}var _a={closed:!0,next:ci,error:wa,complete:ci},Kr=function(){return typeof Symbol=="function"&&Symbol.observable||"@@observable"}();function eo(e){return e}function ka(e){return e.length===0?eo:e.length===1?e[0]:function(n){return e.reduce(function(r,i){return i(r)},n)}}var ae=function(){function e(t){t&&(this._subscribe=t)}return e.prototype.lift=function(t){var n=new e;return n.source=this,n.operator=t,n},e.prototype.subscribe=function(t,n,r){var i=this,o=Sa(t)?t:new tr(t,n,r);return ga(function(){var s=i,a=s.operator,l=s.source;o.add(a?a.call(o,l):l?i._subscribe(o):i._trySubscribe(o))}),o},e.prototype._trySubscribe=function(t){try{return this._subscribe(t)}catch(n){t.error(n)}},e.prototype.forEach=function(t,n){var r=this;return n=ui(n),new n(function(i,o){var s=new tr({next:function(a){try{t(a)}catch(l){o(l),s.unsubscribe()}},error:o,complete:i});r.subscribe(s)})},e.prototype._subscribe=function(t){var n;return(n=this.source)===null||n===void 0?void 0:n.subscribe(t)},e.prototype[Kr]=function(){return this},e.prototype.pipe=function(){for(var t=[],n=0;n<arguments.length;n++)t[n]=arguments[n];return ka(t)(this)},e.prototype.toPromise=function(t){var n=this;return t=ui(t),new t(function(r,i){var o;n.subscribe(function(s){return o=s},function(s){return i(s)},function(){return r(o)})})},e.create=function(t){return new e(t)},e}();function ui(e){var t;return(t=e??Qi.Promise)!==null&&t!==void 0?t:Promise}function xa(e){return e&&Y(e.next)&&Y(e.error)&&Y(e.complete)}function Sa(e){return e&&e instanceof Dr||xa(e)&&Xi(e)}function Ea(e){return Y(e?.lift)}function Qe(e){return function(t){if(Ea(t))return t.lift(function(n){try{return e(n,this)}catch(r){this.error(r)}});throw new TypeError("Unable to lift unknown Observable type")}}function Ye(e,t,n,r,i){return new Ca(e,t,n,r,i)}var Ca=function(e){Br(t,e);function t(n,r,i,o,s,a){var l=e.call(this,n)||this;return l.onFinalize=s,l.shouldUnsubscribe=a,l._next=r?function(c){try{r(c)}catch(u){n.error(u)}}:e.prototype._next,l._error=o?function(c){try{o(c)}catch(u){n.error(u)}finally{this.unsubscribe()}}:e.prototype._error,l._complete=i?function(){try{i()}catch(c){n.error(c)}finally{this.unsubscribe()}}:e.prototype._complete,l}return t.prototype.unsubscribe=function(){var n;if(!this.shouldUnsubscribe||this.shouldUnsubscribe()){var r=this.closed;e.prototype.unsubscribe.call(this),!r&&((n=this.onFinalize)===null||n===void 0||n.call(this))}},t}(Dr),Pa=new ae(function(e){return e.complete()});function $a(e){return e&&Y(e.schedule)}function Aa(e){return e[e.length-1]}function Oa(e){return $a(Aa(e))?e.pop():void 0}var to=function(e){return e&&typeof e.length=="number"&&typeof e!="function"};function no(e){return Y(e?.then)}function ro(e){return Y(e[Kr])}function io(e){return Symbol.asyncIterator&&Y(e?.[Symbol.asyncIterator])}function oo(e){return new TypeError("You provided "+(e!==null&&typeof e=="object"?"an invalid object":"'"+e+"'")+" where a stream was expected. You can provide an Observable, Promise, ReadableStream, Array, AsyncIterable, or Iterable.")}function Ra(){return typeof Symbol!="function"||!Symbol.iterator?"@@iterator":Symbol.iterator}var so=Ra();function ao(e){return Y(e?.[so])}function lo(e){return pa(this,arguments,function(){var n,r,i,o;return Gi(this,function(s){switch(s.label){case 0:n=e.getReader(),s.label=1;case 1:s.trys.push([1,,9,10]),s.label=2;case 2:return[4,st(n.read())];case 3:return r=s.sent(),i=r.value,o=r.done,o?[4,st(void 0)]:[3,5];case 4:return[2,s.sent()];case 5:return[4,st(i)];case 6:return[4,s.sent()];case 7:return s.sent(),[3,2];case 8:return[3,10];case 9:return n.releaseLock(),[7];case 10:return[2]}})})}function co(e){return Y(e?.getReader)}function pt(e){if(e instanceof ae)return e;if(e!=null){if(ro(e))return Ta(e);if(to(e))return Ia(e);if(no(e))return ja(e);if(io(e))return uo(e);if(ao(e))return Ba(e);if(co(e))return La(e)}throw oo(e)}function Ta(e){return new ae(function(t){var n=e[Kr]();if(Y(n.subscribe))return n.subscribe(t);throw new TypeError("Provided object does not correctly implement Symbol.observable")})}function Ia(e){return new ae(function(t){for(var n=0;n<e.length&&!t.closed;n++)t.next(e[n]);t.complete()})}function ja(e){return new ae(function(t){e.then(function(n){t.closed||(t.next(n),t.complete())},function(n){return t.error(n)}).then(null,Zi)})}function Ba(e){return new ae(function(t){var n,r;try{for(var i=$t(e),o=i.next();!o.done;o=i.next()){var s=o.value;if(t.next(s),t.closed)return}}catch(a){n={error:a}}finally{try{o&&!o.done&&(r=i.return)&&r.call(i)}finally{if(n)throw n.error}}t.complete()})}function uo(e){return new ae(function(t){Da(e,t).catch(function(n){return t.error(n)})})}function La(e){return uo(lo(e))}function Da(e,t){var n,r,i,o;return da(this,void 0,void 0,function(){var s,a;return Gi(this,function(l){switch(l.label){case 0:l.trys.push([0,5,6,11]),n=ma(e),l.label=1;case 1:return[4,n.next()];case 2:if(r=l.sent(),!!r.done)return[3,4];if(s=r.value,t.next(s),t.closed)return[2];l.label=3;case 3:return[3,1];case 4:return[3,11];case 5:return a=l.sent(),i={error:a},[3,11];case 6:return l.trys.push([6,,9,10]),r&&!r.done&&(o=n.return)?[4,o.call(n)]:[3,8];case 7:l.sent(),l.label=8;case 8:return[3,10];case 9:if(i)throw i.error;return[7];case 10:return[7];case 11:return t.complete(),[2]}})})}function Te(e,t,n,r,i){r===void 0&&(r=0),i===void 0&&(i=!1);var o=t.schedule(function(){n(),i?e.add(this.schedule(null,r)):this.unsubscribe()},r);if(e.add(o),!i)return o}function fo(e,t){return t===void 0&&(t=0),Qe(function(n,r){n.subscribe(Ye(r,function(i){return Te(r,e,function(){return r.next(i)},t)},function(){return Te(r,e,function(){return r.complete()},t)},function(i){return Te(r,e,function(){return r.error(i)},t)}))})}function ho(e,t){return t===void 0&&(t=0),Qe(function(n,r){r.add(e.schedule(function(){return n.subscribe(r)},t))})}function Ka(e,t){return pt(e).pipe(ho(t),fo(t))}function Na(e,t){return pt(e).pipe(ho(t),fo(t))}function Ma(e,t){return new ae(function(n){var r=0;return t.schedule(function(){r===e.length?n.complete():(n.next(e[r++]),n.closed||this.schedule())})})}function Fa(e,t){return new ae(function(n){var r;return Te(n,t,function(){r=e[so](),Te(n,t,function(){var i,o,s;try{i=r.next(),o=i.value,s=i.done}catch(a){n.error(a);return}s?n.complete():n.next(o)},0,!0)}),function(){return Y(r?.return)&&r.return()}})}function po(e,t){if(!e)throw new Error("Iterable cannot be null");return new ae(function(n){Te(n,t,function(){var r=e[Symbol.asyncIterator]();Te(n,t,function(){r.next().then(function(i){i.done?n.complete():n.next(i.value)})},0,!0)})})}function qa(e,t){return po(lo(e),t)}function Ua(e,t){if(e!=null){if(ro(e))return Ka(e,t);if(to(e))return Ma(e,t);if(no(e))return Na(e,t);if(io(e))return po(e,t);if(ao(e))return Fa(e,t);if(co(e))return qa(e,t)}throw oo(e)}function Va(e,t){return t?Ua(e,t):pt(e)}function ce(){for(var e=[],t=0;t<arguments.length;t++)e[t]=arguments[t];var n=Oa(e);return Va(e,n)}function Wa(e,t){return Qe(function(n,r){var i=0;n.subscribe(Ye(r,function(o){r.next(e.call(t,o,i++))}))})}function za(e,t,n,r,i,o,s,a){var l=[],c=0,u=0,f=!1,h=function(){f&&!l.length&&!c&&t.complete()},p=function(d){return c<r?v(d):l.push(d)},v=function(d){o&&t.next(d),c++;var y=!1;pt(n(d,u++)).subscribe(Ye(t,function(g){i?.(g),o?p(g):t.next(g)},function(){y=!0},void 0,function(){if(y)try{c--;for(var g=function(){var m=l.shift();s?Te(t,s,function(){return v(m)}):v(m)};l.length&&c<r;)g();h()}catch(m){t.error(m)}}))};return e.subscribe(Ye(t,p,function(){f=!0,h()})),function(){a?.()}}function nr(e,t,n){return n===void 0&&(n=1/0),Y(t)?nr(function(r,i){return Wa(function(o,s){return t(r,o,i,s)})(pt(e(r,i)))},n):(typeof t=="number"&&(n=t),Qe(function(r,i){return za(r,i,e,n)}))}function Je(e){return Qe(function(t,n){var r=null,i=!1,o;r=t.subscribe(Ye(n,void 0,void 0,function(s){o=pt(e(s,Je(e)(t))),r?(r.unsubscribe(),r=null,o.subscribe(n)):i=!0})),i&&(r.unsubscribe(),r=null,o.subscribe(n))})}function hn(e,t){return Y(t)?nr(e,t,1):nr(e,1)}function Ie(e){return e<=0?function(){return Pa}:Qe(function(t,n){var r=0;t.subscribe(Ye(n,function(i){++r<=e&&(n.next(i),e<=r&&n.complete())}))})}function je(e,t,n){var r=Y(e)||t||n?{next:e,error:t,complete:n}:e;return r?Qe(function(i,o){var s;(s=r.subscribe)===null||s===void 0||s.call(r);var a=!0;i.subscribe(Ye(o,function(l){var c;(c=r.next)===null||c===void 0||c.call(r,l),o.next(l)},function(){var l;a=!1,(l=r.complete)===null||l===void 0||l.call(r),o.complete()},function(l){var c;a=!1,(c=r.error)===null||c===void 0||c.call(r,l),o.error(l)},function(){var l,c;a&&((l=r.unsubscribe)===null||l===void 0||l.call(r)),(c=r.finalize)===null||c===void 0||c.call(r)}))}):eo}class Ha{loggedIn=L({});tokenValue=L("");set token(t){this.tokenValue[1](t),this.loggedIn[1]({Authorization:`Bearer ${t}`})}get token(){return this.tokenValue[0]}clearToken(){this.loggedIn[1]({Authorization:""})}toObservable(t){return new ae(n=>{t.then(r=>{n.next(r)}).catch(r=>n.error(r)).finally(()=>n.complete())})}post(t,n,r){return new ae(i=>{fetch(t,{method:"POST",headers:{"Content-Type":"application/json",...r?.headers,...this.loggedIn[0]()},body:JSON.stringify(n)}).then(o=>{if(!o.ok)throw new Error(o.statusText);return o.json()}).then(o=>{i.next(o)}).catch(o=>i.error(o)).finally(()=>i.complete())})}get(t,n){return new ae(r=>{fetch(t,{method:"GET",headers:{"Content-Type":"application/json",...n?.headers,...this.loggedIn[0]()}}).then(i=>{if(!i.ok)throw new Error(i.statusText);return i.json()}).then(i=>{r.next(i)}).catch(i=>r.error(i)).finally(()=>r.complete())})}}const Be=new Ha,W=typeof globalThis<"u"?globalThis:typeof self<"u"?self:typeof window<"u"?window:global,X=Object.keys,ee=Array.isArray;function re(e,t){return typeof t!="object"||X(t).forEach(function(n){e[n]=t[n]}),e}typeof Promise>"u"||W.Promise||(W.Promise=Promise);const At=Object.getPrototypeOf,Ya={}.hasOwnProperty;function le(e,t){return Ya.call(e,t)}function ht(e,t){typeof t=="function"&&(t=t(At(e))),(typeof Reflect>"u"?X:Reflect.ownKeys)(t).forEach(n=>{xe(e,n,t[n])})}const mo=Object.defineProperty;function xe(e,t,n,r){mo(e,t,re(n&&le(n,"get")&&typeof n.get=="function"?{get:n.get,set:n.set,configurable:!0}:{value:n,configurable:!0,writable:!0},r))}function at(e){return{from:function(t){return e.prototype=Object.create(t.prototype),xe(e.prototype,"constructor",e),{extend:ht.bind(null,e.prototype)}}}}const Ga=Object.getOwnPropertyDescriptor;function Nr(e,t){let n;return Ga(e,t)||(n=At(e))&&Nr(n,t)}const Xa=[].slice;function dn(e,t,n){return Xa.call(e,t,n)}function yo(e,t){return t(e)}function _t(e){if(!e)throw new Error("Assertion Failed")}function go(e){W.setImmediate?setImmediate(e):setTimeout(e,0)}function vo(e,t){return e.reduce((n,r,i)=>{var o=t(r,i);return o&&(n[o[0]]=o[1]),n},{})}function Se(e,t){if(le(e,t))return e[t];if(!t)return e;if(typeof t!="string"){for(var n=[],r=0,i=t.length;r<i;++r){var o=Se(e,t[r]);n.push(o)}return n}var s=t.indexOf(".");if(s!==-1){var a=e[t.substr(0,s)];return a===void 0?void 0:Se(a,t.substr(s+1))}}function ue(e,t,n){if(e&&t!==void 0&&(!("isFrozen"in Object)||!Object.isFrozen(e)))if(typeof t!="string"&&"length"in t){_t(typeof n!="string"&&"length"in n);for(var r=0,i=t.length;r<i;++r)ue(e,t[r],n[r])}else{var o=t.indexOf(".");if(o!==-1){var s=t.substr(0,o),a=t.substr(o+1);if(a==="")n===void 0?ee(e)&&!isNaN(parseInt(s))?e.splice(s,1):delete e[s]:e[s]=n;else{var l=e[s];l&&le(e,s)||(l=e[s]={}),ue(l,a,n)}}else n===void 0?ee(e)&&!isNaN(parseInt(t))?e.splice(t,1):delete e[t]:e[t]=n}}function bo(e){var t={};for(var n in e)le(e,n)&&(t[n]=e[n]);return t}const Qa=[].concat;function wo(e){return Qa.apply([],e)}const _o="Boolean,String,Date,RegExp,Blob,File,FileList,FileSystemFileHandle,ArrayBuffer,DataView,Uint8ClampedArray,ImageBitmap,ImageData,Map,Set,CryptoKey".split(",").concat(wo([8,16,32,64].map(e=>["Int","Uint","Float"].map(t=>t+e+"Array")))).filter(e=>W[e]),Ja=_o.map(e=>W[e]);vo(_o,e=>[e,!0]);let Ae=null;function Lt(e){Ae=typeof WeakMap<"u"&&new WeakMap;const t=rr(e);return Ae=null,t}function rr(e){if(!e||typeof e!="object")return e;let t=Ae&&Ae.get(e);if(t)return t;if(ee(e)){t=[],Ae&&Ae.set(e,t);for(var n=0,r=e.length;n<r;++n)t.push(rr(e[n]))}else if(Ja.indexOf(e.constructor)>=0)t=e;else{const o=At(e);for(var i in t=o===Object.prototype?{}:Object.create(o),Ae&&Ae.set(e,t),e)le(e,i)&&(t[i]=rr(e[i]))}return t}const{toString:Za}={};function ir(e){return Za.call(e).slice(8,-1)}const or=typeof Symbol<"u"?Symbol.iterator:"@@iterator",el=typeof or=="symbol"?function(e){var t;return e!=null&&(t=e[or])&&t.apply(e)}:function(){return null},it={};function we(e){var t,n,r,i;if(arguments.length===1){if(ee(e))return e.slice();if(this===it&&typeof e=="string")return[e];if(i=el(e)){for(n=[];!(r=i.next()).done;)n.push(r.value);return n}if(e==null)return[e];if(typeof(t=e.length)=="number"){for(n=new Array(t);t--;)n[t]=e[t];return n}return[e]}for(t=arguments.length,n=new Array(t);t--;)n[t]=arguments[t];return n}const Mr=typeof Symbol<"u"?e=>e[Symbol.toStringTag]==="AsyncFunction":()=>!1;var me=typeof location<"u"&&/^(http|https):\/\/(localhost|127\.0\.0\.1)/.test(location.href);function ko(e,t){me=e,xo=t}var xo=()=>!0;const tl=!new Error("").stack;function Ze(){if(tl)try{throw Ze.arguments,new Error}catch(e){return e}return new Error}function sr(e,t){var n=e.stack;return n?(t=t||0,n.indexOf(e.name)===0&&(t+=(e.name+e.message).split(`
-`).length),n.split(`
-`).slice(t).filter(xo).map(r=>`
-`+r).join("")):""}var So=["Unknown","Constraint","Data","TransactionInactive","ReadOnly","Version","NotFound","InvalidState","InvalidAccess","Abort","Timeout","QuotaExceeded","Syntax","DataClone"],Fr=["Modify","Bulk","OpenFailed","VersionChange","Schema","Upgrade","InvalidTable","MissingAPI","NoSuchDatabase","InvalidArgument","SubTransaction","Unsupported","Internal","DatabaseClosed","PrematureCommit","ForeignAwait"].concat(So),nl={VersionChanged:"Database version changed by other database connection",DatabaseClosed:"Database has been closed",Abort:"Transaction aborted",TransactionInactive:"Transaction has already completed or failed",MissingAPI:"IndexedDB API missing. Please visit https://tinyurl.com/y2uuvskb"};function lt(e,t){this._e=Ze(),this.name=e,this.message=t}function Eo(e,t){return e+". Errors: "+Object.keys(t).map(n=>t[n].toString()).filter((n,r,i)=>i.indexOf(n)===r).join(`
-`)}function pn(e,t,n,r){this._e=Ze(),this.failures=t,this.failedKeys=r,this.successCount=n,this.message=Eo(e,t)}function xt(e,t){this._e=Ze(),this.name="BulkError",this.failures=Object.keys(t).map(n=>t[n]),this.failuresByPos=t,this.message=Eo(e,t)}at(lt).from(Error).extend({stack:{get:function(){return this._stack||(this._stack=this.name+": "+this.message+sr(this._e,2))}},toString:function(){return this.name+": "+this.message}}),at(pn).from(lt),at(xt).from(lt);var qr=Fr.reduce((e,t)=>(e[t]=t+"Error",e),{});const rl=lt;var j=Fr.reduce((e,t)=>{var n=t+"Error";function r(i,o){this._e=Ze(),this.name=n,i?typeof i=="string"?(this.message=`${i}${o?`
- `+o:""}`,this.inner=o||null):typeof i=="object"&&(this.message=`${i.name} ${i.message}`,this.inner=i):(this.message=nl[t]||n,this.inner=null)}return at(r).from(rl),e[t]=r,e},{});j.Syntax=SyntaxError,j.Type=TypeError,j.Range=RangeError;var fi=So.reduce((e,t)=>(e[t+"Error"]=j[t],e),{}),Xt=Fr.reduce((e,t)=>(["Syntax","Type","Range"].indexOf(t)===-1&&(e[t+"Error"]=j[t]),e),{});function V(){}function Ot(e){return e}function il(e,t){return e==null||e===Ot?t:function(n){return t(e(n))}}function Ge(e,t){return function(){e.apply(this,arguments),t.apply(this,arguments)}}function ol(e,t){return e===V?t:function(){var n=e.apply(this,arguments);n!==void 0&&(arguments[0]=n);var r=this.onsuccess,i=this.onerror;this.onsuccess=null,this.onerror=null;var o=t.apply(this,arguments);return r&&(this.onsuccess=this.onsuccess?Ge(r,this.onsuccess):r),i&&(this.onerror=this.onerror?Ge(i,this.onerror):i),o!==void 0?o:n}}function sl(e,t){return e===V?t:function(){e.apply(this,arguments);var n=this.onsuccess,r=this.onerror;this.onsuccess=this.onerror=null,t.apply(this,arguments),n&&(this.onsuccess=this.onsuccess?Ge(n,this.onsuccess):n),r&&(this.onerror=this.onerror?Ge(r,this.onerror):r)}}function al(e,t){return e===V?t:function(n){var r=e.apply(this,arguments);re(n,r);var i=this.onsuccess,o=this.onerror;this.onsuccess=null,this.onerror=null;var s=t.apply(this,arguments);return i&&(this.onsuccess=this.onsuccess?Ge(i,this.onsuccess):i),o&&(this.onerror=this.onerror?Ge(o,this.onerror):o),r===void 0?s===void 0?void 0:s:re(r,s)}}function ll(e,t){return e===V?t:function(){return t.apply(this,arguments)!==!1&&e.apply(this,arguments)}}function Ur(e,t){return e===V?t:function(){var n=e.apply(this,arguments);if(n&&typeof n.then=="function"){for(var r=this,i=arguments.length,o=new Array(i);i--;)o[i]=arguments[i];return n.then(function(){return t.apply(r,o)})}return t.apply(this,arguments)}}Xt.ModifyError=pn,Xt.DexieError=lt,Xt.BulkError=xt;var Rt={};const Co=100,[ar,mn,lr]=typeof Promise>"u"?[]:(()=>{let e=Promise.resolve();if(typeof crypto>"u"||!crypto.subtle)return[e,At(e),e];const t=crypto.subtle.digest("SHA-512",new Uint8Array([0]));return[t,At(t),e]})(),Po=mn&&mn.then,Qt=ar&&ar.constructor,Vr=!!lr;var cr=!1,cl=lr?()=>{lr.then(qt)}:W.setImmediate?setImmediate.bind(null,qt):W.MutationObserver?()=>{var e=document.createElement("div");new MutationObserver(()=>{qt(),e=null}).observe(e,{attributes:!0}),e.setAttribute("i","1")}:()=>{setTimeout(qt,0)},St=function(e,t){kt.push([e,t]),yn&&(cl(),yn=!1)},ur=!0,yn=!0,We=[],Jt=[],fr=null,hr=Ot,ct={id:"global",global:!0,ref:0,unhandleds:[],onunhandled:pi,pgp:!1,env:{},finalize:function(){this.unhandleds.forEach(e=>{try{pi(e[0],e[1])}catch{}})}},O=ct,kt=[],ze=0,Zt=[];function E(e){if(typeof this!="object")throw new TypeError("Promises must be constructed via new");this._listeners=[],this.onuncatched=V,this._lib=!1;var t=this._PSD=O;if(me&&(this._stackHolder=Ze(),this._prev=null,this._numPrev=0),typeof e!="function"){if(e!==Rt)throw new TypeError("Not a function");return this._state=arguments[1],this._value=arguments[2],void(this._state===!1&&pr(this,this._value))}this._state=null,this._value=null,++t.ref,Ao(this,e)}const dr={get:function(){var e=O,t=gn;function n(r,i){var o=!e.global&&(e!==O||t!==gn);const s=o&&!Ce();var a=new E((l,c)=>{Wr(this,new $o(vn(r,e,o,s),vn(i,e,o,s),l,c,e))});return me&&To(a,this),a}return n.prototype=Rt,n},set:function(e){xe(this,"then",e&&e.prototype===Rt?dr:{get:function(){return e},set:dr.set})}};function $o(e,t,n,r,i){this.onFulfilled=typeof e=="function"?e:null,this.onRejected=typeof t=="function"?t:null,this.resolve=n,this.reject=r,this.psd=i}function Ao(e,t){try{t(n=>{if(e._state===null){if(n===e)throw new TypeError("A promise cannot be resolved with itself.");var r=e._lib&&Dt();n&&typeof n.then=="function"?Ao(e,(i,o)=>{n instanceof E?n._then(i,o):n.then(i,o)}):(e._state=!0,e._value=n,Oo(e)),r&&Kt()}},pr.bind(null,e))}catch(n){pr(e,n)}}function pr(e,t){if(Jt.push(t),e._state===null){var n=e._lib&&Dt();t=hr(t),e._state=!1,e._value=t,me&&t!==null&&typeof t=="object"&&!t._promise&&function(r,i,o){try{r.apply(null,o)}catch(s){i&&i(s)}}(()=>{var r=Nr(t,"stack");t._promise=e,xe(t,"stack",{get:()=>cr?r&&(r.get?r.get.apply(t):r.value):e.stack})}),function(r){We.some(i=>i._value===r._value)||We.push(r)}(e),Oo(e),n&&Kt()}}function Oo(e){var t=e._listeners;e._listeners=[];for(var n=0,r=t.length;n<r;++n)Wr(e,t[n]);var i=e._PSD;--i.ref||i.finalize(),ze===0&&(++ze,St(()=>{--ze==0&&zr()},[]))}function Wr(e,t){if(e._state!==null){var n=e._state?t.onFulfilled:t.onRejected;if(n===null)return(e._state?t.resolve:t.reject)(e._value);++t.psd.ref,++ze,St(ul,[n,e,t])}else e._listeners.push(t)}function ul(e,t,n){try{fr=t;var r,i=t._value;t._state?r=e(i):(Jt.length&&(Jt=[]),r=e(i),Jt.indexOf(i)===-1&&function(o){for(var s=We.length;s;)if(We[--s]._value===o._value)return void We.splice(s,1)}(t)),n.resolve(r)}catch(o){n.reject(o)}finally{fr=null,--ze==0&&zr(),--n.psd.ref||n.psd.finalize()}}function Ro(e,t,n){if(t.length===n)return t;var r="";if(e._state===!1){var i,o,s=e._value;s!=null?(i=s.name||"Error",o=s.message||s,r=sr(s,0)):(i=s,o=""),t.push(i+(o?": "+o:"")+r)}return me&&((r=sr(e._stackHolder,2))&&t.indexOf(r)===-1&&t.push(r),e._prev&&Ro(e._prev,t,n)),t}function To(e,t){var n=t?t._numPrev+1:0;n<100&&(e._prev=t,e._numPrev=n)}function qt(){Dt()&&Kt()}function Dt(){var e=ur;return ur=!1,yn=!1,e}function Kt(){var e,t,n;do for(;kt.length>0;)for(e=kt,kt=[],n=e.length,t=0;t<n;++t){var r=e[t];r[0].apply(null,r[1])}while(kt.length>0);ur=!0,yn=!0}function zr(){var e=We;We=[],e.forEach(r=>{r._PSD.onunhandled.call(null,r._value,r)});for(var t=Zt.slice(0),n=t.length;n;)t[--n]()}function Ut(e){return new E(Rt,!1,e)}function z(e,t){var n=O;return function(){var r=Dt(),i=O;try{return De(n,!0),e.apply(this,arguments)}catch(o){t&&t(o)}finally{De(i,!1),r&&Kt()}}}ht(E.prototype,{then:dr,_then:function(e,t){Wr(this,new $o(null,null,e,t,O))},catch:function(e){if(arguments.length===1)return this.then(null,e);var t=arguments[0],n=arguments[1];return typeof t=="function"?this.then(null,r=>r instanceof t?n(r):Ut(r)):this.then(null,r=>r&&r.name===t?n(r):Ut(r))},finally:function(e){return this.then(t=>(e(),t),t=>(e(),Ut(t)))},stack:{get:function(){if(this._stack)return this._stack;try{cr=!0;var e=Ro(this,[],20).join(`
-From previous: `);return this._state!==null&&(this._stack=e),e}finally{cr=!1}}},timeout:function(e,t){return e<1/0?new E((n,r)=>{var i=setTimeout(()=>r(new j.Timeout(t)),e);this.then(n,r).finally(clearTimeout.bind(null,i))}):this}}),typeof Symbol<"u"&&Symbol.toStringTag&&xe(E.prototype,Symbol.toStringTag,"Dexie.Promise"),ct.env=Io(),ht(E,{all:function(){var e=we.apply(null,arguments).map(Vt);return new E(function(t,n){e.length===0&&t([]);var r=e.length;e.forEach((i,o)=>E.resolve(i).then(s=>{e[o]=s,--r||t(e)},n))})},resolve:e=>{if(e instanceof E)return e;if(e&&typeof e.then=="function")return new E((n,r)=>{e.then(n,r)});var t=new E(Rt,!0,e);return To(t,fr),t},reject:Ut,race:function(){var e=we.apply(null,arguments).map(Vt);return new E((t,n)=>{e.map(r=>E.resolve(r).then(t,n))})},PSD:{get:()=>O,set:e=>O=e},totalEchoes:{get:()=>gn},newPSD:Le,usePSD:yt,scheduler:{get:()=>St,set:e=>{St=e}},rejectionMapper:{get:()=>hr,set:e=>{hr=e}},follow:(e,t)=>new E((n,r)=>Le((i,o)=>{var s=O;s.unhandleds=[],s.onunhandled=o,s.finalize=Ge(function(){(function(a){function l(){a(),Zt.splice(Zt.indexOf(l),1)}Zt.push(l),++ze,St(()=>{--ze==0&&zr()},[])})(()=>{this.unhandleds.length===0?i():o(this.unhandleds[0])})},s.finalize),e()},t,n,r))}),Qt&&(Qt.allSettled&&xe(E,"allSettled",function(){const e=we.apply(null,arguments).map(Vt);return new E(t=>{e.length===0&&t([]);let n=e.length;const r=new Array(n);e.forEach((i,o)=>E.resolve(i).then(s=>r[o]={status:"fulfilled",value:s},s=>r[o]={status:"rejected",reason:s}).then(()=>--n||t(r)))})}),Qt.any&&typeof AggregateError<"u"&&xe(E,"any",function(){const e=we.apply(null,arguments).map(Vt);return new E((t,n)=>{e.length===0&&n(new AggregateError([]));let r=e.length;const i=new Array(r);e.forEach((o,s)=>E.resolve(o).then(a=>t(a),a=>{i[s]=a,--r||n(new AggregateError(i))}))})}));const Z={awaits:0,echoes:0,id:0};var fl=0,en=[],Dn=0,gn=0,hl=0;function Le(e,t,n,r){var i=O,o=Object.create(i);o.parent=i,o.ref=0,o.global=!1,o.id=++hl;var s=ct.env;o.env=Vr?{Promise:E,PromiseProp:{value:E,configurable:!0,writable:!0},all:E.all,race:E.race,allSettled:E.allSettled,any:E.any,resolve:E.resolve,reject:E.reject,nthen:hi(s.nthen,o),gthen:hi(s.gthen,o)}:{},t&&re(o,t),++i.ref,o.finalize=function(){--this.parent.ref||this.parent.finalize()};var a=yt(o,e,n,r);return o.ref===0&&o.finalize(),a}function mt(){return Z.id||(Z.id=++fl),++Z.awaits,Z.echoes+=Co,Z.id}function Ce(){return!!Z.awaits&&(--Z.awaits==0&&(Z.id=0),Z.echoes=Z.awaits*Co,!0)}function Vt(e){return Z.echoes&&e&&e.constructor===Qt?(mt(),e.then(t=>(Ce(),t),t=>(Ce(),Q(t)))):e}function dl(e){++gn,Z.echoes&&--Z.echoes!=0||(Z.echoes=Z.id=0),en.push(O),De(e,!0)}function pl(){var e=en[en.length-1];en.pop(),De(e,!1)}function De(e,t){var n=O;if((t?!Z.echoes||Dn++&&e===O:!Dn||--Dn&&e===O)||jo(t?dl.bind(null,e):pl),e!==O&&(O=e,n===ct&&(ct.env=Io()),Vr)){var r=ct.env.Promise,i=e.env;mn.then=i.nthen,r.prototype.then=i.gthen,(n.global||e.global)&&(Object.defineProperty(W,"Promise",i.PromiseProp),r.all=i.all,r.race=i.race,r.resolve=i.resolve,r.reject=i.reject,i.allSettled&&(r.allSettled=i.allSettled),i.any&&(r.any=i.any))}}function Io(){var e=W.Promise;return Vr?{Promise:e,PromiseProp:Object.getOwnPropertyDescriptor(W,"Promise"),all:e.all,race:e.race,allSettled:e.allSettled,any:e.any,resolve:e.resolve,reject:e.reject,nthen:mn.then,gthen:e.prototype.then}:{}}function yt(e,t,n,r,i){var o=O;try{return De(e,!0),t(n,r,i)}finally{De(o,!1)}}function jo(e){Po.call(ar,e)}function vn(e,t,n,r){return typeof e!="function"?e:function(){var i=O;n&&mt(),De(t,!0);try{return e.apply(this,arguments)}finally{De(i,!1),r&&jo(Ce)}}}function hi(e,t){return function(n,r){return e.call(this,vn(n,t),vn(r,t))}}(""+Po).indexOf("[native code]")===-1&&(mt=Ce=V);const di="unhandledrejection";function pi(e,t){var n;try{n=t.onuncatched(e)}catch{}if(n!==!1)try{var r,i={promise:t,reason:e};if(W.document&&document.createEvent?((r=document.createEvent("Event")).initEvent(di,!0,!0),re(r,i)):W.CustomEvent&&re(r=new CustomEvent(di,{detail:i}),i),r&&W.dispatchEvent&&(dispatchEvent(r),!W.PromiseRejectionEvent&&W.onunhandledrejection))try{W.onunhandledrejection(r)}catch{}me&&r&&!r.defaultPrevented&&console.warn(`Unhandled rejection: ${e.stack||e}`)}catch{}}var Q=E.reject;function mr(e,t,n,r){if(e.idbdb&&(e._state.openComplete||O.letThrough||e._vip)){var i=e._createTransaction(t,n,e._dbSchema);try{i.create(),e._state.PR1398_maxLoop=3}catch(o){return o.name===qr.InvalidState&&e.isOpen()&&--e._state.PR1398_maxLoop>0?(console.warn("Dexie: Need to reopen db"),e._close(),e.open().then(()=>mr(e,t,n,r))):Q(o)}return i._promise(t,(o,s)=>Le(()=>(O.trans=i,r(o,s,i)))).then(o=>i._completion.then(()=>o))}if(e._state.openComplete)return Q(new j.DatabaseClosed(e._state.dbOpenError));if(!e._state.isBeingOpened){if(!e._options.autoOpen)return Q(new j.DatabaseClosed);e.open().catch(V)}return e._state.dbReadyPromise.then(()=>mr(e,t,n,r))}const mi="3.2.4",Ue="￿",yr=-1/0,ge="Invalid key provided. Keys must be of type string, number, Date or Array<string | number | Date>.",Bo="String expected.",Et=[],$n=typeof navigator<"u"&&/(MSIE|Trident|Edge)/.test(navigator.userAgent),ml=$n,yl=$n,Lo=e=>!/(dexie\.js|dexie\.min\.js)/.test(e),An="__dbnames",Kn="readonly",Nn="readwrite";function Xe(e,t){return e?t?function(){return e.apply(this,arguments)&&t.apply(this,arguments)}:e:t}const Do={type:3,lower:-1/0,lowerOpen:!1,upper:[[]],upperOpen:!1};function Wt(e){return typeof e!="string"||/\./.test(e)?t=>t:t=>(t[e]===void 0&&e in t&&delete(t=Lt(t))[e],t)}class gl{_trans(t,n,r){const i=this._tx||O.trans,o=this.name;function s(l,c,u){if(!u.schema[o])throw new j.NotFound("Table "+o+" not part of transaction");return n(u.idbtrans,u)}const a=Dt();try{return i&&i.db===this.db?i===O.trans?i._promise(t,s,r):Le(()=>i._promise(t,s,r),{trans:i,transless:O.transless||O}):mr(this.db,t,[this.name],s)}finally{a&&Kt()}}get(t,n){return t&&t.constructor===Object?this.where(t).first(n):this._trans("readonly",r=>this.core.get({trans:r,key:t}).then(i=>this.hook.reading.fire(i))).then(n)}where(t){if(typeof t=="string")return new this.db.WhereClause(this,t);if(ee(t))return new this.db.WhereClause(this,`[${t.join("+")}]`);const n=X(t);if(n.length===1)return this.where(n[0]).equals(t[n[0]]);const r=this.schema.indexes.concat(this.schema.primKey).filter(c=>c.compound&&n.every(u=>c.keyPath.indexOf(u)>=0)&&c.keyPath.every(u=>n.indexOf(u)>=0))[0];if(r&&this.db._maxKey!==Ue)return this.where(r.name).equals(r.keyPath.map(c=>t[c]));!r&&me&&console.warn(`The query ${JSON.stringify(t)} on ${this.name} would benefit of a compound index [${n.join("+")}]`);const{idxByName:i}=this.schema,o=this.db._deps.indexedDB;function s(c,u){try{return o.cmp(c,u)===0}catch{return!1}}const[a,l]=n.reduce(([c,u],f)=>{const h=i[f],p=t[f];return[c||h,c||!h?Xe(u,h&&h.multi?v=>{const d=Se(v,f);return ee(d)&&d.some(y=>s(p,y))}:v=>s(p,Se(v,f))):u]},[null,null]);return a?this.where(a.name).equals(t[a.keyPath]).filter(l):r?this.filter(l):this.where(n).equals("")}filter(t){return this.toCollection().and(t)}count(t){return this.toCollection().count(t)}offset(t){return this.toCollection().offset(t)}limit(t){return this.toCollection().limit(t)}each(t){return this.toCollection().each(t)}toArray(t){return this.toCollection().toArray(t)}toCollection(){return new this.db.Collection(new this.db.WhereClause(this))}orderBy(t){return new this.db.Collection(new this.db.WhereClause(this,ee(t)?`[${t.join("+")}]`:t))}reverse(){return this.toCollection().reverse()}mapToClass(t){this.schema.mappedClass=t;const n=r=>{if(!r)return r;const i=Object.create(t.prototype);for(var o in r)if(le(r,o))try{i[o]=r[o]}catch{}return i};return this.schema.readHook&&this.hook.reading.unsubscribe(this.schema.readHook),this.schema.readHook=n,this.hook("reading",n),t}defineClass(){return this.mapToClass(function(t){re(this,t)})}add(t,n){const{auto:r,keyPath:i}=this.schema.primKey;let o=t;return i&&r&&(o=Wt(i)(t)),this._trans("readwrite",s=>this.core.mutate({trans:s,type:"add",keys:n!=null?[n]:null,values:[o]})).then(s=>s.numFailures?E.reject(s.failures[0]):s.lastResult).then(s=>{if(i)try{ue(t,i,s)}catch{}return s})}update(t,n){if(typeof t!="object"||ee(t))return this.where(":id").equals(t).modify(n);{const r=Se(t,this.schema.primKey.keyPath);if(r===void 0)return Q(new j.InvalidArgument("Given object does not contain its primary key"));try{typeof n!="function"?X(n).forEach(i=>{ue(t,i,n[i])}):n(t,{value:t,primKey:r})}catch{}return this.where(":id").equals(r).modify(n)}}put(t,n){const{auto:r,keyPath:i}=this.schema.primKey;let o=t;return i&&r&&(o=Wt(i)(t)),this._trans("readwrite",s=>this.core.mutate({trans:s,type:"put",values:[o],keys:n!=null?[n]:null})).then(s=>s.numFailures?E.reject(s.failures[0]):s.lastResult).then(s=>{if(i)try{ue(t,i,s)}catch{}return s})}delete(t){return this._trans("readwrite",n=>this.core.mutate({trans:n,type:"delete",keys:[t]})).then(n=>n.numFailures?E.reject(n.failures[0]):void 0)}clear(){return this._trans("readwrite",t=>this.core.mutate({trans:t,type:"deleteRange",range:Do})).then(t=>t.numFailures?E.reject(t.failures[0]):void 0)}bulkGet(t){return this._trans("readonly",n=>this.core.getMany({keys:t,trans:n}).then(r=>r.map(i=>this.hook.reading.fire(i))))}bulkAdd(t,n,r){const i=Array.isArray(n)?n:void 0,o=(r=r||(i?void 0:n))?r.allKeys:void 0;return this._trans("readwrite",s=>{const{auto:a,keyPath:l}=this.schema.primKey;if(l&&i)throw new j.InvalidArgument("bulkAdd(): keys argument invalid on tables with inbound keys");if(i&&i.length!==t.length)throw new j.InvalidArgument("Arguments objects and keys must have the same length");const c=t.length;let u=l&&a?t.map(Wt(l)):t;return this.core.mutate({trans:s,type:"add",keys:i,values:u,wantResults:o}).then(({numFailures:f,results:h,lastResult:p,failures:v})=>{if(f===0)return o?h:p;throw new xt(`${this.name}.bulkAdd(): ${f} of ${c} operations failed`,v)})})}bulkPut(t,n,r){const i=Array.isArray(n)?n:void 0,o=(r=r||(i?void 0:n))?r.allKeys:void 0;return this._trans("readwrite",s=>{const{auto:a,keyPath:l}=this.schema.primKey;if(l&&i)throw new j.InvalidArgument("bulkPut(): keys argument invalid on tables with inbound keys");if(i&&i.length!==t.length)throw new j.InvalidArgument("Arguments objects and keys must have the same length");const c=t.length;let u=l&&a?t.map(Wt(l)):t;return this.core.mutate({trans:s,type:"put",keys:i,values:u,wantResults:o}).then(({numFailures:f,results:h,lastResult:p,failures:v})=>{if(f===0)return o?h:p;throw new xt(`${this.name}.bulkPut(): ${f} of ${c} operations failed`,v)})})}bulkDelete(t){const n=t.length;return this._trans("readwrite",r=>this.core.mutate({trans:r,type:"delete",keys:t})).then(({numFailures:r,lastResult:i,failures:o})=>{if(r===0)return i;throw new xt(`${this.name}.bulkDelete(): ${r} of ${n} operations failed`,o)})}}function Ct(e){var t={},n=function(s,a){if(a){for(var l=arguments.length,c=new Array(l-1);--l;)c[l-1]=arguments[l];return t[s].subscribe.apply(null,c),e}if(typeof s=="string")return t[s]};n.addEventType=o;for(var r=1,i=arguments.length;r<i;++r)o(arguments[r]);return n;function o(s,a,l){if(typeof s!="object"){var c;a||(a=ll),l||(l=V);var u={subscribers:[],fire:l,subscribe:function(f){u.subscribers.indexOf(f)===-1&&(u.subscribers.push(f),u.fire=a(u.fire,f))},unsubscribe:function(f){u.subscribers=u.subscribers.filter(function(h){return h!==f}),u.fire=u.subscribers.reduce(a,l)}};return t[s]=n[s]=u,u}X(c=s).forEach(function(f){var h=c[f];if(ee(h))o(f,c[f][0],c[f][1]);else{if(h!=="asap")throw new j.InvalidArgument("Invalid event config");var p=o(f,Ot,function(){for(var v=arguments.length,d=new Array(v);v--;)d[v]=arguments[v];p.subscribers.forEach(function(y){go(function(){y.apply(null,d)})})})}})}}function gt(e,t){return at(t).from({prototype:e}),t}function nt(e,t){return!(e.filter||e.algorithm||e.or)&&(t?e.justLimit:!e.replayFilter)}function Mn(e,t){e.filter=Xe(e.filter,t)}function Fn(e,t,n){var r=e.replayFilter;e.replayFilter=r?()=>Xe(r(),t()):t,e.justLimit=n&&!r}function tn(e,t){if(e.isPrimKey)return t.primaryKey;const n=t.getIndexByKeyPath(e.index);if(!n)throw new j.Schema("KeyPath "+e.index+" on object store "+t.name+" is not indexed");return n}function yi(e,t,n){const r=tn(e,t.schema);return t.openCursor({trans:n,values:!e.keysOnly,reverse:e.dir==="prev",unique:!!e.unique,query:{index:r,range:e.range}})}function zt(e,t,n,r){const i=e.replayFilter?Xe(e.filter,e.replayFilter()):e.filter;if(e.or){const o={},s=(a,l,c)=>{if(!i||i(l,c,h=>l.stop(h),h=>l.fail(h))){var u=l.primaryKey,f=""+u;f==="[object ArrayBuffer]"&&(f=""+new Uint8Array(u)),le(o,f)||(o[f]=!0,t(a,l,c))}};return Promise.all([e.or._iterate(s,n),gi(yi(e,r,n),e.algorithm,s,!e.keysOnly&&e.valueMapper)])}return gi(yi(e,r,n),Xe(e.algorithm,i),t,!e.keysOnly&&e.valueMapper)}function gi(e,t,n,r){var i=z(r?(o,s,a)=>n(r(o),s,a):n);return e.then(o=>{if(o)return o.start(()=>{var s=()=>o.continue();t&&!t(o,a=>s=a,a=>{o.stop(a),s=V},a=>{o.fail(a),s=V})||i(o.value,o,a=>s=a),s()})})}function ne(e,t){try{const n=vi(e),r=vi(t);if(n!==r)return n==="Array"?1:r==="Array"?-1:n==="binary"?1:r==="binary"?-1:n==="string"?1:r==="string"?-1:n==="Date"?1:r!=="Date"?NaN:-1;switch(n){case"number":case"Date":case"string":return e>t?1:e<t?-1:0;case"binary":return function(i,o){const s=i.length,a=o.length,l=s<a?s:a;for(let c=0;c<l;++c)if(i[c]!==o[c])return i[c]<o[c]?-1:1;return s===a?0:s<a?-1:1}(bi(e),bi(t));case"Array":return function(i,o){const s=i.length,a=o.length,l=s<a?s:a;for(let c=0;c<l;++c){const u=ne(i[c],o[c]);if(u!==0)return u}return s===a?0:s<a?-1:1}(e,t)}}catch{}return NaN}function vi(e){const t=typeof e;if(t!=="object")return t;if(ArrayBuffer.isView(e))return"binary";const n=ir(e);return n==="ArrayBuffer"?"binary":n}function bi(e){return e instanceof Uint8Array?e:ArrayBuffer.isView(e)?new Uint8Array(e.buffer,e.byteOffset,e.byteLength):new Uint8Array(e)}class vl{_read(t,n){var r=this._ctx;return r.error?r.table._trans(null,Q.bind(null,r.error)):r.table._trans("readonly",t).then(n)}_write(t){var n=this._ctx;return n.error?n.table._trans(null,Q.bind(null,n.error)):n.table._trans("readwrite",t,"locked")}_addAlgorithm(t){var n=this._ctx;n.algorithm=Xe(n.algorithm,t)}_iterate(t,n){return zt(this._ctx,t,n,this._ctx.table.core)}clone(t){var n=Object.create(this.constructor.prototype),r=Object.create(this._ctx);return t&&re(r,t),n._ctx=r,n}raw(){return this._ctx.valueMapper=null,this}each(t){var n=this._ctx;return this._read(r=>zt(n,t,r,n.table.core))}count(t){return this._read(n=>{const r=this._ctx,i=r.table.core;if(nt(r,!0))return i.count({trans:n,query:{index:tn(r,i.schema),range:r.range}}).then(s=>Math.min(s,r.limit));var o=0;return zt(r,()=>(++o,!1),n,i).then(()=>o)}).then(t)}sortBy(t,n){const r=t.split(".").reverse(),i=r[0],o=r.length-1;function s(c,u){return u?s(c[r[u]],u-1):c[i]}var a=this._ctx.dir==="next"?1:-1;function l(c,u){var f=s(c,o),h=s(u,o);return f<h?-a:f>h?a:0}return this.toArray(function(c){return c.sort(l)}).then(n)}toArray(t){return this._read(n=>{var r=this._ctx;if(r.dir==="next"&&nt(r,!0)&&r.limit>0){const{valueMapper:i}=r,o=tn(r,r.table.core.schema);return r.table.core.query({trans:n,limit:r.limit,values:!0,query:{index:o,range:r.range}}).then(({result:s})=>i?s.map(i):s)}{const i=[];return zt(r,o=>i.push(o),n,r.table.core).then(()=>i)}},t)}offset(t){var n=this._ctx;return t<=0||(n.offset+=t,nt(n)?Fn(n,()=>{var r=t;return(i,o)=>r===0||(r===1?(--r,!1):(o(()=>{i.advance(r),r=0}),!1))}):Fn(n,()=>{var r=t;return()=>--r<0})),this}limit(t){return this._ctx.limit=Math.min(this._ctx.limit,t),Fn(this._ctx,()=>{var n=t;return function(r,i,o){return--n<=0&&i(o),n>=0}},!0),this}until(t,n){return Mn(this._ctx,function(r,i,o){return!t(r.value)||(i(o),n)}),this}first(t){return this.limit(1).toArray(function(n){return n[0]}).then(t)}last(t){return this.reverse().first(t)}filter(t){var n,r;return Mn(this._ctx,function(i){return t(i.value)}),n=this._ctx,r=t,n.isMatch=Xe(n.isMatch,r),this}and(t){return this.filter(t)}or(t){return new this.db.WhereClause(this._ctx.table,t,this)}reverse(){return this._ctx.dir=this._ctx.dir==="prev"?"next":"prev",this._ondirectionchange&&this._ondirectionchange(this._ctx.dir),this}desc(){return this.reverse()}eachKey(t){var n=this._ctx;return n.keysOnly=!n.isMatch,this.each(function(r,i){t(i.key,i)})}eachUniqueKey(t){return this._ctx.unique="unique",this.eachKey(t)}eachPrimaryKey(t){var n=this._ctx;return n.keysOnly=!n.isMatch,this.each(function(r,i){t(i.primaryKey,i)})}keys(t){var n=this._ctx;n.keysOnly=!n.isMatch;var r=[];return this.each(function(i,o){r.push(o.key)}).then(function(){return r}).then(t)}primaryKeys(t){var n=this._ctx;if(n.dir==="next"&&nt(n,!0)&&n.limit>0)return this._read(i=>{var o=tn(n,n.table.core.schema);return n.table.core.query({trans:i,values:!1,limit:n.limit,query:{index:o,range:n.range}})}).then(({result:i})=>i).then(t);n.keysOnly=!n.isMatch;var r=[];return this.each(function(i,o){r.push(o.primaryKey)}).then(function(){return r}).then(t)}uniqueKeys(t){return this._ctx.unique="unique",this.keys(t)}firstKey(t){return this.limit(1).keys(function(n){return n[0]}).then(t)}lastKey(t){return this.reverse().firstKey(t)}distinct(){var t=this._ctx,n=t.index&&t.table.schema.idxByName[t.index];if(!n||!n.multi)return this;var r={};return Mn(this._ctx,function(i){var o=i.primaryKey.toString(),s=le(r,o);return r[o]=!0,!s}),this}modify(t){var n=this._ctx;return this._write(r=>{var i;if(typeof t=="function")i=t;else{var o=X(t),s=o.length;i=function(d){for(var y=!1,g=0;g<s;++g){var m=o[g],b=t[m];Se(d,m)!==b&&(ue(d,m,b),y=!0)}return y}}const a=n.table.core,{outbound:l,extractKey:c}=a.schema.primaryKey,u=this.db._options.modifyChunkSize||200,f=[];let h=0;const p=[],v=(d,y)=>{const{failures:g,numFailures:m}=y;h+=d-m;for(let b of X(g))f.push(g[b])};return this.clone().primaryKeys().then(d=>{const y=g=>{const m=Math.min(u,d.length-g);return a.getMany({trans:r,keys:d.slice(g,g+m),cache:"immutable"}).then(b=>{const k=[],_=[],S=l?[]:null,x=[];for(let w=0;w<m;++w){const R=b[w],P={value:Lt(R),primKey:d[g+w]};i.call(P,P.value,P)!==!1&&(P.value==null?x.push(d[g+w]):l||ne(c(R),c(P.value))===0?(_.push(P.value),l&&S.push(d[g+w])):(x.push(d[g+w]),k.push(P.value)))}const T=nt(n)&&n.limit===1/0&&(typeof t!="function"||t===qn)&&{index:n.index,range:n.range};return Promise.resolve(k.length>0&&a.mutate({trans:r,type:"add",values:k}).then(w=>{for(let R in w.failures)x.splice(parseInt(R),1);v(k.length,w)})).then(()=>(_.length>0||T&&typeof t=="object")&&a.mutate({trans:r,type:"put",keys:S,values:_,criteria:T,changeSpec:typeof t!="function"&&t}).then(w=>v(_.length,w))).then(()=>(x.length>0||T&&t===qn)&&a.mutate({trans:r,type:"delete",keys:x,criteria:T}).then(w=>v(x.length,w))).then(()=>d.length>g+m&&y(g+u))})};return y(0).then(()=>{if(f.length>0)throw new pn("Error modifying one or more objects",f,h,p);return d.length})})})}delete(){var t=this._ctx,n=t.range;return nt(t)&&(t.isPrimKey&&!yl||n.type===3)?this._write(r=>{const{primaryKey:i}=t.table.core.schema,o=n;return t.table.core.count({trans:r,query:{index:i,range:o}}).then(s=>t.table.core.mutate({trans:r,type:"deleteRange",range:o}).then(({failures:a,lastResult:l,results:c,numFailures:u})=>{if(u)throw new pn("Could not delete some values",Object.keys(a).map(f=>a[f]),s-u);return s-u}))}):this.modify(qn)}}const qn=(e,t)=>t.value=null;function bl(e,t){return e<t?-1:e===t?0:1}function wl(e,t){return e>t?-1:e===t?0:1}function se(e,t,n){var r=e instanceof No?new e.Collection(e):e;return r._ctx.error=n?new n(t):new TypeError(t),r}function rt(e){return new e.Collection(e,()=>Ko("")).limit(0)}function _l(e,t,n,r,i,o){for(var s=Math.min(e.length,r.length),a=-1,l=0;l<s;++l){var c=t[l];if(c!==r[l])return i(e[l],n[l])<0?e.substr(0,l)+n[l]+n.substr(l+1):i(e[l],r[l])<0?e.substr(0,l)+r[l]+n.substr(l+1):a>=0?e.substr(0,a)+t[a]+n.substr(a+1):null;i(e[l],c)<0&&(a=l)}return s<r.length&&o==="next"?e+n.substr(e.length):s<e.length&&o==="prev"?e.substr(0,n.length):a<0?null:e.substr(0,a)+r[a]+n.substr(a+1)}function Ht(e,t,n,r){var i,o,s,a,l,c,u,f=n.length;if(!n.every(d=>typeof d=="string"))return se(e,Bo);function h(d){i=function(g){return g==="next"?m=>m.toUpperCase():m=>m.toLowerCase()}(d),o=function(g){return g==="next"?m=>m.toLowerCase():m=>m.toUpperCase()}(d),s=d==="next"?bl:wl;var y=n.map(function(g){return{lower:o(g),upper:i(g)}}).sort(function(g,m){return s(g.lower,m.lower)});a=y.map(function(g){return g.upper}),l=y.map(function(g){return g.lower}),c=d,u=d==="next"?"":r}h("next");var p=new e.Collection(e,()=>$e(a[0],l[f-1]+r));p._ondirectionchange=function(d){h(d)};var v=0;return p._addAlgorithm(function(d,y,g){var m=d.key;if(typeof m!="string")return!1;var b=o(m);if(t(b,l,v))return!0;for(var k=null,_=v;_<f;++_){var S=_l(m,b,a[_],l[_],s,c);S===null&&k===null?v=_+1:(k===null||s(k,S)>0)&&(k=S)}return y(k!==null?function(){d.continue(k+u)}:g),!1}),p}function $e(e,t,n,r){return{type:2,lower:e,upper:t,lowerOpen:n,upperOpen:r}}function Ko(e){return{type:1,lower:e,upper:e}}class No{get Collection(){return this._ctx.table.db.Collection}between(t,n,r,i){r=r!==!1,i=i===!0;try{return this._cmp(t,n)>0||this._cmp(t,n)===0&&(r||i)&&(!r||!i)?rt(this):new this.Collection(this,()=>$e(t,n,!r,!i))}catch{return se(this,ge)}}equals(t){return t==null?se(this,ge):new this.Collection(this,()=>Ko(t))}above(t){return t==null?se(this,ge):new this.Collection(this,()=>$e(t,void 0,!0))}aboveOrEqual(t){return t==null?se(this,ge):new this.Collection(this,()=>$e(t,void 0,!1))}below(t){return t==null?se(this,ge):new this.Collection(this,()=>$e(void 0,t,!1,!0))}belowOrEqual(t){return t==null?se(this,ge):new this.Collection(this,()=>$e(void 0,t))}startsWith(t){return typeof t!="string"?se(this,Bo):this.between(t,t+Ue,!0,!0)}startsWithIgnoreCase(t){return t===""?this.startsWith(t):Ht(this,(n,r)=>n.indexOf(r[0])===0,[t],Ue)}equalsIgnoreCase(t){return Ht(this,(n,r)=>n===r[0],[t],"")}anyOfIgnoreCase(){var t=we.apply(it,arguments);return t.length===0?rt(this):Ht(this,(n,r)=>r.indexOf(n)!==-1,t,"")}startsWithAnyOfIgnoreCase(){var t=we.apply(it,arguments);return t.length===0?rt(this):Ht(this,(n,r)=>r.some(i=>n.indexOf(i)===0),t,Ue)}anyOf(){const t=we.apply(it,arguments);let n=this._cmp;try{t.sort(n)}catch{return se(this,ge)}if(t.length===0)return rt(this);const r=new this.Collection(this,()=>$e(t[0],t[t.length-1]));r._ondirectionchange=o=>{n=o==="next"?this._ascending:this._descending,t.sort(n)};let i=0;return r._addAlgorithm((o,s,a)=>{const l=o.key;for(;n(l,t[i])>0;)if(++i,i===t.length)return s(a),!1;return n(l,t[i])===0||(s(()=>{o.continue(t[i])}),!1)}),r}notEqual(t){return this.inAnyRange([[yr,t],[t,this.db._maxKey]],{includeLowers:!1,includeUppers:!1})}noneOf(){const t=we.apply(it,arguments);if(t.length===0)return new this.Collection(this);try{t.sort(this._ascending)}catch{return se(this,ge)}const n=t.reduce((r,i)=>r?r.concat([[r[r.length-1][1],i]]):[[yr,i]],null);return n.push([t[t.length-1],this.db._maxKey]),this.inAnyRange(n,{includeLowers:!1,includeUppers:!1})}inAnyRange(t,n){const r=this._cmp,i=this._ascending,o=this._descending,s=this._min,a=this._max;if(t.length===0)return rt(this);if(!t.every(m=>m[0]!==void 0&&m[1]!==void 0&&i(m[0],m[1])<=0))return se(this,"First argument to inAnyRange() must be an Array of two-value Arrays [lower,upper] where upper must not be lower than lower",j.InvalidArgument);const l=!n||n.includeLowers!==!1,c=n&&n.includeUppers===!0;let u,f=i;function h(m,b){return f(m[0],b[0])}try{u=t.reduce(function(m,b){let k=0,_=m.length;for(;k<_;++k){const S=m[k];if(r(b[0],S[1])<0&&r(b[1],S[0])>0){S[0]=s(S[0],b[0]),S[1]=a(S[1],b[1]);break}}return k===_&&m.push(b),m},[]),u.sort(h)}catch{return se(this,ge)}let p=0;const v=c?m=>i(m,u[p][1])>0:m=>i(m,u[p][1])>=0,d=l?m=>o(m,u[p][0])>0:m=>o(m,u[p][0])>=0;let y=v;const g=new this.Collection(this,()=>$e(u[0][0],u[u.length-1][1],!l,!c));return g._ondirectionchange=m=>{m==="next"?(y=v,f=i):(y=d,f=o),u.sort(h)},g._addAlgorithm((m,b,k)=>{for(var _=m.key;y(_);)if(++p,p===u.length)return b(k),!1;return!!function(S){return!v(S)&&!d(S)}(_)||(this._cmp(_,u[p][1])===0||this._cmp(_,u[p][0])===0||b(()=>{f===i?m.continue(u[p][0]):m.continue(u[p][1])}),!1)}),g}startsWithAnyOf(){const t=we.apply(it,arguments);return t.every(n=>typeof n=="string")?t.length===0?rt(this):this.inAnyRange(t.map(n=>[n,n+Ue])):se(this,"startsWithAnyOf() only works with strings")}}function de(e){return z(function(t){return Tt(t),e(t.target.error),!1})}function Tt(e){e.stopPropagation&&e.stopPropagation(),e.preventDefault&&e.preventDefault()}const It="storagemutated",Oe="x-storagemutated-1",Ke=Ct(null,It);class kl{_lock(){return _t(!O.global),++this._reculock,this._reculock!==1||O.global||(O.lockOwnerFor=this),this}_unlock(){if(_t(!O.global),--this._reculock==0)for(O.global||(O.lockOwnerFor=null);this._blockedFuncs.length>0&&!this._locked();){var t=this._blockedFuncs.shift();try{yt(t[1],t[0])}catch{}}return this}_locked(){return this._reculock&&O.lockOwnerFor!==this}create(t){if(!this.mode)return this;const n=this.db.idbdb,r=this.db._state.dbOpenError;if(_t(!this.idbtrans),!t&&!n)switch(r&&r.name){case"DatabaseClosedError":throw new j.DatabaseClosed(r);case"MissingAPIError":throw new j.MissingAPI(r.message,r);default:throw new j.OpenFailed(r)}if(!this.active)throw new j.TransactionInactive;return _t(this._completion._state===null),(t=this.idbtrans=t||(this.db.core?this.db.core.transaction(this.storeNames,this.mode,{durability:this.chromeTransactionDurability}):n.transaction(this.storeNames,this.mode,{durability:this.chromeTransactionDurability}))).onerror=z(i=>{Tt(i),this._reject(t.error)}),t.onabort=z(i=>{Tt(i),this.active&&this._reject(new j.Abort(t.error)),this.active=!1,this.on("abort").fire(i)}),t.oncomplete=z(()=>{this.active=!1,this._resolve(),"mutatedParts"in t&&Ke.storagemutated.fire(t.mutatedParts)}),this}_promise(t,n,r){if(t==="readwrite"&&this.mode!=="readwrite")return Q(new j.ReadOnly("Transaction is readonly"));if(!this.active)return Q(new j.TransactionInactive);if(this._locked())return new E((o,s)=>{this._blockedFuncs.push([()=>{this._promise(t,n,r).then(o,s)},O])});if(r)return Le(()=>{var o=new E((s,a)=>{this._lock();const l=n(s,a,this);l&&l.then&&l.then(s,a)});return o.finally(()=>this._unlock()),o._lib=!0,o});var i=new E((o,s)=>{var a=n(o,s,this);a&&a.then&&a.then(o,s)});return i._lib=!0,i}_root(){return this.parent?this.parent._root():this}waitFor(t){var n=this._root();const r=E.resolve(t);if(n._waitingFor)n._waitingFor=n._waitingFor.then(()=>r);else{n._waitingFor=r,n._waitingQueue=[];var i=n.idbtrans.objectStore(n.storeNames[0]);(function s(){for(++n._spinCount;n._waitingQueue.length;)n._waitingQueue.shift()();n._waitingFor&&(i.get(-1/0).onsuccess=s)})()}var o=n._waitingFor;return new E((s,a)=>{r.then(l=>n._waitingQueue.push(z(s.bind(null,l))),l=>n._waitingQueue.push(z(a.bind(null,l)))).finally(()=>{n._waitingFor===o&&(n._waitingFor=null)})})}abort(){this.active&&(this.active=!1,this.idbtrans&&this.idbtrans.abort(),this._reject(new j.Abort))}table(t){const n=this._memoizedTables||(this._memoizedTables={});if(le(n,t))return n[t];const r=this.schema[t];if(!r)throw new j.NotFound("Table "+t+" not part of transaction");const i=new this.db.Table(t,r,this);return i.core=this.db.core.table(t),n[t]=i,i}}function gr(e,t,n,r,i,o,s){return{name:e,keyPath:t,unique:n,multi:r,auto:i,compound:o,src:(n&&!s?"&":"")+(r?"*":"")+(i?"++":"")+Mo(t)}}function Mo(e){return typeof e=="string"?e:e?"["+[].join.call(e,"+")+"]":""}function Fo(e,t,n){return{name:e,primKey:t,indexes:n,mappedClass:null,idxByName:vo(n,r=>[r.name,r])}}let jt=e=>{try{return e.only([[]]),jt=()=>[[]],[[]]}catch{return jt=()=>Ue,Ue}};function vr(e){return e==null?()=>{}:typeof e=="string"?function(t){return t.split(".").length===1?r=>r[t]:r=>Se(r,t)}(e):t=>Se(t,e)}function wi(e){return[].slice.call(e)}let xl=0;function Pt(e){return e==null?":id":typeof e=="string"?e:`[${e.join("+")}]`}function Sl(e,t,n){function r(l){if(l.type===3)return null;if(l.type===4)throw new Error("Cannot convert never type to IDBKeyRange");const{lower:c,upper:u,lowerOpen:f,upperOpen:h}=l;return c===void 0?u===void 0?null:t.upperBound(u,!!h):u===void 0?t.lowerBound(c,!!f):t.bound(c,u,!!f,!!h)}const{schema:i,hasGetAll:o}=function(l,c){const u=wi(l.objectStoreNames);return{schema:{name:l.name,tables:u.map(f=>c.objectStore(f)).map(f=>{const{keyPath:h,autoIncrement:p}=f,v=ee(h),d=h==null,y={},g={name:f.name,primaryKey:{name:null,isPrimaryKey:!0,outbound:d,compound:v,keyPath:h,autoIncrement:p,unique:!0,extractKey:vr(h)},indexes:wi(f.indexNames).map(m=>f.index(m)).map(m=>{const{name:b,unique:k,multiEntry:_,keyPath:S}=m,x={name:b,compound:ee(S),keyPath:S,unique:k,multiEntry:_,extractKey:vr(S)};return y[Pt(S)]=x,x}),getIndexByKeyPath:m=>y[Pt(m)]};return y[":id"]=g.primaryKey,h!=null&&(y[Pt(h)]=g.primaryKey),g})},hasGetAll:u.length>0&&"getAll"in c.objectStore(u[0])&&!(typeof navigator<"u"&&/Safari/.test(navigator.userAgent)&&!/(Chrome\/|Edge\/)/.test(navigator.userAgent)&&[].concat(navigator.userAgent.match(/Safari\/(\d*)/))[1]<604)}}(e,n),s=i.tables.map(l=>function(c){const u=c.name;return{name:u,schema:c,mutate:function({trans:f,type:h,keys:p,values:v,range:d}){return new Promise((y,g)=>{y=z(y);const m=f.objectStore(u),b=m.keyPath==null,k=h==="put"||h==="add";if(!k&&h!=="delete"&&h!=="deleteRange")throw new Error("Invalid operation type: "+h);const{length:_}=p||v||{length:1};if(p&&v&&p.length!==v.length)throw new Error("Given keys array must have same length as given values array.");if(_===0)return y({numFailures:0,failures:{},results:[],lastResult:void 0});let S;const x=[],T=[];let w=0;const R=A=>{++w,Tt(A)};if(h==="deleteRange"){if(d.type===4)return y({numFailures:w,failures:T,results:[],lastResult:void 0});d.type===3?x.push(S=m.clear()):x.push(S=m.delete(r(d)))}else{const[A,I]=k?b?[v,p]:[v,null]:[p,null];if(k)for(let D=0;D<_;++D)x.push(S=I&&I[D]!==void 0?m[h](A[D],I[D]):m[h](A[D])),S.onerror=R;else for(let D=0;D<_;++D)x.push(S=m[h](A[D])),S.onerror=R}const P=A=>{const I=A.target.result;x.forEach((D,fe)=>D.error!=null&&(T[fe]=D.error)),y({numFailures:w,failures:T,results:h==="delete"?p:x.map(D=>D.result),lastResult:I})};S.onerror=A=>{R(A),P(A)},S.onsuccess=P})},getMany:({trans:f,keys:h})=>new Promise((p,v)=>{p=z(p);const d=f.objectStore(u),y=h.length,g=new Array(y);let m,b=0,k=0;const _=x=>{const T=x.target;g[T._pos]=T.result,++k===b&&p(g)},S=de(v);for(let x=0;x<y;++x)h[x]!=null&&(m=d.get(h[x]),m._pos=x,m.onsuccess=_,m.onerror=S,++b);b===0&&p(g)}),get:({trans:f,key:h})=>new Promise((p,v)=>{p=z(p);const d=f.objectStore(u).get(h);d.onsuccess=y=>p(y.target.result),d.onerror=de(v)}),query:function(f){return h=>new Promise((p,v)=>{p=z(p);const{trans:d,values:y,limit:g,query:m}=h,b=g===1/0?void 0:g,{index:k,range:_}=m,S=d.objectStore(u),x=k.isPrimaryKey?S:S.index(k.name),T=r(_);if(g===0)return p({result:[]});if(f){const w=y?x.getAll(T,b):x.getAllKeys(T,b);w.onsuccess=R=>p({result:R.target.result}),w.onerror=de(v)}else{let w=0;const R=y||!("openKeyCursor"in x)?x.openCursor(T):x.openKeyCursor(T),P=[];R.onsuccess=A=>{const I=R.result;return I?(P.push(y?I.value:I.primaryKey),++w===g?p({result:P}):void I.continue()):p({result:P})},R.onerror=de(v)}})}(o),openCursor:function({trans:f,values:h,query:p,reverse:v,unique:d}){return new Promise((y,g)=>{y=z(y);const{index:m,range:b}=p,k=f.objectStore(u),_=m.isPrimaryKey?k:k.index(m.name),S=v?d?"prevunique":"prev":d?"nextunique":"next",x=h||!("openKeyCursor"in _)?_.openCursor(r(b),S):_.openKeyCursor(r(b),S);x.onerror=de(g),x.onsuccess=z(T=>{const w=x.result;if(!w)return void y(null);w.___id=++xl,w.done=!1;const R=w.continue.bind(w);let P=w.continuePrimaryKey;P&&(P=P.bind(w));const A=w.advance.bind(w),I=()=>{throw new Error("Cursor not stopped")};w.trans=f,w.stop=w.continue=w.continuePrimaryKey=w.advance=()=>{throw new Error("Cursor not started")},w.fail=z(g),w.next=function(){let D=1;return this.start(()=>D--?this.continue():this.stop()).then(()=>this)},w.start=D=>{const fe=new Promise((H,oe)=>{H=z(H),x.onerror=de(oe),w.fail=oe,w.stop=ie=>{w.stop=w.continue=w.continuePrimaryKey=w.advance=I,H(ie)}}),he=()=>{if(x.result)try{D()}catch(H){w.fail(H)}else w.done=!0,w.start=()=>{throw new Error("Cursor behind last entry")},w.stop()};return x.onsuccess=z(H=>{x.onsuccess=he,he()}),w.continue=R,w.continuePrimaryKey=P,w.advance=A,he(),fe},y(w)},g)})},count({query:f,trans:h}){const{index:p,range:v}=f;return new Promise((d,y)=>{const g=h.objectStore(u),m=p.isPrimaryKey?g:g.index(p.name),b=r(v),k=b?m.count(b):m.count();k.onsuccess=z(_=>d(_.target.result)),k.onerror=de(y)})}}}(l)),a={};return s.forEach(l=>a[l.name]=l),{stack:"dbcore",transaction:e.transaction.bind(e),table(l){if(!a[l])throw new Error(`Table '${l}' not found`);return a[l]},MIN_KEY:-1/0,MAX_KEY:jt(t),schema:i}}function br({_novip:e},t){const n=t.db,r=function(i,o,{IDBKeyRange:s,indexedDB:a},l){return{dbcore:function(u,f){return f.reduce((h,{create:p})=>({...h,...p(h)}),u)}(Sl(o,s,l),i.dbcore)}}(e._middlewares,n,e._deps,t);e.core=r.dbcore,e.tables.forEach(i=>{const o=i.name;e.core.schema.tables.some(s=>s.name===o)&&(i.core=e.core.table(o),e[o]instanceof e.Table&&(e[o].core=i.core))})}function bn({_novip:e},t,n,r){n.forEach(i=>{const o=r[i];t.forEach(s=>{const a=Nr(s,i);(!a||"value"in a&&a.value===void 0)&&(s===e.Transaction.prototype||s instanceof e.Transaction?xe(s,i,{get(){return this.table(i)},set(l){mo(this,i,{value:l,writable:!0,configurable:!0,enumerable:!0})}}):s[i]=new e.Table(i,o))})})}function wr({_novip:e},t){t.forEach(n=>{for(let r in n)n[r]instanceof e.Table&&delete n[r]})}function El(e,t){return e._cfg.version-t._cfg.version}function Cl(e,t,n,r){const i=e._dbSchema,o=e._createTransaction("readwrite",e._storeNames,i);o.create(n),o._completion.catch(r);const s=o._reject.bind(o),a=O.transless||O;Le(()=>{O.trans=o,O.transless=a,t===0?(X(i).forEach(l=>{Un(n,l,i[l].primKey,i[l].indexes)}),br(e,n),E.follow(()=>e.on.populate.fire(o)).catch(s)):function({_novip:l},c,u,f){const h=[],p=l._versions;let v=l._dbSchema=kr(l,l.idbdb,f),d=!1;const y=p.filter(m=>m._cfg.version>=c);function g(){return h.length?E.resolve(h.shift()(u.idbtrans)).then(g):E.resolve()}return y.forEach(m=>{h.push(()=>{const b=v,k=m._cfg.dbschema;xr(l,b,f),xr(l,k,f),v=l._dbSchema=k;const _=qo(b,k);_.add.forEach(x=>{Un(f,x[0],x[1].primKey,x[1].indexes)}),_.change.forEach(x=>{if(x.recreate)throw new j.Upgrade("Not yet support for changing primary key");{const T=f.objectStore(x.name);x.add.forEach(w=>_r(T,w)),x.change.forEach(w=>{T.deleteIndex(w.name),_r(T,w)}),x.del.forEach(w=>T.deleteIndex(w))}});const S=m._cfg.contentUpgrade;if(S&&m._cfg.version>c){br(l,f),u._memoizedTables={},d=!0;let x=bo(k);_.del.forEach(P=>{x[P]=b[P]}),wr(l,[l.Transaction.prototype]),bn(l,[l.Transaction.prototype],X(x),x),u.schema=x;const T=Mr(S);let w;T&&mt();const R=E.follow(()=>{if(w=S(u),w&&T){var P=Ce.bind(null,null);w.then(P,P)}});return w&&typeof w.then=="function"?E.resolve(w):R.then(()=>w)}}),h.push(b=>{(!d||!ml)&&function(k,_){[].slice.call(_.db.objectStoreNames).forEach(S=>k[S]==null&&_.db.deleteObjectStore(S))}(m._cfg.dbschema,b),wr(l,[l.Transaction.prototype]),bn(l,[l.Transaction.prototype],l._storeNames,l._dbSchema),u.schema=l._dbSchema})}),g().then(()=>{var m,b;b=f,X(m=v).forEach(k=>{b.db.objectStoreNames.contains(k)||Un(b,k,m[k].primKey,m[k].indexes)})})}(e,t,o,n).catch(s)})}function qo(e,t){const n={del:[],add:[],change:[]};let r;for(r in e)t[r]||n.del.push(r);for(r in t){const i=e[r],o=t[r];if(i){const s={name:r,def:o,recreate:!1,del:[],add:[],change:[]};if(""+(i.primKey.keyPath||"")!=""+(o.primKey.keyPath||"")||i.primKey.auto!==o.primKey.auto&&!$n)s.recreate=!0,n.change.push(s);else{const a=i.idxByName,l=o.idxByName;let c;for(c in a)l[c]||s.del.push(c);for(c in l){const u=a[c],f=l[c];u?u.src!==f.src&&s.change.push(f):s.add.push(f)}(s.del.length>0||s.add.length>0||s.change.length>0)&&n.change.push(s)}}else n.add.push([r,o])}return n}function Un(e,t,n,r){const i=e.db.createObjectStore(t,n.keyPath?{keyPath:n.keyPath,autoIncrement:n.auto}:{autoIncrement:n.auto});return r.forEach(o=>_r(i,o)),i}function _r(e,t){e.createIndex(t.name,t.keyPath,{unique:t.unique,multiEntry:t.multi})}function kr(e,t,n){const r={};return dn(t.objectStoreNames,0).forEach(i=>{const o=n.objectStore(i);let s=o.keyPath;const a=gr(Mo(s),s||"",!1,!1,!!o.autoIncrement,s&&typeof s!="string",!0),l=[];for(let u=0;u<o.indexNames.length;++u){const f=o.index(o.indexNames[u]);s=f.keyPath;var c=gr(f.name,s,!!f.unique,!!f.multiEntry,!1,s&&typeof s!="string",!1);l.push(c)}r[i]=Fo(i,a,l)}),r}function xr({_novip:e},t,n){const r=n.db.objectStoreNames;for(let i=0;i<r.length;++i){const o=r[i],s=n.objectStore(o);e._hasGetAll="getAll"in s;for(let a=0;a<s.indexNames.length;++a){const l=s.indexNames[a],c=s.index(l).keyPath,u=typeof c=="string"?c:"["+dn(c).join("+")+"]";if(t[o]){const f=t[o].idxByName[u];f&&(f.name=l,delete t[o].idxByName[u],t[o].idxByName[l]=f)}}}typeof navigator<"u"&&/Safari/.test(navigator.userAgent)&&!/(Chrome\/|Edge\/)/.test(navigator.userAgent)&&W.WorkerGlobalScope&&W instanceof W.WorkerGlobalScope&&[].concat(navigator.userAgent.match(/Safari\/(\d*)/))[1]<604&&(e._hasGetAll=!1)}class Pl{_parseStoresSpec(t,n){X(t).forEach(r=>{if(t[r]!==null){var i=t[r].split(",").map((s,a)=>{const l=(s=s.trim()).replace(/([&*]|\+\+)/g,""),c=/^\[/.test(l)?l.match(/^\[(.*)\]$/)[1].split("+"):l;return gr(l,c||null,/\&/.test(s),/\*/.test(s),/\+\+/.test(s),ee(c),a===0)}),o=i.shift();if(o.multi)throw new j.Schema("Primary key cannot be multi-valued");i.forEach(s=>{if(s.auto)throw new j.Schema("Only primary key can be marked as autoIncrement (++)");if(!s.keyPath)throw new j.Schema("Index must have a name and cannot be an empty string")}),n[r]=Fo(r,o,i)}})}stores(t){const n=this.db;this._cfg.storesSource=this._cfg.storesSource?re(this._cfg.storesSource,t):t;const r=n._versions,i={};let o={};return r.forEach(s=>{re(i,s._cfg.storesSource),o=s._cfg.dbschema={},s._parseStoresSpec(i,o)}),n._dbSchema=o,wr(n,[n._allTables,n,n.Transaction.prototype]),bn(n,[n._allTables,n,n.Transaction.prototype,this._cfg.tables],X(o),o),n._storeNames=X(o),this}upgrade(t){return this._cfg.contentUpgrade=Ur(this._cfg.contentUpgrade||V,t),this}}function Hr(e,t){let n=e._dbNamesDB;return n||(n=e._dbNamesDB=new He(An,{addons:[],indexedDB:e,IDBKeyRange:t}),n.version(1).stores({dbnames:"name"})),n.table("dbnames")}function Yr(e){return e&&typeof e.databases=="function"}function Sr(e){return Le(function(){return O.letThrough=!0,e()})}function $l(){var e;return!navigator.userAgentData&&/Safari\//.test(navigator.userAgent)&&!/Chrom(e|ium)\//.test(navigator.userAgent)&&indexedDB.databases?new Promise(function(t){var n=function(){return indexedDB.databases().finally(t)};e=setInterval(n,100),n()}).finally(function(){return clearInterval(e)}):Promise.resolve()}function Al(e){const t=e._state,{indexedDB:n}=e._deps;if(t.isBeingOpened||e.idbdb)return t.dbReadyPromise.then(()=>t.dbOpenError?Q(t.dbOpenError):e);me&&(t.openCanceller._stackHolder=Ze()),t.isBeingOpened=!0,t.dbOpenError=null,t.openComplete=!1;const r=t.openCanceller;function i(){if(t.openCanceller!==r)throw new j.DatabaseClosed("db.open() was cancelled")}let o=t.dbReadyResolve,s=null,a=!1;return E.race([r,(typeof navigator>"u"?E.resolve():$l()).then(()=>new E((l,c)=>{if(i(),!n)throw new j.MissingAPI;const u=e.name,f=t.autoSchema?n.open(u):n.open(u,Math.round(10*e.verno));if(!f)throw new j.MissingAPI;f.onerror=de(c),f.onblocked=z(e._fireOnBlocked),f.onupgradeneeded=z(h=>{if(s=f.transaction,t.autoSchema&&!e._options.allowEmptyDB){f.onerror=Tt,s.abort(),f.result.close();const v=n.deleteDatabase(u);v.onsuccess=v.onerror=z(()=>{c(new j.NoSuchDatabase(`Database ${u} doesnt exist`))})}else{s.onerror=de(c);var p=h.oldVersion>Math.pow(2,62)?0:h.oldVersion;a=p<1,e._novip.idbdb=f.result,Cl(e,p/10,s,c)}},c),f.onsuccess=z(()=>{s=null;const h=e._novip.idbdb=f.result,p=dn(h.objectStoreNames);if(p.length>0)try{const d=h.transaction((v=p).length===1?v[0]:v,"readonly");t.autoSchema?function({_novip:y},g,m){y.verno=g.version/10;const b=y._dbSchema=kr(0,g,m);y._storeNames=dn(g.objectStoreNames,0),bn(y,[y._allTables],X(b),b)}(e,h,d):(xr(e,e._dbSchema,d),function(y,g){const m=qo(kr(0,y.idbdb,g),y._dbSchema);return!(m.add.length||m.change.some(b=>b.add.length||b.change.length))}(e,d)||console.warn("Dexie SchemaDiff: Schema was extended without increasing the number passed to db.version(). Some queries may fail.")),br(e,d)}catch{}var v;Et.push(e),h.onversionchange=z(d=>{t.vcFired=!0,e.on("versionchange").fire(d)}),h.onclose=z(d=>{e.on("close").fire(d)}),a&&function({indexedDB:d,IDBKeyRange:y},g){!Yr(d)&&g!==An&&Hr(d,y).put({name:g}).catch(V)}(e._deps,u),l()},c)}))]).then(()=>(i(),t.onReadyBeingFired=[],E.resolve(Sr(()=>e.on.ready.fire(e.vip))).then(function l(){if(t.onReadyBeingFired.length>0){let c=t.onReadyBeingFired.reduce(Ur,V);return t.onReadyBeingFired=[],E.resolve(Sr(()=>c(e.vip))).then(l)}}))).finally(()=>{t.onReadyBeingFired=null,t.isBeingOpened=!1}).then(()=>e).catch(l=>{t.dbOpenError=l;try{s&&s.abort()}catch{}return r===t.openCanceller&&e._close(),Q(l)}).finally(()=>{t.openComplete=!0,o()})}function Er(e){var t=o=>e.next(o),n=i(t),r=i(o=>e.throw(o));function i(o){return s=>{var a=o(s),l=a.value;return a.done?l:l&&typeof l.then=="function"?l.then(n,r):ee(l)?Promise.all(l).then(n,r):n(l)}}return i(t)()}function Ol(e,t,n){var r=arguments.length;if(r<2)throw new j.InvalidArgument("Too few arguments");for(var i=new Array(r-1);--r;)i[r-1]=arguments[r];return n=i.pop(),[e,wo(i),n]}function Uo(e,t,n,r,i){return E.resolve().then(()=>{const o=O.transless||O,s=e._createTransaction(t,n,e._dbSchema,r),a={trans:s,transless:o};if(r)s.idbtrans=r.idbtrans;else try{s.create(),e._state.PR1398_maxLoop=3}catch(f){return f.name===qr.InvalidState&&e.isOpen()&&--e._state.PR1398_maxLoop>0?(console.warn("Dexie: Need to reopen db"),e._close(),e.open().then(()=>Uo(e,t,n,null,i))):Q(f)}const l=Mr(i);let c;l&&mt();const u=E.follow(()=>{if(c=i.call(s,s),c)if(l){var f=Ce.bind(null,null);c.then(f,f)}else typeof c.next=="function"&&typeof c.throw=="function"&&(c=Er(c))},a);return(c&&typeof c.then=="function"?E.resolve(c).then(f=>s.active?f:Q(new j.PrematureCommit("Transaction committed too early. See http://bit.ly/2kdckMn"))):u.then(()=>c)).then(f=>(r&&s._resolve(),s._completion.then(()=>f))).catch(f=>(s._reject(f),Q(f)))})}function Yt(e,t,n){const r=ee(e)?e.slice():[e];for(let i=0;i<n;++i)r.push(t);return r}const Rl={stack:"dbcore",name:"VirtualIndexMiddleware",level:1,create:function(e){return{...e,table(t){const n=e.table(t),{schema:r}=n,i={},o=[];function s(u,f,h){const p=Pt(u),v=i[p]=i[p]||[],d=u==null?0:typeof u=="string"?1:u.length,y=f>0,g={...h,isVirtual:y,keyTail:f,keyLength:d,extractKey:vr(u),unique:!y&&h.unique};return v.push(g),g.isPrimaryKey||o.push(g),d>1&&s(d===2?u[0]:u.slice(0,d-1),f+1,h),v.sort((m,b)=>m.keyTail-b.keyTail),g}const a=s(r.primaryKey.keyPath,0,r.primaryKey);i[":id"]=[a];for(const u of r.indexes)s(u.keyPath,0,u);function l(u){const f=u.query.index;return f.isVirtual?{...u,query:{index:f,range:(h=u.query.range,p=f.keyTail,{type:h.type===1?2:h.type,lower:Yt(h.lower,h.lowerOpen?e.MAX_KEY:e.MIN_KEY,p),lowerOpen:!0,upper:Yt(h.upper,h.upperOpen?e.MIN_KEY:e.MAX_KEY,p),upperOpen:!0})}}:u;var h,p}return{...n,schema:{...r,primaryKey:a,indexes:o,getIndexByKeyPath:function(u){const f=i[Pt(u)];return f&&f[0]}},count:u=>n.count(l(u)),query:u=>n.query(l(u)),openCursor(u){const{keyTail:f,isVirtual:h,keyLength:p}=u.query.index;return h?n.openCursor(l(u)).then(v=>v&&function(d){return Object.create(d,{continue:{value:function(g){g!=null?d.continue(Yt(g,u.reverse?e.MAX_KEY:e.MIN_KEY,f)):u.unique?d.continue(d.key.slice(0,p).concat(u.reverse?e.MIN_KEY:e.MAX_KEY,f)):d.continue()}},continuePrimaryKey:{value(g,m){d.continuePrimaryKey(Yt(g,e.MAX_KEY,f),m)}},primaryKey:{get:()=>d.primaryKey},key:{get(){const g=d.key;return p===1?g[0]:g.slice(0,p)}},value:{get:()=>d.value}})}(v)):n.openCursor(u)}}}}}};function Gr(e,t,n,r){return n=n||{},r=r||"",X(e).forEach(i=>{if(le(t,i)){var o=e[i],s=t[i];if(typeof o=="object"&&typeof s=="object"&&o&&s){const a=ir(o);a!==ir(s)?n[r+i]=t[i]:a==="Object"?Gr(o,s,n,r+i+"."):o!==s&&(n[r+i]=t[i])}else o!==s&&(n[r+i]=t[i])}else n[r+i]=void 0}),X(t).forEach(i=>{le(e,i)||(n[r+i]=t[i])}),n}const Tl={stack:"dbcore",name:"HooksMiddleware",level:2,create:e=>({...e,table(t){const n=e.table(t),{primaryKey:r}=n.schema;return{...n,mutate(o){const s=O.trans,{deleting:a,creating:l,updating:c}=s.table(t).hook;switch(o.type){case"add":if(l.fire===V)break;return s._promise("readwrite",()=>u(o),!0);case"put":if(l.fire===V&&c.fire===V)break;return s._promise("readwrite",()=>u(o),!0);case"delete":if(a.fire===V)break;return s._promise("readwrite",()=>u(o),!0);case"deleteRange":if(a.fire===V)break;return s._promise("readwrite",()=>function(h){return f(h.trans,h.range,1e4)}(o),!0)}return n.mutate(o);function u(h){const p=O.trans,v=h.keys||function(d,y){return y.type==="delete"?y.keys:y.keys||y.values.map(d.extractKey)}(r,h);if(!v)throw new Error("Keys missing");return(h=h.type==="add"||h.type==="put"?{...h,keys:v}:{...h}).type!=="delete"&&(h.values=[...h.values]),h.keys&&(h.keys=[...h.keys]),function(d,y,g){return y.type==="add"?Promise.resolve([]):d.getMany({trans:y.trans,keys:g,cache:"immutable"})}(n,h,v).then(d=>{const y=v.map((g,m)=>{const b=d[m],k={onerror:null,onsuccess:null};if(h.type==="delete")a.fire.call(k,g,b,p);else if(h.type==="add"||b===void 0){const _=l.fire.call(k,g,h.values[m],p);g==null&&_!=null&&(g=_,h.keys[m]=g,r.outbound||ue(h.values[m],r.keyPath,g))}else{const _=Gr(b,h.values[m]),S=c.fire.call(k,_,g,b,p);if(S){const x=h.values[m];Object.keys(S).forEach(T=>{le(x,T)?x[T]=S[T]:ue(x,T,S[T])})}}return k});return n.mutate(h).then(({failures:g,results:m,numFailures:b,lastResult:k})=>{for(let _=0;_<v.length;++_){const S=m?m[_]:v[_],x=y[_];S==null?x.onerror&&x.onerror(g[_]):x.onsuccess&&x.onsuccess(h.type==="put"&&d[_]?h.values[_]:S)}return{failures:g,results:m,numFailures:b,lastResult:k}}).catch(g=>(y.forEach(m=>m.onerror&&m.onerror(g)),Promise.reject(g)))})}function f(h,p,v){return n.query({trans:h,values:!1,query:{index:r,range:p},limit:v}).then(({result:d})=>u({type:"delete",keys:d,trans:h}).then(y=>y.numFailures>0?Promise.reject(y.failures[0]):d.length<v?{failures:[],numFailures:0,lastResult:void 0}:f(h,{...p,lower:d[d.length-1],lowerOpen:!0},v)))}}}}})};function Vo(e,t,n){try{if(!t||t.keys.length<e.length)return null;const r=[];for(let i=0,o=0;i<t.keys.length&&o<e.length;++i)ne(t.keys[i],e[o])===0&&(r.push(n?Lt(t.values[i]):t.values[i]),++o);return r.length===e.length?r:null}catch{return null}}const Il={stack:"dbcore",level:-1,create:e=>({table:t=>{const n=e.table(t);return{...n,getMany:r=>{if(!r.cache)return n.getMany(r);const i=Vo(r.keys,r.trans._cache,r.cache==="clone");return i?E.resolve(i):n.getMany(r).then(o=>(r.trans._cache={keys:r.keys,values:r.cache==="clone"?Lt(o):o},o))},mutate:r=>(r.type!=="add"&&(r.trans._cache=null),n.mutate(r))}}})};function Xr(e){return!("from"in e)}const be=function(e,t){if(!this){const n=new be;return e&&"d"in e&&re(n,e),n}re(this,arguments.length?{d:1,from:e,to:arguments.length>1?t:e}:{d:0})};function Bt(e,t,n){const r=ne(t,n);if(isNaN(r))return;if(r>0)throw RangeError();if(Xr(e))return re(e,{from:t,to:n,d:1});const i=e.l,o=e.r;if(ne(n,e.from)<0)return i?Bt(i,t,n):e.l={from:t,to:n,d:1,l:null,r:null},_i(e);if(ne(t,e.to)>0)return o?Bt(o,t,n):e.r={from:t,to:n,d:1,l:null,r:null},_i(e);ne(t,e.from)<0&&(e.from=t,e.l=null,e.d=o?o.d+1:1),ne(n,e.to)>0&&(e.to=n,e.r=null,e.d=e.l?e.l.d+1:1);const s=!e.r;i&&!e.l&&wn(e,i),o&&s&&wn(e,o)}function wn(e,t){Xr(t)||function n(r,{from:i,to:o,l:s,r:a}){Bt(r,i,o),s&&n(r,s),a&&n(r,a)}(e,t)}function jl(e,t){const n=Cr(t);let r=n.next();if(r.done)return!1;let i=r.value;const o=Cr(e);let s=o.next(i.from),a=s.value;for(;!r.done&&!s.done;){if(ne(a.from,i.to)<=0&&ne(a.to,i.from)>=0)return!0;ne(i.from,a.from)<0?i=(r=n.next(a.from)).value:a=(s=o.next(i.from)).value}return!1}function Cr(e){let t=Xr(e)?null:{s:0,n:e};return{next(n){const r=arguments.length>0;for(;t;)switch(t.s){case 0:if(t.s=1,r)for(;t.n.l&&ne(n,t.n.from)<0;)t={up:t,n:t.n.l,s:1};else for(;t.n.l;)t={up:t,n:t.n.l,s:1};case 1:if(t.s=2,!r||ne(n,t.n.to)<=0)return{value:t.n,done:!1};case 2:if(t.n.r){t.s=3,t={up:t,n:t.n.r,s:0};continue}case 3:t=t.up}return{done:!0}}}}function _i(e){var t,n;const r=(((t=e.r)===null||t===void 0?void 0:t.d)||0)-(((n=e.l)===null||n===void 0?void 0:n.d)||0),i=r>1?"r":r<-1?"l":"";if(i){const o=i==="r"?"l":"r",s={...e},a=e[i];e.from=a.from,e.to=a.to,e[i]=a[i],s[i]=a[o],e[o]=s,s.d=ki(s)}e.d=ki(e)}function ki({r:e,l:t}){return(e?t?Math.max(e.d,t.d):e.d:t?t.d:0)+1}ht(be.prototype,{add(e){return wn(this,e),this},addKey(e){return Bt(this,e,e),this},addKeys(e){return e.forEach(t=>Bt(this,t,t)),this},[or](){return Cr(this)}});const Bl={stack:"dbcore",level:0,create:e=>{const t=e.schema.name,n=new be(e.MIN_KEY,e.MAX_KEY);return{...e,table:r=>{const i=e.table(r),{schema:o}=i,{primaryKey:s}=o,{extractKey:a,outbound:l}=s,c={...i,mutate:h=>{const p=h.trans,v=p.mutatedParts||(p.mutatedParts={}),d=S=>{const x=`idb://${t}/${r}/${S}`;return v[x]||(v[x]=new be)},y=d(""),g=d(":dels"),{type:m}=h;let[b,k]=h.type==="deleteRange"?[h.range]:h.type==="delete"?[h.keys]:h.values.length<50?[[],h.values]:[];const _=h.trans._cache;return i.mutate(h).then(S=>{if(ee(b)){m!=="delete"&&(b=S.results),y.addKeys(b);const x=Vo(b,_);x||m==="add"||g.addKeys(b),(x||k)&&function(T,w,R,P){function A(I){const D=T(I.name||"");function fe(H){return H!=null?I.extractKey(H):null}const he=H=>I.multiEntry&&ee(H)?H.forEach(oe=>D.addKey(oe)):D.addKey(H);(R||P).forEach((H,oe)=>{const ie=R&&fe(R[oe]),et=P&&fe(P[oe]);ne(ie,et)!==0&&(ie!=null&&he(ie),et!=null&&he(et))})}w.indexes.forEach(A)}(d,o,x,k)}else if(b){const x={from:b.lower,to:b.upper};g.add(x),y.add(x)}else y.add(n),g.add(n),o.indexes.forEach(x=>d(x.name).add(n));return S})}},u=({query:{index:h,range:p}})=>{var v,d;return[h,new be((v=p.lower)!==null&&v!==void 0?v:e.MIN_KEY,(d=p.upper)!==null&&d!==void 0?d:e.MAX_KEY)]},f={get:h=>[s,new be(h.key)],getMany:h=>[s,new be().addKeys(h.keys)],count:u,query:u,openCursor:u};return X(f).forEach(h=>{c[h]=function(p){const{subscr:v}=O;if(v){const d=k=>{const _=`idb://${t}/${r}/${k}`;return v[_]||(v[_]=new be)},y=d(""),g=d(":dels"),[m,b]=f[h](p);if(d(m.name||"").add(b),!m.isPrimaryKey){if(h!=="count"){const k=h==="query"&&l&&p.values&&i.query({...p,values:!1});return i[h].apply(this,arguments).then(_=>{if(h==="query"){if(l&&p.values)return k.then(({result:x})=>(y.addKeys(x),_));const S=p.values?_.result.map(a):_.result;p.values?y.addKeys(S):g.addKeys(S)}else if(h==="openCursor"){const S=_,x=p.values;return S&&Object.create(S,{key:{get:()=>(g.addKey(S.primaryKey),S.key)},primaryKey:{get(){const T=S.primaryKey;return g.addKey(T),T}},value:{get:()=>(x&&y.addKey(S.primaryKey),S.value)}})}return _})}g.add(n)}}return i[h].apply(this,arguments)}}),c}}}};class He{constructor(t,n){this._middlewares={},this.verno=0;const r=He.dependencies;this._options=n={addons:He.addons,autoOpen:!0,indexedDB:r.indexedDB,IDBKeyRange:r.IDBKeyRange,...n},this._deps={indexedDB:n.indexedDB,IDBKeyRange:n.IDBKeyRange};const{addons:i}=n;this._dbSchema={},this._versions=[],this._storeNames=[],this._allTables={},this.idbdb=null,this._novip=this;const o={dbOpenError:null,isBeingOpened:!1,onReadyBeingFired:null,openComplete:!1,dbReadyResolve:V,dbReadyPromise:null,cancelOpen:V,openCanceller:null,autoSchema:!0,PR1398_maxLoop:3};var s;o.dbReadyPromise=new E(a=>{o.dbReadyResolve=a}),o.openCanceller=new E((a,l)=>{o.cancelOpen=l}),this._state=o,this.name=t,this.on=Ct(this,"populate","blocked","versionchange","close",{ready:[Ur,V]}),this.on.ready.subscribe=yo(this.on.ready.subscribe,a=>(l,c)=>{He.vip(()=>{const u=this._state;if(u.openComplete)u.dbOpenError||E.resolve().then(l),c&&a(l);else if(u.onReadyBeingFired)u.onReadyBeingFired.push(l),c&&a(l);else{a(l);const f=this;c||a(function h(){f.on.ready.unsubscribe(l),f.on.ready.unsubscribe(h)})}})}),this.Collection=(s=this,gt(vl.prototype,function(a,l){this.db=s;let c=Do,u=null;if(l)try{c=l()}catch(v){u=v}const f=a._ctx,h=f.table,p=h.hook.reading.fire;this._ctx={table:h,index:f.index,isPrimKey:!f.index||h.schema.primKey.keyPath&&f.index===h.schema.primKey.name,range:c,keysOnly:!1,dir:"next",unique:"",algorithm:null,filter:null,replayFilter:null,justLimit:!0,isMatch:null,offset:0,limit:1/0,error:u,or:f.or,valueMapper:p!==Ot?p:null}})),this.Table=function(a){return gt(gl.prototype,function(l,c,u){this.db=a,this._tx=u,this.name=l,this.schema=c,this.hook=a._allTables[l]?a._allTables[l].hook:Ct(null,{creating:[ol,V],reading:[il,Ot],updating:[al,V],deleting:[sl,V]})})}(this),this.Transaction=function(a){return gt(kl.prototype,function(l,c,u,f,h){this.db=a,this.mode=l,this.storeNames=c,this.schema=u,this.chromeTransactionDurability=f,this.idbtrans=null,this.on=Ct(this,"complete","error","abort"),this.parent=h||null,this.active=!0,this._reculock=0,this._blockedFuncs=[],this._resolve=null,this._reject=null,this._waitingFor=null,this._waitingQueue=null,this._spinCount=0,this._completion=new E((p,v)=>{this._resolve=p,this._reject=v}),this._completion.then(()=>{this.active=!1,this.on.complete.fire()},p=>{var v=this.active;return this.active=!1,this.on.error.fire(p),this.parent?this.parent._reject(p):v&&this.idbtrans&&this.idbtrans.abort(),Q(p)})})}(this),this.Version=function(a){return gt(Pl.prototype,function(l){this.db=a,this._cfg={version:l,storesSource:null,dbschema:{},tables:{},contentUpgrade:null}})}(this),this.WhereClause=function(a){return gt(No.prototype,function(l,c,u){this.db=a,this._ctx={table:l,index:c===":id"?null:c,or:u};const f=a._deps.indexedDB;if(!f)throw new j.MissingAPI;this._cmp=this._ascending=f.cmp.bind(f),this._descending=(h,p)=>f.cmp(p,h),this._max=(h,p)=>f.cmp(h,p)>0?h:p,this._min=(h,p)=>f.cmp(h,p)<0?h:p,this._IDBKeyRange=a._deps.IDBKeyRange})}(this),this.on("versionchange",a=>{a.newVersion>0?console.warn(`Another connection wants to upgrade database '${this.name}'. Closing db now to resume the upgrade.`):console.warn(`Another connection wants to delete database '${this.name}'. Closing db now to resume the delete request.`),this.close()}),this.on("blocked",a=>{!a.newVersion||a.newVersion<a.oldVersion?console.warn(`Dexie.delete('${this.name}') was blocked`):console.warn(`Upgrade '${this.name}' blocked by other connection holding version ${a.oldVersion/10}`)}),this._maxKey=jt(n.IDBKeyRange),this._createTransaction=(a,l,c,u)=>new this.Transaction(a,l,c,this._options.chromeTransactionDurability,u),this._fireOnBlocked=a=>{this.on("blocked").fire(a),Et.filter(l=>l.name===this.name&&l!==this&&!l._state.vcFired).map(l=>l.on("versionchange").fire(a))},this.use(Rl),this.use(Tl),this.use(Bl),this.use(Il),this.vip=Object.create(this,{_vip:{value:!0}}),i.forEach(a=>a(this))}version(t){if(isNaN(t)||t<.1)throw new j.Type("Given version is not a positive number");if(t=Math.round(10*t)/10,this.idbdb||this._state.isBeingOpened)throw new j.Schema("Cannot add version when database is open");this.verno=Math.max(this.verno,t);const n=this._versions;var r=n.filter(i=>i._cfg.version===t)[0];return r||(r=new this.Version(t),n.push(r),n.sort(El),r.stores({}),this._state.autoSchema=!1,r)}_whenReady(t){return this.idbdb&&(this._state.openComplete||O.letThrough||this._vip)?t():new E((n,r)=>{if(this._state.openComplete)return r(new j.DatabaseClosed(this._state.dbOpenError));if(!this._state.isBeingOpened){if(!this._options.autoOpen)return void r(new j.DatabaseClosed);this.open().catch(V)}this._state.dbReadyPromise.then(n,r)}).then(t)}use({stack:t,create:n,level:r,name:i}){i&&this.unuse({stack:t,name:i});const o=this._middlewares[t]||(this._middlewares[t]=[]);return o.push({stack:t,create:n,level:r??10,name:i}),o.sort((s,a)=>s.level-a.level),this}unuse({stack:t,name:n,create:r}){return t&&this._middlewares[t]&&(this._middlewares[t]=this._middlewares[t].filter(i=>r?i.create!==r:!!n&&i.name!==n)),this}open(){return Al(this)}_close(){const t=this._state,n=Et.indexOf(this);if(n>=0&&Et.splice(n,1),this.idbdb){try{this.idbdb.close()}catch{}this._novip.idbdb=null}t.dbReadyPromise=new E(r=>{t.dbReadyResolve=r}),t.openCanceller=new E((r,i)=>{t.cancelOpen=i})}close(){this._close();const t=this._state;this._options.autoOpen=!1,t.dbOpenError=new j.DatabaseClosed,t.isBeingOpened&&t.cancelOpen(t.dbOpenError)}delete(){const t=arguments.length>0,n=this._state;return new E((r,i)=>{const o=()=>{this.close();var s=this._deps.indexedDB.deleteDatabase(this.name);s.onsuccess=z(()=>{(function({indexedDB:a,IDBKeyRange:l},c){!Yr(a)&&c!==An&&Hr(a,l).delete(c).catch(V)})(this._deps,this.name),r()}),s.onerror=de(i),s.onblocked=this._fireOnBlocked};if(t)throw new j.InvalidArgument("Arguments not allowed in db.delete()");n.isBeingOpened?n.dbReadyPromise.then(o):o()})}backendDB(){return this.idbdb}isOpen(){return this.idbdb!==null}hasBeenClosed(){const t=this._state.dbOpenError;return t&&t.name==="DatabaseClosed"}hasFailed(){return this._state.dbOpenError!==null}dynamicallyOpened(){return this._state.autoSchema}get tables(){return X(this._allTables).map(t=>this._allTables[t])}transaction(){const t=Ol.apply(this,arguments);return this._transaction.apply(this,t)}_transaction(t,n,r){let i=O.trans;i&&i.db===this&&t.indexOf("!")===-1||(i=null);const o=t.indexOf("?")!==-1;let s,a;t=t.replace("!","").replace("?","");try{if(a=n.map(c=>{var u=c instanceof this.Table?c.name:c;if(typeof u!="string")throw new TypeError("Invalid table argument to Dexie.transaction(). Only Table or String are allowed");return u}),t=="r"||t===Kn)s=Kn;else{if(t!="rw"&&t!=Nn)throw new j.InvalidArgument("Invalid transaction mode: "+t);s=Nn}if(i){if(i.mode===Kn&&s===Nn){if(!o)throw new j.SubTransaction("Cannot enter a sub-transaction with READWRITE mode when parent transaction is READONLY");i=null}i&&a.forEach(c=>{if(i&&i.storeNames.indexOf(c)===-1){if(!o)throw new j.SubTransaction("Table "+c+" not included in parent transaction.");i=null}}),o&&i&&!i.active&&(i=null)}}catch(c){return i?i._promise(null,(u,f)=>{f(c)}):Q(c)}const l=Uo.bind(null,this,s,a,i,r);return i?i._promise(s,l,"lock"):O.trans?yt(O.transless,()=>this._whenReady(l)):this._whenReady(l)}table(t){if(!le(this._allTables,t))throw new j.InvalidTable(`Table ${t} does not exist`);return this._allTables[t]}}const Ll=typeof Symbol<"u"&&"observable"in Symbol?Symbol.observable:"@@observable";class Dl{constructor(t){this._subscribe=t}subscribe(t,n,r){return this._subscribe(t&&typeof t!="function"?t:{next:t,error:n,complete:r})}[Ll](){return this}}function Wo(e,t){return X(t).forEach(n=>{wn(e[n]||(e[n]=new be),t[n])}),e}function Kl(e){let t,n=!1;const r=new Dl(i=>{const o=Mr(e);let s=!1,a={},l={};const c={get closed(){return s},unsubscribe:()=>{s=!0,Ke.storagemutated.unsubscribe(p)}};i.start&&i.start(c);let u=!1,f=!1;function h(){return X(l).some(d=>a[d]&&jl(a[d],l[d]))}const p=d=>{Wo(a,d),h()&&v()},v=()=>{if(u||s)return;a={};const d={},y=function(g){o&&mt();const m=()=>Le(e,{subscr:g,trans:null}),b=O.trans?yt(O.transless,m):m();return o&&b.then(Ce,Ce),b}(d);f||(Ke(It,p),f=!0),u=!0,Promise.resolve(y).then(g=>{n=!0,t=g,u=!1,s||(h()?v():(a={},l=d,i.next&&i.next(g)))},g=>{u=!1,n=!1,i.error&&i.error(g),c.unsubscribe()})};return v(),c});return r.hasValue=()=>n,r.getValue=()=>t,r}let Pr;try{Pr={indexedDB:W.indexedDB||W.mozIndexedDB||W.webkitIndexedDB||W.msIndexedDB,IDBKeyRange:W.IDBKeyRange||W.webkitIDBKeyRange}}catch{Pr={indexedDB:null,IDBKeyRange:null}}const Me=He;function nn(e){let t=_e;try{_e=!0,Ke.storagemutated.fire(e)}finally{_e=t}}ht(Me,{...Xt,delete:e=>new Me(e,{addons:[]}).delete(),exists:e=>new Me(e,{addons:[]}).open().then(t=>(t.close(),!0)).catch("NoSuchDatabaseError",()=>!1),getDatabaseNames(e){try{return function({indexedDB:t,IDBKeyRange:n}){return Yr(t)?Promise.resolve(t.databases()).then(r=>r.map(i=>i.name).filter(i=>i!==An)):Hr(t,n).toCollection().primaryKeys()}(Me.dependencies).then(e)}catch{return Q(new j.MissingAPI)}},defineClass:()=>function(e){re(this,e)},ignoreTransaction:e=>O.trans?yt(O.transless,e):e(),vip:Sr,async:function(e){return function(){try{var t=Er(e.apply(this,arguments));return t&&typeof t.then=="function"?t:E.resolve(t)}catch(n){return Q(n)}}},spawn:function(e,t,n){try{var r=Er(e.apply(n,t||[]));return r&&typeof r.then=="function"?r:E.resolve(r)}catch(i){return Q(i)}},currentTransaction:{get:()=>O.trans||null},waitFor:function(e,t){const n=E.resolve(typeof e=="function"?Me.ignoreTransaction(e):e).timeout(t||6e4);return O.trans?O.trans.waitFor(n):n},Promise:E,debug:{get:()=>me,set:e=>{ko(e,e==="dexie"?()=>!0:Lo)}},derive:at,extend:re,props:ht,override:yo,Events:Ct,on:Ke,liveQuery:Kl,extendObservabilitySet:Wo,getByKeyPath:Se,setByKeyPath:ue,delByKeyPath:function(e,t){typeof t=="string"?ue(e,t,void 0):"length"in t&&[].map.call(t,function(n){ue(e,n,void 0)})},shallowClone:bo,deepClone:Lt,getObjectDiff:Gr,cmp:ne,asap:go,minKey:yr,addons:[],connections:Et,errnames:qr,dependencies:Pr,semVer:mi,version:mi.split(".").map(e=>parseInt(e)).reduce((e,t,n)=>e+t/Math.pow(10,2*n))}),Me.maxKey=jt(Me.dependencies.IDBKeyRange),typeof dispatchEvent<"u"&&typeof addEventListener<"u"&&(Ke(It,e=>{if(!_e){let t;$n?(t=document.createEvent("CustomEvent"),t.initCustomEvent(Oe,!0,!0,e)):t=new CustomEvent(Oe,{detail:e}),_e=!0,dispatchEvent(t),_e=!1}}),addEventListener(Oe,({detail:e})=>{_e||nn(e)}));let _e=!1;if(typeof BroadcastChannel<"u"){const e=new BroadcastChannel(Oe);typeof e.unref=="function"&&e.unref(),Ke(It,t=>{_e||e.postMessage(t)}),e.onmessage=t=>{t.data&&nn(t.data)}}else if(typeof self<"u"&&typeof navigator<"u"){Ke(It,t=>{try{_e||(typeof localStorage<"u"&&localStorage.setItem(Oe,JSON.stringify({trig:Math.random(),changedParts:t})),typeof self.clients=="object"&&[...self.clients.matchAll({includeUncontrolled:!0})].forEach(n=>n.postMessage({type:Oe,changedParts:t})))}catch{}}),typeof addEventListener<"u"&&addEventListener("storage",t=>{if(t.key===Oe){const n=JSON.parse(t.newValue);n&&nn(n.changedParts)}});const e=self.document&&navigator.serviceWorker;e&&e.addEventListener("message",function({data:t}){t&&t.type===Oe&&nn(t.changedParts)})}E.rejectionMapper=function(e,t){if(!e||e instanceof lt||e instanceof TypeError||e instanceof SyntaxError||!e.name||!fi[e.name])return e;var n=new fi[e.name](t||e.message,e);return"stack"in e&&xe(n,"stack",{get:function(){return this.inner.stack}}),n},ko(me,Lo);class Nl extends He{spells;classes;races;backgrounds;items;feats;constructor(t){super(t),this.version(1).stores({spells:"name",classes:"name",races:"name",backgrounds:"name",items:"name",feats:"name"})}}const _n=new Nl("dndSrd"),[xi,Ml]=L([]);function Fl(){const e=Be.toObservable(_n.classes.toArray());return xi().length===0&&e.pipe(Ie(1),hn(t=>t.length>0?ce(t):ce([])),hn(t=>t.length===0?Be.post("/api/DnDInfo/Classes",{}).pipe(Ie(1),Je(n=>(console.error("Error: ",n),ce(null))),je(n=>{n&&_n.classes.bulkAdd(n)})):ce(t)),je(t=>t&&t.length>0?Ml(t):null)).subscribe(),xi}const ql="_primary_1a8rj_1",Ul="_accent_1a8rj_12",Vl="_tertiary_1a8rj_65",Wl="_warn_1a8rj_70",zl="_hover_1a8rj_74",Hl="_popup_1a8rj_88",Yl="_table_1a8rj_100",Fe={primary:ql,accent:Ul,tertiary:Vl,warn:Wl,hover:zl,popup:Hl,table:Yl},Gl="_primary_hhp2e_1",Xl="_accent_hhp2e_7",Ql="_warn_hhp2e_13",Jl="_hover_hhp2e_17",Zl="_popup_hhp2e_28",ec="_table_hhp2e_41",tc="_tBody_hhp2e_55",nc="_THead_hhp2e_60",qe={primary:Gl,accent:Xl,warn:Ql,hover:Jl,popup:Zl,table:ec,tBody:tc,THead:nc},rc={primary:qe.primary,accent:qe.accent,tertiary:qe.tertiary,warn:qe.warn,hover:qe.hover,popup:qe.popup,table:qe.table},Si={primary:Fe.primary,accent:Fe.accent,tertiary:Fe.tertiary,warn:Fe.warn,hover:Fe.hover,popup:Fe.popup,table:Fe.table};function ye(e="dark"){switch(e.toLowerCase()){case"light":return rc;case"dark":return document.body.style.backgroundColor="#212121",Si;default:return Si}}const[Ei,ic]=L([]);function zo(){const e=Be.toObservable(_n.spells.toArray());return Ei().length===0&&e.pipe(Ie(1),hn(t=>t.length>0?ce(t):ce([])),hn(t=>t.length===0?Be.post("/api/DnDInfo/Spells",{}).pipe(Ie(1),Je(n=>(console.error("Error: ",n),ce(null))),je(n=>{n&&_n.spells.bulkAdd(n)})):ce(t)),je(t=>t&&t.length>0?ic(t):null)).subscribe(),Ei}const[oc,sc]=L([]);function ac(){return Be.post("/api/DnDInfo/Feats",{}).pipe(Ie(1),Je(e=>(console.error("Error: ",e),ce(null))),je(e=>e?sc(e):null)).subscribe(),oc}const[lc,cc]=L([]);function uc(){return Be.post("/api/DnDInfo/Races",{}).pipe(Ie(1),Je(e=>(console.error("Error: ",e),ce(null))),je(e=>e?cc(e):null)).subscribe(),lc}const[fc,hc]=L([]);function dc(){return Be.post("/api/DnDInfo/Backgrounds",{}).pipe(Ie(1),Je(e=>(console.error("Error: ",e),ce(null))),je(e=>e?hc(e):null)).subscribe(),fc}const[pc,mc]=L([]);function yc(){return Be.post("/api/DnDInfo/Items",{}).pipe(Ie(1),Je(e=>(console.error("Error: ",e),ce(null))),je(e=>e?mc(e):null)).subscribe(),pc}const gc="_AppBody_1ckpc_1",vc={AppBody:gc};function bc(e={}){const{immediate:t=!1,onNeedRefresh:n,onOfflineReady:r,onRegistered:i,onRegisteredSW:o,onRegisterError:s}=e;let a,l;const c=async(f=!0)=>{await l};async function u(){if("serviceWorker"in navigator){if(a=await kn(()=>import("./workbox-window.prod.es5-HOwwy-x7.js"),__vite__mapDeps([])).then(({Workbox:f})=>new f("/sw.js",{scope:"/",type:"classic"})).catch(f=>{s?.(f)}),!a)return;a.addEventListener("activated",f=>{(f.isUpdate||f.isExternal)&&window.location.reload()}),a.addEventListener("installed",f=>{f.isUpdate||r?.()}),a.register({immediate:t}).then(f=>{o?o("/sw.js",f):i?.(f)}).catch(f=>{s?.(f)})}}return l=u(),c}function Ho(e={}){const{immediate:t=!0,onNeedRefresh:n,onOfflineReady:r,onRegistered:i,onRegisteredSW:o,onRegisterError:s}=e,[a,l]=L(!1),[c,u]=L(!1),f=bc({immediate:t,onOfflineReady(){u(!0),r?.()},onNeedRefresh(){l(!0),n?.()},onRegistered:i,onRegisteredSW:o,onRegisterError:s});return{needRefresh:[a,l],offlineReady:[c,u],updateServiceWorker:f}}const wc="_Container_8liaz_1",_c="_Toast_8liaz_15",kc="_ToastMessage_8liaz_43",xc="_ToastButton_8liaz_51",vt={Container:wc,Toast:_c,ToastMessage:kc,ToastButton:xc};var Sc=K("<span>App ready to work offline"),Ec=K("<button>Reload"),Cc=K("<div><div></div><button>Close"),Pc=K("<div>"),$c=K("<span>New content available, click on reload button to update.");const Ac=()=>{const{offlineReady:[e,t],needRefresh:[n,r],updateServiceWorker:i}=Ho({onRegistered(s){console.log("SW Registered: ",s)},onRegisterError(s){console.log("SW registration error",s)}}),o=()=>{t(!1),r(!1)};return(()=>{var s=Pc();return C(s,$(G,{get when(){return e()||n()},get children(){var a=Cc(),l=a.firstChild,c=l.nextSibling;return C(l,$(G,{get fallback(){return $c()},get when(){return e()},get children(){return Sc()}})),C(a,$(G,{get when(){return n()},get children(){var u=Ec();return u.$$click=()=>i(!0),B(()=>F(u,vt.ToastButton)),u}}),c),c.$$click=()=>o(),B(u=>{var f=vt.Toast,h=vt.Message,p=vt.ToastButton;return f!==u.e&&F(a,u.e=f),h!==u.t&&F(l,u.t=h),p!==u.a&&F(c,u.a=p),u},{e:void 0,t:void 0,a:void 0}),a}})),B(()=>F(s,vt.Container)),s})()};pe(["click"]);var Oc=K("<div><h1>Home");const Rc=()=>{const e=ye();return Fl(),zo(),ac(),uc(),yc(),dc(),(()=>{var t=Oc();return t.firstChild,C(t,$(Ac,{}),null),B(()=>F(t,`${e.accent} ${vc.AppBody}`)),t})()},Tc="_active_8eb29_1",Ic="_inactive_8eb29_6",jc="_navbar_8eb29_11",bt={active:Tc,inactive:Ic,navbar:jc};var rn=(e=>(e[e.DnD5e=0]="DnD5e",e[e.TOValient=1]="TOValient",e))(rn||{});function Bc(e){const[t,n]=L([]),r=rn.DnD5e;return B(()=>{switch(!0){case e().startsWith("/info"):const i=[{Name:"All",Link:"/info"},{Name:"Races",Link:"/info/races"},{Name:"Spells",Link:"/info/spells"},{Name:"Feats",Link:"/info/feats"}];r===rn.DnD5e?i.push({Name:"Backgrounds",Link:"/info/backgrounds"}):r===rn.TOValient&&i.push({Name:"Heritages",Link:"/info/heritages"}),n(i);break;case e().startsWith("/characters"):n([{Name:"All",Link:"/characters"},{Name:"View",Link:"/characters/view"},{Name:"Create",Link:"/characters/create"}]);break;case e().startsWith("/homebrew"):n([{Name:"All",Link:"/homebrew"},{Name:"View",Link:"/homebrew/view"},{Name:"Create",Link:"/homebrew/create"}]);break;default:n([]);break}}),t}function Lc(e,t){const n=document.getElementById("root"),r=i=>!e.contains(i.target)&&t()?.();n.addEventListener("click",r),ut(()=>{n.removeEventListener("click",r)})}var Dc=K("<div>");const Kc=e=>{const t=ye(),[n,r]=e.backgroundClick??L(!0);return $(Bs,{get children(){var i=Dc();return $s(Lc,i,()=>()=>r(!n())),C(i,()=>e.children),B(o=>{var s=e.width,a=e.height,l=`translate(${e.translate.x??"-50%"},${e.translate.y??"-50%"})`,c=`${t.popup} ${t.primary}`;return s!==o.e&&((o.e=s)!=null?i.style.setProperty("width",s):i.style.removeProperty("width")),a!==o.t&&((o.t=a)!=null?i.style.setProperty("height",a):i.style.removeProperty("height")),l!==o.a&&((o.a=l)!=null?i.style.setProperty("transform",l):i.style.removeProperty("transform")),c!==o.o&&F(i,o.o=c),o},{e:void 0,t:void 0,a:void 0,o:void 0}),i}})},Nc="_menuButton_frhu5_1",Mc="_menuButtons_frhu5_10",Ci={menuButton:Nc,menuButtons:Mc};var Fc=K('<button type=button><svg width=32pt height=32pt version=1.1 viewBox="0 0 128 128"xmlns=http://www.w3.org/2000/svg><path d="m113.15 40.062c4.9922 0 8.5742-3.5859 8.5742-8.4492s-3.5859-8.7031-8.5742-8.7031h-98.176c-4.8633 0-8.7031 3.8398-8.7031 8.7031s3.8398 8.4492 8.7031 8.4492zm-98.176 47.875c-4.8633 0-8.7031 3.8398-8.7031 8.5742s3.8398 8.7031 8.7031 8.7031h98.305c4.9922 0 8.5742-3.9688 8.5742-8.7031s-3.5859-8.5742-8.5742-8.5742zm0-32.641c-4.8633 0-8.7031 3.8398-8.7031 8.7031s3.8398 8.7031 8.7031 8.7031h98.305c4.9922 0 8.5742-3.8398 8.5742-8.7031s-3.5859-8.7031-8.5742-8.7031z"></path></svg>'),qc=K("<ul>"),Uc=K("<li><button>");const Vc=()=>{const[e,t]=L(!1),n=ye(),[r,i]=L(!1),o=[{name:"Login",condition:()=>!r(),action:()=>{i(!0)}},{name:"Register",condition:()=>!r(),action:()=>{}},{name:"Logout",condition:r,action:()=>{i(!1)}},{name:"Profile",condition:r,action:()=>{}},{name:"Settings",condition:()=>!0,action:()=>{}}];return[(()=>{var s=Fc();return s.$$click=()=>t(!e()),B(()=>F(s,`${n.accent} ${n.hover}`)),s})(),$(G,{get when(){return e()},get children(){return $(Kc,{width:"10%",height:"25%",translate:{x:"36vw",y:"-40vh"},backgroundClick:[e,t],get children(){var s=qc();return C(s,$(Re,{each:o,children:a=>$(G,{get when(){return a.condition()},get children(){var l=Uc(),c=l.firstChild;return Pn(c,"click",a.action,!0),C(c,()=>a.name),B(u=>{var f=`${100/o.filter(v=>v.condition()).length}%`,h=`${n.hover}`,p=`${n.accent} ${Ci.menuButton}`;return f!==u.e&&((u.e=f)!=null?l.style.setProperty("height",f):l.style.removeProperty("height")),h!==u.t&&F(l,u.t=h),p!==u.a&&F(c,u.a=p),u},{e:void 0,t:void 0,a:void 0}),l}})})),B(()=>F(s,`${Ci.menuButtons}`)),s}})}})]};pe(["click"]);var Wc=K("<span>MySite"),zc=K("<li><div>MySite"),Hc=K("<div><ul>"),Yc=K("<div><div><ul>"),Gc=K("<button>"),Pi=K("<li>");const Xc=e=>{const t=ye();L(!1);const[n,r]=L(window.location.pathname),i=[{Name:"Home",Link:"/"},{Name:"Characters",Link:"/characters"},{Name:"Info",Link:"/info"},{Name:"Homebrew",Link:"/homebrew"}];let o=Bc(n);const[s,a]=e.list,[l,c]=L(!1);return B(()=>{c(!s())}),(()=>{var u=Yc(),f=u.firstChild,h=f.firstChild;return C(f,$(G,{get when(){return!s()},get children(){return Wc()}}),h),C(h,$(G,{get when(){return s()},get children(){return zc()}}),null),C(h,$(Re,{each:i,children:(p,v)=>$(G,{get when(){return!s()||v()===0||l()},get children(){var d=Pi();return C(d,$(si,{onClick:()=>r(p.Link),get href(){return p.Link},get children(){return p.Name}}),null),C(d,$(G,{get when(){return U(()=>!!s())()&&v()===0},get children(){var y=Gc();return y.$$click=()=>c(!l()),y.style.setProperty("background","none"),y.style.setProperty("color","white"),y.style.setProperty("font-size","1.5rem"),y.style.setProperty("padding","0 10px"),y.style.setProperty("width","min-content"),y.style.setProperty("margin-right","0"),C(y,()=>l()?"↑":"↓"),y}}),null),B(()=>F(d,`${t.hover} ${n()==="/"&&p.Link==="/"||n().startsWith(p.Link)&&n()!=="/"&&p.Link!=="/"?bt.active:bt.inactive}`)),d}})}),null),C(f,$(Vc,{}),null),C(u,$(G,{get when(){return o().length>0},get children(){var p=Hc(),v=p.firstChild;return C(v,$(Re,{get each(){return o()},children:d=>(()=>{var y=Pi();return C(y,$(si,{onClick:()=>r(d.Link),get href(){return d.Link},get children(){return d.Name}})),B(()=>F(y,`${t.hover} ${n()===d.Link?bt.active:bt.inactive}`)),y})()})),B(()=>F(p,e.style)),p}}),null),B(p=>{var v=bt.navbar,d=s()?{margin:"0 auto",width:"100%"}:{},y=`${e.style}`,g=s()?{margin:"0 auto",width:"min-content"}:{};return v!==p.e&&F(u,p.e=v),p.t=Gn(f,d,p.t),y!==p.a&&F(f,p.a=y),p.o=Gn(h,g,p.o),p},{e:void 0,t:void 0,a:void 0,o:void 0}),u})()};pe(["click"]);var Qc=()=>["%csolid-devtools","color: #fff; background: #2c4f7c; padding: 1px 4px;"];function Jc(...e){console.warn(...Qc(),...e)}Jc('`import "solid-devtools";` is an entry point for the vite plugin.\nIf you\'re seeing this message you probably forgot to add the plugin to your vite config.');const Zc="_SpellsBody_ump1u_1",eu="_materials_needed_ump1u_1",tu="_TableRow_ump1u_32",nu="_headerBar_ump1u_64",Vn={SpellsBody:Zc,materials_needed:eu,TableRow:tu,headerBar:nu},ru="_TableRow_41syk_1",iu={TableRow:ru};var ou=K("<td><strong>Materials:</strong> "),su=K("<td><strong>higher levels:"),au=K("<td><textarea></textarea>"),lu=K("<tr><td><h2><strong></strong></h2></td><td><strong>Level:</strong> </td><td><strong>School:</strong> </td><td><strong>Casting time:</strong> </td><td><strong>Range:</strong> </td><td><strong>Components:</strong> </td><td><strong>Duration:</strong> </td><td><strong>Classes:</strong> </td><td><strong>Subclasses:</strong> </td><td><strong>Description:</strong></td><td><textarea name=spell_desc>");const cu=e=>{const t=e.spell.isMaterial,n=e.spell.isSomatic,r=e.spell.isVerbal;function i(){let s=[],a=[];return r&&s.push("V"),n&&s.push("S"),t&&s.push("M"),a.push(s.join(", ")),t&&a.push(e.spell.materials_Needed),a}let o=i();return(()=>{var s=lu(),a=s.firstChild,l=a.firstChild,c=l.firstChild,u=a.nextSibling,f=u.firstChild;f.nextSibling;var h=u.nextSibling,p=h.firstChild;p.nextSibling;var v=h.nextSibling,d=v.firstChild;d.nextSibling;var y=v.nextSibling,g=y.firstChild;g.nextSibling;var m=y.nextSibling,b=m.firstChild;b.nextSibling;var k=m.nextSibling,_=k.firstChild;_.nextSibling;var S=k.nextSibling,x=S.firstChild;x.nextSibling;var T=S.nextSibling,w=T.firstChild;w.nextSibling;var R=T.nextSibling,P=R.nextSibling,A=P.firstChild;return C(c,()=>e.spell.name),C(u,()=>e.spell.level,null),C(h,()=>e.spell.school,null),C(v,()=>e.spell.castingTime,null),C(y,()=>e.spell.range,null),C(m,()=>o[0],null),C(s,$(G,{when:t,get children(){var I=ou(),D=I.firstChild;return D.nextSibling,C(I,()=>o[1],null),I}}),k),C(k,()=>e.spell.duration,null),C(S,()=>e.spell.classes.join(", "),null),C(T,()=>e.spell.subClasses.join(", "),null),C(A,()=>e.spell.desc),C(s,$(G,{get when(){return!!e.spell.higherLevel},get children(){return su()}}),null),C(s,$(G,{get when(){return!!e.spell.higherLevel},get children(){var I=au(),D=I.firstChild;return C(D,()=>e.spell.higherLevel),I}}),null),B(()=>F(s,`${iu.TableRow}`)),s})()};var uu=K("<div><button>←←</button><button>←</button><select></select><span> / </span><button>→</button><button>→→"),fu=K("<option>");const hu=e=>{const[t,n]=L(1),[r,i]=L(10),o=e.setPaginatedItems,s=ye(),a=U(()=>e.items().slice((t()-1)*r(),t()*r())),l=U(()=>Math.ceil(e.items().length/r())),c=[10,20,50,100];return B(()=>{o(a())}),(()=>{var u=uu(),f=u.firstChild,h=f.nextSibling,p=h.nextSibling,v=p.nextSibling,d=v.firstChild,y=v.nextSibling,g=y.nextSibling;return f.$$click=()=>n(1),h.$$click=()=>n(t()-1),p.addEventListener("change",m=>i(+m.currentTarget.value)),C(p,$(Re,{each:c,children:m=>(()=>{var b=fu();return b.value=m,C(b,m),b})()})),C(v,t,d),C(v,l,null),y.$$click=()=>n(t()+1),g.$$click=()=>n(l()),B(m=>{var b=`${s.accent} `,k=t()===1,_=t()===1,S=t()===l(),x=t()===l();return b!==m.e&&F(u,m.e=b),k!==m.t&&(f.disabled=m.t=k),_!==m.a&&(h.disabled=m.a=_),S!==m.o&&(y.disabled=m.o=S),x!==m.i&&(g.disabled=m.i=x),m},{e:void 0,t:void 0,a:void 0,o:void 0,i:void 0}),u})()};pe(["click"]);const du="_chip_1bifh_1",pu={chip:du};var mu=K("<span> : <!> <button>x");const yu=e=>{const t=ye();return(()=>{var n=mu(),r=n.firstChild,i=r.nextSibling,o=i.nextSibling,s=o.nextSibling;return C(n,()=>e.key,r),C(n,()=>e.value,i),Pn(s,"click",e.clear,!0),B(()=>F(n,`${t.tertiary} ${pu.chip}`)),n})()};pe(["click"]);const gu="_searchBar_3v5r9_1",vu="_booleanSelect_3v5r9_36",bu="_chipBar_3v5r9_61",Wn={searchBar:gu,booleanSelect:vu,chipBar:bu};var wu=K("<button>Clear All");const _u=e=>{const t=ye();return(()=>{var n=wu();return Pn(n,"click",e.clear,!0),B(()=>F(n,`${t.accent}`)),n})()};pe(["click"]);var ku=K("<select>"),$i=K('<input id=searchBar placeholder="Search Spells..."type=text>'),xu=K("<div id=booleanBar><span><label for=falsebox></label><input type=checkbox id=booleanCheckbox>"),Su=K("<div>"),Eu=K("<div><div><span><select id=chipDropdown>"),Ai=K("<option>"),Cu=K('<svg><path d="M 21 3 C 11.621094 3 4 10.621094 4 20 C 4 29.378906 11.621094 37 21 37 C 24.710938 37 28.140625 35.804688 30.9375 33.78125 L 44.09375 46.90625 L 46.90625 44.09375 L 33.90625 31.0625 C 36.460938 28.085938 38 24.222656 38 20 C 38 10.621094 30.378906 3 21 3 Z M 21 5 C 29.296875 5 36 11.703125 36 20 C 36 28.296875 29.296875 35 21 35 C 12.703125 35 6 28.296875 6 20 C 6 11.703125 12.703125 5 21 5 Z">');const Pu=e=>{const[t,n]=L({key:"",value:""}),[r,i]=L("name"),[o,s]=L(""),[a,l]=L(!1),[c,u]=L([]);ye();const f=["Abjuration","Conjuration","Divination","Enchantment","Evocation","Illusion","Necromancy","Transmutation"];return B(()=>{e.setSearchResults(e.spellsSrd().filter(h=>h&&c().length>0?!c().map(v=>{const d=h[v.key];if(d)switch(!0){case typeof d=="string":return d.toLowerCase().includes(v.value.toLowerCase());case Array.isArray(d):return d.join(" ").toLowerCase().includes(v.value.toLowerCase());case typeof d=="boolean":return` ${d} `.toLowerCase().includes(v.value.toLowerCase())}return!1}).includes(!1):!0))}),B(()=>{typeof e.spellsSrd()[0][r()]=="boolean"&&s(`${a()}`)}),B(()=>{u(h=>t().key?[...h,t()]:h),s(""),document.getElementById("searchBar")&&(document.getElementById("searchBar").value="")}),document.getElementById("booleanCheckbox"),(()=>{var h=Eu(),p=h.firstChild,v=p.firstChild,d=v.firstChild;return C(p,$($u,{onClick:()=>n(y=>({key:r(),value:o()}))}),v),d.addEventListener("change",y=>i(y.target.value)),C(d,$(Re,{get each(){return Object.keys(e.spellsSrd()[0])},children:y=>(()=>{var g=Ai();return g.value=y,C(g,y),g})()})),C(p,$(ei,{get children(){return[$(Mt,{get when(){return typeof e.spellsSrd()[0][r()]=="string"},get children(){return $(ei,{get fallback(){return(()=>{var y=$i();return y.$$keydown=g=>{g.key==="Enter"&&n(m=>({key:r(),value:g.currentTarget.value.trim()}))},y.addEventListener("change",g=>s(g.currentTarget.value)),B(()=>y.value=o()),y})()},get children(){return $(Mt,{get when(){return r()==="school"},get children(){var y=ku();return y.addEventListener("change",g=>s(g.currentTarget.value)),C(y,$(Re,{each:f,children:g=>(()=>{var m=Ai();return m.value=g,C(m,g),m})()})),y}})}})}}),$(Mt,{get when(){return Array.isArray(e.spellsSrd()[0][r()])},get children(){var y=$i();return y.$$keydown=g=>{g.key==="Enter"&&n(m=>({key:r(),value:g.currentTarget.value.trim()}))},y.addEventListener("change",g=>s(g.currentTarget.value)),B(()=>y.value=o()),y}}),$(Mt,{get when(){return typeof e.spellsSrd()[0][r()]=="boolean"},get children(){var y=xu(),g=y.firstChild,m=g.firstChild,b=m.nextSibling;return C(m,$(G,{get when(){return a()===!1},children:"false"}),null),C(m,$(G,{get when(){return a()===!0},children:"true"}),null),b.addEventListener("change",()=>{l(!a()),s(`${a()}`)}),B(()=>F(y,`${Wn.booleanSelect}`)),B(()=>b.checked=a()),y}})]}}),null),C(h,$(G,{get when(){return c().length>0},get children(){var y=Su();return C(y,$(_u,{clear:()=>u([])}),null),C(y,$(Re,{get each(){return c()},children:(g,m)=>$(yu,{get key(){return g.key},get value(){return g.value},clear:()=>u(b=>b.filter((k,_)=>_!==m()))})}),null),B(()=>F(y,`${Wn.chipBar}`)),y}}),null),B(()=>F(p,`${Wn.searchBar}`)),h})()},$u=e=>(()=>{var t=Cu();return Mi(t,fn(e,{xmlns:"http://www.w3.org/2000/svg",x:"0px",y:"0px",width:"100",height:"100",viewBox:"0 0 50 50"}),!0,!0),t})();pe(["keydown"]);var Au=K("<div><h1>Spells</h1><table><tbody></tbody></table><div id=pagginator>"),Ou=K("<span><i>"),Ru=K("<tr><td><span><span></span><span><i></i></span></span><span></span></td><td><button>"),Tu=K("<hr>");const Iu=()=>{const e=ye(),t=zo(),[n,r]=L([]),[i,o]=L([]),[s,a]=L([]),l=f=>c(f)?r(n().filter(h=>h!==f)):r([...n(),f]),c=f=>n().includes(f),u=f=>{switch(f){case"0":return"Cantrip";case"1":return"1st";case"2":return"2nd";case"3":return"3rd";default:return`${f}th`}};return(()=>{var f=Au(),h=f.firstChild,p=h.nextSibling,v=p.firstChild,d=p.nextSibling;return C(f,$(Pu,{searchResults:s,setSearchResults:a,spellsSrd:t}),p),C(v,$(Re,{get each(){return i()},children:(y,g)=>[(()=>{var m=Ru(),b=m.firstChild,k=b.firstChild,_=k.firstChild,S=_.nextSibling,x=S.firstChild,T=k.nextSibling,w=b.nextSibling,R=w.firstChild;return C(_,()=>y.name),C(x,()=>y.school),C(k,$(G,{get when(){return y.concentration},get children(){var P=Ou(),A=P.firstChild;return C(A,()=>y.duration),P}}),null),C(T,()=>u(y.level)),R.$$click=()=>l(g()),C(R,()=>c(g())?"↑":"↓"),B(P=>{var A=`${Vn.TableRow}`,I=Vn.headerBar;return A!==P.e&&F(m,P.e=A),I!==P.t&&F(b,P.t=I),P},{e:void 0,t:void 0}),m})(),$(G,{get when(){return c(g())},get children(){return $(cu,{spell:y})}}),Tu()]})),C(d,$(hu,{items:s,setPaginatedItems:o})),B(y=>{var g=`${e.accent} ${Vn.SpellsBody}`,m=`${e.table}`;return g!==y.e&&F(f,y.e=g),m!==y.t&&F(p,y.t=m),y},{e:void 0,t:void 0}),f})()};pe(["click"]);var ju=K("<div>");const Yo=Rr(()=>kn(()=>import("./characters-lq3nEyhS.js"),__vite__mapDeps([0,1]))),Go=Rr(()=>kn(()=>import("./create-ZPnJCx-R.js"),__vite__mapDeps([2,1]))),Xo=Rr(()=>kn(()=>import("./view-Ae3TD1QO.js"),__vite__mapDeps([3,1]))),Bu=document.getElementById("root"),Lu=20*1e3,{needRefresh:Du,offlineReady:Ku,updateServiceWorker:Mu}=Ho({onRegisteredSW(e,t){t&&setInterval(()=>{Yo.preload(),Go.preload(),Xo.preload(),console.log("onRegisteredSW: ",t),console.log("onRegisteredSW.url: ",e),t.update()},Lu)},onRegisterError(e){console.log("SW registration error",e)},onNeedRefresh(){console.log("onNeedRefresh")},onOfflineReady(){console.log("onOfflineReady")}}),Nu=e=>{const t=ye(),n=function(){let s=!1;return function(a){(/(android|bb\d+|meego).+mobile|avantgo|bada\/|blackberry|blazer|compal|elaine|fennec|hiptop|iemobile|ip(hone|od)|iris|kindle|lge |maemo|midp|mmp|mobile.+firefox|netfront|opera m(ob|in)i|palm( os)?|phone|p(ixi|re)\/|plucker|pocket|psp|series(4|6)0|symbian|treo|up\.(browser|link)|vodafone|wap|windows ce|xda|xiino/i.test(a)||/1207|6310|6590|3gso|4thp|50[1-6]i|770s|802s|a wa|abac|ac(er|oo|s\-)|ai(ko|rn)|al(av|ca|co)|amoi|an(ex|ny|yw)|aptu|ar(ch|go)|as(te|us)|attw|au(di|\-m|r |s )|avan|be(ck|ll|nq)|bi(lb|rd)|bl(ac|az)|br(e|v)w|bumb|bw\-(n|u)|c55\/|capi|ccwa|cdm\-|cell|chtm|cldc|cmd\-|co(mp|nd)|craw|da(it|ll|ng)|dbte|dc\-s|devi|dica|dmob|do(c|p)o|ds(12|\-d)|el(49|ai)|em(l2|ul)|er(ic|k0)|esl8|ez([4-7]0|os|wa|ze)|fetc|fly(\-|_)|g1 u|g560|gene|gf\-5|g\-mo|go(\.w|od)|gr(ad|un)|haie|hcit|hd\-(m|p|t)|hei\-|hi(pt|ta)|hp( i|ip)|hs\-c|ht(c(\-| |_|a|g|p|s|t)|tp)|hu(aw|tc)|i\-(20|go|ma)|i230|iac( |\-|\/)|ibro|idea|ig01|ikom|im1k|inno|ipaq|iris|ja(t|v)a|jbro|jemu|jigs|kddi|keji|kgt( |\/)|klon|kpt |kwc\-|kyo(c|k)|le(no|xi)|lg( g|\/(k|l|u)|50|54|\-[a-w])|libw|lynx|m1\-w|m3ga|m50\/|ma(te|ui|xo)|mc(01|21|ca)|m\-cr|me(rc|ri)|mi(o8|oa|ts)|mmef|mo(01|02|bi|de|do|t(\-| |o|v)|zz)|mt(50|p1|v )|mwbp|mywa|n10[0-2]|n20[2-3]|n30(0|2)|n50(0|2|5)|n7(0(0|1)|10)|ne((c|m)\-|on|tf|wf|wg|wt)|nok(6|i)|nzph|o2im|op(ti|wv)|oran|owg1|p800|pan(a|d|t)|pdxg|pg(13|\-([1-8]|c))|phil|pire|pl(ay|uc)|pn\-2|po(ck|rt|se)|prox|psio|pt\-g|qa\-a|qc(07|12|21|32|60|\-[2-7]|i\-)|qtek|r380|r600|raks|rim9|ro(ve|zo)|s55\/|sa(ge|ma|mm|ms|ny|va)|sc(01|h\-|oo|p\-)|sdk\/|se(c(\-|0|1)|47|mc|nd|ri)|sgh\-|shar|sie(\-|m)|sk\-0|sl(45|id)|sm(al|ar|b3|it|t5)|so(ft|ny)|sp(01|h\-|v\-|v )|sy(01|mb)|t2(18|50)|t6(00|10|18)|ta(gt|lk)|tcl\-|tdg\-|tel(i|m)|tim\-|t\-mo|to(pl|sh)|ts(70|m\-|m3|m5)|tx\-9|up(\.b|g1|si)|utst|v400|v750|veri|vi(rg|te)|vk(40|5[0-3]|\-v)|vm40|voda|vulc|vx(52|53|60|61|70|80|81|83|85|98)|w3c(\-| )|webc|whit|wi(g |nc|nw)|wmlb|wonu|x700|yas\-|your|zeto|zte\-/i.test(a.substr(0,4)))&&(s=!0)}(navigator.userAgent||navigator.vendor),s},r=()=>o(window.matchMedia("only screen and (max-width: 768px)").matches);window.addEventListener("resize",r),B(()=>{console.log("updateServiceWorker.needRefresh: ",Du[0]())}),B(()=>{console.log("updateServiceWorker.offlineReady: ",Ku[0]())});const[i,o]=L(n());return B(()=>{const s=!!(n()||window.innerWidth<=768);o(s)}),(()=>{var s=ju();return s.style.setProperty("height","100vh"),C(s,$(Xc,{list:[i,o],get style(){return t.accent}}),null),C(s,()=>e.children,null),B(()=>F(s,t.primary)),s})()};Es(()=>$(fa,{root:Nu,get children(){return[$(Ne,{path:"/",component:Rc}),$(Ne,{path:"/characters",get children(){return[$(Ne,{path:"/",component:Yo}),$(Ne,{path:"/view",component:Xo}),$(Ne,{path:"/create",component:Go})]}}),$(Ne,{path:"/info",get children(){return $(Ne,{path:"/spells",component:Iu})}})]}}),Bu);export{F as a,B as c,K as t,ye as u};
-function __vite__mapDeps(indexes) {
-  if (!__vite__mapDeps.viteFileDeps) {
-    __vite__mapDeps.viteFileDeps = ["assets/characters-lq3nEyhS.js","assets/characters-i2fzY9Kk.css","assets/create-ZPnJCx-R.js","assets/view-Ae3TD1QO.js"]
-  }
-  return indexes.map((i) => __vite__mapDeps.viteFileDeps[i])
-}
