+++ wwwroot/assets/index-QW1j2Poi.js
@@ -1,0 +1,15 @@
+(function(){const t=document.createElement("link").relList;if(t&&t.supports&&t.supports("modulepreload"))return;for(const i of document.querySelectorAll('link[rel="modulepreload"]'))r(i);new MutationObserver(i=>{for(const o of i)if(o.type==="childList")for(const s of o.addedNodes)s.tagName==="LINK"&&s.rel==="modulepreload"&&r(s)}).observe(document,{childList:!0,subtree:!0});function n(i){const o={};return i.integrity&&(o.integrity=i.integrity),i.referrerPolicy&&(o.referrerPolicy=i.referrerPolicy),i.crossOrigin==="use-credentials"?o.credentials="include":i.crossOrigin==="anonymous"?o.credentials="omit":o.credentials="same-origin",o}function r(i){if(i.ep)return;i.ep=!0;const o=n(i);fetch(i.href,o)}})();const Zo="modulepreload",es=function(e){return"/"+e},Jr={},kn=function(t,n,r){let i=Promise.resolve();if(n&&n.length>0){const o=document.getElementsByTagName("link");i=Promise.all(n.map(s=>{if(s=es(s),s in Jr)return;Jr[s]=!0;const a=s.endsWith(".css"),l=a?'[rel="stylesheet"]':"";if(!!r)for(let f=o.length-1;f>=0;f--){const h=o[f];if(h.href===s&&(!a||h.rel==="stylesheet"))return}else if(document.querySelector(`link[href="${s}"]${l}`))return;const u=document.createElement("link");if(u.rel=a?"stylesheet":Zo,a||(u.as="script",u.crossOrigin=""),u.href=s,document.head.appendChild(u),a)return new Promise((f,h)=>{u.addEventListener("load",f),u.addEventListener("error",()=>h(new Error(`Unable to preload CSS for ${s}`)))})}))}return i.then(()=>t()).catch(o=>{const s=new Event("vite:preloadError",{cancelable:!0});if(s.payload=o,window.dispatchEvent(s),!s.defaultPrevented)throw o})},N={context:void 0,registry:void 0};function _t(e){N.context=e}const ts=(e,t)=>e===t,on=Symbol("solid-proxy"),ns=Symbol("solid-track"),sn={equals:ts};let Ri=Di;const Pe=1,an=2,Ti={owned:null,cleanups:null,context:null,owner:null},On={};var M=null;let Rn=null,rs=null,q=null,te=null,ke=null,xn=0;function ot(e,t){const n=q,r=M,i=e.length===0,o=t===void 0?r:t,s=i?Ti:{owned:null,cleanups:null,context:o?o.context:null,owner:o},a=i?e:()=>e(()=>J(()=>Cn(s)));M=s,q=null;try{return Ee(a,!0)}finally{q=n,M=r}}function L(e,t){t=t?Object.assign({},sn,t):sn;const n={value:e,observers:null,observerSlots:null,comparator:t.equals||void 0},r=i=>(typeof i=="function"&&(i=i(n.value)),Li(n,i));return[Bi.bind(n),r]}function Zr(e,t,n){const r=En(e,t,!0,Pe);dt(r)}function B(e,t,n){const r=En(e,t,!1,Pe);dt(r)}function is(e,t,n){Ri=fs;const r=En(e,t,!1,Pe);(!n||!n.render)&&(r.user=!0),ke?ke.push(r):dt(r)}function U(e,t,n){n=n?Object.assign({},sn,n):sn;const r=En(e,t,!0,0);return r.observers=null,r.observerSlots=null,r.comparator=n.equals||void 0,dt(r),Bi.bind(r)}function os(e){return e&&typeof e=="object"&&"then"in e}function ss(e,t,n){let r,i,o;arguments.length===2&&typeof t=="object"||arguments.length===1?(r=!0,i=e,o=t||{}):(r=e,i=t,o=n||{});let s=null,a=On,l=null,c=!1,u="initialValue"in o,f=typeof r=="function"&&U(r);const h=new Set,[d,v]=(o.storage||L)(o.initialValue),[m,g]=L(void 0),[y,p]=L(void 0,{equals:!1}),[b,w]=L(u?"ready":"unresolved");if(N.context){l=`${N.context.id}${N.context.count++}`;let _;o.ssrLoadFrom==="initial"?a=o.initialValue:N.load&&(_=N.load(l))&&(a=_)}function k(_,R,$,A){return s===_&&(s=null,A!==void 0&&(u=!0),(_===a||R===a)&&o.onHydrated&&queueMicrotask(()=>o.onHydrated(A,{value:R})),a=On,S(R,$)),R}function S(_,R){Ee(()=>{R===void 0&&v(()=>_),w(R!==void 0?"errored":u?"ready":"unresolved"),g(R);for(const $ of h.keys())$.decrement();h.clear()},!1)}function x(){const _=ls,R=d(),$=m();if($!==void 0&&!s)throw $;return q&&!q.user&&_&&Zr(()=>{y(),s&&(_.resolved||h.has(_)||(_.increment(),h.add(_)))}),R}function T(_=!0){if(_!==!1&&c)return;c=!1;const R=f?f():r;if(R==null||R===!1){k(s,J(d));return}const $=a!==On?a:J(()=>i(R,{value:d(),refetching:_}));return os($)?(s=$,"value"in $?($.status==="success"?k(s,$.value,void 0,R):k(s,void 0,void 0,R),$):(c=!0,queueMicrotask(()=>c=!1),Ee(()=>{w(u?"refreshing":"pending"),p()},!1),$.then(A=>k($,A,void 0,R),A=>k($,void 0,Ni(A),R)))):(k(s,$,void 0,R),$)}return Object.defineProperties(x,{state:{get:()=>b()},error:{get:()=>m()},loading:{get(){const _=b();return _==="pending"||_==="refreshing"}},latest:{get(){if(!u)return x();const _=m();if(_&&!s)throw _;return d()}}}),f?Zr(()=>T(!1)):T(!1),[x,{refetch:T,mutate:v}]}function J(e){if(q===null)return e();const t=q;q=null;try{return e()}finally{q=t}}function Ii(e,t,n){const r=Array.isArray(e);let i,o=n&&n.defer;return s=>{let a;if(r){a=Array(e.length);for(let c=0;c<e.length;c++)a[c]=e[c]()}else a=e();if(o){o=!1;return}const l=J(()=>t(a,i,s));return i=a,l}}function ut(e){return M===null||(M.cleanups===null?M.cleanups=[e]:M.cleanups.push(e)),e}function ln(){return M}function Ar(e,t){const n=M,r=q;M=e,q=null;try{return Ee(t,!0)}catch(i){Rr(i)}finally{M=n,q=r}}function as(e){const t=q,n=M;return Promise.resolve().then(()=>{q=t,M=n;let r;return Ee(e,!1),q=M=null,r?r.done:void 0})}function ji(e,t){const n=Symbol("context");return{id:n,Provider:hs(n),defaultValue:e}}function Or(e){return M&&M.context&&M.context[e.id]!==void 0?M.context[e.id]:e.defaultValue}function Sn(e){const t=U(e),n=U(()=>Hn(t()));return n.toArray=()=>{const r=n();return Array.isArray(r)?r:r!=null?[r]:[]},n}let ls;function Bi(){if(this.sources&&this.state)if(this.state===Pe)dt(this);else{const e=te;te=null,Ee(()=>un(this),!1),te=e}if(q){const e=this.observers?this.observers.length:0;q.sources?(q.sources.push(this),q.sourceSlots.push(e)):(q.sources=[this],q.sourceSlots=[e]),this.observers?(this.observers.push(q),this.observerSlots.push(q.sources.length-1)):(this.observers=[q],this.observerSlots=[q.sources.length-1])}return this.value}function Li(e,t,n){let r=e.value;return(!e.comparator||!e.comparator(r,t))&&(e.value=t,e.observers&&e.observers.length&&Ee(()=>{for(let i=0;i<e.observers.length;i+=1){const o=e.observers[i],s=Rn&&Rn.running;s&&Rn.disposed.has(o),(s?!o.tState:!o.state)&&(o.pure?te.push(o):ke.push(o),o.observers&&Ki(o)),s||(o.state=Pe)}if(te.length>1e6)throw te=[],new Error},!1)),t}function dt(e){if(!e.fn)return;Cn(e);const t=xn;cs(e,e.value,t)}function cs(e,t,n){let r;const i=M,o=q;q=M=e;try{r=e.fn(t)}catch(s){return e.pure&&(e.state=Pe,e.owned&&e.owned.forEach(Cn),e.owned=null),e.updatedAt=n+1,Rr(s)}finally{q=o,M=i}(!e.updatedAt||e.updatedAt<=n)&&(e.updatedAt!=null&&"observers"in e?Li(e,r):e.value=r,e.updatedAt=n)}function En(e,t,n,r=Pe,i){const o={fn:e,state:r,updatedAt:null,owned:null,sources:null,sourceSlots:null,cleanups:null,value:t,owner:M,context:M?M.context:null,pure:n};return M===null||M!==Ti&&(M.owned?M.owned.push(o):M.owned=[o]),o}function cn(e){if(e.state===0)return;if(e.state===an)return un(e);if(e.suspense&&J(e.suspense.inFallback))return e.suspense.effects.push(e);const t=[e];for(;(e=e.owner)&&(!e.updatedAt||e.updatedAt<xn);)e.state&&t.push(e);for(let n=t.length-1;n>=0;n--)if(e=t[n],e.state===Pe)dt(e);else if(e.state===an){const r=te;te=null,Ee(()=>un(e,t[0]),!1),te=r}}function Ee(e,t){if(te)return e();let n=!1;t||(te=[]),ke?n=!0:ke=[],xn++;try{const r=e();return us(n),r}catch(r){n||(ke=null),te=null,Rr(r)}}function us(e){if(te&&(Di(te),te=null),e)return;const t=ke;ke=null,t.length&&Ee(()=>Ri(t),!1)}function Di(e){for(let t=0;t<e.length;t++)cn(e[t])}function fs(e){let t,n=0;for(t=0;t<e.length;t++){const r=e[t];r.user?e[n++]=r:cn(r)}if(N.context){if(N.count){N.effects||(N.effects=[]),N.effects.push(...e.slice(0,n));return}else N.effects&&(e=[...N.effects,...e],n+=N.effects.length,delete N.effects);_t()}for(t=0;t<n;t++)cn(e[t])}function un(e,t){e.state=0;for(let n=0;n<e.sources.length;n+=1){const r=e.sources[n];if(r.sources){const i=r.state;i===Pe?r!==t&&(!r.updatedAt||r.updatedAt<xn)&&cn(r):i===an&&un(r,t)}}}function Ki(e){for(let t=0;t<e.observers.length;t+=1){const n=e.observers[t];n.state||(n.state=an,n.pure?te.push(n):ke.push(n),n.observers&&Ki(n))}}function Cn(e){let t;if(e.sources)for(;e.sources.length;){const n=e.sources.pop(),r=e.sourceSlots.pop(),i=n.observers;if(i&&i.length){const o=i.pop(),s=n.observerSlots.pop();r<i.length&&(o.sourceSlots[s]=r,i[r]=o,n.observerSlots[r]=s)}}if(e.owned){for(t=e.owned.length-1;t>=0;t--)Cn(e.owned[t]);e.owned=null}if(e.cleanups){for(t=e.cleanups.length-1;t>=0;t--)e.cleanups[t]();e.cleanups=null}e.state=0}function Ni(e){return e instanceof Error?e:new Error(typeof e=="string"?e:"Unknown error",{cause:e})}function Rr(e,t=M){throw Ni(e)}function Hn(e){if(typeof e=="function"&&!e.length)return Hn(e());if(Array.isArray(e)){const t=[];for(let n=0;n<e.length;n++){const r=Hn(e[n]);Array.isArray(r)?t.push.apply(t,r):t.push(r)}return t}return e}function hs(e,t){return function(r){let i;return B(()=>i=J(()=>(M.context={...M.context,[e]:r.value},Sn(()=>r.children))),void 0),i}}const ds=Symbol("fallback");function ei(e){for(let t=0;t<e.length;t++)e[t]()}function ps(e,t,n={}){let r=[],i=[],o=[],s=0,a=t.length>1?[]:null;return ut(()=>ei(o)),()=>{let l=e()||[],c,u;return l[ns],J(()=>{let h=l.length,d,v,m,g,y,p,b,w,k;if(h===0)s!==0&&(ei(o),o=[],r=[],i=[],s=0,a&&(a=[])),n.fallback&&(r=[ds],i[0]=ot(S=>(o[0]=S,n.fallback())),s=1);else if(s===0){for(i=new Array(h),u=0;u<h;u++)r[u]=l[u],i[u]=ot(f);s=h}else{for(m=new Array(h),g=new Array(h),a&&(y=new Array(h)),p=0,b=Math.min(s,h);p<b&&r[p]===l[p];p++);for(b=s-1,w=h-1;b>=p&&w>=p&&r[b]===l[w];b--,w--)m[w]=i[b],g[w]=o[b],a&&(y[w]=a[b]);for(d=new Map,v=new Array(w+1),u=w;u>=p;u--)k=l[u],c=d.get(k),v[u]=c===void 0?-1:c,d.set(k,u);for(c=p;c<=b;c++)k=r[c],u=d.get(k),u!==void 0&&u!==-1?(m[u]=i[c],g[u]=o[c],a&&(y[u]=a[c]),u=v[u],d.set(k,u)):o[c]();for(u=p;u<h;u++)u in m?(i[u]=m[u],o[u]=g[u],a&&(a[u]=y[u],a[u](u))):i[u]=ot(f);i=i.slice(0,s=h),r=l.slice(0)}return i});function f(h){if(o[u]=h,a){const[d,v]=L(u);return a[u]=v,t(l[u],d)}return t(l[u])}}}function P(e,t){return J(()=>e(t||{}))}function Mt(){return!0}const Yn={get(e,t,n){return t===on?n:e.get(t)},has(e,t){return t===on?!0:e.has(t)},set:Mt,deleteProperty:Mt,getOwnPropertyDescriptor(e,t){return{configurable:!0,enumerable:!0,get(){return e.get(t)},set:Mt,deleteProperty:Mt}},ownKeys(e){return e.keys()}};function Tn(e){return(e=typeof e=="function"?e():e)?e:{}}function ms(){for(let e=0,t=this.length;e<t;++e){const n=this[e]();if(n!==void 0)return n}}function fn(...e){let t=!1;for(let s=0;s<e.length;s++){const a=e[s];t=t||!!a&&on in a,e[s]=typeof a=="function"?(t=!0,U(a)):a}if(t)return new Proxy({get(s){for(let a=e.length-1;a>=0;a--){const l=Tn(e[a])[s];if(l!==void 0)return l}},has(s){for(let a=e.length-1;a>=0;a--)if(s in Tn(e[a]))return!0;return!1},keys(){const s=[];for(let a=0;a<e.length;a++)s.push(...Object.keys(Tn(e[a])));return[...new Set(s)]}},Yn);const n={},r=Object.create(null);for(let s=e.length-1;s>=0;s--){const a=e[s];if(!a)continue;const l=Object.getOwnPropertyNames(a);for(let c=l.length-1;c>=0;c--){const u=l[c];if(u==="__proto__"||u==="constructor")continue;const f=Object.getOwnPropertyDescriptor(a,u);if(!r[u])r[u]=f.get?{enumerable:!0,configurable:!0,get:ms.bind(n[u]=[f.get.bind(a)])}:f.value!==void 0?f:void 0;else{const h=n[u];h&&(f.get?h.push(f.get.bind(a)):f.value!==void 0&&h.push(()=>f.value))}}}const i={},o=Object.keys(r);for(let s=o.length-1;s>=0;s--){const a=o[s],l=r[a];l&&l.get?Object.defineProperty(i,a,l):i[a]=l?l.value:void 0}return i}function ys(e,...t){if(on in e){const i=new Set(t.length>1?t.flat():t[0]),o=t.map(s=>new Proxy({get(a){return s.includes(a)?e[a]:void 0},has(a){return s.includes(a)&&a in e},keys(){return s.filter(a=>a in e)}},Yn));return o.push(new Proxy({get(s){return i.has(s)?void 0:e[s]},has(s){return i.has(s)?!1:s in e},keys(){return Object.keys(e).filter(s=>!i.has(s))}},Yn)),o}const n={},r=t.map(()=>({}));for(const i of Object.getOwnPropertyNames(e)){const o=Object.getOwnPropertyDescriptor(e,i),s=!o.get&&!o.set&&o.enumerable&&o.writable&&o.configurable;let a=!1,l=0;for(const c of t)c.includes(i)&&(a=!0,s?r[l][i]=o.value:Object.defineProperty(r[l],i,o)),++l;a||(s?n[i]=o.value:Object.defineProperty(n,i,o))}return[...r,n]}function Tr(e){let t,n;const r=i=>{const o=N.context;if(o){const[a,l]=L();N.count||(N.count=0),N.count++,(n||(n=e())).then(c=>{_t(o),N.count--,l(()=>c.default),_t()}),t=a}else if(!t){const[a]=ss(()=>(n||(n=e())).then(l=>l.default));t=a}let s;return U(()=>(s=t())&&J(()=>{if(!o)return s(i);const a=N.context;_t(o);const l=s(i);return _t(a),l}))};return r.preload=()=>n||((n=e()).then(i=>t=()=>i.default),n),r}const Mi=e=>`Stale read from <${e}>.`;function Re(e){const t="fallback"in e&&{fallback:()=>e.fallback};return U(ps(()=>e.each,e.children,t||void 0))}function G(e){const t=e.keyed,n=U(()=>e.when,void 0,{equals:(r,i)=>t?r===i:!r==!i});return U(()=>{const r=n();if(r){const i=e.children;return typeof i=="function"&&i.length>0?J(()=>i(t?r:()=>{if(!J(n))throw Mi("Show");return e.when})):i}return e.fallback},void 0,void 0)}function ti(e){let t=!1;const n=(o,s)=>(t?o[1]===s[1]:!o[1]==!s[1])&&o[2]===s[2],r=Sn(()=>e.children),i=U(()=>{let o=r();Array.isArray(o)||(o=[o]);for(let s=0;s<o.length;s++){const a=o[s].when;if(a)return t=!!o[s].keyed,[s,a,o[s]]}return[-1]},void 0,{equals:n});return U(()=>{const[o,s,a]=i();if(o<0)return e.fallback;const l=a.children;return typeof l=="function"&&l.length>0?J(()=>l(t?s:()=>{if(J(i)[0]!==o)throw Mi("Match");return a.when})):l},void 0,void 0)}function gt(e){return e}const gs=["allowfullscreen","async","autofocus","autoplay","checked","controls","default","disabled","formnovalidate","hidden","indeterminate","inert","ismap","loop","multiple","muted","nomodule","novalidate","open","playsinline","readonly","required","reversed","seamless","selected"],vs=new Set(["className","value","readOnly","formNoValidate","isMap","noModule","playsInline",...gs]),bs=new Set(["innerHTML","textContent","innerText","children"]),ws=Object.assign(Object.create(null),{className:"class",htmlFor:"for"}),_s=Object.assign(Object.create(null),{class:"className",formnovalidate:{$:"formNoValidate",BUTTON:1,INPUT:1},ismap:{$:"isMap",IMG:1},nomodule:{$:"noModule",SCRIPT:1},playsinline:{$:"playsInline",VIDEO:1},readonly:{$:"readOnly",INPUT:1,TEXTAREA:1}});function ks(e,t){const n=_s[e];return typeof n=="object"?n[t]?n.$:void 0:n}const xs=new Set(["beforeinput","click","dblclick","contextmenu","focusin","focusout","input","keydown","keyup","mousedown","mousemove","mouseout","mouseover","mouseup","pointerdown","pointermove","pointerout","pointerover","pointerup","touchend","touchmove","touchstart"]),Ss={xlink:"http://www.w3.org/1999/xlink",xml:"http://www.w3.org/XML/1998/namespace"};function Es(e,t,n){let r=n.length,i=t.length,o=r,s=0,a=0,l=t[i-1].nextSibling,c=null;for(;s<i||a<o;){if(t[s]===n[a]){s++,a++;continue}for(;t[i-1]===n[o-1];)i--,o--;if(i===s){const u=o<r?a?n[a-1].nextSibling:n[o-a]:l;for(;a<o;)e.insertBefore(n[a++],u)}else if(o===a)for(;s<i;)(!c||!c.has(t[s]))&&t[s].remove(),s++;else if(t[s]===n[o-1]&&n[a]===t[i-1]){const u=t[--i].nextSibling;e.insertBefore(n[a++],t[s++].nextSibling),e.insertBefore(n[--o],u),t[i]=n[o]}else{if(!c){c=new Map;let f=a;for(;f<o;)c.set(n[f],f++)}const u=c.get(t[s]);if(u!=null)if(a<u&&u<o){let f=s,h=1,d;for(;++f<i&&f<o&&!((d=c.get(t[f]))==null||d!==u+h);)h++;if(h>u-a){const v=t[s];for(;a<u;)e.insertBefore(n[a++],v)}else e.replaceChild(n[a++],t[s++])}else s++;else t[s++].remove()}}}const ni="_$DX_DELEGATE";function Cs(e,t,n,r={}){let i;return ot(o=>{i=o,t===document?e():C(t,e(),t.firstChild?null:void 0,n)},r.owner),()=>{i(),t.textContent=""}}function K(e,t,n){let r;const i=()=>{const s=document.createElement("template");return s.innerHTML=e,n?s.content.firstChild.firstChild:s.content.firstChild},o=t?()=>J(()=>document.importNode(r||(r=i()),!0)):()=>(r||(r=i())).cloneNode(!0);return o.cloneNode=o,o}function pe(e,t=window.document){const n=t[ni]||(t[ni]=new Set);for(let r=0,i=e.length;r<i;r++){const o=e[r];n.has(o)||(n.add(o),t.addEventListener(o,Ts))}}function Gn(e,t,n){N.context||(n==null?e.removeAttribute(t):e.setAttribute(t,n))}function Ps(e,t,n,r){N.context||(r==null?e.removeAttributeNS(t,n):e.setAttributeNS(t,n,r))}function F(e,t){N.context||(t==null?e.removeAttribute("class"):e.className=t)}function Pn(e,t,n,r){if(r)Array.isArray(n)?(e[`$$${t}`]=n[0],e[`$$${t}Data`]=n[1]):e[`$$${t}`]=n;else if(Array.isArray(n)){const i=n[0];e.addEventListener(t,n[0]=o=>i.call(e,n[1],o))}else e.addEventListener(t,n)}function $s(e,t,n={}){const r=Object.keys(t||{}),i=Object.keys(n);let o,s;for(o=0,s=i.length;o<s;o++){const a=i[o];!a||a==="undefined"||t[a]||(ri(e,a,!1),delete n[a])}for(o=0,s=r.length;o<s;o++){const a=r[o],l=!!t[a];!a||a==="undefined"||n[a]===l||!l||(ri(e,a,!0),n[a]=l)}return n}function Xn(e,t,n){if(!t)return n?Gn(e,"style"):t;const r=e.style;if(typeof t=="string")return r.cssText=t;typeof n=="string"&&(r.cssText=n=void 0),n||(n={}),t||(t={});let i,o;for(o in n)t[o]==null&&r.removeProperty(o),delete n[o];for(o in t)i=t[o],i!==n[o]&&(r.setProperty(o,i),n[o]=i);return n}function Fi(e,t={},n,r){const i={};return r||B(()=>i.children=ft(e,t.children,i.children)),B(()=>t.ref&&t.ref(e)),B(()=>Os(e,t,n,!0,i,!0)),i}function As(e,t,n){return J(()=>e(t,n))}function C(e,t,n,r){if(n!==void 0&&!r&&(r=[]),typeof t!="function")return ft(e,t,r,n);B(i=>ft(e,t(),i,n),r)}function Os(e,t,n,r,i={},o=!1){t||(t={});for(const s in i)if(!(s in t)){if(s==="children")continue;i[s]=ii(e,s,null,i[s],n,o)}for(const s in t){if(s==="children"){r||ft(e,t.children);continue}const a=t[s];i[s]=ii(e,s,a,i[s],n,o)}}function Rs(e){return e.toLowerCase().replace(/-([a-z])/g,(t,n)=>n.toUpperCase())}function ri(e,t,n){const r=t.trim().split(/\s+/);for(let i=0,o=r.length;i<o;i++)e.classList.toggle(r[i],n)}function ii(e,t,n,r,i,o){let s,a,l,c,u;if(t==="style")return Xn(e,n,r);if(t==="classList")return $s(e,n,r);if(n===r)return r;if(t==="ref")o||n(e);else if(t.slice(0,3)==="on:"){const f=t.slice(3);r&&e.removeEventListener(f,r),n&&e.addEventListener(f,n)}else if(t.slice(0,10)==="oncapture:"){const f=t.slice(10);r&&e.removeEventListener(f,r,!0),n&&e.addEventListener(f,n,!0)}else if(t.slice(0,2)==="on"){const f=t.slice(2).toLowerCase(),h=xs.has(f);if(!h&&r){const d=Array.isArray(r)?r[0]:r;e.removeEventListener(f,d)}(h||n)&&(Pn(e,f,n,h),h&&pe([f]))}else if(t.slice(0,5)==="attr:")Gn(e,t.slice(5),n);else if((u=t.slice(0,5)==="prop:")||(l=bs.has(t))||!i&&((c=ks(t,e.tagName))||(a=vs.has(t)))||(s=e.nodeName.includes("-"))){if(u)t=t.slice(5),a=!0;else if(N.context)return n;t==="class"||t==="className"?F(e,n):s&&!a&&!l?e[Rs(t)]=n:e[c||t]=n}else{const f=i&&t.indexOf(":")>-1&&Ss[t.split(":")[0]];f?Ps(e,f,t,n):Gn(e,ws[t]||t,n)}return n}function Ts(e){const t=`$$${e.type}`;let n=e.composedPath&&e.composedPath()[0]||e.target;for(e.target!==n&&Object.defineProperty(e,"target",{configurable:!0,value:n}),Object.defineProperty(e,"currentTarget",{configurable:!0,get(){return n||document}}),N.registry&&!N.done&&(N.done=_$HY.done=!0);n;){const r=n[t];if(r&&!n.disabled){const i=n[`${t}Data`];if(i!==void 0?r.call(n,i,e):r.call(n,e),e.cancelBubble)return}n=n._$host||n.parentNode||n.host}}function ft(e,t,n,r,i){if(N.context){!n&&(n=[...e.childNodes]);let a=[];for(let l=0;l<n.length;l++){const c=n[l];c.nodeType===8&&c.data.slice(0,2)==="!$"?c.remove():a.push(c)}n=a}for(;typeof n=="function";)n=n();if(t===n)return n;const o=typeof t,s=r!==void 0;if(e=s&&n[0]&&n[0].parentNode||e,o==="string"||o==="number"){if(N.context)return n;if(o==="number"&&(t=t.toString()),s){let a=n[0];a&&a.nodeType===3?a.data!==t&&(a.data=t):a=document.createTextNode(t),n=tt(e,n,r,a)}else n!==""&&typeof n=="string"?n=e.firstChild.data=t:n=e.textContent=t}else if(t==null||o==="boolean"){if(N.context)return n;n=tt(e,n,r)}else{if(o==="function")return B(()=>{let a=t();for(;typeof a=="function";)a=a();n=ft(e,a,n,r)}),()=>n;if(Array.isArray(t)){const a=[],l=n&&Array.isArray(n);if(Qn(a,t,n,i))return B(()=>n=ft(e,a,n,r,!0)),()=>n;if(N.context){if(!a.length)return n;if(r===void 0)return[...e.childNodes];let c=a[0],u=[c];for(;(c=c.nextSibling)!==r;)u.push(c);return n=u}if(a.length===0){if(n=tt(e,n,r),s)return n}else l?n.length===0?oi(e,a,r):Es(e,n,a):(n&&tt(e),oi(e,a));n=a}else if(t.nodeType){if(N.context&&t.parentNode)return n=s?[t]:t;if(Array.isArray(n)){if(s)return n=tt(e,n,r,t);tt(e,n,null,t)}else n==null||n===""||!e.firstChild?e.appendChild(t):e.replaceChild(t,e.firstChild);n=t}}return n}function Qn(e,t,n,r){let i=!1;for(let o=0,s=t.length;o<s;o++){let a=t[o],l=n&&n[o],c;if(!(a==null||a===!0||a===!1))if((c=typeof a)=="object"&&a.nodeType)e.push(a);else if(Array.isArray(a))i=Qn(e,a,l)||i;else if(c==="function")if(r){for(;typeof a=="function";)a=a();i=Qn(e,Array.isArray(a)?a:[a],Array.isArray(l)?l:[l])||i}else e.push(a),i=!0;else{const u=String(a);l&&l.nodeType===3&&l.data===u?e.push(l):e.push(document.createTextNode(u))}}return i}function oi(e,t,n=null){for(let r=0,i=t.length;r<i;r++)e.insertBefore(t[r],n)}function tt(e,t,n,r){if(n===void 0)return e.textContent="";const i=r||document.createTextNode("");if(t.length){let o=!1;for(let s=t.length-1;s>=0;s--){const a=t[s];if(i!==a){const l=a.parentNode===e;!o&&!s?l?e.replaceChild(i,a):e.insertBefore(i,n):l&&a.remove()}else o=!0}}else e.insertBefore(i,n);return[i]}const Is=!1,js="http://www.w3.org/2000/svg";function Bs(e,t=!1){return t?document.createElementNS(js,e):document.createElement(e)}function Ls(e){const{useShadow:t}=e,n=document.createTextNode(""),r=()=>e.mount||document.body,i=ln();let o,s=!!N.context;return is(()=>{s&&(ln().user=s=!1),o||(o=Ar(i,()=>U(()=>e.children)));const a=r();if(a instanceof HTMLHeadElement){const[l,c]=L(!1),u=()=>c(!0);ot(f=>C(a,()=>l()?f():o(),null)),ut(u)}else{const l=Bs(e.isSVG?"g":"div",e.isSVG),c=t&&l.attachShadow?l.attachShadow({mode:"open"}):l;Object.defineProperty(l,"_$host",{get(){return n.parentNode},configurable:!0}),C(c,o),a.appendChild(l),e.ref&&e.ref(l),ut(()=>a.removeChild(l))}},void 0,{render:!s}),n}function qi(){let e=new Set;function t(i){return e.add(i),()=>e.delete(i)}let n=!1;function r(i,o){if(n)return!(n=!1);const s={to:i,options:o,defaultPrevented:!1,preventDefault:()=>s.defaultPrevented=!0};for(const a of e)a.listener({...s,from:a.location,retry:l=>{l&&(n=!0),a.navigate(i,{...o,resolve:!1})}});return!s.defaultPrevented}return{subscribe:t,confirm:r}}let Jn;function Ir(){(!window.history.state||window.history.state._depth==null)&&window.history.replaceState({...window.history.state,_depth:window.history.length-1},""),Jn=window.history.state._depth}Ir();function Ds(e){return{...e,_depth:window.history.state&&window.history.state._depth}}function Ks(e,t){let n=!1;return()=>{const r=Jn;Ir();const i=r==null?null:Jn-r;if(n){n=!1;return}i&&t(i)?(n=!0,window.history.go(-i)):e()}}const Ns=/^(?:[a-z0-9]+:)?\/\//i,Ms=/^\/+|(\/)\/+$/g,Ui="http://sr";function Ve(e,t=!1){const n=e.replace(Ms,"$1");return n?t||/^[?#]/.test(n)?n:"/"+n:""}function Gt(e,t,n){if(Ns.test(t))return;const r=Ve(e),i=n&&Ve(n);let o="";return!i||t.startsWith("/")?o=r:i.toLowerCase().indexOf(r.toLowerCase())!==0?o=r+i:o=i,(o||"/")+Ve(t,!o)}function Fs(e,t){if(e==null)throw new Error(t);return e}function qs(e,t){return Ve(e).replace(/\/*(\*.*)?$/g,"")+Ve(t)}function Vi(e){const t={};return e.searchParams.forEach((n,r)=>{t[r]=n}),t}function Us(e,t,n){const[r,i]=e.split("/*",2),o=r.split("/").filter(Boolean),s=o.length;return a=>{const l=a.split("/").filter(Boolean),c=l.length-s;if(c<0||c>0&&i===void 0&&!t)return null;const u={path:s?"":"/",params:{}},f=h=>n===void 0?void 0:n[h];for(let h=0;h<s;h++){const d=o[h],v=l[h],m=d[0]===":",g=m?d.slice(1):d;if(m&&In(v,f(g)))u.params[g]=v;else if(m||!In(v,d))return null;u.path+=`/${v}`}if(i){const h=c?l.slice(-c).join("/"):"";if(In(h,f(i)))u.params[i]=h;else return null}return u}}function In(e,t){const n=r=>r.localeCompare(e,void 0,{sensitivity:"base"})===0;return t===void 0?!0:typeof t=="string"?n(t):typeof t=="function"?t(e):Array.isArray(t)?t.some(n):t instanceof RegExp?t.test(e):!1}function Vs(e){const[t,n]=e.pattern.split("/*",2),r=t.split("/").filter(Boolean);return r.reduce((i,o)=>i+(o.startsWith(":")?2:3),r.length-(n===void 0?0:1))}function Wi(e){const t=new Map,n=ln();return new Proxy({},{get(r,i){return t.has(i)||Ar(n,()=>t.set(i,U(()=>e()[i]))),t.get(i)()},getOwnPropertyDescriptor(){return{enumerable:!0,configurable:!0}},ownKeys(){return Reflect.ownKeys(e())}})}function zi(e){let t=/(\/?\:[^\/]+)\?/.exec(e);if(!t)return[e];let n=e.slice(0,t.index),r=e.slice(t.index+t[0].length);const i=[n,n+=t[1]];for(;t=/^(\/\:[^\/]+)\?/.exec(r);)i.push(n+=t[1]),r=r.slice(t[0].length);return zi(r).reduce((o,s)=>[...o,...i.map(a=>a+s)],[])}const Ws=100,Hi=ji(),jr=ji(),Br=()=>Fs(Or(Hi),"<A> and 'use' router primitives can be only used inside a Route."),zs=()=>Or(jr)||Br().base,Hs=e=>{const t=zs();return U(()=>t.resolvePath(e()))},Ys=e=>{const t=Br();return U(()=>{const n=e();return n!==void 0?t.renderPath(n):n})},Gs=()=>Br().location;function Xs(e,t=""){const{component:n,load:r,children:i,info:o}=e,s=!i||Array.isArray(i)&&!i.length,a={key:e,component:n,load:r,info:o};return Yi(e.path).reduce((l,c)=>{for(const u of zi(c)){const f=qs(t,u);let h=s?f:f.split("/*",1)[0];h=h.split("/").map(d=>d.startsWith(":")||d.startsWith("*")?d:encodeURIComponent(d)).join("/"),l.push({...a,originalPath:c,pattern:h,matcher:Us(h,!s,e.matchFilters)})}return l},[])}function Qs(e,t=0){return{routes:e,score:Vs(e[e.length-1])*1e4-t,matcher(n){const r=[];for(let i=e.length-1;i>=0;i--){const o=e[i],s=o.matcher(n);if(!s)return null;r.unshift({...s,route:o})}return r}}}function Yi(e){return Array.isArray(e)?e:[e]}function Gi(e,t="",n=[],r=[]){const i=Yi(e);for(let o=0,s=i.length;o<s;o++){const a=i[o];if(a&&typeof a=="object"){a.hasOwnProperty("path")||(a.path="");const l=Xs(a,t);for(const c of l){n.push(c);const u=Array.isArray(a.children)&&a.children.length===0;if(a.children&&!u)Gi(a.children,c.pattern,n,r);else{const f=Qs([...n],r.length);r.push(f)}n.pop()}}}return n.length?r:r.sort((o,s)=>s.score-o.score)}function jn(e,t){for(let n=0,r=e.length;n<r;n++){const i=e[n].matcher(t);if(i)return i}return[]}function Js(e,t){const n=new URL(Ui),r=U(l=>{const c=e();try{return new URL(c,n)}catch{return console.error(`Invalid path ${c}`),l}},n,{equals:(l,c)=>l.href===c.href}),i=U(()=>r().pathname),o=U(()=>r().search,!0),s=U(()=>r().hash),a=()=>"";return{get pathname(){return i()},get search(){return o()},get hash(){return s()},get state(){return t()},get key(){return a()},query:Wi(Ii(o,()=>Vi(r())))}}let ve;function Zs(){return ve}function ea(e,t,n,r={}){const{signal:[i,o],utils:s={}}=e,a=s.parsePath||(A=>A),l=s.renderPath||(A=>A),c=s.beforeLeave||qi(),u=Gt("",r.base||"");if(u===void 0)throw new Error(`${u} is not a valid base path`);u&&!i().value&&o({value:u,replace:!0,scroll:!1});const[f,h]=L(!1),d=async A=>{h(!0);try{await as(A)}finally{h(!1)}},[v,m]=L(i().value),[g,y]=L(i().state),p=Js(v,g),b=[],w=L([]),k=U(()=>typeof r.transformUrl=="function"?jn(t(),r.transformUrl(p.pathname)):jn(t(),p.pathname)),S=Wi(()=>{const A=k(),I={};for(let D=0;D<A.length;D++)Object.assign(I,A[D].params);return I}),x={pattern:u,path:()=>u,outlet:()=>null,resolvePath(A){return Gt(u,A)}};return B(()=>{const{value:A,state:I}=i();J(()=>{d(()=>{ve="native",A!==v()&&m(A),y(I),w[1]([])}).then(()=>{ve=void 0})})}),{base:x,location:p,params:S,isRouting:f,renderPath:l,parsePath:a,navigatorFactory:_,matches:k,beforeLeave:c,preloadRoute:$,singleFlight:r.singleFlight===void 0?!0:r.singleFlight,submissions:w};function T(A,I,D){J(()=>{if(typeof I=="number"){I&&(s.go?s.go(I):console.warn("Router integration does not support relative routing"));return}const{replace:fe,resolve:he,scroll:H,state:oe}={replace:!1,resolve:!0,scroll:!0,...D},ie=he?A.resolvePath(I):Gt("",I);if(ie===void 0)throw new Error(`Path '${I}' is not a routable path`);if(b.length>=Ws)throw new Error("Too many redirects");const et=v();if((ie!==et||oe!==g())&&!Is){if(c.confirm(ie,D)){const Jo=b.push({value:et,replace:fe,scroll:H,state:g()});d(()=>{ve="navigate",m(ie),y(oe),w[1]([])}).then(()=>{b.length===Jo&&(ve=void 0,R({value:ie,state:oe}))})}}})}function _(A){return A=A||Or(jr)||x,(I,D)=>T(A,I,D)}function R(A){const I=b[0];I&&((A.value!==I.value||A.state!==I.state)&&o({...A,replace:I.replace,scroll:I.scroll}),b.length=0)}function $(A,I){const D=jn(t(),A.pathname),fe=ve;ve="preload";for(let he in D){const{route:H,params:oe}=D[he];H.component&&H.component.preload&&H.component.preload();const{load:ie}=H;I&&ie&&Ar(n(),()=>ie({params:oe,location:{pathname:A.pathname,search:A.search,hash:A.hash,query:Vi(A),state:null,key:""},intent:"preload"}))}ve=fe}}function ta(e,t,n,r){const{base:i,location:o,params:s}=e,{pattern:a,component:l,load:c}=r().route,u=U(()=>r().path);l&&l.preload&&l.preload();const f=c?c({params:s,location:o,intent:ve||"initial"}):void 0;return{parent:t,pattern:a,path:u,outlet:()=>l?P(l,{params:s,location:o,data:f,get children(){return n()}}):n(),resolvePath(d){return Gt(i.path(),d,u())}}}const na=e=>t=>{const{base:n}=t,r=Sn(()=>t.children),i=U(()=>Gi(r(),t.base||""));let o;const s=ea(e,i,()=>o,{base:n,singleFlight:t.singleFlight,transformUrl:t.transformUrl});return e.create&&e.create(s),P(Hi.Provider,{value:s,get children(){return P(ra,{routerState:s,get root(){return t.root},get load(){return t.rootLoad},get children(){return[U(()=>(o=ln())&&null),P(ia,{routerState:s,get branches(){return i()}})]}})}})};function ra(e){const t=e.routerState.location,n=e.routerState.params,r=U(()=>e.load&&J(()=>{e.load({params:n,location:t,intent:Zs()||"initial"})}));return P(G,{get when(){return e.root},keyed:!0,get fallback(){return e.children},children:i=>P(i,{params:n,location:t,get data(){return r()},get children(){return e.children}})})}function ia(e){const t=[];let n;const r=U(Ii(e.routerState.matches,(i,o,s)=>{let a=o&&i.length===o.length;const l=[];for(let c=0,u=i.length;c<u;c++){const f=o&&o[c],h=i[c];s&&f&&h.route.key===f.route.key?l[c]=s[c]:(a=!1,t[c]&&t[c](),ot(d=>{t[c]=d,l[c]=ta(e.routerState,l[c-1]||e.routerState.base,si(()=>r()[c+1]),()=>e.routerState.matches()[c])}))}return t.splice(i.length).forEach(c=>c()),s&&a?s:(n=l[0],l)}));return si(()=>r()&&n)()}const si=e=>()=>P(G,{get when(){return e()},keyed:!0,children:t=>P(jr.Provider,{value:t,get children(){return t.outlet()}})}),Ne=e=>{const t=Sn(()=>e.children);return fn(e,{get children(){return t()}})};function oa([e,t],n,r){return[n?()=>n(e()):e,r?i=>t(r(i)):t]}function sa(e){if(e==="#")return null;try{return document.querySelector(e)}catch{return null}}function aa(e){let t=!1;const n=i=>typeof i=="string"?{value:i}:i,r=oa(L(n(e.get()),{equals:(i,o)=>i.value===o.value&&i.state===o.state}),void 0,i=>(!t&&e.set(i),i));return e.init&&ut(e.init((i=e.get())=>{t=!0,r[1](n(i)),t=!1})),na({signal:r,create:e.create,utils:e.utils})}function la(e,t,n){return e.addEventListener(t,n),()=>e.removeEventListener(t,n)}function ca(e,t){const n=sa(`#${e}`);n?n.scrollIntoView():t&&window.scrollTo(0,0)}const ua=new Map;function fa(e=!0,t=!1,n="/_server",r){return i=>{const o=i.base.path(),s=i.navigatorFactory(i.base);let a={};function l(m){return m.namespaceURI==="http://www.w3.org/2000/svg"}function c(m){if(m.defaultPrevented||m.button!==0||m.metaKey||m.altKey||m.ctrlKey||m.shiftKey)return;const g=m.composedPath().find(S=>S instanceof Node&&S.nodeName.toUpperCase()==="A");if(!g||t&&!g.hasAttribute("link"))return;const y=l(g),p=y?g.href.baseVal:g.href;if((y?g.target.baseVal:g.target)||!p&&!g.hasAttribute("state"))return;const w=(g.getAttribute("rel")||"").split(/\s+/);if(g.hasAttribute("download")||w&&w.includes("external"))return;const k=y?new URL(p,document.baseURI):new URL(p);if(!(k.origin!==window.location.origin||o&&k.pathname&&!k.pathname.toLowerCase().startsWith(o.toLowerCase())))return[g,k]}function u(m){const g=c(m);if(!g)return;const[y,p]=g,b=i.parsePath(p.pathname+p.search+p.hash),w=y.getAttribute("state");m.preventDefault(),s(b,{resolve:!1,replace:y.hasAttribute("replace"),scroll:!y.hasAttribute("noscroll"),state:w&&JSON.parse(w)})}function f(m){const g=c(m);if(!g)return;const[y,p]=g;typeof r=="function"&&(p.pathname=r(p.pathname)),a[p.pathname]||i.preloadRoute(p,y.getAttribute("preload")!=="false")}function h(m){const g=c(m);if(!g)return;const[y,p]=g;typeof r=="function"&&(p.pathname=r(p.pathname)),!a[p.pathname]&&(a[p.pathname]=setTimeout(()=>{i.preloadRoute(p,y.getAttribute("preload")!=="false"),delete a[p.pathname]},200))}function d(m){const g=c(m);if(!g)return;const[,y]=g;typeof r=="function"&&(y.pathname=r(y.pathname)),a[y.pathname]&&(clearTimeout(a[y.pathname]),delete a[y.pathname])}function v(m){if(m.defaultPrevented)return;let g=m.submitter&&m.submitter.hasAttribute("formaction")?m.submitter.getAttribute("formaction"):m.target.getAttribute("action");if(!g)return;if(!g.startsWith("https://action/")){const p=new URL(g,Ui);if(g=i.parsePath(p.pathname+p.search),!g.startsWith(n))return}if(m.target.method.toUpperCase()!=="POST")throw new Error("Only POST forms are supported for Actions");const y=ua.get(g);if(y){m.preventDefault();const p=new FormData(m.target);m.submitter&&m.submitter.name&&p.append(m.submitter.name,m.submitter.value),y.call({r:i,f:m.target},p)}}pe(["click","submit"]),document.addEventListener("click",u),e&&(document.addEventListener("mouseover",h),document.addEventListener("mouseout",d),document.addEventListener("focusin",f),document.addEventListener("touchstart",f)),document.addEventListener("submit",v),ut(()=>{document.removeEventListener("click",u),e&&(document.removeEventListener("mouseover",h),document.removeEventListener("mouseout",d),document.removeEventListener("focusin",f),document.removeEventListener("touchstart",f)),document.removeEventListener("submit",v)})}}function ha(e){const t=()=>{const r=window.location.pathname+window.location.search;return{value:e.transformUrl?e.transformUrl(r)+window.location.hash:r+window.location.hash,state:window.history.state}},n=qi();return aa({get:t,set({value:r,replace:i,scroll:o,state:s}){i?window.history.replaceState(Ds(s),"",r):window.history.pushState(s,"",r),ca(window.location.hash.slice(1),o),Ir()},init:r=>la(window,"popstate",Ks(r,i=>{if(i&&i<0)return!n.confirm(i);{const o=t();return!n.confirm(o.value,{state:o.state})}})),create:fa(e.preload,e.explicitLinks,e.actionBase,e.transformUrl),utils:{go:r=>window.history.go(r),beforeLeave:n}})(e)}var da=K("<a>");function ai(e){e=fn({inactiveClass:"inactive",activeClass:"active"},e);const[,t]=ys(e,["href","state","class","activeClass","inactiveClass","end"]),n=Hs(()=>e.href),r=Ys(n),i=Gs(),o=U(()=>{const s=n();if(s===void 0)return[!1,!1];const a=Ve(s.split(/[?#]/,1)[0]).toLowerCase(),l=Ve(i.pathname).toLowerCase();return[e.end?a===l:l.startsWith(a),a===l]});return(()=>{var s=da();return Fi(s,fn(t,{get href(){return r()||e.href},get state(){return JSON.stringify(e.state)},get classList(){return{...e.class&&{[e.class]:!0},[e.inactiveClass]:!o()[0],[e.activeClass]:o()[0],...t.classList}},link:"",get"aria-current"(){return o()[1]?"page":void 0}}),!1,!1),s})()}var Zn=function(e,t){return Zn=Object.setPrototypeOf||{__proto__:[]}instanceof Array&&function(n,r){n.__proto__=r}||function(n,r){for(var i in r)Object.prototype.hasOwnProperty.call(r,i)&&(n[i]=r[i])},Zn(e,t)};function Lr(e,t){if(typeof t!="function"&&t!==null)throw new TypeError("Class extends value "+String(t)+" is not a constructor or null");Zn(e,t);function n(){this.constructor=e}e.prototype=t===null?Object.create(t):(n.prototype=t.prototype,new n)}function pa(e,t,n,r){function i(o){return o instanceof n?o:new n(function(s){s(o)})}return new(n||(n=Promise))(function(o,s){function a(u){try{c(r.next(u))}catch(f){s(f)}}function l(u){try{c(r.throw(u))}catch(f){s(f)}}function c(u){u.done?o(u.value):i(u.value).then(a,l)}c((r=r.apply(e,t||[])).next())})}function Xi(e,t){var n={label:0,sent:function(){if(o[0]&1)throw o[1];return o[1]},trys:[],ops:[]},r,i,o,s;return s={next:a(0),throw:a(1),return:a(2)},typeof Symbol=="function"&&(s[Symbol.iterator]=function(){return this}),s;function a(c){return function(u){return l([c,u])}}function l(c){if(r)throw new TypeError("Generator is already executing.");for(;s&&(s=0,c[0]&&(n=0)),n;)try{if(r=1,i&&(o=c[0]&2?i.return:c[0]?i.throw||((o=i.return)&&o.call(i),0):i.next)&&!(o=o.call(i,c[1])).done)return o;switch(i=0,o&&(c=[c[0]&2,o.value]),c[0]){case 0:case 1:o=c;break;case 4:return n.label++,{value:c[1],done:!1};case 5:n.label++,i=c[1],c=[0];continue;case 7:c=n.ops.pop(),n.trys.pop();continue;default:if(o=n.trys,!(o=o.length>0&&o[o.length-1])&&(c[0]===6||c[0]===2)){n=0;continue}if(c[0]===3&&(!o||c[1]>o[0]&&c[1]<o[3])){n.label=c[1];break}if(c[0]===6&&n.label<o[1]){n.label=o[1],o=c;break}if(o&&n.label<o[2]){n.label=o[2],n.ops.push(c);break}o[2]&&n.ops.pop(),n.trys.pop();continue}c=t.call(e,n)}catch(u){c=[6,u],i=0}finally{r=o=0}if(c[0]&5)throw c[1];return{value:c[0]?c[1]:void 0,done:!0}}}function At(e){var t=typeof Symbol=="function"&&Symbol.iterator,n=t&&e[t],r=0;if(n)return n.call(e);if(e&&typeof e.length=="number")return{next:function(){return e&&r>=e.length&&(e=void 0),{value:e&&e[r++],done:!e}}};throw new TypeError(t?"Object is not iterable.":"Symbol.iterator is not defined.")}function er(e,t){var n=typeof Symbol=="function"&&e[Symbol.iterator];if(!n)return e;var r=n.call(e),i,o=[],s;try{for(;(t===void 0||t-- >0)&&!(i=r.next()).done;)o.push(i.value)}catch(a){s={error:a}}finally{try{i&&!i.done&&(n=r.return)&&n.call(r)}finally{if(s)throw s.error}}return o}function tr(e,t,n){if(n||arguments.length===2)for(var r=0,i=t.length,o;r<i;r++)(o||!(r in t))&&(o||(o=Array.prototype.slice.call(t,0,r)),o[r]=t[r]);return e.concat(o||Array.prototype.slice.call(t))}function st(e){return this instanceof st?(this.v=e,this):new st(e)}function ma(e,t,n){if(!Symbol.asyncIterator)throw new TypeError("Symbol.asyncIterator is not defined.");var r=n.apply(e,t||[]),i,o=[];return i={},s("next"),s("throw"),s("return"),i[Symbol.asyncIterator]=function(){return this},i;function s(h){r[h]&&(i[h]=function(d){return new Promise(function(v,m){o.push([h,d,v,m])>1||a(h,d)})})}function a(h,d){try{l(r[h](d))}catch(v){f(o[0][3],v)}}function l(h){h.value instanceof st?Promise.resolve(h.value.v).then(c,u):f(o[0][2],h)}function c(h){a("next",h)}function u(h){a("throw",h)}function f(h,d){h(d),o.shift(),o.length&&a(o[0][0],o[0][1])}}function ya(e){if(!Symbol.asyncIterator)throw new TypeError("Symbol.asyncIterator is not defined.");var t=e[Symbol.asyncIterator],n;return t?t.call(e):(e=typeof At=="function"?At(e):e[Symbol.iterator](),n={},r("next"),r("throw"),r("return"),n[Symbol.asyncIterator]=function(){return this},n);function r(o){n[o]=e[o]&&function(s){return new Promise(function(a,l){s=e[o](s),i(a,l,s.done,s.value)})}}function i(o,s,a,l){Promise.resolve(l).then(function(c){o({value:c,done:a})},s)}}function Y(e){return typeof e=="function"}function ga(e){var t=function(r){Error.call(r),r.stack=new Error().stack},n=e(t);return n.prototype=Object.create(Error.prototype),n.prototype.constructor=n,n}var Bn=ga(function(e){return function(n){e(this),this.message=n?n.length+` errors occurred during unsubscription:
+`+n.map(function(r,i){return i+1+") "+r.toString()}).join(`
+  `):"",this.name="UnsubscriptionError",this.errors=n}});function li(e,t){if(e){var n=e.indexOf(t);0<=n&&e.splice(n,1)}}var Dr=function(){function e(t){this.initialTeardown=t,this.closed=!1,this._parentage=null,this._finalizers=null}return e.prototype.unsubscribe=function(){var t,n,r,i,o;if(!this.closed){this.closed=!0;var s=this._parentage;if(s)if(this._parentage=null,Array.isArray(s))try{for(var a=At(s),l=a.next();!l.done;l=a.next()){var c=l.value;c.remove(this)}}catch(m){t={error:m}}finally{try{l&&!l.done&&(n=a.return)&&n.call(a)}finally{if(t)throw t.error}}else s.remove(this);var u=this.initialTeardown;if(Y(u))try{u()}catch(m){o=m instanceof Bn?m.errors:[m]}var f=this._finalizers;if(f){this._finalizers=null;try{for(var h=At(f),d=h.next();!d.done;d=h.next()){var v=d.value;try{ci(v)}catch(m){o=o??[],m instanceof Bn?o=tr(tr([],er(o)),er(m.errors)):o.push(m)}}}catch(m){r={error:m}}finally{try{d&&!d.done&&(i=h.return)&&i.call(h)}finally{if(r)throw r.error}}}if(o)throw new Bn(o)}},e.prototype.add=function(t){var n;if(t&&t!==this)if(this.closed)ci(t);else{if(t instanceof e){if(t.closed||t._hasParent(this))return;t._addParent(this)}(this._finalizers=(n=this._finalizers)!==null&&n!==void 0?n:[]).push(t)}},e.prototype._hasParent=function(t){var n=this._parentage;return n===t||Array.isArray(n)&&n.includes(t)},e.prototype._addParent=function(t){var n=this._parentage;this._parentage=Array.isArray(n)?(n.push(t),n):n?[n,t]:t},e.prototype._removeParent=function(t){var n=this._parentage;n===t?this._parentage=null:Array.isArray(n)&&li(n,t)},e.prototype.remove=function(t){var n=this._finalizers;n&&li(n,t),t instanceof e&&t._removeParent(this)},e.EMPTY=function(){var t=new e;return t.closed=!0,t}(),e}();Dr.EMPTY;function Qi(e){return e instanceof Dr||e&&"closed"in e&&Y(e.remove)&&Y(e.add)&&Y(e.unsubscribe)}function ci(e){Y(e)?e():e.unsubscribe()}var Ji={onUnhandledError:null,onStoppedNotification:null,Promise:void 0,useDeprecatedSynchronousErrorHandling:!1,useDeprecatedNextContext:!1},Zi={setTimeout:function(e,t){for(var n=[],r=2;r<arguments.length;r++)n[r-2]=arguments[r];return setTimeout.apply(void 0,tr([e,t],er(n)))},clearTimeout:function(e){var t=Zi.delegate;return(t?.clearTimeout||clearTimeout)(e)},delegate:void 0};function eo(e){Zi.setTimeout(function(){throw e})}function ui(){}function va(e){e()}var Kr=function(e){Lr(t,e);function t(n){var r=e.call(this)||this;return r.isStopped=!1,n?(r.destination=n,Qi(n)&&n.add(r)):r.destination=ka,r}return t.create=function(n,r,i){return new nr(n,r,i)},t.prototype.next=function(n){this.isStopped||this._next(n)},t.prototype.error=function(n){this.isStopped||(this.isStopped=!0,this._error(n))},t.prototype.complete=function(){this.isStopped||(this.isStopped=!0,this._complete())},t.prototype.unsubscribe=function(){this.closed||(this.isStopped=!0,e.prototype.unsubscribe.call(this),this.destination=null)},t.prototype._next=function(n){this.destination.next(n)},t.prototype._error=function(n){try{this.destination.error(n)}finally{this.unsubscribe()}},t.prototype._complete=function(){try{this.destination.complete()}finally{this.unsubscribe()}},t}(Dr),ba=Function.prototype.bind;function Ln(e,t){return ba.call(e,t)}var wa=function(){function e(t){this.partialObserver=t}return e.prototype.next=function(t){var n=this.partialObserver;if(n.next)try{n.next(t)}catch(r){Ft(r)}},e.prototype.error=function(t){var n=this.partialObserver;if(n.error)try{n.error(t)}catch(r){Ft(r)}else Ft(t)},e.prototype.complete=function(){var t=this.partialObserver;if(t.complete)try{t.complete()}catch(n){Ft(n)}},e}(),nr=function(e){Lr(t,e);function t(n,r,i){var o=e.call(this)||this,s;if(Y(n)||!n)s={next:n??void 0,error:r??void 0,complete:i??void 0};else{var a;o&&Ji.useDeprecatedNextContext?(a=Object.create(n),a.unsubscribe=function(){return o.unsubscribe()},s={next:n.next&&Ln(n.next,a),error:n.error&&Ln(n.error,a),complete:n.complete&&Ln(n.complete,a)}):s=n}return o.destination=new wa(s),o}return t}(Kr);function Ft(e){eo(e)}function _a(e){throw e}var ka={closed:!0,next:ui,error:_a,complete:ui},Nr=function(){return typeof Symbol=="function"&&Symbol.observable||"@@observable"}();function to(e){return e}function xa(e){return e.length===0?to:e.length===1?e[0]:function(n){return e.reduce(function(r,i){return i(r)},n)}}var ae=function(){function e(t){t&&(this._subscribe=t)}return e.prototype.lift=function(t){var n=new e;return n.source=this,n.operator=t,n},e.prototype.subscribe=function(t,n,r){var i=this,o=Ea(t)?t:new nr(t,n,r);return va(function(){var s=i,a=s.operator,l=s.source;o.add(a?a.call(o,l):l?i._subscribe(o):i._trySubscribe(o))}),o},e.prototype._trySubscribe=function(t){try{return this._subscribe(t)}catch(n){t.error(n)}},e.prototype.forEach=function(t,n){var r=this;return n=fi(n),new n(function(i,o){var s=new nr({next:function(a){try{t(a)}catch(l){o(l),s.unsubscribe()}},error:o,complete:i});r.subscribe(s)})},e.prototype._subscribe=function(t){var n;return(n=this.source)===null||n===void 0?void 0:n.subscribe(t)},e.prototype[Nr]=function(){return this},e.prototype.pipe=function(){for(var t=[],n=0;n<arguments.length;n++)t[n]=arguments[n];return xa(t)(this)},e.prototype.toPromise=function(t){var n=this;return t=fi(t),new t(function(r,i){var o;n.subscribe(function(s){return o=s},function(s){return i(s)},function(){return r(o)})})},e.create=function(t){return new e(t)},e}();function fi(e){var t;return(t=e??Ji.Promise)!==null&&t!==void 0?t:Promise}function Sa(e){return e&&Y(e.next)&&Y(e.error)&&Y(e.complete)}function Ea(e){return e&&e instanceof Kr||Sa(e)&&Qi(e)}function Ca(e){return Y(e?.lift)}function Qe(e){return function(t){if(Ca(t))return t.lift(function(n){try{return e(n,this)}catch(r){this.error(r)}});throw new TypeError("Unable to lift unknown Observable type")}}function Ye(e,t,n,r,i){return new Pa(e,t,n,r,i)}var Pa=function(e){Lr(t,e);function t(n,r,i,o,s,a){var l=e.call(this,n)||this;return l.onFinalize=s,l.shouldUnsubscribe=a,l._next=r?function(c){try{r(c)}catch(u){n.error(u)}}:e.prototype._next,l._error=o?function(c){try{o(c)}catch(u){n.error(u)}finally{this.unsubscribe()}}:e.prototype._error,l._complete=i?function(){try{i()}catch(c){n.error(c)}finally{this.unsubscribe()}}:e.prototype._complete,l}return t.prototype.unsubscribe=function(){var n;if(!this.shouldUnsubscribe||this.shouldUnsubscribe()){var r=this.closed;e.prototype.unsubscribe.call(this),!r&&((n=this.onFinalize)===null||n===void 0||n.call(this))}},t}(Kr),$a=new ae(function(e){return e.complete()});function Aa(e){return e&&Y(e.schedule)}function Oa(e){return e[e.length-1]}function Ra(e){return Aa(Oa(e))?e.pop():void 0}var no=function(e){return e&&typeof e.length=="number"&&typeof e!="function"};function ro(e){return Y(e?.then)}function io(e){return Y(e[Nr])}function oo(e){return Symbol.asyncIterator&&Y(e?.[Symbol.asyncIterator])}function so(e){return new TypeError("You provided "+(e!==null&&typeof e=="object"?"an invalid object":"'"+e+"'")+" where a stream was expected. You can provide an Observable, Promise, ReadableStream, Array, AsyncIterable, or Iterable.")}function Ta(){return typeof Symbol!="function"||!Symbol.iterator?"@@iterator":Symbol.iterator}var ao=Ta();function lo(e){return Y(e?.[ao])}function co(e){return ma(this,arguments,function(){var n,r,i,o;return Xi(this,function(s){switch(s.label){case 0:n=e.getReader(),s.label=1;case 1:s.trys.push([1,,9,10]),s.label=2;case 2:return[4,st(n.read())];case 3:return r=s.sent(),i=r.value,o=r.done,o?[4,st(void 0)]:[3,5];case 4:return[2,s.sent()];case 5:return[4,st(i)];case 6:return[4,s.sent()];case 7:return s.sent(),[3,2];case 8:return[3,10];case 9:return n.releaseLock(),[7];case 10:return[2]}})})}function uo(e){return Y(e?.getReader)}function pt(e){if(e instanceof ae)return e;if(e!=null){if(io(e))return Ia(e);if(no(e))return ja(e);if(ro(e))return Ba(e);if(oo(e))return fo(e);if(lo(e))return La(e);if(uo(e))return Da(e)}throw so(e)}function Ia(e){return new ae(function(t){var n=e[Nr]();if(Y(n.subscribe))return n.subscribe(t);throw new TypeError("Provided object does not correctly implement Symbol.observable")})}function ja(e){return new ae(function(t){for(var n=0;n<e.length&&!t.closed;n++)t.next(e[n]);t.complete()})}function Ba(e){return new ae(function(t){e.then(function(n){t.closed||(t.next(n),t.complete())},function(n){return t.error(n)}).then(null,eo)})}function La(e){return new ae(function(t){var n,r;try{for(var i=At(e),o=i.next();!o.done;o=i.next()){var s=o.value;if(t.next(s),t.closed)return}}catch(a){n={error:a}}finally{try{o&&!o.done&&(r=i.return)&&r.call(i)}finally{if(n)throw n.error}}t.complete()})}function fo(e){return new ae(function(t){Ka(e,t).catch(function(n){return t.error(n)})})}function Da(e){return fo(co(e))}function Ka(e,t){var n,r,i,o;return pa(this,void 0,void 0,function(){var s,a;return Xi(this,function(l){switch(l.label){case 0:l.trys.push([0,5,6,11]),n=ya(e),l.label=1;case 1:return[4,n.next()];case 2:if(r=l.sent(),!!r.done)return[3,4];if(s=r.value,t.next(s),t.closed)return[2];l.label=3;case 3:return[3,1];case 4:return[3,11];case 5:return a=l.sent(),i={error:a},[3,11];case 6:return l.trys.push([6,,9,10]),r&&!r.done&&(o=n.return)?[4,o.call(n)]:[3,8];case 7:l.sent(),l.label=8;case 8:return[3,10];case 9:if(i)throw i.error;return[7];case 10:return[7];case 11:return t.complete(),[2]}})})}function Te(e,t,n,r,i){r===void 0&&(r=0),i===void 0&&(i=!1);var o=t.schedule(function(){n(),i?e.add(this.schedule(null,r)):this.unsubscribe()},r);if(e.add(o),!i)return o}function ho(e,t){return t===void 0&&(t=0),Qe(function(n,r){n.subscribe(Ye(r,function(i){return Te(r,e,function(){return r.next(i)},t)},function(){return Te(r,e,function(){return r.complete()},t)},function(i){return Te(r,e,function(){return r.error(i)},t)}))})}function po(e,t){return t===void 0&&(t=0),Qe(function(n,r){r.add(e.schedule(function(){return n.subscribe(r)},t))})}function Na(e,t){return pt(e).pipe(po(t),ho(t))}function Ma(e,t){return pt(e).pipe(po(t),ho(t))}function Fa(e,t){return new ae(function(n){var r=0;return t.schedule(function(){r===e.length?n.complete():(n.next(e[r++]),n.closed||this.schedule())})})}function qa(e,t){return new ae(function(n){var r;return Te(n,t,function(){r=e[ao](),Te(n,t,function(){var i,o,s;try{i=r.next(),o=i.value,s=i.done}catch(a){n.error(a);return}s?n.complete():n.next(o)},0,!0)}),function(){return Y(r?.return)&&r.return()}})}function mo(e,t){if(!e)throw new Error("Iterable cannot be null");return new ae(function(n){Te(n,t,function(){var r=e[Symbol.asyncIterator]();Te(n,t,function(){r.next().then(function(i){i.done?n.complete():n.next(i.value)})},0,!0)})})}function Ua(e,t){return mo(co(e),t)}function Va(e,t){if(e!=null){if(io(e))return Na(e,t);if(no(e))return Fa(e,t);if(ro(e))return Ma(e,t);if(oo(e))return mo(e,t);if(lo(e))return qa(e,t);if(uo(e))return Ua(e,t)}throw so(e)}function Wa(e,t){return t?Va(e,t):pt(e)}function ce(){for(var e=[],t=0;t<arguments.length;t++)e[t]=arguments[t];var n=Ra(e);return Wa(e,n)}function za(e,t){return Qe(function(n,r){var i=0;n.subscribe(Ye(r,function(o){r.next(e.call(t,o,i++))}))})}function Ha(e,t,n,r,i,o,s,a){var l=[],c=0,u=0,f=!1,h=function(){f&&!l.length&&!c&&t.complete()},d=function(m){return c<r?v(m):l.push(m)},v=function(m){o&&t.next(m),c++;var g=!1;pt(n(m,u++)).subscribe(Ye(t,function(y){i?.(y),o?d(y):t.next(y)},function(){g=!0},void 0,function(){if(g)try{c--;for(var y=function(){var p=l.shift();s?Te(t,s,function(){return v(p)}):v(p)};l.length&&c<r;)y();h()}catch(p){t.error(p)}}))};return e.subscribe(Ye(t,d,function(){f=!0,h()})),function(){a?.()}}function rr(e,t,n){return n===void 0&&(n=1/0),Y(t)?rr(function(r,i){return za(function(o,s){return t(r,o,i,s)})(pt(e(r,i)))},n):(typeof t=="number"&&(n=t),Qe(function(r,i){return Ha(r,i,e,n)}))}function Je(e){return Qe(function(t,n){var r=null,i=!1,o;r=t.subscribe(Ye(n,void 0,void 0,function(s){o=pt(e(s,Je(e)(t))),r?(r.unsubscribe(),r=null,o.subscribe(n)):i=!0})),i&&(r.unsubscribe(),r=null,o.subscribe(n))})}function hn(e,t){return Y(t)?rr(e,t,1):rr(e,1)}function Ie(e){return e<=0?function(){return $a}:Qe(function(t,n){var r=0;t.subscribe(Ye(n,function(i){++r<=e&&(n.next(i),e<=r&&n.complete())}))})}function je(e,t,n){var r=Y(e)||t||n?{next:e,error:t,complete:n}:e;return r?Qe(function(i,o){var s;(s=r.subscribe)===null||s===void 0||s.call(r);var a=!0;i.subscribe(Ye(o,function(l){var c;(c=r.next)===null||c===void 0||c.call(r,l),o.next(l)},function(){var l;a=!1,(l=r.complete)===null||l===void 0||l.call(r),o.complete()},function(l){var c;a=!1,(c=r.error)===null||c===void 0||c.call(r,l),o.error(l)},function(){var l,c;a&&((l=r.unsubscribe)===null||l===void 0||l.call(r)),(c=r.finalize)===null||c===void 0||c.call(r)}))}):to}class Ya{loggedIn=L({});tokenValue=L("");set token(t){this.tokenValue[1](t),this.loggedIn[1]({Authorization:`Bearer ${t}`})}get token(){return this.tokenValue[0]}clearToken(){this.loggedIn[1]({Authorization:""})}toObservable(t){return new ae(n=>{t.then(r=>{n.next(r)}).catch(r=>n.error(r)).finally(()=>n.complete())})}post(t,n,r){return new ae(i=>{fetch(t,{method:"POST",headers:{"Content-Type":"application/json",...r?.headers,...this.loggedIn[0]()},body:JSON.stringify(n)}).then(o=>{if(!o.ok)throw new Error(o.statusText);return o.json()}).then(o=>{i.next(o)}).catch(o=>i.error(o)).finally(()=>i.complete())})}get(t,n){return new ae(r=>{fetch(t,{method:"GET",headers:{"Content-Type":"application/json",...n?.headers,...this.loggedIn[0]()}}).then(i=>{if(!i.ok)throw new Error(i.statusText);return i.json()}).then(i=>{r.next(i)}).catch(i=>r.error(i)).finally(()=>r.complete())})}}const Be=new Ya,W=typeof globalThis<"u"?globalThis:typeof self<"u"?self:typeof window<"u"?window:global,X=Object.keys,ee=Array.isArray;function re(e,t){return typeof t!="object"||X(t).forEach(function(n){e[n]=t[n]}),e}typeof Promise>"u"||W.Promise||(W.Promise=Promise);const Ot=Object.getPrototypeOf,Ga={}.hasOwnProperty;function le(e,t){return Ga.call(e,t)}function ht(e,t){typeof t=="function"&&(t=t(Ot(e))),(typeof Reflect>"u"?X:Reflect.ownKeys)(t).forEach(n=>{xe(e,n,t[n])})}const yo=Object.defineProperty;function xe(e,t,n,r){yo(e,t,re(n&&le(n,"get")&&typeof n.get=="function"?{get:n.get,set:n.set,configurable:!0}:{value:n,configurable:!0,writable:!0},r))}function at(e){return{from:function(t){return e.prototype=Object.create(t.prototype),xe(e.prototype,"constructor",e),{extend:ht.bind(null,e.prototype)}}}}const Xa=Object.getOwnPropertyDescriptor;function Mr(e,t){let n;return Xa(e,t)||(n=Ot(e))&&Mr(n,t)}const Qa=[].slice;function dn(e,t,n){return Qa.call(e,t,n)}function go(e,t){return t(e)}function kt(e){if(!e)throw new Error("Assertion Failed")}function vo(e){W.setImmediate?setImmediate(e):setTimeout(e,0)}function bo(e,t){return e.reduce((n,r,i)=>{var o=t(r,i);return o&&(n[o[0]]=o[1]),n},{})}function Se(e,t){if(le(e,t))return e[t];if(!t)return e;if(typeof t!="string"){for(var n=[],r=0,i=t.length;r<i;++r){var o=Se(e,t[r]);n.push(o)}return n}var s=t.indexOf(".");if(s!==-1){var a=e[t.substr(0,s)];return a===void 0?void 0:Se(a,t.substr(s+1))}}function ue(e,t,n){if(e&&t!==void 0&&(!("isFrozen"in Object)||!Object.isFrozen(e)))if(typeof t!="string"&&"length"in t){kt(typeof n!="string"&&"length"in n);for(var r=0,i=t.length;r<i;++r)ue(e,t[r],n[r])}else{var o=t.indexOf(".");if(o!==-1){var s=t.substr(0,o),a=t.substr(o+1);if(a==="")n===void 0?ee(e)&&!isNaN(parseInt(s))?e.splice(s,1):delete e[s]:e[s]=n;else{var l=e[s];l&&le(e,s)||(l=e[s]={}),ue(l,a,n)}}else n===void 0?ee(e)&&!isNaN(parseInt(t))?e.splice(t,1):delete e[t]:e[t]=n}}function wo(e){var t={};for(var n in e)le(e,n)&&(t[n]=e[n]);return t}const Ja=[].concat;function _o(e){return Ja.apply([],e)}const ko="Boolean,String,Date,RegExp,Blob,File,FileList,FileSystemFileHandle,ArrayBuffer,DataView,Uint8ClampedArray,ImageBitmap,ImageData,Map,Set,CryptoKey".split(",").concat(_o([8,16,32,64].map(e=>["Int","Uint","Float"].map(t=>t+e+"Array")))).filter(e=>W[e]),Za=ko.map(e=>W[e]);bo(ko,e=>[e,!0]);let Ae=null;function Dt(e){Ae=typeof WeakMap<"u"&&new WeakMap;const t=ir(e);return Ae=null,t}function ir(e){if(!e||typeof e!="object")return e;let t=Ae&&Ae.get(e);if(t)return t;if(ee(e)){t=[],Ae&&Ae.set(e,t);for(var n=0,r=e.length;n<r;++n)t.push(ir(e[n]))}else if(Za.indexOf(e.constructor)>=0)t=e;else{const o=Ot(e);for(var i in t=o===Object.prototype?{}:Object.create(o),Ae&&Ae.set(e,t),e)le(e,i)&&(t[i]=ir(e[i]))}return t}const{toString:el}={};function or(e){return el.call(e).slice(8,-1)}const sr=typeof Symbol<"u"?Symbol.iterator:"@@iterator",tl=typeof sr=="symbol"?function(e){var t;return e!=null&&(t=e[sr])&&t.apply(e)}:function(){return null},it={};function we(e){var t,n,r,i;if(arguments.length===1){if(ee(e))return e.slice();if(this===it&&typeof e=="string")return[e];if(i=tl(e)){for(n=[];!(r=i.next()).done;)n.push(r.value);return n}if(e==null)return[e];if(typeof(t=e.length)=="number"){for(n=new Array(t);t--;)n[t]=e[t];return n}return[e]}for(t=arguments.length,n=new Array(t);t--;)n[t]=arguments[t];return n}const Fr=typeof Symbol<"u"?e=>e[Symbol.toStringTag]==="AsyncFunction":()=>!1;var me=typeof location<"u"&&/^(http|https):\/\/(localhost|127\.0\.0\.1)/.test(location.href);function xo(e,t){me=e,So=t}var So=()=>!0;const nl=!new Error("").stack;function Ze(){if(nl)try{throw Ze.arguments,new Error}catch(e){return e}return new Error}function ar(e,t){var n=e.stack;return n?(t=t||0,n.indexOf(e.name)===0&&(t+=(e.name+e.message).split(`
+`).length),n.split(`
+`).slice(t).filter(So).map(r=>`
+`+r).join("")):""}var Eo=["Unknown","Constraint","Data","TransactionInactive","ReadOnly","Version","NotFound","InvalidState","InvalidAccess","Abort","Timeout","QuotaExceeded","Syntax","DataClone"],qr=["Modify","Bulk","OpenFailed","VersionChange","Schema","Upgrade","InvalidTable","MissingAPI","NoSuchDatabase","InvalidArgument","SubTransaction","Unsupported","Internal","DatabaseClosed","PrematureCommit","ForeignAwait"].concat(Eo),rl={VersionChanged:"Database version changed by other database connection",DatabaseClosed:"Database has been closed",Abort:"Transaction aborted",TransactionInactive:"Transaction has already completed or failed",MissingAPI:"IndexedDB API missing. Please visit https://tinyurl.com/y2uuvskb"};function lt(e,t){this._e=Ze(),this.name=e,this.message=t}function Co(e,t){return e+". Errors: "+Object.keys(t).map(n=>t[n].toString()).filter((n,r,i)=>i.indexOf(n)===r).join(`
+`)}function pn(e,t,n,r){this._e=Ze(),this.failures=t,this.failedKeys=r,this.successCount=n,this.message=Co(e,t)}function St(e,t){this._e=Ze(),this.name="BulkError",this.failures=Object.keys(t).map(n=>t[n]),this.failuresByPos=t,this.message=Co(e,t)}at(lt).from(Error).extend({stack:{get:function(){return this._stack||(this._stack=this.name+": "+this.message+ar(this._e,2))}},toString:function(){return this.name+": "+this.message}}),at(pn).from(lt),at(St).from(lt);var Ur=qr.reduce((e,t)=>(e[t]=t+"Error",e),{});const il=lt;var j=qr.reduce((e,t)=>{var n=t+"Error";function r(i,o){this._e=Ze(),this.name=n,i?typeof i=="string"?(this.message=`${i}${o?`
+ `+o:""}`,this.inner=o||null):typeof i=="object"&&(this.message=`${i.name} ${i.message}`,this.inner=i):(this.message=rl[t]||n,this.inner=null)}return at(r).from(il),e[t]=r,e},{});j.Syntax=SyntaxError,j.Type=TypeError,j.Range=RangeError;var hi=Eo.reduce((e,t)=>(e[t+"Error"]=j[t],e),{}),Xt=qr.reduce((e,t)=>(["Syntax","Type","Range"].indexOf(t)===-1&&(e[t+"Error"]=j[t]),e),{});function V(){}function Rt(e){return e}function ol(e,t){return e==null||e===Rt?t:function(n){return t(e(n))}}function Ge(e,t){return function(){e.apply(this,arguments),t.apply(this,arguments)}}function sl(e,t){return e===V?t:function(){var n=e.apply(this,arguments);n!==void 0&&(arguments[0]=n);var r=this.onsuccess,i=this.onerror;this.onsuccess=null,this.onerror=null;var o=t.apply(this,arguments);return r&&(this.onsuccess=this.onsuccess?Ge(r,this.onsuccess):r),i&&(this.onerror=this.onerror?Ge(i,this.onerror):i),o!==void 0?o:n}}function al(e,t){return e===V?t:function(){e.apply(this,arguments);var n=this.onsuccess,r=this.onerror;this.onsuccess=this.onerror=null,t.apply(this,arguments),n&&(this.onsuccess=this.onsuccess?Ge(n,this.onsuccess):n),r&&(this.onerror=this.onerror?Ge(r,this.onerror):r)}}function ll(e,t){return e===V?t:function(n){var r=e.apply(this,arguments);re(n,r);var i=this.onsuccess,o=this.onerror;this.onsuccess=null,this.onerror=null;var s=t.apply(this,arguments);return i&&(this.onsuccess=this.onsuccess?Ge(i,this.onsuccess):i),o&&(this.onerror=this.onerror?Ge(o,this.onerror):o),r===void 0?s===void 0?void 0:s:re(r,s)}}function cl(e,t){return e===V?t:function(){return t.apply(this,arguments)!==!1&&e.apply(this,arguments)}}function Vr(e,t){return e===V?t:function(){var n=e.apply(this,arguments);if(n&&typeof n.then=="function"){for(var r=this,i=arguments.length,o=new Array(i);i--;)o[i]=arguments[i];return n.then(function(){return t.apply(r,o)})}return t.apply(this,arguments)}}Xt.ModifyError=pn,Xt.DexieError=lt,Xt.BulkError=St;var Tt={};const Po=100,[lr,mn,cr]=typeof Promise>"u"?[]:(()=>{let e=Promise.resolve();if(typeof crypto>"u"||!crypto.subtle)return[e,Ot(e),e];const t=crypto.subtle.digest("SHA-512",new Uint8Array([0]));return[t,Ot(t),e]})(),$o=mn&&mn.then,Qt=lr&&lr.constructor,Wr=!!cr;var ur=!1,ul=cr?()=>{cr.then(qt)}:W.setImmediate?setImmediate.bind(null,qt):W.MutationObserver?()=>{var e=document.createElement("div");new MutationObserver(()=>{qt(),e=null}).observe(e,{attributes:!0}),e.setAttribute("i","1")}:()=>{setTimeout(qt,0)},Et=function(e,t){xt.push([e,t]),yn&&(ul(),yn=!1)},fr=!0,yn=!0,We=[],Jt=[],hr=null,dr=Rt,ct={id:"global",global:!0,ref:0,unhandleds:[],onunhandled:mi,pgp:!1,env:{},finalize:function(){this.unhandleds.forEach(e=>{try{mi(e[0],e[1])}catch{}})}},O=ct,xt=[],ze=0,Zt=[];function E(e){if(typeof this!="object")throw new TypeError("Promises must be constructed via new");this._listeners=[],this.onuncatched=V,this._lib=!1;var t=this._PSD=O;if(me&&(this._stackHolder=Ze(),this._prev=null,this._numPrev=0),typeof e!="function"){if(e!==Tt)throw new TypeError("Not a function");return this._state=arguments[1],this._value=arguments[2],void(this._state===!1&&mr(this,this._value))}this._state=null,this._value=null,++t.ref,Oo(this,e)}const pr={get:function(){var e=O,t=gn;function n(r,i){var o=!e.global&&(e!==O||t!==gn);const s=o&&!Ce();var a=new E((l,c)=>{zr(this,new Ao(vn(r,e,o,s),vn(i,e,o,s),l,c,e))});return me&&Io(a,this),a}return n.prototype=Tt,n},set:function(e){xe(this,"then",e&&e.prototype===Tt?pr:{get:function(){return e},set:pr.set})}};function Ao(e,t,n,r,i){this.onFulfilled=typeof e=="function"?e:null,this.onRejected=typeof t=="function"?t:null,this.resolve=n,this.reject=r,this.psd=i}function Oo(e,t){try{t(n=>{if(e._state===null){if(n===e)throw new TypeError("A promise cannot be resolved with itself.");var r=e._lib&&Kt();n&&typeof n.then=="function"?Oo(e,(i,o)=>{n instanceof E?n._then(i,o):n.then(i,o)}):(e._state=!0,e._value=n,Ro(e)),r&&Nt()}},mr.bind(null,e))}catch(n){mr(e,n)}}function mr(e,t){if(Jt.push(t),e._state===null){var n=e._lib&&Kt();t=dr(t),e._state=!1,e._value=t,me&&t!==null&&typeof t=="object"&&!t._promise&&function(r,i,o){try{r.apply(null,o)}catch(s){i&&i(s)}}(()=>{var r=Mr(t,"stack");t._promise=e,xe(t,"stack",{get:()=>ur?r&&(r.get?r.get.apply(t):r.value):e.stack})}),function(r){We.some(i=>i._value===r._value)||We.push(r)}(e),Ro(e),n&&Nt()}}function Ro(e){var t=e._listeners;e._listeners=[];for(var n=0,r=t.length;n<r;++n)zr(e,t[n]);var i=e._PSD;--i.ref||i.finalize(),ze===0&&(++ze,Et(()=>{--ze==0&&Hr()},[]))}function zr(e,t){if(e._state!==null){var n=e._state?t.onFulfilled:t.onRejected;if(n===null)return(e._state?t.resolve:t.reject)(e._value);++t.psd.ref,++ze,Et(fl,[n,e,t])}else e._listeners.push(t)}function fl(e,t,n){try{hr=t;var r,i=t._value;t._state?r=e(i):(Jt.length&&(Jt=[]),r=e(i),Jt.indexOf(i)===-1&&function(o){for(var s=We.length;s;)if(We[--s]._value===o._value)return void We.splice(s,1)}(t)),n.resolve(r)}catch(o){n.reject(o)}finally{hr=null,--ze==0&&Hr(),--n.psd.ref||n.psd.finalize()}}function To(e,t,n){if(t.length===n)return t;var r="";if(e._state===!1){var i,o,s=e._value;s!=null?(i=s.name||"Error",o=s.message||s,r=ar(s,0)):(i=s,o=""),t.push(i+(o?": "+o:"")+r)}return me&&((r=ar(e._stackHolder,2))&&t.indexOf(r)===-1&&t.push(r),e._prev&&To(e._prev,t,n)),t}function Io(e,t){var n=t?t._numPrev+1:0;n<100&&(e._prev=t,e._numPrev=n)}function qt(){Kt()&&Nt()}function Kt(){var e=fr;return fr=!1,yn=!1,e}function Nt(){var e,t,n;do for(;xt.length>0;)for(e=xt,xt=[],n=e.length,t=0;t<n;++t){var r=e[t];r[0].apply(null,r[1])}while(xt.length>0);fr=!0,yn=!0}function Hr(){var e=We;We=[],e.forEach(r=>{r._PSD.onunhandled.call(null,r._value,r)});for(var t=Zt.slice(0),n=t.length;n;)t[--n]()}function Ut(e){return new E(Tt,!1,e)}function z(e,t){var n=O;return function(){var r=Kt(),i=O;try{return De(n,!0),e.apply(this,arguments)}catch(o){t&&t(o)}finally{De(i,!1),r&&Nt()}}}ht(E.prototype,{then:pr,_then:function(e,t){zr(this,new Ao(null,null,e,t,O))},catch:function(e){if(arguments.length===1)return this.then(null,e);var t=arguments[0],n=arguments[1];return typeof t=="function"?this.then(null,r=>r instanceof t?n(r):Ut(r)):this.then(null,r=>r&&r.name===t?n(r):Ut(r))},finally:function(e){return this.then(t=>(e(),t),t=>(e(),Ut(t)))},stack:{get:function(){if(this._stack)return this._stack;try{ur=!0;var e=To(this,[],20).join(`
+From previous: `);return this._state!==null&&(this._stack=e),e}finally{ur=!1}}},timeout:function(e,t){return e<1/0?new E((n,r)=>{var i=setTimeout(()=>r(new j.Timeout(t)),e);this.then(n,r).finally(clearTimeout.bind(null,i))}):this}}),typeof Symbol<"u"&&Symbol.toStringTag&&xe(E.prototype,Symbol.toStringTag,"Dexie.Promise"),ct.env=jo(),ht(E,{all:function(){var e=we.apply(null,arguments).map(Vt);return new E(function(t,n){e.length===0&&t([]);var r=e.length;e.forEach((i,o)=>E.resolve(i).then(s=>{e[o]=s,--r||t(e)},n))})},resolve:e=>{if(e instanceof E)return e;if(e&&typeof e.then=="function")return new E((n,r)=>{e.then(n,r)});var t=new E(Tt,!0,e);return Io(t,hr),t},reject:Ut,race:function(){var e=we.apply(null,arguments).map(Vt);return new E((t,n)=>{e.map(r=>E.resolve(r).then(t,n))})},PSD:{get:()=>O,set:e=>O=e},totalEchoes:{get:()=>gn},newPSD:Le,usePSD:yt,scheduler:{get:()=>Et,set:e=>{Et=e}},rejectionMapper:{get:()=>dr,set:e=>{dr=e}},follow:(e,t)=>new E((n,r)=>Le((i,o)=>{var s=O;s.unhandleds=[],s.onunhandled=o,s.finalize=Ge(function(){(function(a){function l(){a(),Zt.splice(Zt.indexOf(l),1)}Zt.push(l),++ze,Et(()=>{--ze==0&&Hr()},[])})(()=>{this.unhandleds.length===0?i():o(this.unhandleds[0])})},s.finalize),e()},t,n,r))}),Qt&&(Qt.allSettled&&xe(E,"allSettled",function(){const e=we.apply(null,arguments).map(Vt);return new E(t=>{e.length===0&&t([]);let n=e.length;const r=new Array(n);e.forEach((i,o)=>E.resolve(i).then(s=>r[o]={status:"fulfilled",value:s},s=>r[o]={status:"rejected",reason:s}).then(()=>--n||t(r)))})}),Qt.any&&typeof AggregateError<"u"&&xe(E,"any",function(){const e=we.apply(null,arguments).map(Vt);return new E((t,n)=>{e.length===0&&n(new AggregateError([]));let r=e.length;const i=new Array(r);e.forEach((o,s)=>E.resolve(o).then(a=>t(a),a=>{i[s]=a,--r||n(new AggregateError(i))}))})}));const Z={awaits:0,echoes:0,id:0};var hl=0,en=[],Dn=0,gn=0,dl=0;function Le(e,t,n,r){var i=O,o=Object.create(i);o.parent=i,o.ref=0,o.global=!1,o.id=++dl;var s=ct.env;o.env=Wr?{Promise:E,PromiseProp:{value:E,configurable:!0,writable:!0},all:E.all,race:E.race,allSettled:E.allSettled,any:E.any,resolve:E.resolve,reject:E.reject,nthen:di(s.nthen,o),gthen:di(s.gthen,o)}:{},t&&re(o,t),++i.ref,o.finalize=function(){--this.parent.ref||this.parent.finalize()};var a=yt(o,e,n,r);return o.ref===0&&o.finalize(),a}function mt(){return Z.id||(Z.id=++hl),++Z.awaits,Z.echoes+=Po,Z.id}function Ce(){return!!Z.awaits&&(--Z.awaits==0&&(Z.id=0),Z.echoes=Z.awaits*Po,!0)}function Vt(e){return Z.echoes&&e&&e.constructor===Qt?(mt(),e.then(t=>(Ce(),t),t=>(Ce(),Q(t)))):e}function pl(e){++gn,Z.echoes&&--Z.echoes!=0||(Z.echoes=Z.id=0),en.push(O),De(e,!0)}function ml(){var e=en[en.length-1];en.pop(),De(e,!1)}function De(e,t){var n=O;if((t?!Z.echoes||Dn++&&e===O:!Dn||--Dn&&e===O)||Bo(t?pl.bind(null,e):ml),e!==O&&(O=e,n===ct&&(ct.env=jo()),Wr)){var r=ct.env.Promise,i=e.env;mn.then=i.nthen,r.prototype.then=i.gthen,(n.global||e.global)&&(Object.defineProperty(W,"Promise",i.PromiseProp),r.all=i.all,r.race=i.race,r.resolve=i.resolve,r.reject=i.reject,i.allSettled&&(r.allSettled=i.allSettled),i.any&&(r.any=i.any))}}function jo(){var e=W.Promise;return Wr?{Promise:e,PromiseProp:Object.getOwnPropertyDescriptor(W,"Promise"),all:e.all,race:e.race,allSettled:e.allSettled,any:e.any,resolve:e.resolve,reject:e.reject,nthen:mn.then,gthen:e.prototype.then}:{}}function yt(e,t,n,r,i){var o=O;try{return De(e,!0),t(n,r,i)}finally{De(o,!1)}}function Bo(e){$o.call(lr,e)}function vn(e,t,n,r){return typeof e!="function"?e:function(){var i=O;n&&mt(),De(t,!0);try{return e.apply(this,arguments)}finally{De(i,!1),r&&Bo(Ce)}}}function di(e,t){return function(n,r){return e.call(this,vn(n,t),vn(r,t))}}(""+$o).indexOf("[native code]")===-1&&(mt=Ce=V);const pi="unhandledrejection";function mi(e,t){var n;try{n=t.onuncatched(e)}catch{}if(n!==!1)try{var r,i={promise:t,reason:e};if(W.document&&document.createEvent?((r=document.createEvent("Event")).initEvent(pi,!0,!0),re(r,i)):W.CustomEvent&&re(r=new CustomEvent(pi,{detail:i}),i),r&&W.dispatchEvent&&(dispatchEvent(r),!W.PromiseRejectionEvent&&W.onunhandledrejection))try{W.onunhandledrejection(r)}catch{}me&&r&&!r.defaultPrevented&&console.warn(`Unhandled rejection: ${e.stack||e}`)}catch{}}var Q=E.reject;function yr(e,t,n,r){if(e.idbdb&&(e._state.openComplete||O.letThrough||e._vip)){var i=e._createTransaction(t,n,e._dbSchema);try{i.create(),e._state.PR1398_maxLoop=3}catch(o){return o.name===Ur.InvalidState&&e.isOpen()&&--e._state.PR1398_maxLoop>0?(console.warn("Dexie: Need to reopen db"),e._close(),e.open().then(()=>yr(e,t,n,r))):Q(o)}return i._promise(t,(o,s)=>Le(()=>(O.trans=i,r(o,s,i)))).then(o=>i._completion.then(()=>o))}if(e._state.openComplete)return Q(new j.DatabaseClosed(e._state.dbOpenError));if(!e._state.isBeingOpened){if(!e._options.autoOpen)return Q(new j.DatabaseClosed);e.open().catch(V)}return e._state.dbReadyPromise.then(()=>yr(e,t,n,r))}const yi="3.2.4",Ue="￿",gr=-1/0,ge="Invalid key provided. Keys must be of type string, number, Date or Array<string | number | Date>.",Lo="String expected.",Ct=[],$n=typeof navigator<"u"&&/(MSIE|Trident|Edge)/.test(navigator.userAgent),yl=$n,gl=$n,Do=e=>!/(dexie\.js|dexie\.min\.js)/.test(e),An="__dbnames",Kn="readonly",Nn="readwrite";function Xe(e,t){return e?t?function(){return e.apply(this,arguments)&&t.apply(this,arguments)}:e:t}const Ko={type:3,lower:-1/0,lowerOpen:!1,upper:[[]],upperOpen:!1};function Wt(e){return typeof e!="string"||/\./.test(e)?t=>t:t=>(t[e]===void 0&&e in t&&delete(t=Dt(t))[e],t)}class vl{_trans(t,n,r){const i=this._tx||O.trans,o=this.name;function s(l,c,u){if(!u.schema[o])throw new j.NotFound("Table "+o+" not part of transaction");return n(u.idbtrans,u)}const a=Kt();try{return i&&i.db===this.db?i===O.trans?i._promise(t,s,r):Le(()=>i._promise(t,s,r),{trans:i,transless:O.transless||O}):yr(this.db,t,[this.name],s)}finally{a&&Nt()}}get(t,n){return t&&t.constructor===Object?this.where(t).first(n):this._trans("readonly",r=>this.core.get({trans:r,key:t}).then(i=>this.hook.reading.fire(i))).then(n)}where(t){if(typeof t=="string")return new this.db.WhereClause(this,t);if(ee(t))return new this.db.WhereClause(this,`[${t.join("+")}]`);const n=X(t);if(n.length===1)return this.where(n[0]).equals(t[n[0]]);const r=this.schema.indexes.concat(this.schema.primKey).filter(c=>c.compound&&n.every(u=>c.keyPath.indexOf(u)>=0)&&c.keyPath.every(u=>n.indexOf(u)>=0))[0];if(r&&this.db._maxKey!==Ue)return this.where(r.name).equals(r.keyPath.map(c=>t[c]));!r&&me&&console.warn(`The query ${JSON.stringify(t)} on ${this.name} would benefit of a compound index [${n.join("+")}]`);const{idxByName:i}=this.schema,o=this.db._deps.indexedDB;function s(c,u){try{return o.cmp(c,u)===0}catch{return!1}}const[a,l]=n.reduce(([c,u],f)=>{const h=i[f],d=t[f];return[c||h,c||!h?Xe(u,h&&h.multi?v=>{const m=Se(v,f);return ee(m)&&m.some(g=>s(d,g))}:v=>s(d,Se(v,f))):u]},[null,null]);return a?this.where(a.name).equals(t[a.keyPath]).filter(l):r?this.filter(l):this.where(n).equals("")}filter(t){return this.toCollection().and(t)}count(t){return this.toCollection().count(t)}offset(t){return this.toCollection().offset(t)}limit(t){return this.toCollection().limit(t)}each(t){return this.toCollection().each(t)}toArray(t){return this.toCollection().toArray(t)}toCollection(){return new this.db.Collection(new this.db.WhereClause(this))}orderBy(t){return new this.db.Collection(new this.db.WhereClause(this,ee(t)?`[${t.join("+")}]`:t))}reverse(){return this.toCollection().reverse()}mapToClass(t){this.schema.mappedClass=t;const n=r=>{if(!r)return r;const i=Object.create(t.prototype);for(var o in r)if(le(r,o))try{i[o]=r[o]}catch{}return i};return this.schema.readHook&&this.hook.reading.unsubscribe(this.schema.readHook),this.schema.readHook=n,this.hook("reading",n),t}defineClass(){return this.mapToClass(function(t){re(this,t)})}add(t,n){const{auto:r,keyPath:i}=this.schema.primKey;let o=t;return i&&r&&(o=Wt(i)(t)),this._trans("readwrite",s=>this.core.mutate({trans:s,type:"add",keys:n!=null?[n]:null,values:[o]})).then(s=>s.numFailures?E.reject(s.failures[0]):s.lastResult).then(s=>{if(i)try{ue(t,i,s)}catch{}return s})}update(t,n){if(typeof t!="object"||ee(t))return this.where(":id").equals(t).modify(n);{const r=Se(t,this.schema.primKey.keyPath);if(r===void 0)return Q(new j.InvalidArgument("Given object does not contain its primary key"));try{typeof n!="function"?X(n).forEach(i=>{ue(t,i,n[i])}):n(t,{value:t,primKey:r})}catch{}return this.where(":id").equals(r).modify(n)}}put(t,n){const{auto:r,keyPath:i}=this.schema.primKey;let o=t;return i&&r&&(o=Wt(i)(t)),this._trans("readwrite",s=>this.core.mutate({trans:s,type:"put",values:[o],keys:n!=null?[n]:null})).then(s=>s.numFailures?E.reject(s.failures[0]):s.lastResult).then(s=>{if(i)try{ue(t,i,s)}catch{}return s})}delete(t){return this._trans("readwrite",n=>this.core.mutate({trans:n,type:"delete",keys:[t]})).then(n=>n.numFailures?E.reject(n.failures[0]):void 0)}clear(){return this._trans("readwrite",t=>this.core.mutate({trans:t,type:"deleteRange",range:Ko})).then(t=>t.numFailures?E.reject(t.failures[0]):void 0)}bulkGet(t){return this._trans("readonly",n=>this.core.getMany({keys:t,trans:n}).then(r=>r.map(i=>this.hook.reading.fire(i))))}bulkAdd(t,n,r){const i=Array.isArray(n)?n:void 0,o=(r=r||(i?void 0:n))?r.allKeys:void 0;return this._trans("readwrite",s=>{const{auto:a,keyPath:l}=this.schema.primKey;if(l&&i)throw new j.InvalidArgument("bulkAdd(): keys argument invalid on tables with inbound keys");if(i&&i.length!==t.length)throw new j.InvalidArgument("Arguments objects and keys must have the same length");const c=t.length;let u=l&&a?t.map(Wt(l)):t;return this.core.mutate({trans:s,type:"add",keys:i,values:u,wantResults:o}).then(({numFailures:f,results:h,lastResult:d,failures:v})=>{if(f===0)return o?h:d;throw new St(`${this.name}.bulkAdd(): ${f} of ${c} operations failed`,v)})})}bulkPut(t,n,r){const i=Array.isArray(n)?n:void 0,o=(r=r||(i?void 0:n))?r.allKeys:void 0;return this._trans("readwrite",s=>{const{auto:a,keyPath:l}=this.schema.primKey;if(l&&i)throw new j.InvalidArgument("bulkPut(): keys argument invalid on tables with inbound keys");if(i&&i.length!==t.length)throw new j.InvalidArgument("Arguments objects and keys must have the same length");const c=t.length;let u=l&&a?t.map(Wt(l)):t;return this.core.mutate({trans:s,type:"put",keys:i,values:u,wantResults:o}).then(({numFailures:f,results:h,lastResult:d,failures:v})=>{if(f===0)return o?h:d;throw new St(`${this.name}.bulkPut(): ${f} of ${c} operations failed`,v)})})}bulkDelete(t){const n=t.length;return this._trans("readwrite",r=>this.core.mutate({trans:r,type:"delete",keys:t})).then(({numFailures:r,lastResult:i,failures:o})=>{if(r===0)return i;throw new St(`${this.name}.bulkDelete(): ${r} of ${n} operations failed`,o)})}}function Pt(e){var t={},n=function(s,a){if(a){for(var l=arguments.length,c=new Array(l-1);--l;)c[l-1]=arguments[l];return t[s].subscribe.apply(null,c),e}if(typeof s=="string")return t[s]};n.addEventType=o;for(var r=1,i=arguments.length;r<i;++r)o(arguments[r]);return n;function o(s,a,l){if(typeof s!="object"){var c;a||(a=cl),l||(l=V);var u={subscribers:[],fire:l,subscribe:function(f){u.subscribers.indexOf(f)===-1&&(u.subscribers.push(f),u.fire=a(u.fire,f))},unsubscribe:function(f){u.subscribers=u.subscribers.filter(function(h){return h!==f}),u.fire=u.subscribers.reduce(a,l)}};return t[s]=n[s]=u,u}X(c=s).forEach(function(f){var h=c[f];if(ee(h))o(f,c[f][0],c[f][1]);else{if(h!=="asap")throw new j.InvalidArgument("Invalid event config");var d=o(f,Rt,function(){for(var v=arguments.length,m=new Array(v);v--;)m[v]=arguments[v];d.subscribers.forEach(function(g){vo(function(){g.apply(null,m)})})})}})}}function vt(e,t){return at(t).from({prototype:e}),t}function nt(e,t){return!(e.filter||e.algorithm||e.or)&&(t?e.justLimit:!e.replayFilter)}function Mn(e,t){e.filter=Xe(e.filter,t)}function Fn(e,t,n){var r=e.replayFilter;e.replayFilter=r?()=>Xe(r(),t()):t,e.justLimit=n&&!r}function tn(e,t){if(e.isPrimKey)return t.primaryKey;const n=t.getIndexByKeyPath(e.index);if(!n)throw new j.Schema("KeyPath "+e.index+" on object store "+t.name+" is not indexed");return n}function gi(e,t,n){const r=tn(e,t.schema);return t.openCursor({trans:n,values:!e.keysOnly,reverse:e.dir==="prev",unique:!!e.unique,query:{index:r,range:e.range}})}function zt(e,t,n,r){const i=e.replayFilter?Xe(e.filter,e.replayFilter()):e.filter;if(e.or){const o={},s=(a,l,c)=>{if(!i||i(l,c,h=>l.stop(h),h=>l.fail(h))){var u=l.primaryKey,f=""+u;f==="[object ArrayBuffer]"&&(f=""+new Uint8Array(u)),le(o,f)||(o[f]=!0,t(a,l,c))}};return Promise.all([e.or._iterate(s,n),vi(gi(e,r,n),e.algorithm,s,!e.keysOnly&&e.valueMapper)])}return vi(gi(e,r,n),Xe(e.algorithm,i),t,!e.keysOnly&&e.valueMapper)}function vi(e,t,n,r){var i=z(r?(o,s,a)=>n(r(o),s,a):n);return e.then(o=>{if(o)return o.start(()=>{var s=()=>o.continue();t&&!t(o,a=>s=a,a=>{o.stop(a),s=V},a=>{o.fail(a),s=V})||i(o.value,o,a=>s=a),s()})})}function ne(e,t){try{const n=bi(e),r=bi(t);if(n!==r)return n==="Array"?1:r==="Array"?-1:n==="binary"?1:r==="binary"?-1:n==="string"?1:r==="string"?-1:n==="Date"?1:r!=="Date"?NaN:-1;switch(n){case"number":case"Date":case"string":return e>t?1:e<t?-1:0;case"binary":return function(i,o){const s=i.length,a=o.length,l=s<a?s:a;for(let c=0;c<l;++c)if(i[c]!==o[c])return i[c]<o[c]?-1:1;return s===a?0:s<a?-1:1}(wi(e),wi(t));case"Array":return function(i,o){const s=i.length,a=o.length,l=s<a?s:a;for(let c=0;c<l;++c){const u=ne(i[c],o[c]);if(u!==0)return u}return s===a?0:s<a?-1:1}(e,t)}}catch{}return NaN}function bi(e){const t=typeof e;if(t!=="object")return t;if(ArrayBuffer.isView(e))return"binary";const n=or(e);return n==="ArrayBuffer"?"binary":n}function wi(e){return e instanceof Uint8Array?e:ArrayBuffer.isView(e)?new Uint8Array(e.buffer,e.byteOffset,e.byteLength):new Uint8Array(e)}class bl{_read(t,n){var r=this._ctx;return r.error?r.table._trans(null,Q.bind(null,r.error)):r.table._trans("readonly",t).then(n)}_write(t){var n=this._ctx;return n.error?n.table._trans(null,Q.bind(null,n.error)):n.table._trans("readwrite",t,"locked")}_addAlgorithm(t){var n=this._ctx;n.algorithm=Xe(n.algorithm,t)}_iterate(t,n){return zt(this._ctx,t,n,this._ctx.table.core)}clone(t){var n=Object.create(this.constructor.prototype),r=Object.create(this._ctx);return t&&re(r,t),n._ctx=r,n}raw(){return this._ctx.valueMapper=null,this}each(t){var n=this._ctx;return this._read(r=>zt(n,t,r,n.table.core))}count(t){return this._read(n=>{const r=this._ctx,i=r.table.core;if(nt(r,!0))return i.count({trans:n,query:{index:tn(r,i.schema),range:r.range}}).then(s=>Math.min(s,r.limit));var o=0;return zt(r,()=>(++o,!1),n,i).then(()=>o)}).then(t)}sortBy(t,n){const r=t.split(".").reverse(),i=r[0],o=r.length-1;function s(c,u){return u?s(c[r[u]],u-1):c[i]}var a=this._ctx.dir==="next"?1:-1;function l(c,u){var f=s(c,o),h=s(u,o);return f<h?-a:f>h?a:0}return this.toArray(function(c){return c.sort(l)}).then(n)}toArray(t){return this._read(n=>{var r=this._ctx;if(r.dir==="next"&&nt(r,!0)&&r.limit>0){const{valueMapper:i}=r,o=tn(r,r.table.core.schema);return r.table.core.query({trans:n,limit:r.limit,values:!0,query:{index:o,range:r.range}}).then(({result:s})=>i?s.map(i):s)}{const i=[];return zt(r,o=>i.push(o),n,r.table.core).then(()=>i)}},t)}offset(t){var n=this._ctx;return t<=0||(n.offset+=t,nt(n)?Fn(n,()=>{var r=t;return(i,o)=>r===0||(r===1?(--r,!1):(o(()=>{i.advance(r),r=0}),!1))}):Fn(n,()=>{var r=t;return()=>--r<0})),this}limit(t){return this._ctx.limit=Math.min(this._ctx.limit,t),Fn(this._ctx,()=>{var n=t;return function(r,i,o){return--n<=0&&i(o),n>=0}},!0),this}until(t,n){return Mn(this._ctx,function(r,i,o){return!t(r.value)||(i(o),n)}),this}first(t){return this.limit(1).toArray(function(n){return n[0]}).then(t)}last(t){return this.reverse().first(t)}filter(t){var n,r;return Mn(this._ctx,function(i){return t(i.value)}),n=this._ctx,r=t,n.isMatch=Xe(n.isMatch,r),this}and(t){return this.filter(t)}or(t){return new this.db.WhereClause(this._ctx.table,t,this)}reverse(){return this._ctx.dir=this._ctx.dir==="prev"?"next":"prev",this._ondirectionchange&&this._ondirectionchange(this._ctx.dir),this}desc(){return this.reverse()}eachKey(t){var n=this._ctx;return n.keysOnly=!n.isMatch,this.each(function(r,i){t(i.key,i)})}eachUniqueKey(t){return this._ctx.unique="unique",this.eachKey(t)}eachPrimaryKey(t){var n=this._ctx;return n.keysOnly=!n.isMatch,this.each(function(r,i){t(i.primaryKey,i)})}keys(t){var n=this._ctx;n.keysOnly=!n.isMatch;var r=[];return this.each(function(i,o){r.push(o.key)}).then(function(){return r}).then(t)}primaryKeys(t){var n=this._ctx;if(n.dir==="next"&&nt(n,!0)&&n.limit>0)return this._read(i=>{var o=tn(n,n.table.core.schema);return n.table.core.query({trans:i,values:!1,limit:n.limit,query:{index:o,range:n.range}})}).then(({result:i})=>i).then(t);n.keysOnly=!n.isMatch;var r=[];return this.each(function(i,o){r.push(o.primaryKey)}).then(function(){return r}).then(t)}uniqueKeys(t){return this._ctx.unique="unique",this.keys(t)}firstKey(t){return this.limit(1).keys(function(n){return n[0]}).then(t)}lastKey(t){return this.reverse().firstKey(t)}distinct(){var t=this._ctx,n=t.index&&t.table.schema.idxByName[t.index];if(!n||!n.multi)return this;var r={};return Mn(this._ctx,function(i){var o=i.primaryKey.toString(),s=le(r,o);return r[o]=!0,!s}),this}modify(t){var n=this._ctx;return this._write(r=>{var i;if(typeof t=="function")i=t;else{var o=X(t),s=o.length;i=function(m){for(var g=!1,y=0;y<s;++y){var p=o[y],b=t[p];Se(m,p)!==b&&(ue(m,p,b),g=!0)}return g}}const a=n.table.core,{outbound:l,extractKey:c}=a.schema.primaryKey,u=this.db._options.modifyChunkSize||200,f=[];let h=0;const d=[],v=(m,g)=>{const{failures:y,numFailures:p}=g;h+=m-p;for(let b of X(y))f.push(y[b])};return this.clone().primaryKeys().then(m=>{const g=y=>{const p=Math.min(u,m.length-y);return a.getMany({trans:r,keys:m.slice(y,y+p),cache:"immutable"}).then(b=>{const w=[],k=[],S=l?[]:null,x=[];for(let _=0;_<p;++_){const R=b[_],$={value:Dt(R),primKey:m[y+_]};i.call($,$.value,$)!==!1&&($.value==null?x.push(m[y+_]):l||ne(c(R),c($.value))===0?(k.push($.value),l&&S.push(m[y+_])):(x.push(m[y+_]),w.push($.value)))}const T=nt(n)&&n.limit===1/0&&(typeof t!="function"||t===qn)&&{index:n.index,range:n.range};return Promise.resolve(w.length>0&&a.mutate({trans:r,type:"add",values:w}).then(_=>{for(let R in _.failures)x.splice(parseInt(R),1);v(w.length,_)})).then(()=>(k.length>0||T&&typeof t=="object")&&a.mutate({trans:r,type:"put",keys:S,values:k,criteria:T,changeSpec:typeof t!="function"&&t}).then(_=>v(k.length,_))).then(()=>(x.length>0||T&&t===qn)&&a.mutate({trans:r,type:"delete",keys:x,criteria:T}).then(_=>v(x.length,_))).then(()=>m.length>y+p&&g(y+u))})};return g(0).then(()=>{if(f.length>0)throw new pn("Error modifying one or more objects",f,h,d);return m.length})})})}delete(){var t=this._ctx,n=t.range;return nt(t)&&(t.isPrimKey&&!gl||n.type===3)?this._write(r=>{const{primaryKey:i}=t.table.core.schema,o=n;return t.table.core.count({trans:r,query:{index:i,range:o}}).then(s=>t.table.core.mutate({trans:r,type:"deleteRange",range:o}).then(({failures:a,lastResult:l,results:c,numFailures:u})=>{if(u)throw new pn("Could not delete some values",Object.keys(a).map(f=>a[f]),s-u);return s-u}))}):this.modify(qn)}}const qn=(e,t)=>t.value=null;function wl(e,t){return e<t?-1:e===t?0:1}function _l(e,t){return e>t?-1:e===t?0:1}function se(e,t,n){var r=e instanceof Mo?new e.Collection(e):e;return r._ctx.error=n?new n(t):new TypeError(t),r}function rt(e){return new e.Collection(e,()=>No("")).limit(0)}function kl(e,t,n,r,i,o){for(var s=Math.min(e.length,r.length),a=-1,l=0;l<s;++l){var c=t[l];if(c!==r[l])return i(e[l],n[l])<0?e.substr(0,l)+n[l]+n.substr(l+1):i(e[l],r[l])<0?e.substr(0,l)+r[l]+n.substr(l+1):a>=0?e.substr(0,a)+t[a]+n.substr(a+1):null;i(e[l],c)<0&&(a=l)}return s<r.length&&o==="next"?e+n.substr(e.length):s<e.length&&o==="prev"?e.substr(0,n.length):a<0?null:e.substr(0,a)+r[a]+n.substr(a+1)}function Ht(e,t,n,r){var i,o,s,a,l,c,u,f=n.length;if(!n.every(m=>typeof m=="string"))return se(e,Lo);function h(m){i=function(y){return y==="next"?p=>p.toUpperCase():p=>p.toLowerCase()}(m),o=function(y){return y==="next"?p=>p.toLowerCase():p=>p.toUpperCase()}(m),s=m==="next"?wl:_l;var g=n.map(function(y){return{lower:o(y),upper:i(y)}}).sort(function(y,p){return s(y.lower,p.lower)});a=g.map(function(y){return y.upper}),l=g.map(function(y){return y.lower}),c=m,u=m==="next"?"":r}h("next");var d=new e.Collection(e,()=>$e(a[0],l[f-1]+r));d._ondirectionchange=function(m){h(m)};var v=0;return d._addAlgorithm(function(m,g,y){var p=m.key;if(typeof p!="string")return!1;var b=o(p);if(t(b,l,v))return!0;for(var w=null,k=v;k<f;++k){var S=kl(p,b,a[k],l[k],s,c);S===null&&w===null?v=k+1:(w===null||s(w,S)>0)&&(w=S)}return g(w!==null?function(){m.continue(w+u)}:y),!1}),d}function $e(e,t,n,r){return{type:2,lower:e,upper:t,lowerOpen:n,upperOpen:r}}function No(e){return{type:1,lower:e,upper:e}}class Mo{get Collection(){return this._ctx.table.db.Collection}between(t,n,r,i){r=r!==!1,i=i===!0;try{return this._cmp(t,n)>0||this._cmp(t,n)===0&&(r||i)&&(!r||!i)?rt(this):new this.Collection(this,()=>$e(t,n,!r,!i))}catch{return se(this,ge)}}equals(t){return t==null?se(this,ge):new this.Collection(this,()=>No(t))}above(t){return t==null?se(this,ge):new this.Collection(this,()=>$e(t,void 0,!0))}aboveOrEqual(t){return t==null?se(this,ge):new this.Collection(this,()=>$e(t,void 0,!1))}below(t){return t==null?se(this,ge):new this.Collection(this,()=>$e(void 0,t,!1,!0))}belowOrEqual(t){return t==null?se(this,ge):new this.Collection(this,()=>$e(void 0,t))}startsWith(t){return typeof t!="string"?se(this,Lo):this.between(t,t+Ue,!0,!0)}startsWithIgnoreCase(t){return t===""?this.startsWith(t):Ht(this,(n,r)=>n.indexOf(r[0])===0,[t],Ue)}equalsIgnoreCase(t){return Ht(this,(n,r)=>n===r[0],[t],"")}anyOfIgnoreCase(){var t=we.apply(it,arguments);return t.length===0?rt(this):Ht(this,(n,r)=>r.indexOf(n)!==-1,t,"")}startsWithAnyOfIgnoreCase(){var t=we.apply(it,arguments);return t.length===0?rt(this):Ht(this,(n,r)=>r.some(i=>n.indexOf(i)===0),t,Ue)}anyOf(){const t=we.apply(it,arguments);let n=this._cmp;try{t.sort(n)}catch{return se(this,ge)}if(t.length===0)return rt(this);const r=new this.Collection(this,()=>$e(t[0],t[t.length-1]));r._ondirectionchange=o=>{n=o==="next"?this._ascending:this._descending,t.sort(n)};let i=0;return r._addAlgorithm((o,s,a)=>{const l=o.key;for(;n(l,t[i])>0;)if(++i,i===t.length)return s(a),!1;return n(l,t[i])===0||(s(()=>{o.continue(t[i])}),!1)}),r}notEqual(t){return this.inAnyRange([[gr,t],[t,this.db._maxKey]],{includeLowers:!1,includeUppers:!1})}noneOf(){const t=we.apply(it,arguments);if(t.length===0)return new this.Collection(this);try{t.sort(this._ascending)}catch{return se(this,ge)}const n=t.reduce((r,i)=>r?r.concat([[r[r.length-1][1],i]]):[[gr,i]],null);return n.push([t[t.length-1],this.db._maxKey]),this.inAnyRange(n,{includeLowers:!1,includeUppers:!1})}inAnyRange(t,n){const r=this._cmp,i=this._ascending,o=this._descending,s=this._min,a=this._max;if(t.length===0)return rt(this);if(!t.every(p=>p[0]!==void 0&&p[1]!==void 0&&i(p[0],p[1])<=0))return se(this,"First argument to inAnyRange() must be an Array of two-value Arrays [lower,upper] where upper must not be lower than lower",j.InvalidArgument);const l=!n||n.includeLowers!==!1,c=n&&n.includeUppers===!0;let u,f=i;function h(p,b){return f(p[0],b[0])}try{u=t.reduce(function(p,b){let w=0,k=p.length;for(;w<k;++w){const S=p[w];if(r(b[0],S[1])<0&&r(b[1],S[0])>0){S[0]=s(S[0],b[0]),S[1]=a(S[1],b[1]);break}}return w===k&&p.push(b),p},[]),u.sort(h)}catch{return se(this,ge)}let d=0;const v=c?p=>i(p,u[d][1])>0:p=>i(p,u[d][1])>=0,m=l?p=>o(p,u[d][0])>0:p=>o(p,u[d][0])>=0;let g=v;const y=new this.Collection(this,()=>$e(u[0][0],u[u.length-1][1],!l,!c));return y._ondirectionchange=p=>{p==="next"?(g=v,f=i):(g=m,f=o),u.sort(h)},y._addAlgorithm((p,b,w)=>{for(var k=p.key;g(k);)if(++d,d===u.length)return b(w),!1;return!!function(S){return!v(S)&&!m(S)}(k)||(this._cmp(k,u[d][1])===0||this._cmp(k,u[d][0])===0||b(()=>{f===i?p.continue(u[d][0]):p.continue(u[d][1])}),!1)}),y}startsWithAnyOf(){const t=we.apply(it,arguments);return t.every(n=>typeof n=="string")?t.length===0?rt(this):this.inAnyRange(t.map(n=>[n,n+Ue])):se(this,"startsWithAnyOf() only works with strings")}}function de(e){return z(function(t){return It(t),e(t.target.error),!1})}function It(e){e.stopPropagation&&e.stopPropagation(),e.preventDefault&&e.preventDefault()}const jt="storagemutated",Oe="x-storagemutated-1",Ke=Pt(null,jt);class xl{_lock(){return kt(!O.global),++this._reculock,this._reculock!==1||O.global||(O.lockOwnerFor=this),this}_unlock(){if(kt(!O.global),--this._reculock==0)for(O.global||(O.lockOwnerFor=null);this._blockedFuncs.length>0&&!this._locked();){var t=this._blockedFuncs.shift();try{yt(t[1],t[0])}catch{}}return this}_locked(){return this._reculock&&O.lockOwnerFor!==this}create(t){if(!this.mode)return this;const n=this.db.idbdb,r=this.db._state.dbOpenError;if(kt(!this.idbtrans),!t&&!n)switch(r&&r.name){case"DatabaseClosedError":throw new j.DatabaseClosed(r);case"MissingAPIError":throw new j.MissingAPI(r.message,r);default:throw new j.OpenFailed(r)}if(!this.active)throw new j.TransactionInactive;return kt(this._completion._state===null),(t=this.idbtrans=t||(this.db.core?this.db.core.transaction(this.storeNames,this.mode,{durability:this.chromeTransactionDurability}):n.transaction(this.storeNames,this.mode,{durability:this.chromeTransactionDurability}))).onerror=z(i=>{It(i),this._reject(t.error)}),t.onabort=z(i=>{It(i),this.active&&this._reject(new j.Abort(t.error)),this.active=!1,this.on("abort").fire(i)}),t.oncomplete=z(()=>{this.active=!1,this._resolve(),"mutatedParts"in t&&Ke.storagemutated.fire(t.mutatedParts)}),this}_promise(t,n,r){if(t==="readwrite"&&this.mode!=="readwrite")return Q(new j.ReadOnly("Transaction is readonly"));if(!this.active)return Q(new j.TransactionInactive);if(this._locked())return new E((o,s)=>{this._blockedFuncs.push([()=>{this._promise(t,n,r).then(o,s)},O])});if(r)return Le(()=>{var o=new E((s,a)=>{this._lock();const l=n(s,a,this);l&&l.then&&l.then(s,a)});return o.finally(()=>this._unlock()),o._lib=!0,o});var i=new E((o,s)=>{var a=n(o,s,this);a&&a.then&&a.then(o,s)});return i._lib=!0,i}_root(){return this.parent?this.parent._root():this}waitFor(t){var n=this._root();const r=E.resolve(t);if(n._waitingFor)n._waitingFor=n._waitingFor.then(()=>r);else{n._waitingFor=r,n._waitingQueue=[];var i=n.idbtrans.objectStore(n.storeNames[0]);(function s(){for(++n._spinCount;n._waitingQueue.length;)n._waitingQueue.shift()();n._waitingFor&&(i.get(-1/0).onsuccess=s)})()}var o=n._waitingFor;return new E((s,a)=>{r.then(l=>n._waitingQueue.push(z(s.bind(null,l))),l=>n._waitingQueue.push(z(a.bind(null,l)))).finally(()=>{n._waitingFor===o&&(n._waitingFor=null)})})}abort(){this.active&&(this.active=!1,this.idbtrans&&this.idbtrans.abort(),this._reject(new j.Abort))}table(t){const n=this._memoizedTables||(this._memoizedTables={});if(le(n,t))return n[t];const r=this.schema[t];if(!r)throw new j.NotFound("Table "+t+" not part of transaction");const i=new this.db.Table(t,r,this);return i.core=this.db.core.table(t),n[t]=i,i}}function vr(e,t,n,r,i,o,s){return{name:e,keyPath:t,unique:n,multi:r,auto:i,compound:o,src:(n&&!s?"&":"")+(r?"*":"")+(i?"++":"")+Fo(t)}}function Fo(e){return typeof e=="string"?e:e?"["+[].join.call(e,"+")+"]":""}function qo(e,t,n){return{name:e,primKey:t,indexes:n,mappedClass:null,idxByName:bo(n,r=>[r.name,r])}}let Bt=e=>{try{return e.only([[]]),Bt=()=>[[]],[[]]}catch{return Bt=()=>Ue,Ue}};function br(e){return e==null?()=>{}:typeof e=="string"?function(t){return t.split(".").length===1?r=>r[t]:r=>Se(r,t)}(e):t=>Se(t,e)}function _i(e){return[].slice.call(e)}let Sl=0;function $t(e){return e==null?":id":typeof e=="string"?e:`[${e.join("+")}]`}function El(e,t,n){function r(l){if(l.type===3)return null;if(l.type===4)throw new Error("Cannot convert never type to IDBKeyRange");const{lower:c,upper:u,lowerOpen:f,upperOpen:h}=l;return c===void 0?u===void 0?null:t.upperBound(u,!!h):u===void 0?t.lowerBound(c,!!f):t.bound(c,u,!!f,!!h)}const{schema:i,hasGetAll:o}=function(l,c){const u=_i(l.objectStoreNames);return{schema:{name:l.name,tables:u.map(f=>c.objectStore(f)).map(f=>{const{keyPath:h,autoIncrement:d}=f,v=ee(h),m=h==null,g={},y={name:f.name,primaryKey:{name:null,isPrimaryKey:!0,outbound:m,compound:v,keyPath:h,autoIncrement:d,unique:!0,extractKey:br(h)},indexes:_i(f.indexNames).map(p=>f.index(p)).map(p=>{const{name:b,unique:w,multiEntry:k,keyPath:S}=p,x={name:b,compound:ee(S),keyPath:S,unique:w,multiEntry:k,extractKey:br(S)};return g[$t(S)]=x,x}),getIndexByKeyPath:p=>g[$t(p)]};return g[":id"]=y.primaryKey,h!=null&&(g[$t(h)]=y.primaryKey),y})},hasGetAll:u.length>0&&"getAll"in c.objectStore(u[0])&&!(typeof navigator<"u"&&/Safari/.test(navigator.userAgent)&&!/(Chrome\/|Edge\/)/.test(navigator.userAgent)&&[].concat(navigator.userAgent.match(/Safari\/(\d*)/))[1]<604)}}(e,n),s=i.tables.map(l=>function(c){const u=c.name;return{name:u,schema:c,mutate:function({trans:f,type:h,keys:d,values:v,range:m}){return new Promise((g,y)=>{g=z(g);const p=f.objectStore(u),b=p.keyPath==null,w=h==="put"||h==="add";if(!w&&h!=="delete"&&h!=="deleteRange")throw new Error("Invalid operation type: "+h);const{length:k}=d||v||{length:1};if(d&&v&&d.length!==v.length)throw new Error("Given keys array must have same length as given values array.");if(k===0)return g({numFailures:0,failures:{},results:[],lastResult:void 0});let S;const x=[],T=[];let _=0;const R=A=>{++_,It(A)};if(h==="deleteRange"){if(m.type===4)return g({numFailures:_,failures:T,results:[],lastResult:void 0});m.type===3?x.push(S=p.clear()):x.push(S=p.delete(r(m)))}else{const[A,I]=w?b?[v,d]:[v,null]:[d,null];if(w)for(let D=0;D<k;++D)x.push(S=I&&I[D]!==void 0?p[h](A[D],I[D]):p[h](A[D])),S.onerror=R;else for(let D=0;D<k;++D)x.push(S=p[h](A[D])),S.onerror=R}const $=A=>{const I=A.target.result;x.forEach((D,fe)=>D.error!=null&&(T[fe]=D.error)),g({numFailures:_,failures:T,results:h==="delete"?d:x.map(D=>D.result),lastResult:I})};S.onerror=A=>{R(A),$(A)},S.onsuccess=$})},getMany:({trans:f,keys:h})=>new Promise((d,v)=>{d=z(d);const m=f.objectStore(u),g=h.length,y=new Array(g);let p,b=0,w=0;const k=x=>{const T=x.target;y[T._pos]=T.result,++w===b&&d(y)},S=de(v);for(let x=0;x<g;++x)h[x]!=null&&(p=m.get(h[x]),p._pos=x,p.onsuccess=k,p.onerror=S,++b);b===0&&d(y)}),get:({trans:f,key:h})=>new Promise((d,v)=>{d=z(d);const m=f.objectStore(u).get(h);m.onsuccess=g=>d(g.target.result),m.onerror=de(v)}),query:function(f){return h=>new Promise((d,v)=>{d=z(d);const{trans:m,values:g,limit:y,query:p}=h,b=y===1/0?void 0:y,{index:w,range:k}=p,S=m.objectStore(u),x=w.isPrimaryKey?S:S.index(w.name),T=r(k);if(y===0)return d({result:[]});if(f){const _=g?x.getAll(T,b):x.getAllKeys(T,b);_.onsuccess=R=>d({result:R.target.result}),_.onerror=de(v)}else{let _=0;const R=g||!("openKeyCursor"in x)?x.openCursor(T):x.openKeyCursor(T),$=[];R.onsuccess=A=>{const I=R.result;return I?($.push(g?I.value:I.primaryKey),++_===y?d({result:$}):void I.continue()):d({result:$})},R.onerror=de(v)}})}(o),openCursor:function({trans:f,values:h,query:d,reverse:v,unique:m}){return new Promise((g,y)=>{g=z(g);const{index:p,range:b}=d,w=f.objectStore(u),k=p.isPrimaryKey?w:w.index(p.name),S=v?m?"prevunique":"prev":m?"nextunique":"next",x=h||!("openKeyCursor"in k)?k.openCursor(r(b),S):k.openKeyCursor(r(b),S);x.onerror=de(y),x.onsuccess=z(T=>{const _=x.result;if(!_)return void g(null);_.___id=++Sl,_.done=!1;const R=_.continue.bind(_);let $=_.continuePrimaryKey;$&&($=$.bind(_));const A=_.advance.bind(_),I=()=>{throw new Error("Cursor not stopped")};_.trans=f,_.stop=_.continue=_.continuePrimaryKey=_.advance=()=>{throw new Error("Cursor not started")},_.fail=z(y),_.next=function(){let D=1;return this.start(()=>D--?this.continue():this.stop()).then(()=>this)},_.start=D=>{const fe=new Promise((H,oe)=>{H=z(H),x.onerror=de(oe),_.fail=oe,_.stop=ie=>{_.stop=_.continue=_.continuePrimaryKey=_.advance=I,H(ie)}}),he=()=>{if(x.result)try{D()}catch(H){_.fail(H)}else _.done=!0,_.start=()=>{throw new Error("Cursor behind last entry")},_.stop()};return x.onsuccess=z(H=>{x.onsuccess=he,he()}),_.continue=R,_.continuePrimaryKey=$,_.advance=A,he(),fe},g(_)},y)})},count({query:f,trans:h}){const{index:d,range:v}=f;return new Promise((m,g)=>{const y=h.objectStore(u),p=d.isPrimaryKey?y:y.index(d.name),b=r(v),w=b?p.count(b):p.count();w.onsuccess=z(k=>m(k.target.result)),w.onerror=de(g)})}}}(l)),a={};return s.forEach(l=>a[l.name]=l),{stack:"dbcore",transaction:e.transaction.bind(e),table(l){if(!a[l])throw new Error(`Table '${l}' not found`);return a[l]},MIN_KEY:-1/0,MAX_KEY:Bt(t),schema:i}}function wr({_novip:e},t){const n=t.db,r=function(i,o,{IDBKeyRange:s,indexedDB:a},l){return{dbcore:function(u,f){return f.reduce((h,{create:d})=>({...h,...d(h)}),u)}(El(o,s,l),i.dbcore)}}(e._middlewares,n,e._deps,t);e.core=r.dbcore,e.tables.forEach(i=>{const o=i.name;e.core.schema.tables.some(s=>s.name===o)&&(i.core=e.core.table(o),e[o]instanceof e.Table&&(e[o].core=i.core))})}function bn({_novip:e},t,n,r){n.forEach(i=>{const o=r[i];t.forEach(s=>{const a=Mr(s,i);(!a||"value"in a&&a.value===void 0)&&(s===e.Transaction.prototype||s instanceof e.Transaction?xe(s,i,{get(){return this.table(i)},set(l){yo(this,i,{value:l,writable:!0,configurable:!0,enumerable:!0})}}):s[i]=new e.Table(i,o))})})}function _r({_novip:e},t){t.forEach(n=>{for(let r in n)n[r]instanceof e.Table&&delete n[r]})}function Cl(e,t){return e._cfg.version-t._cfg.version}function Pl(e,t,n,r){const i=e._dbSchema,o=e._createTransaction("readwrite",e._storeNames,i);o.create(n),o._completion.catch(r);const s=o._reject.bind(o),a=O.transless||O;Le(()=>{O.trans=o,O.transless=a,t===0?(X(i).forEach(l=>{Un(n,l,i[l].primKey,i[l].indexes)}),wr(e,n),E.follow(()=>e.on.populate.fire(o)).catch(s)):function({_novip:l},c,u,f){const h=[],d=l._versions;let v=l._dbSchema=xr(l,l.idbdb,f),m=!1;const g=d.filter(p=>p._cfg.version>=c);function y(){return h.length?E.resolve(h.shift()(u.idbtrans)).then(y):E.resolve()}return g.forEach(p=>{h.push(()=>{const b=v,w=p._cfg.dbschema;Sr(l,b,f),Sr(l,w,f),v=l._dbSchema=w;const k=Uo(b,w);k.add.forEach(x=>{Un(f,x[0],x[1].primKey,x[1].indexes)}),k.change.forEach(x=>{if(x.recreate)throw new j.Upgrade("Not yet support for changing primary key");{const T=f.objectStore(x.name);x.add.forEach(_=>kr(T,_)),x.change.forEach(_=>{T.deleteIndex(_.name),kr(T,_)}),x.del.forEach(_=>T.deleteIndex(_))}});const S=p._cfg.contentUpgrade;if(S&&p._cfg.version>c){wr(l,f),u._memoizedTables={},m=!0;let x=wo(w);k.del.forEach($=>{x[$]=b[$]}),_r(l,[l.Transaction.prototype]),bn(l,[l.Transaction.prototype],X(x),x),u.schema=x;const T=Fr(S);let _;T&&mt();const R=E.follow(()=>{if(_=S(u),_&&T){var $=Ce.bind(null,null);_.then($,$)}});return _&&typeof _.then=="function"?E.resolve(_):R.then(()=>_)}}),h.push(b=>{(!m||!yl)&&function(w,k){[].slice.call(k.db.objectStoreNames).forEach(S=>w[S]==null&&k.db.deleteObjectStore(S))}(p._cfg.dbschema,b),_r(l,[l.Transaction.prototype]),bn(l,[l.Transaction.prototype],l._storeNames,l._dbSchema),u.schema=l._dbSchema})}),y().then(()=>{var p,b;b=f,X(p=v).forEach(w=>{b.db.objectStoreNames.contains(w)||Un(b,w,p[w].primKey,p[w].indexes)})})}(e,t,o,n).catch(s)})}function Uo(e,t){const n={del:[],add:[],change:[]};let r;for(r in e)t[r]||n.del.push(r);for(r in t){const i=e[r],o=t[r];if(i){const s={name:r,def:o,recreate:!1,del:[],add:[],change:[]};if(""+(i.primKey.keyPath||"")!=""+(o.primKey.keyPath||"")||i.primKey.auto!==o.primKey.auto&&!$n)s.recreate=!0,n.change.push(s);else{const a=i.idxByName,l=o.idxByName;let c;for(c in a)l[c]||s.del.push(c);for(c in l){const u=a[c],f=l[c];u?u.src!==f.src&&s.change.push(f):s.add.push(f)}(s.del.length>0||s.add.length>0||s.change.length>0)&&n.change.push(s)}}else n.add.push([r,o])}return n}function Un(e,t,n,r){const i=e.db.createObjectStore(t,n.keyPath?{keyPath:n.keyPath,autoIncrement:n.auto}:{autoIncrement:n.auto});return r.forEach(o=>kr(i,o)),i}function kr(e,t){e.createIndex(t.name,t.keyPath,{unique:t.unique,multiEntry:t.multi})}function xr(e,t,n){const r={};return dn(t.objectStoreNames,0).forEach(i=>{const o=n.objectStore(i);let s=o.keyPath;const a=vr(Fo(s),s||"",!1,!1,!!o.autoIncrement,s&&typeof s!="string",!0),l=[];for(let u=0;u<o.indexNames.length;++u){const f=o.index(o.indexNames[u]);s=f.keyPath;var c=vr(f.name,s,!!f.unique,!!f.multiEntry,!1,s&&typeof s!="string",!1);l.push(c)}r[i]=qo(i,a,l)}),r}function Sr({_novip:e},t,n){const r=n.db.objectStoreNames;for(let i=0;i<r.length;++i){const o=r[i],s=n.objectStore(o);e._hasGetAll="getAll"in s;for(let a=0;a<s.indexNames.length;++a){const l=s.indexNames[a],c=s.index(l).keyPath,u=typeof c=="string"?c:"["+dn(c).join("+")+"]";if(t[o]){const f=t[o].idxByName[u];f&&(f.name=l,delete t[o].idxByName[u],t[o].idxByName[l]=f)}}}typeof navigator<"u"&&/Safari/.test(navigator.userAgent)&&!/(Chrome\/|Edge\/)/.test(navigator.userAgent)&&W.WorkerGlobalScope&&W instanceof W.WorkerGlobalScope&&[].concat(navigator.userAgent.match(/Safari\/(\d*)/))[1]<604&&(e._hasGetAll=!1)}class $l{_parseStoresSpec(t,n){X(t).forEach(r=>{if(t[r]!==null){var i=t[r].split(",").map((s,a)=>{const l=(s=s.trim()).replace(/([&*]|\+\+)/g,""),c=/^\[/.test(l)?l.match(/^\[(.*)\]$/)[1].split("+"):l;return vr(l,c||null,/\&/.test(s),/\*/.test(s),/\+\+/.test(s),ee(c),a===0)}),o=i.shift();if(o.multi)throw new j.Schema("Primary key cannot be multi-valued");i.forEach(s=>{if(s.auto)throw new j.Schema("Only primary key can be marked as autoIncrement (++)");if(!s.keyPath)throw new j.Schema("Index must have a name and cannot be an empty string")}),n[r]=qo(r,o,i)}})}stores(t){const n=this.db;this._cfg.storesSource=this._cfg.storesSource?re(this._cfg.storesSource,t):t;const r=n._versions,i={};let o={};return r.forEach(s=>{re(i,s._cfg.storesSource),o=s._cfg.dbschema={},s._parseStoresSpec(i,o)}),n._dbSchema=o,_r(n,[n._allTables,n,n.Transaction.prototype]),bn(n,[n._allTables,n,n.Transaction.prototype,this._cfg.tables],X(o),o),n._storeNames=X(o),this}upgrade(t){return this._cfg.contentUpgrade=Vr(this._cfg.contentUpgrade||V,t),this}}function Yr(e,t){let n=e._dbNamesDB;return n||(n=e._dbNamesDB=new He(An,{addons:[],indexedDB:e,IDBKeyRange:t}),n.version(1).stores({dbnames:"name"})),n.table("dbnames")}function Gr(e){return e&&typeof e.databases=="function"}function Er(e){return Le(function(){return O.letThrough=!0,e()})}function Al(){var e;return!navigator.userAgentData&&/Safari\//.test(navigator.userAgent)&&!/Chrom(e|ium)\//.test(navigator.userAgent)&&indexedDB.databases?new Promise(function(t){var n=function(){return indexedDB.databases().finally(t)};e=setInterval(n,100),n()}).finally(function(){return clearInterval(e)}):Promise.resolve()}function Ol(e){const t=e._state,{indexedDB:n}=e._deps;if(t.isBeingOpened||e.idbdb)return t.dbReadyPromise.then(()=>t.dbOpenError?Q(t.dbOpenError):e);me&&(t.openCanceller._stackHolder=Ze()),t.isBeingOpened=!0,t.dbOpenError=null,t.openComplete=!1;const r=t.openCanceller;function i(){if(t.openCanceller!==r)throw new j.DatabaseClosed("db.open() was cancelled")}let o=t.dbReadyResolve,s=null,a=!1;return E.race([r,(typeof navigator>"u"?E.resolve():Al()).then(()=>new E((l,c)=>{if(i(),!n)throw new j.MissingAPI;const u=e.name,f=t.autoSchema?n.open(u):n.open(u,Math.round(10*e.verno));if(!f)throw new j.MissingAPI;f.onerror=de(c),f.onblocked=z(e._fireOnBlocked),f.onupgradeneeded=z(h=>{if(s=f.transaction,t.autoSchema&&!e._options.allowEmptyDB){f.onerror=It,s.abort(),f.result.close();const v=n.deleteDatabase(u);v.onsuccess=v.onerror=z(()=>{c(new j.NoSuchDatabase(`Database ${u} doesnt exist`))})}else{s.onerror=de(c);var d=h.oldVersion>Math.pow(2,62)?0:h.oldVersion;a=d<1,e._novip.idbdb=f.result,Pl(e,d/10,s,c)}},c),f.onsuccess=z(()=>{s=null;const h=e._novip.idbdb=f.result,d=dn(h.objectStoreNames);if(d.length>0)try{const m=h.transaction((v=d).length===1?v[0]:v,"readonly");t.autoSchema?function({_novip:g},y,p){g.verno=y.version/10;const b=g._dbSchema=xr(0,y,p);g._storeNames=dn(y.objectStoreNames,0),bn(g,[g._allTables],X(b),b)}(e,h,m):(Sr(e,e._dbSchema,m),function(g,y){const p=Uo(xr(0,g.idbdb,y),g._dbSchema);return!(p.add.length||p.change.some(b=>b.add.length||b.change.length))}(e,m)||console.warn("Dexie SchemaDiff: Schema was extended without increasing the number passed to db.version(). Some queries may fail.")),wr(e,m)}catch{}var v;Ct.push(e),h.onversionchange=z(m=>{t.vcFired=!0,e.on("versionchange").fire(m)}),h.onclose=z(m=>{e.on("close").fire(m)}),a&&function({indexedDB:m,IDBKeyRange:g},y){!Gr(m)&&y!==An&&Yr(m,g).put({name:y}).catch(V)}(e._deps,u),l()},c)}))]).then(()=>(i(),t.onReadyBeingFired=[],E.resolve(Er(()=>e.on.ready.fire(e.vip))).then(function l(){if(t.onReadyBeingFired.length>0){let c=t.onReadyBeingFired.reduce(Vr,V);return t.onReadyBeingFired=[],E.resolve(Er(()=>c(e.vip))).then(l)}}))).finally(()=>{t.onReadyBeingFired=null,t.isBeingOpened=!1}).then(()=>e).catch(l=>{t.dbOpenError=l;try{s&&s.abort()}catch{}return r===t.openCanceller&&e._close(),Q(l)}).finally(()=>{t.openComplete=!0,o()})}function Cr(e){var t=o=>e.next(o),n=i(t),r=i(o=>e.throw(o));function i(o){return s=>{var a=o(s),l=a.value;return a.done?l:l&&typeof l.then=="function"?l.then(n,r):ee(l)?Promise.all(l).then(n,r):n(l)}}return i(t)()}function Rl(e,t,n){var r=arguments.length;if(r<2)throw new j.InvalidArgument("Too few arguments");for(var i=new Array(r-1);--r;)i[r-1]=arguments[r];return n=i.pop(),[e,_o(i),n]}function Vo(e,t,n,r,i){return E.resolve().then(()=>{const o=O.transless||O,s=e._createTransaction(t,n,e._dbSchema,r),a={trans:s,transless:o};if(r)s.idbtrans=r.idbtrans;else try{s.create(),e._state.PR1398_maxLoop=3}catch(f){return f.name===Ur.InvalidState&&e.isOpen()&&--e._state.PR1398_maxLoop>0?(console.warn("Dexie: Need to reopen db"),e._close(),e.open().then(()=>Vo(e,t,n,null,i))):Q(f)}const l=Fr(i);let c;l&&mt();const u=E.follow(()=>{if(c=i.call(s,s),c)if(l){var f=Ce.bind(null,null);c.then(f,f)}else typeof c.next=="function"&&typeof c.throw=="function"&&(c=Cr(c))},a);return(c&&typeof c.then=="function"?E.resolve(c).then(f=>s.active?f:Q(new j.PrematureCommit("Transaction committed too early. See http://bit.ly/2kdckMn"))):u.then(()=>c)).then(f=>(r&&s._resolve(),s._completion.then(()=>f))).catch(f=>(s._reject(f),Q(f)))})}function Yt(e,t,n){const r=ee(e)?e.slice():[e];for(let i=0;i<n;++i)r.push(t);return r}const Tl={stack:"dbcore",name:"VirtualIndexMiddleware",level:1,create:function(e){return{...e,table(t){const n=e.table(t),{schema:r}=n,i={},o=[];function s(u,f,h){const d=$t(u),v=i[d]=i[d]||[],m=u==null?0:typeof u=="string"?1:u.length,g=f>0,y={...h,isVirtual:g,keyTail:f,keyLength:m,extractKey:br(u),unique:!g&&h.unique};return v.push(y),y.isPrimaryKey||o.push(y),m>1&&s(m===2?u[0]:u.slice(0,m-1),f+1,h),v.sort((p,b)=>p.keyTail-b.keyTail),y}const a=s(r.primaryKey.keyPath,0,r.primaryKey);i[":id"]=[a];for(const u of r.indexes)s(u.keyPath,0,u);function l(u){const f=u.query.index;return f.isVirtual?{...u,query:{index:f,range:(h=u.query.range,d=f.keyTail,{type:h.type===1?2:h.type,lower:Yt(h.lower,h.lowerOpen?e.MAX_KEY:e.MIN_KEY,d),lowerOpen:!0,upper:Yt(h.upper,h.upperOpen?e.MIN_KEY:e.MAX_KEY,d),upperOpen:!0})}}:u;var h,d}return{...n,schema:{...r,primaryKey:a,indexes:o,getIndexByKeyPath:function(u){const f=i[$t(u)];return f&&f[0]}},count:u=>n.count(l(u)),query:u=>n.query(l(u)),openCursor(u){const{keyTail:f,isVirtual:h,keyLength:d}=u.query.index;return h?n.openCursor(l(u)).then(v=>v&&function(m){return Object.create(m,{continue:{value:function(y){y!=null?m.continue(Yt(y,u.reverse?e.MAX_KEY:e.MIN_KEY,f)):u.unique?m.continue(m.key.slice(0,d).concat(u.reverse?e.MIN_KEY:e.MAX_KEY,f)):m.continue()}},continuePrimaryKey:{value(y,p){m.continuePrimaryKey(Yt(y,e.MAX_KEY,f),p)}},primaryKey:{get:()=>m.primaryKey},key:{get(){const y=m.key;return d===1?y[0]:y.slice(0,d)}},value:{get:()=>m.value}})}(v)):n.openCursor(u)}}}}}};function Xr(e,t,n,r){return n=n||{},r=r||"",X(e).forEach(i=>{if(le(t,i)){var o=e[i],s=t[i];if(typeof o=="object"&&typeof s=="object"&&o&&s){const a=or(o);a!==or(s)?n[r+i]=t[i]:a==="Object"?Xr(o,s,n,r+i+"."):o!==s&&(n[r+i]=t[i])}else o!==s&&(n[r+i]=t[i])}else n[r+i]=void 0}),X(t).forEach(i=>{le(e,i)||(n[r+i]=t[i])}),n}const Il={stack:"dbcore",name:"HooksMiddleware",level:2,create:e=>({...e,table(t){const n=e.table(t),{primaryKey:r}=n.schema;return{...n,mutate(o){const s=O.trans,{deleting:a,creating:l,updating:c}=s.table(t).hook;switch(o.type){case"add":if(l.fire===V)break;return s._promise("readwrite",()=>u(o),!0);case"put":if(l.fire===V&&c.fire===V)break;return s._promise("readwrite",()=>u(o),!0);case"delete":if(a.fire===V)break;return s._promise("readwrite",()=>u(o),!0);case"deleteRange":if(a.fire===V)break;return s._promise("readwrite",()=>function(h){return f(h.trans,h.range,1e4)}(o),!0)}return n.mutate(o);function u(h){const d=O.trans,v=h.keys||function(m,g){return g.type==="delete"?g.keys:g.keys||g.values.map(m.extractKey)}(r,h);if(!v)throw new Error("Keys missing");return(h=h.type==="add"||h.type==="put"?{...h,keys:v}:{...h}).type!=="delete"&&(h.values=[...h.values]),h.keys&&(h.keys=[...h.keys]),function(m,g,y){return g.type==="add"?Promise.resolve([]):m.getMany({trans:g.trans,keys:y,cache:"immutable"})}(n,h,v).then(m=>{const g=v.map((y,p)=>{const b=m[p],w={onerror:null,onsuccess:null};if(h.type==="delete")a.fire.call(w,y,b,d);else if(h.type==="add"||b===void 0){const k=l.fire.call(w,y,h.values[p],d);y==null&&k!=null&&(y=k,h.keys[p]=y,r.outbound||ue(h.values[p],r.keyPath,y))}else{const k=Xr(b,h.values[p]),S=c.fire.call(w,k,y,b,d);if(S){const x=h.values[p];Object.keys(S).forEach(T=>{le(x,T)?x[T]=S[T]:ue(x,T,S[T])})}}return w});return n.mutate(h).then(({failures:y,results:p,numFailures:b,lastResult:w})=>{for(let k=0;k<v.length;++k){const S=p?p[k]:v[k],x=g[k];S==null?x.onerror&&x.onerror(y[k]):x.onsuccess&&x.onsuccess(h.type==="put"&&m[k]?h.values[k]:S)}return{failures:y,results:p,numFailures:b,lastResult:w}}).catch(y=>(g.forEach(p=>p.onerror&&p.onerror(y)),Promise.reject(y)))})}function f(h,d,v){return n.query({trans:h,values:!1,query:{index:r,range:d},limit:v}).then(({result:m})=>u({type:"delete",keys:m,trans:h}).then(g=>g.numFailures>0?Promise.reject(g.failures[0]):m.length<v?{failures:[],numFailures:0,lastResult:void 0}:f(h,{...d,lower:m[m.length-1],lowerOpen:!0},v)))}}}}})};function Wo(e,t,n){try{if(!t||t.keys.length<e.length)return null;const r=[];for(let i=0,o=0;i<t.keys.length&&o<e.length;++i)ne(t.keys[i],e[o])===0&&(r.push(n?Dt(t.values[i]):t.values[i]),++o);return r.length===e.length?r:null}catch{return null}}const jl={stack:"dbcore",level:-1,create:e=>({table:t=>{const n=e.table(t);return{...n,getMany:r=>{if(!r.cache)return n.getMany(r);const i=Wo(r.keys,r.trans._cache,r.cache==="clone");return i?E.resolve(i):n.getMany(r).then(o=>(r.trans._cache={keys:r.keys,values:r.cache==="clone"?Dt(o):o},o))},mutate:r=>(r.type!=="add"&&(r.trans._cache=null),n.mutate(r))}}})};function Qr(e){return!("from"in e)}const be=function(e,t){if(!this){const n=new be;return e&&"d"in e&&re(n,e),n}re(this,arguments.length?{d:1,from:e,to:arguments.length>1?t:e}:{d:0})};function Lt(e,t,n){const r=ne(t,n);if(isNaN(r))return;if(r>0)throw RangeError();if(Qr(e))return re(e,{from:t,to:n,d:1});const i=e.l,o=e.r;if(ne(n,e.from)<0)return i?Lt(i,t,n):e.l={from:t,to:n,d:1,l:null,r:null},ki(e);if(ne(t,e.to)>0)return o?Lt(o,t,n):e.r={from:t,to:n,d:1,l:null,r:null},ki(e);ne(t,e.from)<0&&(e.from=t,e.l=null,e.d=o?o.d+1:1),ne(n,e.to)>0&&(e.to=n,e.r=null,e.d=e.l?e.l.d+1:1);const s=!e.r;i&&!e.l&&wn(e,i),o&&s&&wn(e,o)}function wn(e,t){Qr(t)||function n(r,{from:i,to:o,l:s,r:a}){Lt(r,i,o),s&&n(r,s),a&&n(r,a)}(e,t)}function Bl(e,t){const n=Pr(t);let r=n.next();if(r.done)return!1;let i=r.value;const o=Pr(e);let s=o.next(i.from),a=s.value;for(;!r.done&&!s.done;){if(ne(a.from,i.to)<=0&&ne(a.to,i.from)>=0)return!0;ne(i.from,a.from)<0?i=(r=n.next(a.from)).value:a=(s=o.next(i.from)).value}return!1}function Pr(e){let t=Qr(e)?null:{s:0,n:e};return{next(n){const r=arguments.length>0;for(;t;)switch(t.s){case 0:if(t.s=1,r)for(;t.n.l&&ne(n,t.n.from)<0;)t={up:t,n:t.n.l,s:1};else for(;t.n.l;)t={up:t,n:t.n.l,s:1};case 1:if(t.s=2,!r||ne(n,t.n.to)<=0)return{value:t.n,done:!1};case 2:if(t.n.r){t.s=3,t={up:t,n:t.n.r,s:0};continue}case 3:t=t.up}return{done:!0}}}}function ki(e){var t,n;const r=(((t=e.r)===null||t===void 0?void 0:t.d)||0)-(((n=e.l)===null||n===void 0?void 0:n.d)||0),i=r>1?"r":r<-1?"l":"";if(i){const o=i==="r"?"l":"r",s={...e},a=e[i];e.from=a.from,e.to=a.to,e[i]=a[i],s[i]=a[o],e[o]=s,s.d=xi(s)}e.d=xi(e)}function xi({r:e,l:t}){return(e?t?Math.max(e.d,t.d):e.d:t?t.d:0)+1}ht(be.prototype,{add(e){return wn(this,e),this},addKey(e){return Lt(this,e,e),this},addKeys(e){return e.forEach(t=>Lt(this,t,t)),this},[sr](){return Pr(this)}});const Ll={stack:"dbcore",level:0,create:e=>{const t=e.schema.name,n=new be(e.MIN_KEY,e.MAX_KEY);return{...e,table:r=>{const i=e.table(r),{schema:o}=i,{primaryKey:s}=o,{extractKey:a,outbound:l}=s,c={...i,mutate:h=>{const d=h.trans,v=d.mutatedParts||(d.mutatedParts={}),m=S=>{const x=`idb://${t}/${r}/${S}`;return v[x]||(v[x]=new be)},g=m(""),y=m(":dels"),{type:p}=h;let[b,w]=h.type==="deleteRange"?[h.range]:h.type==="delete"?[h.keys]:h.values.length<50?[[],h.values]:[];const k=h.trans._cache;return i.mutate(h).then(S=>{if(ee(b)){p!=="delete"&&(b=S.results),g.addKeys(b);const x=Wo(b,k);x||p==="add"||y.addKeys(b),(x||w)&&function(T,_,R,$){function A(I){const D=T(I.name||"");function fe(H){return H!=null?I.extractKey(H):null}const he=H=>I.multiEntry&&ee(H)?H.forEach(oe=>D.addKey(oe)):D.addKey(H);(R||$).forEach((H,oe)=>{const ie=R&&fe(R[oe]),et=$&&fe($[oe]);ne(ie,et)!==0&&(ie!=null&&he(ie),et!=null&&he(et))})}_.indexes.forEach(A)}(m,o,x,w)}else if(b){const x={from:b.lower,to:b.upper};y.add(x),g.add(x)}else g.add(n),y.add(n),o.indexes.forEach(x=>m(x.name).add(n));return S})}},u=({query:{index:h,range:d}})=>{var v,m;return[h,new be((v=d.lower)!==null&&v!==void 0?v:e.MIN_KEY,(m=d.upper)!==null&&m!==void 0?m:e.MAX_KEY)]},f={get:h=>[s,new be(h.key)],getMany:h=>[s,new be().addKeys(h.keys)],count:u,query:u,openCursor:u};return X(f).forEach(h=>{c[h]=function(d){const{subscr:v}=O;if(v){const m=w=>{const k=`idb://${t}/${r}/${w}`;return v[k]||(v[k]=new be)},g=m(""),y=m(":dels"),[p,b]=f[h](d);if(m(p.name||"").add(b),!p.isPrimaryKey){if(h!=="count"){const w=h==="query"&&l&&d.values&&i.query({...d,values:!1});return i[h].apply(this,arguments).then(k=>{if(h==="query"){if(l&&d.values)return w.then(({result:x})=>(g.addKeys(x),k));const S=d.values?k.result.map(a):k.result;d.values?g.addKeys(S):y.addKeys(S)}else if(h==="openCursor"){const S=k,x=d.values;return S&&Object.create(S,{key:{get:()=>(y.addKey(S.primaryKey),S.key)},primaryKey:{get(){const T=S.primaryKey;return y.addKey(T),T}},value:{get:()=>(x&&g.addKey(S.primaryKey),S.value)}})}return k})}y.add(n)}}return i[h].apply(this,arguments)}}),c}}}};class He{constructor(t,n){this._middlewares={},this.verno=0;const r=He.dependencies;this._options=n={addons:He.addons,autoOpen:!0,indexedDB:r.indexedDB,IDBKeyRange:r.IDBKeyRange,...n},this._deps={indexedDB:n.indexedDB,IDBKeyRange:n.IDBKeyRange};const{addons:i}=n;this._dbSchema={},this._versions=[],this._storeNames=[],this._allTables={},this.idbdb=null,this._novip=this;const o={dbOpenError:null,isBeingOpened:!1,onReadyBeingFired:null,openComplete:!1,dbReadyResolve:V,dbReadyPromise:null,cancelOpen:V,openCanceller:null,autoSchema:!0,PR1398_maxLoop:3};var s;o.dbReadyPromise=new E(a=>{o.dbReadyResolve=a}),o.openCanceller=new E((a,l)=>{o.cancelOpen=l}),this._state=o,this.name=t,this.on=Pt(this,"populate","blocked","versionchange","close",{ready:[Vr,V]}),this.on.ready.subscribe=go(this.on.ready.subscribe,a=>(l,c)=>{He.vip(()=>{const u=this._state;if(u.openComplete)u.dbOpenError||E.resolve().then(l),c&&a(l);else if(u.onReadyBeingFired)u.onReadyBeingFired.push(l),c&&a(l);else{a(l);const f=this;c||a(function h(){f.on.ready.unsubscribe(l),f.on.ready.unsubscribe(h)})}})}),this.Collection=(s=this,vt(bl.prototype,function(a,l){this.db=s;let c=Ko,u=null;if(l)try{c=l()}catch(v){u=v}const f=a._ctx,h=f.table,d=h.hook.reading.fire;this._ctx={table:h,index:f.index,isPrimKey:!f.index||h.schema.primKey.keyPath&&f.index===h.schema.primKey.name,range:c,keysOnly:!1,dir:"next",unique:"",algorithm:null,filter:null,replayFilter:null,justLimit:!0,isMatch:null,offset:0,limit:1/0,error:u,or:f.or,valueMapper:d!==Rt?d:null}})),this.Table=function(a){return vt(vl.prototype,function(l,c,u){this.db=a,this._tx=u,this.name=l,this.schema=c,this.hook=a._allTables[l]?a._allTables[l].hook:Pt(null,{creating:[sl,V],reading:[ol,Rt],updating:[ll,V],deleting:[al,V]})})}(this),this.Transaction=function(a){return vt(xl.prototype,function(l,c,u,f,h){this.db=a,this.mode=l,this.storeNames=c,this.schema=u,this.chromeTransactionDurability=f,this.idbtrans=null,this.on=Pt(this,"complete","error","abort"),this.parent=h||null,this.active=!0,this._reculock=0,this._blockedFuncs=[],this._resolve=null,this._reject=null,this._waitingFor=null,this._waitingQueue=null,this._spinCount=0,this._completion=new E((d,v)=>{this._resolve=d,this._reject=v}),this._completion.then(()=>{this.active=!1,this.on.complete.fire()},d=>{var v=this.active;return this.active=!1,this.on.error.fire(d),this.parent?this.parent._reject(d):v&&this.idbtrans&&this.idbtrans.abort(),Q(d)})})}(this),this.Version=function(a){return vt($l.prototype,function(l){this.db=a,this._cfg={version:l,storesSource:null,dbschema:{},tables:{},contentUpgrade:null}})}(this),this.WhereClause=function(a){return vt(Mo.prototype,function(l,c,u){this.db=a,this._ctx={table:l,index:c===":id"?null:c,or:u};const f=a._deps.indexedDB;if(!f)throw new j.MissingAPI;this._cmp=this._ascending=f.cmp.bind(f),this._descending=(h,d)=>f.cmp(d,h),this._max=(h,d)=>f.cmp(h,d)>0?h:d,this._min=(h,d)=>f.cmp(h,d)<0?h:d,this._IDBKeyRange=a._deps.IDBKeyRange})}(this),this.on("versionchange",a=>{a.newVersion>0?console.warn(`Another connection wants to upgrade database '${this.name}'. Closing db now to resume the upgrade.`):console.warn(`Another connection wants to delete database '${this.name}'. Closing db now to resume the delete request.`),this.close()}),this.on("blocked",a=>{!a.newVersion||a.newVersion<a.oldVersion?console.warn(`Dexie.delete('${this.name}') was blocked`):console.warn(`Upgrade '${this.name}' blocked by other connection holding version ${a.oldVersion/10}`)}),this._maxKey=Bt(n.IDBKeyRange),this._createTransaction=(a,l,c,u)=>new this.Transaction(a,l,c,this._options.chromeTransactionDurability,u),this._fireOnBlocked=a=>{this.on("blocked").fire(a),Ct.filter(l=>l.name===this.name&&l!==this&&!l._state.vcFired).map(l=>l.on("versionchange").fire(a))},this.use(Tl),this.use(Il),this.use(Ll),this.use(jl),this.vip=Object.create(this,{_vip:{value:!0}}),i.forEach(a=>a(this))}version(t){if(isNaN(t)||t<.1)throw new j.Type("Given version is not a positive number");if(t=Math.round(10*t)/10,this.idbdb||this._state.isBeingOpened)throw new j.Schema("Cannot add version when database is open");this.verno=Math.max(this.verno,t);const n=this._versions;var r=n.filter(i=>i._cfg.version===t)[0];return r||(r=new this.Version(t),n.push(r),n.sort(Cl),r.stores({}),this._state.autoSchema=!1,r)}_whenReady(t){return this.idbdb&&(this._state.openComplete||O.letThrough||this._vip)?t():new E((n,r)=>{if(this._state.openComplete)return r(new j.DatabaseClosed(this._state.dbOpenError));if(!this._state.isBeingOpened){if(!this._options.autoOpen)return void r(new j.DatabaseClosed);this.open().catch(V)}this._state.dbReadyPromise.then(n,r)}).then(t)}use({stack:t,create:n,level:r,name:i}){i&&this.unuse({stack:t,name:i});const o=this._middlewares[t]||(this._middlewares[t]=[]);return o.push({stack:t,create:n,level:r??10,name:i}),o.sort((s,a)=>s.level-a.level),this}unuse({stack:t,name:n,create:r}){return t&&this._middlewares[t]&&(this._middlewares[t]=this._middlewares[t].filter(i=>r?i.create!==r:!!n&&i.name!==n)),this}open(){return Ol(this)}_close(){const t=this._state,n=Ct.indexOf(this);if(n>=0&&Ct.splice(n,1),this.idbdb){try{this.idbdb.close()}catch{}this._novip.idbdb=null}t.dbReadyPromise=new E(r=>{t.dbReadyResolve=r}),t.openCanceller=new E((r,i)=>{t.cancelOpen=i})}close(){this._close();const t=this._state;this._options.autoOpen=!1,t.dbOpenError=new j.DatabaseClosed,t.isBeingOpened&&t.cancelOpen(t.dbOpenError)}delete(){const t=arguments.length>0,n=this._state;return new E((r,i)=>{const o=()=>{this.close();var s=this._deps.indexedDB.deleteDatabase(this.name);s.onsuccess=z(()=>{(function({indexedDB:a,IDBKeyRange:l},c){!Gr(a)&&c!==An&&Yr(a,l).delete(c).catch(V)})(this._deps,this.name),r()}),s.onerror=de(i),s.onblocked=this._fireOnBlocked};if(t)throw new j.InvalidArgument("Arguments not allowed in db.delete()");n.isBeingOpened?n.dbReadyPromise.then(o):o()})}backendDB(){return this.idbdb}isOpen(){return this.idbdb!==null}hasBeenClosed(){const t=this._state.dbOpenError;return t&&t.name==="DatabaseClosed"}hasFailed(){return this._state.dbOpenError!==null}dynamicallyOpened(){return this._state.autoSchema}get tables(){return X(this._allTables).map(t=>this._allTables[t])}transaction(){const t=Rl.apply(this,arguments);return this._transaction.apply(this,t)}_transaction(t,n,r){let i=O.trans;i&&i.db===this&&t.indexOf("!")===-1||(i=null);const o=t.indexOf("?")!==-1;let s,a;t=t.replace("!","").replace("?","");try{if(a=n.map(c=>{var u=c instanceof this.Table?c.name:c;if(typeof u!="string")throw new TypeError("Invalid table argument to Dexie.transaction(). Only Table or String are allowed");return u}),t=="r"||t===Kn)s=Kn;else{if(t!="rw"&&t!=Nn)throw new j.InvalidArgument("Invalid transaction mode: "+t);s=Nn}if(i){if(i.mode===Kn&&s===Nn){if(!o)throw new j.SubTransaction("Cannot enter a sub-transaction with READWRITE mode when parent transaction is READONLY");i=null}i&&a.forEach(c=>{if(i&&i.storeNames.indexOf(c)===-1){if(!o)throw new j.SubTransaction("Table "+c+" not included in parent transaction.");i=null}}),o&&i&&!i.active&&(i=null)}}catch(c){return i?i._promise(null,(u,f)=>{f(c)}):Q(c)}const l=Vo.bind(null,this,s,a,i,r);return i?i._promise(s,l,"lock"):O.trans?yt(O.transless,()=>this._whenReady(l)):this._whenReady(l)}table(t){if(!le(this._allTables,t))throw new j.InvalidTable(`Table ${t} does not exist`);return this._allTables[t]}}const Dl=typeof Symbol<"u"&&"observable"in Symbol?Symbol.observable:"@@observable";class Kl{constructor(t){this._subscribe=t}subscribe(t,n,r){return this._subscribe(t&&typeof t!="function"?t:{next:t,error:n,complete:r})}[Dl](){return this}}function zo(e,t){return X(t).forEach(n=>{wn(e[n]||(e[n]=new be),t[n])}),e}function Nl(e){let t,n=!1;const r=new Kl(i=>{const o=Fr(e);let s=!1,a={},l={};const c={get closed(){return s},unsubscribe:()=>{s=!0,Ke.storagemutated.unsubscribe(d)}};i.start&&i.start(c);let u=!1,f=!1;function h(){return X(l).some(m=>a[m]&&Bl(a[m],l[m]))}const d=m=>{zo(a,m),h()&&v()},v=()=>{if(u||s)return;a={};const m={},g=function(y){o&&mt();const p=()=>Le(e,{subscr:y,trans:null}),b=O.trans?yt(O.transless,p):p();return o&&b.then(Ce,Ce),b}(m);f||(Ke(jt,d),f=!0),u=!0,Promise.resolve(g).then(y=>{n=!0,t=y,u=!1,s||(h()?v():(a={},l=m,i.next&&i.next(y)))},y=>{u=!1,n=!1,i.error&&i.error(y),c.unsubscribe()})};return v(),c});return r.hasValue=()=>n,r.getValue=()=>t,r}let $r;try{$r={indexedDB:W.indexedDB||W.mozIndexedDB||W.webkitIndexedDB||W.msIndexedDB,IDBKeyRange:W.IDBKeyRange||W.webkitIDBKeyRange}}catch{$r={indexedDB:null,IDBKeyRange:null}}const Me=He;function nn(e){let t=_e;try{_e=!0,Ke.storagemutated.fire(e)}finally{_e=t}}ht(Me,{...Xt,delete:e=>new Me(e,{addons:[]}).delete(),exists:e=>new Me(e,{addons:[]}).open().then(t=>(t.close(),!0)).catch("NoSuchDatabaseError",()=>!1),getDatabaseNames(e){try{return function({indexedDB:t,IDBKeyRange:n}){return Gr(t)?Promise.resolve(t.databases()).then(r=>r.map(i=>i.name).filter(i=>i!==An)):Yr(t,n).toCollection().primaryKeys()}(Me.dependencies).then(e)}catch{return Q(new j.MissingAPI)}},defineClass:()=>function(e){re(this,e)},ignoreTransaction:e=>O.trans?yt(O.transless,e):e(),vip:Er,async:function(e){return function(){try{var t=Cr(e.apply(this,arguments));return t&&typeof t.then=="function"?t:E.resolve(t)}catch(n){return Q(n)}}},spawn:function(e,t,n){try{var r=Cr(e.apply(n,t||[]));return r&&typeof r.then=="function"?r:E.resolve(r)}catch(i){return Q(i)}},currentTransaction:{get:()=>O.trans||null},waitFor:function(e,t){const n=E.resolve(typeof e=="function"?Me.ignoreTransaction(e):e).timeout(t||6e4);return O.trans?O.trans.waitFor(n):n},Promise:E,debug:{get:()=>me,set:e=>{xo(e,e==="dexie"?()=>!0:Do)}},derive:at,extend:re,props:ht,override:go,Events:Pt,on:Ke,liveQuery:Nl,extendObservabilitySet:zo,getByKeyPath:Se,setByKeyPath:ue,delByKeyPath:function(e,t){typeof t=="string"?ue(e,t,void 0):"length"in t&&[].map.call(t,function(n){ue(e,n,void 0)})},shallowClone:wo,deepClone:Dt,getObjectDiff:Xr,cmp:ne,asap:vo,minKey:gr,addons:[],connections:Ct,errnames:Ur,dependencies:$r,semVer:yi,version:yi.split(".").map(e=>parseInt(e)).reduce((e,t,n)=>e+t/Math.pow(10,2*n))}),Me.maxKey=Bt(Me.dependencies.IDBKeyRange),typeof dispatchEvent<"u"&&typeof addEventListener<"u"&&(Ke(jt,e=>{if(!_e){let t;$n?(t=document.createEvent("CustomEvent"),t.initCustomEvent(Oe,!0,!0,e)):t=new CustomEvent(Oe,{detail:e}),_e=!0,dispatchEvent(t),_e=!1}}),addEventListener(Oe,({detail:e})=>{_e||nn(e)}));let _e=!1;if(typeof BroadcastChannel<"u"){const e=new BroadcastChannel(Oe);typeof e.unref=="function"&&e.unref(),Ke(jt,t=>{_e||e.postMessage(t)}),e.onmessage=t=>{t.data&&nn(t.data)}}else if(typeof self<"u"&&typeof navigator<"u"){Ke(jt,t=>{try{_e||(typeof localStorage<"u"&&localStorage.setItem(Oe,JSON.stringify({trig:Math.random(),changedParts:t})),typeof self.clients=="object"&&[...self.clients.matchAll({includeUncontrolled:!0})].forEach(n=>n.postMessage({type:Oe,changedParts:t})))}catch{}}),typeof addEventListener<"u"&&addEventListener("storage",t=>{if(t.key===Oe){const n=JSON.parse(t.newValue);n&&nn(n.changedParts)}});const e=self.document&&navigator.serviceWorker;e&&e.addEventListener("message",function({data:t}){t&&t.type===Oe&&nn(t.changedParts)})}E.rejectionMapper=function(e,t){if(!e||e instanceof lt||e instanceof TypeError||e instanceof SyntaxError||!e.name||!hi[e.name])return e;var n=new hi[e.name](t||e.message,e);return"stack"in e&&xe(n,"stack",{get:function(){return this.inner.stack}}),n},xo(me,Do);class Ml extends He{spells;classes;races;backgrounds;items;feats;constructor(t){super(t),this.version(1).stores({spells:"name",classes:"name",races:"name",backgrounds:"name",items:"name",feats:"name"})}}const _n=new Ml("dndSrd"),[Si,Fl]=L([]);function ql(){const e=Be.toObservable(_n.classes.toArray());return Si().length===0&&e.pipe(Ie(1),hn(t=>t.length>0?ce(t):ce([])),hn(t=>t.length===0?Be.post("/api/DnDInfo/Classes",{}).pipe(Ie(1),Je(n=>(console.error("Error: ",n),ce(null))),je(n=>{n&&_n.classes.bulkAdd(n)})):ce(t)),je(t=>t&&t.length>0?Fl(t):null)).subscribe(),Si}const Ul="_primary_1a8rj_1",Vl="_accent_1a8rj_12",Wl="_tertiary_1a8rj_65",zl="_warn_1a8rj_70",Hl="_hover_1a8rj_74",Yl="_popup_1a8rj_88",Gl="_table_1a8rj_100",Fe={primary:Ul,accent:Vl,tertiary:Wl,warn:zl,hover:Hl,popup:Yl,table:Gl},Xl="_primary_hhp2e_1",Ql="_accent_hhp2e_7",Jl="_warn_hhp2e_13",Zl="_hover_hhp2e_17",ec="_popup_hhp2e_28",tc="_table_hhp2e_41",nc="_tBody_hhp2e_55",rc="_THead_hhp2e_60",qe={primary:Xl,accent:Ql,warn:Jl,hover:Zl,popup:ec,table:tc,tBody:nc,THead:rc},ic={primary:qe.primary,accent:qe.accent,tertiary:qe.tertiary,warn:qe.warn,hover:qe.hover,popup:qe.popup,table:qe.table},Ei={primary:Fe.primary,accent:Fe.accent,tertiary:Fe.tertiary,warn:Fe.warn,hover:Fe.hover,popup:Fe.popup,table:Fe.table};function ye(e="dark"){switch(e.toLowerCase()){case"light":return ic;case"dark":return document.body.style.backgroundColor="#212121",Ei;default:return Ei}}const[Ci,oc]=L([]);function Ho(){const e=Be.toObservable(_n.spells.toArray());return Ci().length===0&&e.pipe(Ie(1),hn(t=>t.length>0?ce(t):ce([])),hn(t=>t.length===0?Be.post("/api/DnDInfo/Spells",{}).pipe(Ie(1),Je(n=>(console.error("Error: ",n),ce(null))),je(n=>{n&&_n.spells.bulkAdd(n)})):ce(t)),je(t=>t&&t.length>0?oc(t):null)).subscribe(),Ci}const[sc,ac]=L([]);function lc(){return Be.post("/api/DnDInfo/Feats",{}).pipe(Ie(1),Je(e=>(console.error("Error: ",e),ce(null))),je(e=>e?ac(e):null)).subscribe(),sc}const[cc,uc]=L([]);function fc(){return Be.post("/api/DnDInfo/Races",{}).pipe(Ie(1),Je(e=>(console.error("Error: ",e),ce(null))),je(e=>e?uc(e):null)).subscribe(),cc}const[hc,dc]=L([]);function pc(){return Be.post("/api/DnDInfo/Backgrounds",{}).pipe(Ie(1),Je(e=>(console.error("Error: ",e),ce(null))),je(e=>e?dc(e):null)).subscribe(),hc}const[mc,yc]=L([]);function gc(){return Be.post("/api/DnDInfo/Items",{}).pipe(Ie(1),Je(e=>(console.error("Error: ",e),ce(null))),je(e=>e?yc(e):null)).subscribe(),mc}const vc="_AppBody_1b0bi_1",bc={AppBody:vc};function wc(e={}){const{immediate:t=!1,onNeedRefresh:n,onOfflineReady:r,onRegistered:i,onRegisteredSW:o,onRegisterError:s}=e;let a,l;const c=async(f=!0)=>{await l};async function u(){if("serviceWorker"in navigator){if(a=await kn(()=>import("./workbox-window.prod.es5-HOwwy-x7.js"),__vite__mapDeps([])).then(({Workbox:f})=>new f("/sw.js",{scope:"/",type:"classic"})).catch(f=>{s?.(f)}),!a)return;a.addEventListener("activated",f=>{(f.isUpdate||f.isExternal)&&window.location.reload()}),a.addEventListener("installed",f=>{f.isUpdate||r?.()}),a.register({immediate:t}).then(f=>{o?o("/sw.js",f):i?.(f)}).catch(f=>{s?.(f)})}}return l=u(),c}function Yo(e={}){const{immediate:t=!0,onNeedRefresh:n,onOfflineReady:r,onRegistered:i,onRegisteredSW:o,onRegisterError:s}=e,[a,l]=L(!1),[c,u]=L(!1),f=wc({immediate:t,onOfflineReady(){u(!0),r?.()},onNeedRefresh(){l(!0),n?.()},onRegistered:i,onRegisteredSW:o,onRegisterError:s});return{needRefresh:[a,l],offlineReady:[c,u],updateServiceWorker:f}}const _c="_Container_8liaz_1",kc="_Toast_8liaz_15",xc="_ToastMessage_8liaz_43",Sc="_ToastButton_8liaz_51",bt={Container:_c,Toast:kc,ToastMessage:xc,ToastButton:Sc};var Ec=K("<span>App ready to work offline"),Cc=K("<button>Reload"),Pc=K("<div><div></div><button>Close"),$c=K("<div>"),Ac=K("<span>New content available, click on reload button to update.");const Oc=()=>{const{offlineReady:[e,t],needRefresh:[n,r],updateServiceWorker:i}=Yo({onRegistered(s){console.log("SW Registered: ",s)},onRegisterError(s){console.log("SW registration error",s)}}),o=()=>{t(!1),r(!1)};return(()=>{var s=$c();return C(s,P(G,{get when(){return e()||n()},get children(){var a=Pc(),l=a.firstChild,c=l.nextSibling;return C(l,P(G,{get fallback(){return Ac()},get when(){return e()},get children(){return Ec()}})),C(a,P(G,{get when(){return n()},get children(){var u=Cc();return u.$$click=()=>i(!0),B(()=>F(u,bt.ToastButton)),u}}),c),c.$$click=()=>o(),B(u=>{var f=bt.Toast,h=bt.Message,d=bt.ToastButton;return f!==u.e&&F(a,u.e=f),h!==u.t&&F(l,u.t=h),d!==u.a&&F(c,u.a=d),u},{e:void 0,t:void 0,a:void 0}),a}})),B(()=>F(s,bt.Container)),s})()};pe(["click"]);var Rc=K("<div><h1>Home");const Tc=()=>{const e=ye();return ql(),Ho(),lc(),fc(),gc(),pc(),(()=>{var t=Rc();return t.firstChild,C(t,P(Oc,{}),null),B(()=>F(t,`${e.accent} ${bc.AppBody}`)),t})()},Ic="_active_8eb29_1",jc="_inactive_8eb29_6",Bc="_navbar_8eb29_11",wt={active:Ic,inactive:jc,navbar:Bc};var rn=(e=>(e[e.DnD5e=0]="DnD5e",e[e.TOValient=1]="TOValient",e))(rn||{});function Lc(e){const[t,n]=L([]),r=rn.DnD5e;return B(()=>{switch(!0){case e().startsWith("/info"):const i=[{Name:"All",Link:"/info"},{Name:"Races",Link:"/info/races"},{Name:"Spells",Link:"/info/spells"},{Name:"Feats",Link:"/info/feats"}];r===rn.DnD5e?i.push({Name:"Backgrounds",Link:"/info/backgrounds"}):r===rn.TOValient&&i.push({Name:"Heritages",Link:"/info/heritages"}),n(i);break;case e().startsWith("/characters"):n([{Name:"All",Link:"/characters"},{Name:"View",Link:"/characters/view"},{Name:"Create",Link:"/characters/create"}]);break;case e().startsWith("/homebrew"):n([{Name:"All",Link:"/homebrew"},{Name:"View",Link:"/homebrew/view"},{Name:"Create",Link:"/homebrew/create"}]);break;default:n([]);break}}),t}function Dc(e,t){const n=document.getElementById("root"),r=i=>!e.contains(i.target)&&t()?.();n.addEventListener("click",r),ut(()=>{n.removeEventListener("click",r)})}var Kc=K("<div>");const Nc=e=>{const t=ye(),[n,r]=e.backgroundClick??L(!0);return P(Ls,{get children(){var i=Kc();return As(Dc,i,()=>()=>r(!n())),C(i,()=>e.children),B(o=>{var s=e.width,a=e.height,l=`translate(${e.translate.x??"-50%"},${e.translate.y??"-50%"})`,c=`${t.popup} ${t.primary}`;return s!==o.e&&((o.e=s)!=null?i.style.setProperty("width",s):i.style.removeProperty("width")),a!==o.t&&((o.t=a)!=null?i.style.setProperty("height",a):i.style.removeProperty("height")),l!==o.a&&((o.a=l)!=null?i.style.setProperty("transform",l):i.style.removeProperty("transform")),c!==o.o&&F(i,o.o=c),o},{e:void 0,t:void 0,a:void 0,o:void 0}),i}})},Mc="_menuButton_frhu5_1",Fc="_menuButtons_frhu5_10",Pi={menuButton:Mc,menuButtons:Fc};var qc=K('<button type=button><svg width=32pt height=32pt version=1.1 viewBox="0 0 128 128"xmlns=http://www.w3.org/2000/svg><path d="m113.15 40.062c4.9922 0 8.5742-3.5859 8.5742-8.4492s-3.5859-8.7031-8.5742-8.7031h-98.176c-4.8633 0-8.7031 3.8398-8.7031 8.7031s3.8398 8.4492 8.7031 8.4492zm-98.176 47.875c-4.8633 0-8.7031 3.8398-8.7031 8.5742s3.8398 8.7031 8.7031 8.7031h98.305c4.9922 0 8.5742-3.9688 8.5742-8.7031s-3.5859-8.5742-8.5742-8.5742zm0-32.641c-4.8633 0-8.7031 3.8398-8.7031 8.7031s3.8398 8.7031 8.7031 8.7031h98.305c4.9922 0 8.5742-3.8398 8.5742-8.7031s-3.5859-8.7031-8.5742-8.7031z"></path></svg>'),Uc=K("<ul>"),Vc=K("<li><button>");const Wc=()=>{const[e,t]=L(!1),n=ye(),[r,i]=L(!1),o=[{name:"Login",condition:()=>!r(),action:()=>{i(!0)}},{name:"Register",condition:()=>!r(),action:()=>{}},{name:"Logout",condition:r,action:()=>{i(!1)}},{name:"Profile",condition:r,action:()=>{}},{name:"Settings",condition:()=>!0,action:()=>{}}];return[(()=>{var s=qc();return s.$$click=()=>t(!e()),B(()=>F(s,`${n.accent} ${n.hover}`)),s})(),P(G,{get when(){return e()},get children(){return P(Nc,{width:"10%",height:"25%",translate:{x:"36vw",y:"-40vh"},backgroundClick:[e,t],get children(){var s=Uc();return C(s,P(Re,{each:o,children:a=>P(G,{get when(){return a.condition()},get children(){var l=Vc(),c=l.firstChild;return Pn(c,"click",a.action,!0),C(c,()=>a.name),B(u=>{var f=`${100/o.filter(v=>v.condition()).length}%`,h=`${n.hover}`,d=`${n.accent} ${Pi.menuButton}`;return f!==u.e&&((u.e=f)!=null?l.style.setProperty("height",f):l.style.removeProperty("height")),h!==u.t&&F(l,u.t=h),d!==u.a&&F(c,u.a=d),u},{e:void 0,t:void 0,a:void 0}),l}})})),B(()=>F(s,`${Pi.menuButtons}`)),s}})}})]};pe(["click"]);var zc=K("<span>MySite"),Hc=K("<li><div>MySite"),Yc=K("<div><ul>"),Gc=K("<div><div><ul>"),Xc=K("<button>"),$i=K("<li>");const Qc=e=>{const t=ye();L(!1);const[n,r]=L(window.location.pathname),i=[{Name:"Home",Link:"/"},{Name:"Characters",Link:"/characters"},{Name:"Info",Link:"/info"},{Name:"Homebrew",Link:"/homebrew"}];let o=Lc(n);const[s,a]=e.list,[l,c]=L(!1);return B(()=>{c(!s())}),(()=>{var u=Gc(),f=u.firstChild,h=f.firstChild;return C(f,P(G,{get when(){return!s()},get children(){return zc()}}),h),C(h,P(G,{get when(){return s()},get children(){return Hc()}}),null),C(h,P(Re,{each:i,children:(d,v)=>P(G,{get when(){return!s()||v()===0||l()},get children(){var m=$i();return C(m,P(ai,{onClick:()=>r(d.Link),get href(){return d.Link},get children(){return d.Name}}),null),C(m,P(G,{get when(){return U(()=>!!s())()&&v()===0},get children(){var g=Xc();return g.$$click=()=>c(!l()),g.style.setProperty("background","none"),g.style.setProperty("color","white"),g.style.setProperty("font-size","1.5rem"),g.style.setProperty("padding","0 10px"),g.style.setProperty("width","min-content"),g.style.setProperty("margin-right","0"),C(g,()=>l()?"↑":"↓"),g}}),null),B(()=>F(m,`${t.hover} ${n()==="/"&&d.Link==="/"||n().startsWith(d.Link)&&n()!=="/"&&d.Link!=="/"?wt.active:wt.inactive}`)),m}})}),null),C(f,P(Wc,{}),null),C(u,P(G,{get when(){return o().length>0},get children(){var d=Yc(),v=d.firstChild;return C(v,P(Re,{get each(){return o()},children:m=>(()=>{var g=$i();return C(g,P(ai,{onClick:()=>r(m.Link),get href(){return m.Link},get children(){return m.Name}})),B(()=>F(g,`${t.hover} ${n()===m.Link?wt.active:wt.inactive}`)),g})()})),B(()=>F(d,e.style)),d}}),null),B(d=>{var v=wt.navbar,m=s()?{margin:"0 auto",width:"100%"}:{},g=`${e.style}`,y=s()?{margin:"0 auto",width:"min-content"}:{};return v!==d.e&&F(u,d.e=v),d.t=Xn(f,m,d.t),g!==d.a&&F(f,d.a=g),d.o=Xn(h,y,d.o),d},{e:void 0,t:void 0,a:void 0,o:void 0}),u})()};pe(["click"]);const Jc="_SpellsBody_ump1u_1",Zc="_materials_needed_ump1u_1",eu="_TableRow_ump1u_32",tu="_headerBar_ump1u_64",Vn={SpellsBody:Jc,materials_needed:Zc,TableRow:eu,headerBar:tu},nu="_TableRow_41syk_1",ru={TableRow:nu};var iu=K("<td><strong>Materials:</strong> "),ou=K("<td><strong>higher levels:"),su=K("<td><textarea></textarea>"),au=K("<tr><td><h2><strong></strong></h2></td><td><strong>Level:</strong> </td><td><strong>School:</strong> </td><td><strong>Casting time:</strong> </td><td><strong>Range:</strong> </td><td><strong>Components:</strong> </td><td><strong>Duration:</strong> </td><td><strong>Classes:</strong> </td><td><strong>Subclasses:</strong> </td><td><strong>Description:</strong></td><td><textarea name=spell_desc>");const lu=e=>{const t=e.spell.isMaterial,n=e.spell.isSomatic,r=e.spell.isVerbal;function i(){let s=[],a=[];return r&&s.push("V"),n&&s.push("S"),t&&s.push("M"),a.push(s.join(", ")),t&&a.push(e.spell.materials_Needed),a}let o=i();return(()=>{var s=au(),a=s.firstChild,l=a.firstChild,c=l.firstChild,u=a.nextSibling,f=u.firstChild;f.nextSibling;var h=u.nextSibling,d=h.firstChild;d.nextSibling;var v=h.nextSibling,m=v.firstChild;m.nextSibling;var g=v.nextSibling,y=g.firstChild;y.nextSibling;var p=g.nextSibling,b=p.firstChild;b.nextSibling;var w=p.nextSibling,k=w.firstChild;k.nextSibling;var S=w.nextSibling,x=S.firstChild;x.nextSibling;var T=S.nextSibling,_=T.firstChild;_.nextSibling;var R=T.nextSibling,$=R.nextSibling,A=$.firstChild;return C(c,()=>e.spell.name),C(u,()=>e.spell.level,null),C(h,()=>e.spell.school,null),C(v,()=>e.spell.castingTime,null),C(g,()=>e.spell.range,null),C(p,()=>o[0],null),C(s,P(G,{when:t,get children(){var I=iu(),D=I.firstChild;return D.nextSibling,C(I,()=>o[1],null),I}}),w),C(w,()=>e.spell.duration,null),C(S,()=>e.spell.classes.join(", "),null),C(T,()=>e.spell.subClasses.join(", "),null),C(A,()=>e.spell.desc),C(s,P(G,{get when(){return!!e.spell.higherLevel},get children(){return ou()}}),null),C(s,P(G,{get when(){return!!e.spell.higherLevel},get children(){var I=su(),D=I.firstChild;return C(D,()=>e.spell.higherLevel),I}}),null),B(()=>F(s,`${ru.TableRow}`)),s})()};var cu=K("<div><button>←←</button><button>←</button><select></select><span> / </span><button>→</button><button>→→"),uu=K("<option>");const fu=e=>{const[t,n]=L(1),[r,i]=L(10),o=e.setPaginatedItems,s=ye(),a=U(()=>e.items().slice((t()-1)*r(),t()*r())),l=U(()=>Math.ceil(e.items().length/r())),c=[10,20,50,100];return B(()=>{o(a())}),(()=>{var u=cu(),f=u.firstChild,h=f.nextSibling,d=h.nextSibling,v=d.nextSibling,m=v.firstChild,g=v.nextSibling,y=g.nextSibling;return f.$$click=()=>n(1),h.$$click=()=>n(t()-1),d.addEventListener("change",p=>i(+p.currentTarget.value)),C(d,P(Re,{each:c,children:p=>(()=>{var b=uu();return b.value=p,C(b,p),b})()})),C(v,t,m),C(v,l,null),g.$$click=()=>n(t()+1),y.$$click=()=>n(l()),B(p=>{var b=`${s.accent} `,w=t()===1,k=t()===1,S=t()===l(),x=t()===l();return b!==p.e&&F(u,p.e=b),w!==p.t&&(f.disabled=p.t=w),k!==p.a&&(h.disabled=p.a=k),S!==p.o&&(g.disabled=p.o=S),x!==p.i&&(y.disabled=p.i=x),p},{e:void 0,t:void 0,a:void 0,o:void 0,i:void 0}),u})()};pe(["click"]);const hu="_chip_1bifh_1",du={chip:hu};var pu=K("<span> : <!> <button>x");const mu=e=>{const t=ye();return(()=>{var n=pu(),r=n.firstChild,i=r.nextSibling,o=i.nextSibling,s=o.nextSibling;return C(n,()=>e.key,r),C(n,()=>e.value,i),Pn(s,"click",e.clear,!0),B(()=>F(n,`${t.tertiary} ${du.chip}`)),n})()};pe(["click"]);const yu="_searchBar_176b5_1",gu="_booleanSelect_176b5_39",vu="_chipBar_176b5_67",Wn={searchBar:yu,booleanSelect:gu,chipBar:vu};var bu=K("<button>Clear All");const wu=e=>{const t=ye();return(()=>{var n=bu();return Pn(n,"click",e.clear,!0),B(()=>F(n,`${t.accent}`)),n})()};pe(["click"]);function zn(e){return e.replace(/[_-]/g," ").split(" ").map(t=>t.charAt(0).toUpperCase()+t.slice(1)).join(" ")}var _u=K("<select>"),ku=K("<select><option value=0>0</option><option value=1>1</option><option value=2>2</option><option value=3>3</option><option value=4>4</option><option value=5>5</option><option value=6>6</option><option value=7>7</option><option value=8>8</option><option value=9>9"),Ai=K('<input id=searchBar placeholder="Search Spells..."type=text>'),xu=K("<div id=booleanBar><span><label for=falsebox></label><input type=checkbox id=booleanCheckbox>"),Su=K("<div>"),Eu=K("<div><div><span><select id=chipDropdown>"),Oi=K("<option>"),Cu=K('<svg><path d="M 21 3 C 11.621094 3 4 10.621094 4 20 C 4 29.378906 11.621094 37 21 37 C 24.710938 37 28.140625 35.804688 30.9375 33.78125 L 44.09375 46.90625 L 46.90625 44.09375 L 33.90625 31.0625 C 36.460938 28.085938 38 24.222656 38 20 C 38 10.621094 30.378906 3 21 3 Z M 21 5 C 29.296875 5 36 11.703125 36 20 C 36 28.296875 29.296875 35 21 35 C 12.703125 35 6 28.296875 6 20 C 6 11.703125 12.703125 5 21 5 Z">');const Pu=e=>{const[t,n]=L({key:"",value:""}),[r,i]=L("name"),[o,s]=L(""),[a,l]=L(!1),[c,u]=L([]);L(!1),ye(),document.getElementById("booleanCheckbox");const f=["Abjuration","Conjuration","Divination","Enchantment","Evocation","Illusion","Necromancy","Transmutation"];B(()=>{e.setSearchResults(e.spellsSrd().filter(d=>d&&c().length>0?!c().map(m=>{const g=d[m.key];if(g)switch(!0){case typeof g=="string":return g.toLowerCase().includes(m.value.toLowerCase());case Array.isArray(g):return g.join(" ").toLowerCase().includes(m.value.toLowerCase());case typeof g=="boolean":return` ${g} `.toLowerCase().includes(m.value.toLowerCase())}return!1}).includes(!1):!0))}),B(()=>{typeof e.spellsSrd()[0][r()]=="boolean"&&s(`${a()}`)}),B(()=>{u(d=>t().key?[...d,t()]:d),s(""),document.getElementById("searchBar")&&(document.getElementById("searchBar").value="")});const h=d=>{switch(d){case"name":return"Name";case"level":return"Level";case"school":return"School";case"castingTime":return"Casting Time";case"range":return"Range";case"duration":return"Duration";case"concentration":return"Conc";case"ritual":return"Ritual";case"damageType":return"Dmg Type";case"isMaterial":return"Material";case"isSomatic":return"Somatic";case"isVerbal":return"Verbal";case"classes":return"Classes";case"desc":return"Desc";case"subClasses":return"Subclasses"}};return(()=>{var d=Eu(),v=d.firstChild,m=v.firstChild,g=m.firstChild;return C(v,P($u,{onClick:()=>n(y=>({key:r(),value:zn(o())}))}),m),g.addEventListener("change",y=>i(y.target.value)),C(g,P(Re,{get each(){return Object.keys(e.spellsSrd()[0]).filter(y=>!["materials_Needed","higherLevel","page"].includes(y))},children:y=>(()=>{var p=Oi();return p.value=y,C(p,()=>h(y)),p})()})),C(v,P(ti,{get children(){return[P(gt,{get when(){return typeof e.spellsSrd()[0][r()]=="string"},get children(){return P(ti,{get fallback(){return(()=>{var y=Ai();return y.$$keydown=p=>{p.key==="Enter"&&n(b=>({key:r(),value:zn(p.currentTarget.value)}))},y.addEventListener("change",p=>s(p.currentTarget.value)),B(()=>y.value=o()),y})()},get children(){return[P(gt,{get when(){return r()==="school"},get children(){var y=_u();return y.addEventListener("change",p=>s(p.currentTarget.value)),C(y,P(Re,{each:f,children:p=>(()=>{var b=Oi();return b.value=p,C(b,p),b})()})),y}}),P(gt,{get when(){return r()==="level"},get children(){var y=ku();return y.addEventListener("change",p=>s(p.currentTarget.value)),y}})]}})}}),P(gt,{get when(){return Array.isArray(e.spellsSrd()[0][r()])},get children(){var y=Ai();return y.$$keydown=p=>{p.key==="Enter"&&n(b=>({key:r(),value:zn(p.currentTarget.value)}))},y.addEventListener("change",p=>s(p.currentTarget.value)),B(()=>y.value=o()),y}}),P(gt,{get when(){return typeof e.spellsSrd()[0][r()]=="boolean"},get children(){var y=xu(),p=y.firstChild,b=p.firstChild,w=b.nextSibling;return C(b,P(G,{get when(){return a()===!1},children:"false"}),null),C(b,P(G,{get when(){return a()===!0},children:"true"}),null),w.addEventListener("change",()=>{l(!a()),s(`${a()}`)}),B(()=>F(y,`${Wn.booleanSelect}`)),B(()=>w.checked=a()),y}})]}}),null),C(d,P(G,{get when(){return c().length>0},get children(){var y=Su();return C(y,P(wu,{clear:()=>u([])}),null),C(y,P(Re,{get each(){return c()},children:(p,b)=>P(mu,{get key(){return h(p.key)??""},get value(){return p.value},clear:()=>u(w=>w.filter((k,S)=>S!==b()))})}),null),B(()=>F(y,`${Wn.chipBar}`)),y}}),null),B(()=>F(v,`${Wn.searchBar}`)),d})()},$u=e=>(()=>{var t=Cu();return Fi(t,fn(e,{xmlns:"http://www.w3.org/2000/svg",x:"0px",y:"0px",width:"100",height:"100",viewBox:"0 0 50 50"}),!0,!0),t})();pe(["keydown"]);var Au=K("<div><h1>Spells</h1><table><tbody></tbody></table><div id=pagginator>"),Ou=K("<span><i>"),Ru=K("<tr><td><span><span></span><span><i></i></span></span><span></span></td><td><button>"),Tu=K("<hr>");const Iu=()=>{const e=ye(),t=Ho(),[n,r]=L([]),[i,o]=L([]),[s,a]=L([]),l=f=>c(f)?r(n().filter(h=>h!==f)):r([...n(),f]),c=f=>n().includes(f),u=f=>{switch(f){case"0":return"Cantrip";case"1":return"1st";case"2":return"2nd";case"3":return"3rd";default:return`${f}th`}};return(()=>{var f=Au(),h=f.firstChild,d=h.nextSibling,v=d.firstChild,m=d.nextSibling;return C(f,P(Pu,{searchResults:s,setSearchResults:a,spellsSrd:t}),d),C(v,P(Re,{get each(){return i()},children:(g,y)=>[(()=>{var p=Ru(),b=p.firstChild,w=b.firstChild,k=w.firstChild,S=k.nextSibling,x=S.firstChild,T=w.nextSibling,_=b.nextSibling,R=_.firstChild;return C(k,()=>g.name),C(x,()=>g.school),C(w,P(G,{get when(){return g.concentration},get children(){var $=Ou(),A=$.firstChild;return C(A,()=>g.duration),$}}),null),C(T,()=>u(g.level)),R.$$click=()=>l(y()),C(R,()=>c(y())?"↑":"↓"),B($=>{var A=`${Vn.TableRow}`,I=Vn.headerBar;return A!==$.e&&F(p,$.e=A),I!==$.t&&F(b,$.t=I),$},{e:void 0,t:void 0}),p})(),P(G,{get when(){return c(y())},get children(){return P(lu,{spell:g})}}),Tu()]})),C(m,P(fu,{items:s,setPaginatedItems:o})),B(g=>{var y=`${e.accent} ${Vn.SpellsBody}`,p=`${e.table}`;return y!==g.e&&F(f,g.e=y),p!==g.t&&F(d,g.t=p),g},{e:void 0,t:void 0}),f})()};pe(["click"]);var ju=K("<div>");const Go=Tr(()=>kn(()=>import("./characters-mfshZVVg.js"),__vite__mapDeps([0,1]))),Xo=Tr(()=>kn(()=>import("./create-Mq4Hk_jr.js"),__vite__mapDeps([2,1]))),Qo=Tr(()=>kn(()=>import("./view-2bb1VBdY.js"),__vite__mapDeps([3,1]))),Bu=document.getElementById("root"),Lu=20*1e3,{needRefresh:Du,offlineReady:Ku,updateServiceWorker:Mu}=Yo({onRegisteredSW(e,t){t&&setInterval(()=>{Go.preload(),Xo.preload(),Qo.preload(),console.log("onRegisteredSW: ",t),console.log("onRegisteredSW.url: ",e),t.update()},Lu)},onRegisterError(e){console.log("SW registration error",e)},onNeedRefresh(){console.log("onNeedRefresh")},onOfflineReady(){console.log("onOfflineReady")}}),Nu=e=>{const t=ye(),n=function(){let s=!1;return function(a){(/(android|bb\d+|meego).+mobile|avantgo|bada\/|blackberry|blazer|compal|elaine|fennec|hiptop|iemobile|ip(hone|od)|iris|kindle|lge |maemo|midp|mmp|mobile.+firefox|netfront|opera m(ob|in)i|palm( os)?|phone|p(ixi|re)\/|plucker|pocket|psp|series(4|6)0|symbian|treo|up\.(browser|link)|vodafone|wap|windows ce|xda|xiino/i.test(a)||/1207|6310|6590|3gso|4thp|50[1-6]i|770s|802s|a wa|abac|ac(er|oo|s\-)|ai(ko|rn)|al(av|ca|co)|amoi|an(ex|ny|yw)|aptu|ar(ch|go)|as(te|us)|attw|au(di|\-m|r |s )|avan|be(ck|ll|nq)|bi(lb|rd)|bl(ac|az)|br(e|v)w|bumb|bw\-(n|u)|c55\/|capi|ccwa|cdm\-|cell|chtm|cldc|cmd\-|co(mp|nd)|craw|da(it|ll|ng)|dbte|dc\-s|devi|dica|dmob|do(c|p)o|ds(12|\-d)|el(49|ai)|em(l2|ul)|er(ic|k0)|esl8|ez([4-7]0|os|wa|ze)|fetc|fly(\-|_)|g1 u|g560|gene|gf\-5|g\-mo|go(\.w|od)|gr(ad|un)|haie|hcit|hd\-(m|p|t)|hei\-|hi(pt|ta)|hp( i|ip)|hs\-c|ht(c(\-| |_|a|g|p|s|t)|tp)|hu(aw|tc)|i\-(20|go|ma)|i230|iac( |\-|\/)|ibro|idea|ig01|ikom|im1k|inno|ipaq|iris|ja(t|v)a|jbro|jemu|jigs|kddi|keji|kgt( |\/)|klon|kpt |kwc\-|kyo(c|k)|le(no|xi)|lg( g|\/(k|l|u)|50|54|\-[a-w])|libw|lynx|m1\-w|m3ga|m50\/|ma(te|ui|xo)|mc(01|21|ca)|m\-cr|me(rc|ri)|mi(o8|oa|ts)|mmef|mo(01|02|bi|de|do|t(\-| |o|v)|zz)|mt(50|p1|v )|mwbp|mywa|n10[0-2]|n20[2-3]|n30(0|2)|n50(0|2|5)|n7(0(0|1)|10)|ne((c|m)\-|on|tf|wf|wg|wt)|nok(6|i)|nzph|o2im|op(ti|wv)|oran|owg1|p800|pan(a|d|t)|pdxg|pg(13|\-([1-8]|c))|phil|pire|pl(ay|uc)|pn\-2|po(ck|rt|se)|prox|psio|pt\-g|qa\-a|qc(07|12|21|32|60|\-[2-7]|i\-)|qtek|r380|r600|raks|rim9|ro(ve|zo)|s55\/|sa(ge|ma|mm|ms|ny|va)|sc(01|h\-|oo|p\-)|sdk\/|se(c(\-|0|1)|47|mc|nd|ri)|sgh\-|shar|sie(\-|m)|sk\-0|sl(45|id)|sm(al|ar|b3|it|t5)|so(ft|ny)|sp(01|h\-|v\-|v )|sy(01|mb)|t2(18|50)|t6(00|10|18)|ta(gt|lk)|tcl\-|tdg\-|tel(i|m)|tim\-|t\-mo|to(pl|sh)|ts(70|m\-|m3|m5)|tx\-9|up(\.b|g1|si)|utst|v400|v750|veri|vi(rg|te)|vk(40|5[0-3]|\-v)|vm40|voda|vulc|vx(52|53|60|61|70|80|81|83|85|98)|w3c(\-| )|webc|whit|wi(g |nc|nw)|wmlb|wonu|x700|yas\-|your|zeto|zte\-/i.test(a.substr(0,4)))&&(s=!0)}(navigator.userAgent||navigator.vendor),s},r=()=>o(window.matchMedia("only screen and (max-width: 768px)").matches);window.addEventListener("resize",r),B(()=>{console.log("updateServiceWorker.needRefresh: ",Du[0]())}),B(()=>{console.log("updateServiceWorker.offlineReady: ",Ku[0]())});const[i,o]=L(n());return B(()=>{const s=!!(n()||window.innerWidth<=768);o(s)}),(()=>{var s=ju();return s.style.setProperty("height","100vh"),C(s,P(Qc,{list:[i,o],get style(){return t.accent}}),null),C(s,()=>e.children,null),B(()=>F(s,t.primary)),s})()};Cs(()=>P(ha,{root:Nu,get children(){return[P(Ne,{path:"/",component:Tc}),P(Ne,{path:"/characters",get children(){return[P(Ne,{path:"/",component:Go}),P(Ne,{path:"/view",component:Qo}),P(Ne,{path:"/create",component:Xo})]}}),P(Ne,{path:"/info",get children(){return P(Ne,{path:"/spells",component:Iu})}})]}}),Bu);export{F as a,B as c,K as t,ye as u};
+function __vite__mapDeps(indexes) {
+  if (!__vite__mapDeps.viteFileDeps) {
+    __vite__mapDeps.viteFileDeps = ["assets/characters-mfshZVVg.js","assets/characters-i2fzY9Kk.css","assets/create-Mq4Hk_jr.js","assets/view-2bb1VBdY.js"]
+  }
+  return indexes.map((i) => __vite__mapDeps.viteFileDeps[i])
+}
