+++ wwwroot/index.html
@@ -8,8 +8,8 @@
     <link rel="apple-touch-icon" href="/apple-touch-icon-180x180.png">
     <link rel="manifest" href="/assets/manifest-814D2Q7q.json" />
     <title>Solid App</title>
-    <script type="module" crossorigin src="/assets/index-5Mdl4-AY.js"></script>
-    <link rel="stylesheet" crossorigin href="/assets/index-_MYHaj3D.css">
+    <script type="module" crossorigin src="/assets/index-QW1j2Poi.js"></script>
+    <link rel="stylesheet" crossorigin href="/assets/index-6khB0nzR.css">
   <link rel="manifest" href="/manifest.webmanifest"></head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
