+++ wwwroot/sw.js
@@ -1,1 +1,1 @@
-if(!self.define){let e,i={};const n=(n,s)=>(n=new URL(n+".js",s).href,i[n]||new Promise((i=>{if("document"in self){const e=document.createElement("script");e.src=n,e.onload=i,document.head.appendChild(e)}else e=n,importScripts(n),i()})).then((()=>{let e=i[n];if(!e)throw new Error(`Module ${n} didn’t register its module`);return e})));self.define=(s,c)=>{const o=e||("document"in self?document.currentScript.src:"")||location.href;if(i[o])return;let r={};const a=e=>n(e,o),l={module:{uri:o},exports:r,require:a};i[o]=Promise.all(s.map((e=>l[e]||a(e)))).then((e=>(c(...e),r)))}}define(["./workbox-3e911b1d"],(function(e){"use strict";self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"assets/characters-i2fzY9Kk.css",revision:null},{url:"assets/characters-lq3nEyhS.js",revision:null},{url:"assets/create-ZPnJCx-R.js",revision:null},{url:"assets/index-_MYHaj3D.css",revision:null},{url:"assets/index-5Mdl4-AY.js",revision:null},{url:"assets/view-Ae3TD1QO.js",revision:null},{url:"assets/workbox-window.prod.es5-HOwwy-x7.js",revision:null},{url:"claims-sw.ts",revision:"ae906878678588d70a937b450a16d9c1"},{url:"index.html",revision:"9514357728e5d91ce7841d6522a01510"},{url:"prompt-sw.ts",revision:"e8a81682b8df37c440e7cf1982b6794f"},{url:"icons/icon-96x96.png",revision:"c226c489425e6a865378cc5cf8158f51"},{url:"cloud-moon-svgrepo-com.svg",revision:"19e2331308a2a97f1cfd929740a06c2b"},{url:"icons/apple-touch-icon-180x180.png",revision:"1932150af7aaf6106a0ddaa3c4cc9c79"},{url:"icons/favicon.ico",revision:"bff39e61386509149df27ee458d7fd0f"},{url:"icons/icon-128x128.png",revision:"8b5af456a56884cff2959108a3fd79c9"},{url:"icons/icon-144x144.png",revision:"fab98ae34f7934fb7f137d3e247ec8b5"},{url:"icons/icon-152x152.png",revision:"f48876b8bb6c6df5e28ee9dcc481ad97"},{url:"icons/icon-180x180.png",revision:"4c4aa9cc61f481ffa63412ec4e7457c7"},{url:"icons/icon-192x192.png",revision:"13c83070c93520a614c49dc807f06937"},{url:"icons/icon-384x384.png",revision:"39dd627a2f4a0d739f2b28211885f05f"},{url:"icons/icon-512x512.png",revision:"0e2671f2e64ab6edc9c067621c98fba5"},{url:"icons/icon-64x64.png",revision:"97422a2e356a6a331b5e1203be90fe43"},{url:"icons/icon512_maskable.png",revision:"d35b7044af3d18df64c88d069302c61e"},{url:"icons/icon512_rounded.png",revision:"e4915e9dbc89c6a140226ce515d14159"},{url:"icons/maskable-icon-512x512.png",revision:"4709319b0d4636a19d9fb7d8ba80d38f"},{url:"icons/pwa-192x192.png",revision:"a1f44904818851b3115c802fb4a3d963"},{url:"icons/pwa-512x512.png",revision:"4bb35928a2da893434a80738a96cddc7"},{url:"icons/pwa-64x64.png",revision:"a27356990eafb0c5c31a1c0c4cb7cc9d"},{url:"manifest.webmanifest",revision:"de228873a276796948756c249652bb56"}],{}),e.cleanupOutdatedCaches(),e.registerRoute(new e.NavigationRoute(e.createHandlerBoundToURL("index.html")))}));
+if(!self.define){let i,e={};const n=(n,s)=>(n=new URL(n+".js",s).href,e[n]||new Promise((e=>{if("document"in self){const i=document.createElement("script");i.src=n,i.onload=e,document.head.appendChild(i)}else i=n,importScripts(n),e()})).then((()=>{let i=e[n];if(!i)throw new Error(`Module ${n} didn’t register its module`);return i})));self.define=(s,c)=>{const o=i||("document"in self?document.currentScript.src:"")||location.href;if(e[o])return;let r={};const a=i=>n(i,o),l={module:{uri:o},exports:r,require:a};e[o]=Promise.all(s.map((i=>l[i]||a(i)))).then((i=>(c(...i),r)))}}define(["./workbox-3e911b1d"],(function(i){"use strict";self.skipWaiting(),i.clientsClaim(),i.precacheAndRoute([{url:"assets/characters-i2fzY9Kk.css",revision:null},{url:"assets/characters-mfshZVVg.js",revision:null},{url:"assets/create-Mq4Hk_jr.js",revision:null},{url:"assets/index-6khB0nzR.css",revision:null},{url:"assets/index-QW1j2Poi.js",revision:null},{url:"assets/view-2bb1VBdY.js",revision:null},{url:"assets/workbox-window.prod.es5-HOwwy-x7.js",revision:null},{url:"claims-sw.ts",revision:"ae906878678588d70a937b450a16d9c1"},{url:"index.html",revision:"0a155cc7953225e99792adb50da763bf"},{url:"prompt-sw.ts",revision:"e8a81682b8df37c440e7cf1982b6794f"},{url:"icons/icon-96x96.png",revision:"c226c489425e6a865378cc5cf8158f51"},{url:"cloud-moon-svgrepo-com.svg",revision:"19e2331308a2a97f1cfd929740a06c2b"},{url:"icons/apple-touch-icon-180x180.png",revision:"1932150af7aaf6106a0ddaa3c4cc9c79"},{url:"icons/favicon.ico",revision:"bff39e61386509149df27ee458d7fd0f"},{url:"icons/icon-128x128.png",revision:"8b5af456a56884cff2959108a3fd79c9"},{url:"icons/icon-144x144.png",revision:"fab98ae34f7934fb7f137d3e247ec8b5"},{url:"icons/icon-152x152.png",revision:"f48876b8bb6c6df5e28ee9dcc481ad97"},{url:"icons/icon-180x180.png",revision:"4c4aa9cc61f481ffa63412ec4e7457c7"},{url:"icons/icon-192x192.png",revision:"13c83070c93520a614c49dc807f06937"},{url:"icons/icon-384x384.png",revision:"39dd627a2f4a0d739f2b28211885f05f"},{url:"icons/icon-512x512.png",revision:"0e2671f2e64ab6edc9c067621c98fba5"},{url:"icons/icon-64x64.png",revision:"97422a2e356a6a331b5e1203be90fe43"},{url:"icons/icon512_maskable.png",revision:"d35b7044af3d18df64c88d069302c61e"},{url:"icons/icon512_rounded.png",revision:"e4915e9dbc89c6a140226ce515d14159"},{url:"icons/maskable-icon-512x512.png",revision:"4709319b0d4636a19d9fb7d8ba80d38f"},{url:"icons/pwa-192x192.png",revision:"a1f44904818851b3115c802fb4a3d963"},{url:"icons/pwa-512x512.png",revision:"4bb35928a2da893434a80738a96cddc7"},{url:"icons/pwa-64x64.png",revision:"a27356990eafb0c5c31a1c0c4cb7cc9d"},{url:"manifest.webmanifest",revision:"de228873a276796948756c249652bb56"}],{}),i.cleanupOutdatedCaches(),i.registerRoute(new i.NavigationRoute(i.createHandlerBoundToURL("index.html")))}));
